Math.Graph.Digraph

Description

This module supports digraphs and reversible graphs with clearly separated vertex and edge sets. This is often needed in practical applications of graphs.

Documentation

type Digraph v e = Graph Three (GraphElem v e)

type ReversibleGraph v e = Graph Four (GraphElem v e)

type GenericGraph mon v e = Graph mon (GraphElem v e)

vertexGE :: v -> Graph m (GraphElem v e)

verticesFromSetGE :: (Ord e, Ord v) => Set v -> GenericGraph m v e

verticesGE :: (Ord v, Ord e) => [v] -> GenericGraph m v e

edgeGE :: (Ord v, Ord e) => e -> v -> v -> Digraph v e

completeGE :: (Ord v, Ord e) => [v] -> (v -> v -> e) -> Digraph v e

edgesFromMapGE :: (Ord e, Ord v) => Map e (v, v) -> Digraph v e

edgesGE :: (Ord v, Ord e) => [(e, (v, v))] -> Digraph v e

loopGE :: (Ord v, Ord e) => [(v, e)] -> Digraph v e

reversibleEdgeGE :: (Ord v, Ord e) => e -> e -> v -> v -> ReversibleGraph v e

reversibleCompleteGE :: (Ord v, Ord e) => [v] -> (v -> v -> (e, e)) -> ReversibleGraph v e

reversibleEdgesGE :: (Ord v, Ord e) => [((e, e), (v, v))] -> ReversibleGraph v e

subobjectClassifierGraphGE :: ReversibleGraph Text Text

Orphan instances

(Ord v, Ord e) => DigraphFactory (Graph Three (GraphElem v e)) v e

Instance details

Methods

vertexGraph :: v -> Graph Three (GraphElem v e)

edgeGraph :: e -> v -> v -> Graph Three (GraphElem v e)

verticesFromSetGraph :: Set v -> Graph Three (GraphElem v e)

verticesGraph :: [v] -> Graph Three (GraphElem v e)

completeGraph :: [v] -> (v -> v -> e) -> Graph Three (GraphElem v e)

edgeGraphFromMap :: Map e (v, v) -> Graph Three (GraphElem v e)

edgesGraph :: [(e, (v, v))] -> Graph Three (GraphElem v e)

loopGraph :: [(v, e)] -> Graph Three (GraphElem v e)