Math.Graph.Graph

Description

This graph representation is based on the book "Lawvere: Sets for mathematics". vertices are such edges where e act gdom == e == e act gcod

Note it's probably better to use Reversible instead.

Documentation

type Vertex n = Bool -> n

type Edge n = Bool -> n

type Graph n = Set (Edge n)

source :: Edge n -> n

target :: Edge n -> n

act :: (b -> x) -> Endo b -> b -> x

x `act` (b . a) == (x `act` b) `act` a

x `act` id == x

toAction :: Edge n -> Action n Bool

showVertex :: Show n => Vertex n -> String

vertex :: n -> Vertex n

edge :: n -> n -> Edge n

actionGraphOf :: Ord n => Graph n -> Set (Edge n, Edge n, Edge n)

reversibleGraphOf :: Ord n => Graph n -> Set (Edge n, Edge n, Edge n, Edge n)

edgesFrom :: Ord n => Graph n -> n -> Set (Edge n)

edgesTo :: Ord n => Graph n -> n -> Set (Edge n)

targetsFrom :: Ord n => Graph n -> n -> Set n

sourcesTo :: Ord n => Graph n -> n -> Set n

split :: Ord n => Graph n -> (Set (Vertex n), Set (Edge n))

vertices :: Ord n => Graph n -> Set (Vertex n)

edges :: Ord n => Graph n -> Set (Edge n)

vertexSubset :: Eq n => Prop (Edge n)

loopSubset :: Eq n => Prop (Edge n)

oneLaneLoopSubset :: Eq n => Prop (Edge n)

degenerateLoopSubset :: Eq n => Prop (Edge n)

edgeFromSubset :: Eq n => Vertex n -> Prop (Edge n)

edgeToSubset :: Eq n => Vertex n -> Prop (Edge n)

edgeSubset :: Eq n => Vertex n -> Vertex n -> Prop (Edge n)

toVertex :: (Eq n, MonadPlus m) => Edge n -> m (Vertex n)

emptyG :: Graph n

vertexG :: n -> Graph n

edgeG :: Ord n => n -> n -> Graph n

verticesG :: Ord n => Set n -> Graph n

edgesG :: Ord n => Set (n, n) -> Graph n

unionG :: Ord n => Graph n -> Graph n -> Graph n

composeG :: Ord n => Graph n -> Graph n -> Graph n