Math.Graph.InGraphMonad

Description

This module provides ability to inspect graph data in a specialized monad.

Examples: let example1 = reversibleCompleteG ["a","b","c"] (\a b -> (a ++ b, b ++ a))

show example1 == "([a, b, c] ; [aa = a <-> a, ba / ab = b <-> a, bb = b <-> b, ca / ac = c <-> a, cb / bc = c <-> b, cc = c <-> c])"

example1 `inGraphM` edgesEndingToM "a" == return (fromList ["aa","ba","ca"])

example1 `inGraphM` verticesM == return (fromList ["a","b","c"])

example1 `inGraphM` edgesM == return (fromList ["aa","ab","ac","ba","bb","bc","ca","cb","cc"])

example1 `inGraphM` do { x <- sourceM "ab" ; y <- targetM "ab" ; return (x,y) } == return ("a","b")

Documentation

newtype InGraphM mon e m a

Constructors

InGraphM

Fields

runInGraphM :: ReaderT (Graph mon e) m a

inGraphM :: Graph mon e -> InGraphM mon e m a -> m a

currentGraphM :: Monad m => InGraphM mon e m (Graph mon e)

actM :: Monad m => e -> mon -> InGraphM mon e m e

elementsM :: Monad m => InGraphM mon e m (Set e)

verticesM :: (Ord e, GraphMonoid mon, Monad m) => InGraphM mon e m (Set e)

edgesM :: (Ord e, GraphMonoid mon, Monad m) => InGraphM mon e m (Set e)

nodesLinkedFromM :: (Monad m, Ord e, GraphMonoid mon) => e -> InGraphM mon e m (Set e)

edgesStartingFromM :: (Monad m, Ord e, GraphMonoid mon) => e -> InGraphM mon e m (Set e)

edgesEndingToM :: (Monad m, Ord e, GraphMonoid mon) => e -> InGraphM mon e m (Set e)

linksM :: (Monad m, Ord e, GraphMonoid mon) => InGraphM mon e m (Set (e, e, e))

reversibleLinksM :: (Monad m, Ord e, ReversibleGraphMonoid mon) => InGraphM mon e m (Set ((e, e), (e, e)))

sourceM :: (GraphMonoid mon, Monad m) => e -> InGraphM mon e m e

targetM :: (GraphMonoid mon, Monad m) => e -> InGraphM mon e m e

inverseM :: (ReversibleGraphMonoid mon, Monad m) => e -> InGraphM mon e m e

isSourceVertexM :: (GraphMonoid mon, Monad m, Eq e) => e -> InGraphM mon e m Bool

isTargetVertexM :: (GraphMonoid mon, Monad m, Eq e) => e -> InGraphM mon e m Bool

equally valid implementation. These are equivalent in normal graphs because Bool -> Bool structure has 4 elements, which are related by not operation which switches source and target. From Lawvere,Rosebrugh: Sets for Mathematics, Exercise 10.29. Note however that we are not requiring ReversibleGraphMonoid, which actually contains this not operation. This is a reason to separate these three operations (isSourceVertexM, isTargetVertexM, isVertexM) when digraphs are used.

The left action of the monoid causes the following constraints:

source . target == source

target . source == target

source . source == source

target . target == target

Due to the first two, it's not necessary to distinguish these operations.

isVertexM :: (GraphMonoid mon, Monad m, Eq e) => e -> InGraphM mon e m Bool

isEdgeM :: (GraphMonoid mon, Monad m, Eq e) => e -> InGraphM mon e m Bool

hasPathM :: (GraphMonoid mon, Monad m, Eq e) => [e] -> InGraphM mon e m Bool

isEdgeBetweenM :: (Monad m, GraphMonoid mon, Eq e) => e -> e -> e -> InGraphM mon e m Bool

isLoopM :: (Eq e, Monad m, GraphMonoid mon) => e -> InGraphM mon e m Bool

isOneLaneLoopM :: (Eq e, Monad m, ReversibleGraphMonoid mon) => e -> InGraphM mon e m Bool

liftReaderGraph :: (Monad m, MonadTrans t, MonadReader (Graph mon e) (t m)) => InGraphM mon e m a -> t m a