Math.Matrix.Covector

Description

License: LGPL

For background material, see K. Chandrasekhara Rao: Functional Analysis.

See also: http://theoreticalminimum.com/courses/advanced-quantum-mechanics/2013/fall/lecture-1

Documentation

contract :: (Scalar (v a) ~ a, LinearTransform v Vector1 a) => (Dual :*: v) a -> (Vector1 :*: v) a -> Scalar (v a)

dual_apply :: LinearTransform v Vector1 a => Dual (v a) -> v a -> a

(*><) :: Dual (v a) -> v a :-> Vector1 a

x2_op :: (Num s, ConjugateSymmetric s) => Dual (Vector2 s)

y2_op :: (Num s, ConjugateSymmetric s) => Dual (Vector2 s)

(∇) :: VectorDerivative v d arr => d v -> arr v v

unicode support for gradient. Use with parenthesis.

(∇·) :: VectorDerivative v d arr => arr v v -> d v

unicode support for divergence. Use with parenthesis.

(·∇) :: (VectorDerivative v d arr, InnerProductSpace v) => v -> d v -> d v

unicode support for directional derivative. Binary operator.

(∇×) :: VectorCrossProduct v arr => arr v v -> arr v v

unicode support for curl. Use with parenthesis.

(∇·∇) :: VectorDerivative v d arr => d v -> d v

unicode support for laplace. Use with parenthesis.

directional_derivative_endo :: (VectorDerivative v Dual LinearMap, InnerProductSpace v) => v -> Endo (Dual v)

dconst :: (Linearizable LinearMap (:*:) f Vector1 v, Diagonalizable f v) => v -> Dual (f v)

outer_vector :: (Linearizable LinearMap (:*:) f g w, DualNum f w, DualNum g w) => Dual (f w) -> g w -> f w :-> g w

conjugate_transpose :: (Diagonalizable m a, Linearizable LinearMap (:*:) m m a, Linearizable LinearMap (:*:) m n a, Linearizable LinearMap (:*:) n m a, Diagonalizable n a, LinearIso n m a, Transposable m n a, ConjugateSymmetric (m a)) => (m a :-> n a) -> n a :-> m a

http://en.wikipedia.org/wiki/Conjugate_transpose

(⊹) :: (Linearizable LinearMap (:*:) m m a, Linearizable LinearMap (:*:) m n a, Linearizable LinearMap (:*:) n m a, Diagonalizable m a, Diagonalizable n a, LinearIso n m a, Transposable m n a, ConjugateSymmetric (m a)) => (m a :-> n a) -> n a :-> m a

unicode alias (unicode HERMITIAN CONJUGATE MATRIX character)

is_hermitian :: (Eq (m (m a)), ConjugateSymmetric (m a), Linearizable LinearMap (:*:) m m a, Transposable m m a, Diagonalizable m a, LinearTransform m m a, Applicative m, Foldable m) => (m a :-> m a) -> Bool

https://en.wikipedia.org/wiki/Conjugate_transpose

is_skew_hermitian :: (Num (Scalar (m a)), Eq (m (m a)), ConjugateSymmetric (m a), Applicative m, Linearizable LinearMap (:*:) m m a, VectorSpace (m a), Diagonalizable m a, LinearTransform m m a, Transposable m m a, Foldable m) => (m a :-> m a) -> Bool

https://en.wikipedia.org/wiki/Conjugate_transpose

is_normal :: (Num a, Eq (m (m a)), SupportsMatrixMultiplication m m m a, Eq (Scalar (m a)), ConjugateSymmetric (m a), Foldable m, Applicative m, Transposable m m a, Linearizable LinearMap (:*:) m m a, InnerProductSpace (m a)) => (m a :-> m a) -> Bool

https://en.wikipedia.org/wiki/Conjugate_transpose

bravector :: (DualNum f v, InnerProductSpace (f v)) => f v -> Dual (f v)

partial application of dot product on second argument (- %. v).

ketvector :: (DualNum f v, InnerProductSpace (f v)) => f v -> Dual (f v)

partial application of dot product on first argument: (v %. -).

kronecker :: (Eq i, Functor n, Functor m, Num c) => m i -> n i -> (m :*: n) c

dual_bilinear :: DualNum f v => (f v -> f v -> v) -> f v -> Dual (f v)

kronecker_dual :: (DualNum f a, Eq (f a), Num a) => f a -> Dual (f a)

covector_kernel :: (DualNum f v, Num v, Eq v) => Dual (f v) -> f v -> Bool

https://en.wikipedia.org/wiki/Dual_space#Injection_into_the_double-dual

adjoint :: (f a :-> f a) -> Dual (f a) -> Dual (f a)

would like to make adjoint as instance of CoFunctor, but the constraint prevents. Instead we declare LinearMap.

scalar_map :: (Vector1 a :-> Vector1 a) -> Dual (f a) -> Dual (f a)

operator_map :: (Diagonalizable w a, LinearTransform v Vector1 a, Linearizable LinearMap (:*:) w Vector1 a) => ((v a -> a) -> w a -> a) -> Dual (v a) -> Dual (w a)

type DualNum f v = (LinearTransform f Vector1 v, ConjugateSymmetric v, Diagonalizable f v, Linearizable LinearMap (:*:) f Vector1 v, VectorSpace (f v), v ~ Scalar (f v))

dual_derivative :: (VectorDerivative v Dual LinearMap, VectorDerivative w Dual LinearMap, InnerProductSpace v, InnerProductSpace w) => (Dual v -> Dual w -> c) -> v -> w -> Dual v -> Dual w -> c

This computes \(DDM(\otimes,{\bar{x}},{\bar{y}}) = [({\bar{x}} \cdot \nabla) \times ({\bar{y}} \cdot \nabla)]\) where \(\times\) is the outer product for linear operators derived from product of dual spaces \(\otimes\). Notice \({\bar{x}} \cdot \nabla\) is normally called the directional derivative, so this is really a two dimensional version of directional derivative with some possibility to choose how two dual vectors are combined.

dual_differential_outer_product :: (LinearTransform f Vector1 v, ConjugateSymmetric v, Diagonalizable f v, v ~ Scalar (f v), Linearizable LinearMap (:*:) f Vector1 v, VectorDerivative (f v) Dual LinearMap, InnerProductSpace (f v)) => f v -> f v -> Dual (f v) -> Dual (f v) -> Dual (f v)

dual_differential_dot_product :: (DualNum f v, VectorDerivative (f v) Dual LinearMap, InnerProductSpace (f v), StandardBasis (f v)) => f v -> f v -> Dual (f v) -> Dual (f v) -> Scalar (Dual (f v))

norm_covector :: (v ~ Scalar (f v), Diagonalizable f v, Linearizable LinearMap (:*:) f Vector1 v, NormedSpace (f v)) => Dual (f v)

del2 :: (ConjugateSymmetric v, Closed v) => Vector2 (Dual (Vector2 v) -> Dual (Vector2 v))

hessian2 :: (ConjugateSymmetric a, Closed a) => Dual (Vector2 a) -> (Vector2 :*: Vector2) (Dual (Vector2 a))

Orphan instances

(LinearTransform f Vector1 v, ConjugateSymmetric v, Linearizable LinearMap (:*:) f Vector1 v, FractionalSpace (f v), v ~ Scalar (f v), Floating v, Diagonalizable f v) => Floating (Dual (f v))

Instance details

Methods

pi :: Dual (f v)

exp :: Dual (f v) -> Dual (f v)

log :: Dual (f v) -> Dual (f v)

sqrt :: Dual (f v) -> Dual (f v)

(**) :: Dual (f v) -> Dual (f v) -> Dual (f v)

logBase :: Dual (f v) -> Dual (f v) -> Dual (f v)

sin :: Dual (f v) -> Dual (f v)

cos :: Dual (f v) -> Dual (f v)

tan :: Dual (f v) -> Dual (f v)

asin :: Dual (f v) -> Dual (f v)

acos :: Dual (f v) -> Dual (f v)

atan :: Dual (f v) -> Dual (f v)

sinh :: Dual (f v) -> Dual (f v)

cosh :: Dual (f v) -> Dual (f v)

tanh :: Dual (f v) -> Dual (f v)

asinh :: Dual (f v) -> Dual (f v)

acosh :: Dual (f v) -> Dual (f v)

atanh :: Dual (f v) -> Dual (f v)

log1p :: Dual (f v) -> Dual (f v)

expm1 :: Dual (f v) -> Dual (f v)

log1pexp :: Dual (f v) -> Dual (f v)

log1mexp :: Dual (f v) -> Dual (f v)

(DualNum f v, FractionalSpace (f v)) => Fractional (Dual (f v))

Instance details

Methods

(/) :: Dual (f v) -> Dual (f v) -> Dual (f v)

recip :: Dual (f v) -> Dual (f v)

fromRational :: Rational -> Dual (f v)

DualNum f v => Num (Dual (f v))

Instance details

Methods

(+) :: Dual (f v) -> Dual (f v) -> Dual (f v)

(-) :: Dual (f v) -> Dual (f v) -> Dual (f v)

(*) :: Dual (f v) -> Dual (f v) -> Dual (f v)

negate :: Dual (f v) -> Dual (f v)

abs :: Dual (f v) -> Dual (f v)

signum :: Dual (f v) -> Dual (f v)

fromInteger :: Integer -> Dual (f v)

(DualNum f v, StandardBasis (f v), Show v) => Show (Dual (f v))

Instance details

Methods

showsPrec :: Int -> Dual (f v) -> ShowS

show :: Dual (f v) -> String

showList :: [Dual (f v)] -> ShowS

(v ~ Scalar (f v), VectorSpace (f v), DualNum f v) => LieAlgebra (Dual (f v))

Instance details

Methods

(%<>%) :: Dual (f v) -> Dual (f v) -> Dual (f v)

(StandardBasis (f v), Num v, VectorSpace (f v), DualNum f v) => InnerProductSpace (Dual (f v))

Instance details

Methods

(%.) :: Dual (f v) -> Dual (f v) -> Scalar (Dual (f v))

(Diagonalizable f a, Linearizable LinearMap (:*:) f g a, LinearTransform f g a, Show ((f :*: g) a)) => Show (LinearMap (f a) (g a))

Instance details

Methods

showsPrec :: Int -> LinearMap (f a) (g a) -> ShowS

show :: LinearMap (f a) (g a) -> String

showList :: [LinearMap (f a) (g a)] -> ShowS

Semigroup v => Semigroup (LinearMap v v)

Instance details

Methods

(<>) :: LinearMap v v -> LinearMap v v -> LinearMap v v

sconcat :: NonEmpty (LinearMap v v) -> LinearMap v v

stimes :: Integral b => b -> LinearMap v v -> LinearMap v v

(Semigroup v, VectorSpace v) => Monoid (LinearMap v v)

Instance details

Methods

mempty :: LinearMap v v

mappend :: LinearMap v v -> LinearMap v v -> LinearMap v v

mconcat :: [LinearMap v v] -> LinearMap v v

(PpShowVerticalF f, LinearTransform f g a, Linearizable LinearMap (:*:) f g a, PpShowF g, PpShow a, Diagonalizable f a) => PpShow (LinearMap (f a) (g a))

Instance details

Methods

pp :: LinearMap (f a) (g a) -> Doc

(ConjugateSymmetric a, Num a) => StandardBasis ((Dual :*: Vector2) a)
https://en.wikipedia.org/wiki/Dual_space

Instance details

Methods

unit_vectors :: [(Dual :*: Vector2) a]