Math.Matrix.Indexable

Documentation

indexableMatrixProjections :: (forall b. Indexable m b, forall c. Indexable n c) => (m :*: n) (Index (m :*: n) a)

applyIndex :: Indexable m a => Index m a -> m a -> a

with :: (Indexable f a, Num a) => (Index f a -> b) -> f b

with2 :: (Indexable f a, Indexable g b, Num a, Num b) => (Index f a -> Index g b -> c) -> (f :*: g) c

with3 :: (Indexable f a, Indexable g b, Indexable h c, Num a, Num b, Num c) => (Index f a -> Index g b -> Index h c -> d) -> (f :*: (g :*: h)) d

equalIndices :: Indexable f Int => Index f Int -> Index f Int -> Bool

sumIndices :: (Foldable f, Indexable f b, Num b) => (f :*: f) b -> b

Orphan instances

Indexable I a

Instance details

Methods

diagonalProjections :: I (Index I a)

basisVector :: Index I a -> a -> I a -> I a

indexProject :: Index I a -> I a -> a

indexableIndices :: I a

(forall b. Indexable f b, forall c. Indexable g c) => Indexable (f :*: g) a

Instance details

Methods

diagonalProjections :: (f :*: g) (Index (f :*: g) a)

basisVector :: Index (f :*: g) a -> a -> (f :*: g) a -> (f :*: g) a

indexProject :: Index (f :*: g) a -> (f :*: g) a -> a

indexableIndices :: (f :*: g) a