Math.Matrix.Interface

Description

These should match standard definitions of vector spaces. Used for reference: K. Chandrasekhara Rao: Functional Analysis. also see Warner: Modern algebra.

Documentation

newtype (f :*: g) a infixr 5

The primary data type for matrices. Note that indices are represented in the functors, If you want to use numeric indices, use Simple. This can be viewed as a bicategory with one object (a). https://ncatlab.org/nlab/show/bicategory or as a monoidal category https://ncatlab.org/nlab/show/monoidal+category or as a matrix over a base field a https://ncatlab.org/nlab/show/matrix or as a tensor product of functors f and g https://ncatlab.org/nlab/show/tensor+product or as a composition of functors f and g

Constructors

Matrix

Fields

cells :: f (g a)

matrix :: (Functor m, Functor n) => (a -> b -> c) -> m a -> n b -> (m :*: n) c

This method of matrix construction is especially nice. This is the functoriality of the tensor product.

inverseMatrix :: (Contravariant m, Contravariant n) => (g a -> c -> b) -> (m :*: n) c -> n b -> (m :*: g) a

leftInverseMatrix :: (Contravariant m, Functor n) => (g a -> c -> b) -> (m :*: n) b -> n c -> (m :*: g) a

rightInverseMatrix :: (Functor m, Contravariant n) => (t -> a -> b) -> m t -> n b -> (m :*: n) a

inverseFix :: Contravariant p => (a -> a) -> p a

inverseFix2 :: Contravariant p => (a -> b) -> (b -> a) -> p a

matrixCompose :: (Functor m, Functor n, Category cat) => m (cat b c) -> n (cat a b) -> (m :*: n) (cat a c)

tensorProduct :: (Num a, Functor m, Functor n) => m a -> n a -> (m :*: n) a

tensorProductLin :: (Linearizable arr (:*:) f g a, Num a, Functor f, Functor g) => f a -> g a -> arr (f a) (g a)

bilinearImpl :: (VectorSpace (g c), Scalar (g c) ~ c, Indexable f c, Indexable g c, Integral c, VectorSpace ((f :*: g) c)) => (f c -> f c -> g c) -> f c -> f c -> (f :*: g) c

https://en.wikipedia.org/wiki/Tensor_product

bilinearity : \[f (a + b,c) = f(a, c) + f(b, c)\] \[f (k b, c) = k f(b, c)\] \[f (a, b + c) = f(a, b) + f(a,c)\] \[f (a, k c) = k f(a, c)\] https://en.wikipedia.org/wiki/Bilinear_map

applicativeMatrix :: (Applicative f, Functor m, Functor n) => f (a -> b -> c) -> (m :*: f) a -> (n :*: f) b -> (m :*: n) (f c)

(>*<) :: (Applicative f, Functor m, Functor n) => f (a -> b -> c) -> ((m :*: f) a, (n :*: f) b) -> (m :*: n) (f c)

class Num (Scalar v) => VectorSpace v where

https://en.wikipedia.org/wiki/Matrix_%28mathematics%29

Associated Types

type Scalar v

Methods

vzero :: v

vnegate :: v -> v

(%+) :: v -> v -> v infixl 6

(%*) :: Scalar v -> v -> v infix 7

class VectorSpace v => DecomposableVectorSpace v cofree | v -> cofree where

cofree is right adjoint to Scalar

Minimal complete definition

decompose

Methods

decompose :: (Scalar v -> res) -> v -> cofree res

project :: v -> cofree (Scalar v)

type VectorSpaceOver v a = (VectorSpace v, Scalar v ~ a)

type PrimitiveSpace v = (v ~ Scalar v, VectorSpace v)

type ComplexVectorSpace v a = VectorSpaceOver v (Complex a)

type Linear a b = (VectorSpace a, VectorSpace b, Scalar a ~ Scalar b)

type LinearInnerProductSpace a b = (Linear a b, InnerProductSpace a, InnerProductSpace b)

type SupportsMatrixMultiplication f g h a = (VectorSpace (f a), VectorSpace (h a), InnerProductSpace (g a), Scalar (f a) ~ a, Scalar (h a) ~ a, Scalar (g a) ~ a, Num a, ConjugateSymmetric a, Functor f, Transposable g h a)

type LinearIso f g a = (LinearTransform f g a, LinearTransform g f a)

class VectorSpace v => BilinearVectorSpace v where

Methods

biLin :: v -> v -> Scalar v

class VectorSpace m => InnerProductSpace m where

https://en.wikipedia.org/wiki/Dot_product | https://en.wikipedia.org/wiki/Outer_product

Methods

(%.) :: m -> m -> Scalar m infix 7

outer :: (InnerProductSpace m, Scalar m ~ Scalar v, VectorSpace v) => m -> v -> m -> v

class VectorSpace m => LieAlgebra m where

https://en.wikipedia.org/wiki/Lie_algebra

Methods

(%<>%) :: m -> m -> m

class VectorSpace s => MetricSpace s where

Methods

distance :: s -> s -> Scalar s

class (VectorSpace m, Num (Scalar m)) => NormedSpace m where

norm_squared is an optimization that often avoids computing square root

Minimal complete definition

norm | normSquared

Methods

norm :: m -> Scalar m

default norm :: Floating (Scalar m) => m -> Scalar m

normSquared :: m -> Scalar m

matrix_norm :: (Functor f, NormedSpace (g a), NormedSpace (f (Scalar (g a)))) => (f :*: g) a -> Scalar (f (Scalar (g a)))

This computes norm of each row, then computes the norm of the resulting column vector.

class CompleteSpace m

class ConjugateSymmetric m where

Methods

conj :: m -> m

class (Scalar (m a) ~ Scalar (n a), Functor m, Functor n) => LinearTransform m n a where

Methods

(<*>>)

Arguments

:: n a

-> (m :*: n) a

-> m a
vector times matrix

(<<*>)

Arguments

:: (m :*: n) a

-> m a

-> n a
matrix times vector

data Lin b c where

Constructors

Lin :: (f :*: g) a -> Lin (f a) (g a)

class (Functor m, Functor n, Scalar (m a) ~ Scalar (n a)) => Transposable m n a where

Methods

transposeImpl :: (m :*: n) a -> (n :*: m) a

transpose :: (Transposable g f a, Linearizable arr (:*:) f g a, Linearizable arr (:*:) g f a) => arr (g a) (f a) -> arr (f a) (g a)

indexableTranspose :: (Functor n, Indexable m a) => (n :*: m) a -> (m :*: n) a

updateColumn :: Applicative h => (a -> f b -> g c) -> h a -> (h :*: f) b -> (h :*: g) c

updateRow :: (a -> f (g b) -> f' (g' b')) -> a -> (f :*: g) b -> (f' :*: g') b'

class UpdateableMatrixDimension f where

Example use:

write_column (Vector3 3 4 5) `ycoord3` identity3 == [[1,3,0],[0,4,0],[0,5,1]]

Methods

writeRow :: Applicative h => h a -> f ((f :*: h) a -> (f :*: h) a)

writeColumn :: Applicative h => h a -> f ((h :*: f) a -> (h :*: f) a)

solveMatrix :: (Traceable m b, Fractional b, UpdateableMatrixDimension m) => (m :*: m) b -> m b -> m b

Cramer's rule https://en.wikipedia.org/wiki/Cramer%27s_rule.

Solves \({\mathbf x}\) from matrix equation \(A{\mathbf x} = {\mathbf b}\), where \(A\) is first parameter and \({\mathbf b}\) is second parameter. Returns vector \({\mathbf x}\). Satisfies requirements:

a <<*> solve_matrix a b == b

\[{\mathbf x}_i = {{\det(A[A_{ki} := {\mathbf b}_k])} \over {\det(A)}}\]

(<!-!>) :: (m a -> a) -> (a -> m a) -> Index m a

runIndex :: Index m a -> m a -> a

appIndex :: Applicative f => f (Index m a) -> f (m a) -> f a

type Index m a = m a :==: I a

class (Applicative m, Num a) => Indexable m a where

Minimal complete definition

diagonalProjections, indexableIndices

Methods

diagonalProjections :: m (Index m a)

basisVector :: Index m a -> m a

indexProject :: Index m a -> m a -> a

indexableIndices :: m a

class (Num a, Indexable m a, Transposable m m a) => Diagonalizable m a where

https://en.wikipedia.org/wiki/Square_matrix

Minimal complete definition

identity, diagonalImpl, diagonalMatrixImpl

Methods

identityImpl :: m Integer -> (m :*: m) a

argument to identityImpl is dimension of the matrix

identity :: (m :*: m) a

diagonalImpl :: (m :*: m) a -> m a

diagonalMatrixImpl :: m a -> (m :*: m) a

basis :: Diagonalizable m a => m (m a)

coefficients :: (Foldable m, Applicative m, VectorSpace v) => m (Scalar v) -> m v -> v

bilinearMap :: (VectorSpace a, Foldable m, Foldable n, Diagonalizable m b, Diagonalizable n c) => (m b -> n c -> a) -> m (Scalar a) -> n (Scalar a) -> a

linearMap_ :: (Foldable m, Diagonalizable m a, VectorSpace b) => (m a -> b) -> m (Scalar b) -> b

https://en.wikipedia.org/wiki/Linear_map#Matrices

linearMap' :: (b ~ Scalar (n b), Foldable m, Diagonalizable m a, VectorSpace (n b)) => (m a -> n b) -> m b -> n b

linearIdentity :: (Linearizable arr (:*:) m m a, LinearTransform m m a, Diagonalizable m a) => arr (m a) (m a)

diagonal :: (Linearizable arr (:*:) m m a, Diagonalizable m a) => arr (m a) (m a) -> m a

class (Functor m, Functor n) => ProjectionSpace (m :: Type -> Type) (n :: Type -> Type) where

Associated Types

data (m \\\ n) a

Methods

projectFirst :: m a -> n a

projectSecond :: m a -> (m \\\ n) a

joinVector :: n a -> (m \\\ n) a -> m a

class CodiagonalMatrix m a where

CodiagonalMatrix represents a matrix that can be split along the diagonal. The Codiagonal type represents a matrix without its diagonal. The ProjectionVector type represents a vector down from first element of diagonal when the diagonal is removed. This vector often has less elements than the original vector. Similarly for vector right from the first element of diagonal.

Associated Types

data Codiagonal m a

type m \\ a

Methods

codiagonalImpl :: (m :*: m) a -> Codiagonal m a

(|\|) :: m a -> Codiagonal m a -> (m :*: m) a

downProject :: Codiagonal m a -> m \\ a

rightProject :: Codiagonal m a -> m \\ a

class Category arr => Linearizable arr prod f g a | arr -> prod where

NOTICE: Linearizable instances for matrices that have similar dimensions are special.

Methods

fromLinear :: arr (f a) (g a) -> prod f g a

linear :: prod f g a -> arr (f a) (g a)

class Diagonalizable m a => Traceable m a where

https://ncatlab.org/nlab/show/dimension

Minimal complete definition

traceImpl, determinantImpl

Methods

traceImpl :: (m :*: m) a -> a

determinantImpl :: (m :*: m) a -> a

vectorDimension :: m a -> a

class (Category arr, Traceable m a) => LinearTraceable arr m a | m a -> arr where

Minimal complete definition

Nothing

Methods

determinant :: arr (m a) (m a) -> a

default determinant :: Linearizable arr (:*:) m m a => arr (m a) (m a) -> a

trace :: arr (m a) (m a) -> a

default trace :: Linearizable arr (:*:) m m a => arr (m a) (m a) -> a

class Traceable m a => Invertible m a where

http://en.wikipedia.org/wiki/Adjugate \({\mathsf{cofactor}}(A) = |A|(A^{-1})^{\top}\) https://en.wikipedia.org/wiki/Cross_product

Methods

cofactorImpl :: (m :*: m) a -> (m :*: m) a

adjucateImpl :: (m :*: m) a -> (m :*: m) a

inverseImpl :: (m :*: m) a -> (m :*: m) a

class LinearTraceable arr m a => LinearInvertible arr m a where

Minimal complete definition

Nothing

Methods

cofactor :: arr (m a) (m a) -> arr (m a) (m a)

default cofactor :: (Invertible m a, Linearizable arr (:*:) m m a) => arr (m a) (m a) -> arr (m a) (m a)

adjucate :: arr (m a) (m a) -> arr (m a) (m a)

default adjucate :: (Invertible m a, Linearizable arr (:*:) m m a) => arr (m a) (m a) -> arr (m a) (m a)

inverse :: arr (m a) (m a) -> arr (m a) (m a)

default inverse :: (Invertible m a, Linearizable arr (:*:) m m a) => arr (m a) (m a) -> arr (m a) (m a)

dualBasisImpl :: Invertible m a => (m :*: m) a -> (m :*: m) a

this computes \[f(A) = (A^{-1})^{\top}\] it's used to compute dual basis for a set of basis vectors.

dualBasis :: (Invertible m a, Linearizable arr (:*:) m m a) => arr (m a) (m a) -> arr (m a) (m a)

class Functor m => EigenDecomposable m a where

Methods

eigenvalues :: (m :*: m) a -> m a

class EigenDecomposable m a => EigenVectorable m a where

Methods

eigenvectors :: (m :*: m) a -> (m :*: m) a

class (Applicative m, Applicative n) => AppendableVector m n where

Associated Types

type m :+: n :: Type -> Type

Methods

(||>>) :: m a -> n a -> (m :+: n) a

class AppendableVector m n => SplittableVector m n where

Methods

vsplit :: (m :+: n) a -> (m a, n a)

class Conditional a where

Iverson bracket: http://en.wikipedia.org/wiki/Iverson_bracket

Methods

fromBoolean :: Bool -> a

class StandardBasis m where

Methods

unitVectors :: [m]

class VectorSpace v => CoordinateSpace v where

Associated Types

type Coordinate v

Methods

index :: Coordinate v -> v -> Scalar v

listVector :: [Scalar v] -> v

dimensionSize :: v -> Int

coordinates :: v -> [Coordinate v]

class (Num (Scalar v), VectorSpace v) => NumSpace v

vector space with scalars in Num class

class (Fractional (Scalar v), NumSpace v) => FractionalSpace v

vector space with fractional scalars

class VectorSpace v => FiniteDimensional v d i arr where

https://en.wikipedia.org/wiki/Dual_space#Injection_into_the_double-dual

Methods

finite :: arr ((d :*: d) v) (i v)

class HasIdentityLinear v arr where

Methods

matIdentity :: (Num a, ConjugateSymmetric a) => arr (v a) (v a)

class VectorSpace v => Dualizable v d where

Methods

covector :: (v -> Scalar v) -> d v

bracket :: d v -> v -> Scalar v

class VectorSpace v => VectorDerivative v d arr | d -> arr, v arr -> d where

https://en.wikipedia.org/wiki/Laplace_operator https://en.wikipedia.org/wiki/Divergence https://en.wikipedia.org/wiki/Gradient https://en.wikipedia.org/wiki/Directional_derivative Notice: for directional derivative, the direction is not automatically normalized, since that needs NormedSpace

Minimal complete definition

divergence, grad, directionalDerivative

Methods

divergence :: arr v v -> d v

grad :: d v -> arr v v

directionalDerivative :: v -> d v -> d v

laplace :: d v -> d v

normalizedDirectionalDerivative :: (VectorDerivative v d arr, NormedSpace v, Fractional (Scalar v)) => v -> d v -> d v

version of directional derivative that normalizes the direction: https://mathworld.wolfram.com/DirectionalDerivative.html

class VectorCrossProduct v arr where

https://en.wikipedia.org/wiki/Curl_(mathematics)

Methods

curl :: arr v v -> arr v v

class VectorLaplacian v arr where

https://en.wikipedia.org/wiki/Vector_Laplacian

Methods

vectorLaplace :: arr v v -> arr v v

class Functor f => ProjectionDual f d a where

Methods

projectionDual :: f (d (f a))

matrixM :: (Traversable f, Traversable g, Monad m) => (a -> b -> m c) -> f a -> g b -> m ((f :*: g) c)

matrixMatrix :: (Functor m, Functor n, Functor m', Functor n') => (a -> b -> c) -> (m :*: m') a -> (n :*: n') b -> ((m :*: n) :*: (m' :*: n')) c

newtype Basis m

Constructors

Basis [m]

(%-) :: VectorSpace v => v -> v -> v infixl 6

(%/) :: (Fractional (Scalar v), VectorSpace v) => v -> Scalar v -> v

innerproductspaceCosAngle :: (InnerProductSpace m, Floating (Scalar m)) => m -> m -> Scalar m

https://en.wikipedia.org/wiki/Angle

normedLieAlgebraSinAngle :: (LieAlgebra m, NormedSpace m, Floating (Scalar m)) => m -> m -> Scalar m

angle :: (InnerProductSpace m, Floating (Scalar m)) => m -> m -> Scalar m

https://en.wikipedia.org/wiki/Angle

iVec :: StandardBasis v => v

jVec :: StandardBasis v => v

kVec :: StandardBasis v => v

lVec :: StandardBasis v => v

innerproductspaceNorm :: (Floating (Scalar m), InnerProductSpace m) => m -> Scalar m

innerproductspaceNormSquared :: (Floating (Scalar m), InnerProductSpace m) => m -> Scalar m

vsum :: (Foldable t, VectorSpace a) => t a -> a

indexUnit :: StandardBasis v => Int -> v

projection :: (Fractional (Scalar v), VectorSpace v, InnerProductSpace v) => v -> v -> v

normalize :: (Fractional (Scalar a), NormedSpace a) => a -> a

vectorLength :: (Floating (Scalar m), InnerProductSpace m) => m -> Scalar m

divide :: (Fractional (Scalar v), VectorSpace v) => v -> Scalar v -> v

vaverage :: (Num v, Fractional (Scalar v), VectorSpace v) => [v] -> v

toScalarList :: (StandardBasis m, InnerProductSpace m) => m -> [Scalar m]

fromScalarList :: (VectorSpace a, StandardBasis a) => [Scalar a] -> a

toListCS :: CoordinateSpace v => v -> [Scalar v]

toListCS2 :: (CoordinateSpace m, CoordinateSpace (Scalar m)) => m -> [[Scalar (Scalar m)]]

coordinates2 :: (CoordinateSpace m, CoordinateSpace (Scalar m)) => m -> [[(Coordinate m, Coordinate (Scalar m))]]

basisOf :: StandardBasis m => Basis m

listMatrix :: (CoordinateSpace n, CoordinateSpace (Scalar n)) => [[Scalar (Scalar n)]] -> n

(%*%) :: SupportsMatrixMultiplication f g h a => (f :*: g) a -> (g :*: h) a -> (f :*: h) (Scalar (g a))

generalized implementation of matrix multiplication see http://en.wikipedia.org/wiki/Matrix_multiplication

(%**%) :: (SupportsMatrixMultiplication f g h a, Linearizable arr (:*:) f h a, Linearizable arr (:*:) f g a, Linearizable arr (:*:) g h a) => arr (f a) (g a) -> arr (g a) (h a) -> arr (f a) (h a)

type MatrixNorm arr h m a = (LinearTraceable arr m (Scalar (h a)), InnerProductSpace (h a), ConjugateSymmetric a, Transposable h m a)

In this version, we must assume VectorSpaceOver (h a) a constraint, but the result type is nicer.

(%^%) :: (SupportsMatrixMultiplication f f f a, Diagonalizable f (f a), Diagonalizable f a) => (f :*: f) a -> Integer -> (f :*: f) a

(|><|) :: (Functor m, Functor n, InnerProductSpace a) => m a -> n a -> (m :*: n) (Scalar a)

identityCS :: (CoordinateSpace m, CoordinateSpace (Scalar m), Num (Scalar (Scalar m))) => (Int, Int) -> m

(%.%) :: (Num (Scalar m), CoordinateSpace m) => m -> m -> Scalar m infixl 7

basisCoordinates :: InnerProductSpace v => Basis v -> v -> [Scalar v]

coordinateSpaceFunctionMatrix :: (CoordinateSpace m, StandardBasis v) => (v -> Scalar m) -> m

functionMatrix :: Diagonalizable f b => (f b -> g b) -> (f :*: g) b

This is the linearity condition:

convex_combination :: (VectorSpace v, Fractional (Scalar v), Foldable t, Applicative t) => t (Scalar v) -> t v -> v

https://en.wikipedia.org/wiki/Convex_combination This computes \[f([a_0,a_1,...,a_n], [{\mathbf b}_0,{\mathbf b}_1,...,{\mathbf b}_n]) = {{\sum_{j=0}^n{a_j{\mathbf b}_j}} \over \sum_{i=0}^n{a_i}}\]

vec2Cast :: (a :~: b) -> (a :~: b) -> a :~: b

hilbertSpace :: (Num m, NormedSpace m) => m -> m -> Scalar m

lieAdjoint :: LieAlgebra v => v -> Endo v

diagonalProjectionsList :: Monoid a => [Index List a]

isOnUnitCircle :: (NormedSpace v, Eq (Scalar v)) => v -> Bool

https://en.wikipedia.org/wiki/Norm_(mathematics)

isInsideUnitCircle :: (NormedSpace v, Ord (Scalar v)) => v -> Bool

https://en.wikipedia.org/wiki/Norm_(mathematics)

leftUnitor :: (I :*: f) a -> f a

rightUnitor :: Functor f => (f :*: I) a -> f a

associator :: Functor f => ((f :*: g) :*: h) a -> (f :*: (g :*: h)) a

unassociator :: Functor f => (f :*: (g :*: h)) a -> ((f :*: g) :*: h) a

diagonalProjectionsComplex :: Num a => Complex (Index Complex a)

(<!>) :: (Functor f, Functor g) => (g :*: f) a -> (g c -> b, f a -> c) -> b

reduceI :: (I :*: I) a -> I a

sumCoordinates :: (Foldable t, Num a) => t a -> a

joinMatrix :: (Monad g, Monad f, forall b. Transposable g f b) => (f :*: g) ((f :*: g) a) -> (f :*: g) a

type Projection f = forall a. f a -> a

type ElementRemovals f g g' = forall a. f (g a -> g' a)

data Determinance f g f' g' gg a' a res

Constructors

Determinance

Fields

determinanceRemoves1 :: g (g a -> gg a)

determinanceRemoves2 :: g (f (gg a) -> f' (g' a'))

determinanceCoordProj1 :: Projection f

determinanceCoordProj2 :: Projection g

determinanceCombine :: g a -> res

determinanceNestedDeterminant :: (f' :*: g') a' -> a

genericDeterminant :: (Functor f, Applicative g, Num a) => Determinance f g f' g' gg a' a res -> (f :*: g) a -> res

Orphan instances

PpShowVerticalF Complex

Instance details

Methods

ppfVertical :: PpShow a => Complex a -> Doc