Math.Matrix.SIMD

Description

This modules provides SIMD optimized versions of vectors. Relies heavily on GHC SIMD support. SIMD works only when using LLVM and GHC 8.x The only combinations of vectors supported are 4 * Int32, 8 * Int16 and 16 * Int8, 4*Float, 2*Double Conversions to and from optimized types and Vector4 Int32, (Vector2 :*: Vector2) Int32, (Vector4 Int16, Vector4 Int16) Vector4 Float, (Vector2 :*: Vector2) Float, and (Vector4 :*: Vector4) Int8 can be performed as appropriate for each type.

Documentation

data SVec4 a

data SVec8 a

data SVec16 a

data FVec4 a

data FVec2 a

class Optimal a where

Minimal complete definition

toO, fromO, zipO, mapO, constantO

Methods

toO :: a -> Optimized a

fromO :: Optimized a -> a

zipO :: (Scalar a -> Scalar a -> Scalar a) -> Optimized a -> Optimized a -> Optimized a

mapO :: (Scalar a -> Scalar a) -> Optimized a -> Optimized a

constantO :: Scalar a -> Optimized a

type family Optimized a where ...

Equations

Optimized (Vector4 Int32) = SVec4 Int32

Optimized (Vector4 Float) = FVec4 Float

Optimized (Vector2 Double) = FVec2 Double

Optimized (Vector4 Int16, Vector4 Int16) = SVec8 Int16

Optimized ((Vector2 :*: Vector4) Int16) = SVec8 Int16

Optimized ((Vector4 :*: Vector2) Int16) = SVec8 Int16

Optimized ((Vector4 :*: Vector4) Int8) = SVec16 Int8

Optimized ((Vector2 :*: Vector2) Float) = FVec4 Float

makeSVec4 :: Int32 -> Int32 -> Int32 -> Int32 -> SVec4 Int32

makeSVec8 :: Int16 -> Int16 -> Int16 -> Int16 -> Int16 -> Int16 -> Int16 -> Int16 -> SVec8 Int16

makeSVec16 :: Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> SVec16 Int8

makeFVec4 :: Float -> Float -> Float -> Float -> FVec4 Float

fromInt32Vector4 :: Vector4 Int32 -> SVec4 Int32

toInt32Vector4 :: SVec4 Int32 -> Vector4 Int32

to2x2Matrix :: SVec4 Int32 -> (Vector2 :*: Vector2) Int32

fromInt16Vector8 :: Vector4 Int16 -> Vector4 Int16 -> SVec8 Int16

toInt16Vector8 :: SVec8 Int16 -> (Vector4 Int16, Vector4 Int16)

fromInt8Vector16 :: SVec16 Int8 -> (Vector4 :*: Vector4) Int8

toInt8Vector16 :: (Vector4 :*: Vector4) Int8 -> SVec16 Int8

fromDoubleVector2 :: Vector2 Double -> FVec2 Double

toDoubleVector2 :: FVec2 Double -> Vector2 Double

fromFloatVector4 :: Vector4 Float -> FVec4 Float

toFloatVector4 :: FVec4 Float -> Vector4 Float

from2x2FloatVector4 :: (Vector2 :*: Vector2) Float -> FVec4 Float

to2x2FloatVector4 :: FVec4 Float -> (Vector2 :*: Vector2) Float

transposeFloatVector4 :: FVec4 Float -> FVec4 Float

eqSVec4 :: SVec4 Int32 -> SVec4 Int32 -> Bool

eqSVec8 :: SVec8 Int16 -> SVec8 Int16 -> Bool

eqSVec16 :: SVec16 Int8 -> SVec16 Int8 -> Bool

zipSVec4 :: (Int32 -> Int32 -> Int32) -> SVec4 Int32 -> SVec4 Int32 -> SVec4 Int32

zipSVec8 :: (Int16 -> Int16 -> Int16) -> SVec8 Int16 -> SVec8 Int16 -> SVec8 Int16

zipSVec16 :: (Int8 -> Int8 -> Int8) -> SVec16 Int8 -> SVec16 Int8 -> SVec16 Int8

zipFVec4 :: (Float -> Float -> Float) -> FVec4 Float -> FVec4 Float -> FVec4 Float

mapSVec4 :: (Int32 -> Int32) -> SVec4 Int32 -> SVec4 Int32

mapSVec8 :: (Int16 -> Int16) -> SVec8 Int16 -> SVec8 Int16

mapSVec16 :: (Int8 -> Int8) -> SVec16 Int8 -> SVec16 Int8

mapFVec4 :: (Float -> Float) -> FVec4 Float -> FVec4 Float

constantSVec4 :: Int32 -> SVec4 Int32

constantSVec8 :: Int16 -> SVec8 Int16

constantSVec16 :: Int8 -> SVec16 Int8

constantFVec4 :: Float -> FVec4 Float

fvec4_x :: FVec4 Float -> Float

fvec4_y :: FVec4 Float -> Float

fvec4_z :: FVec4 Float -> Float

fvec4_t :: FVec4 Float -> Float

svec4_x :: SVec4 Int32 -> Int32

svec4_y :: SVec4 Int32 -> Int32

svec4_z :: SVec4 Int32 -> Int32

svec4_t :: SVec4 Int32 -> Int32

svec8_1 :: SVec8 Int16 -> Int16

svec8_2 :: SVec8 Int16 -> Int16

svec8_3 :: SVec8 Int16 -> Int16

svec8_4 :: SVec8 Int16 -> Int16

svec8_5 :: SVec8 Int16 -> Int16

svec8_6 :: SVec8 Int16 -> Int16

svec8_7 :: SVec8 Int16 -> Int16

svec8_8 :: SVec8 Int16 -> Int16

svec16_1 :: SVec16 Int8 -> Int8

svec16_2 :: SVec16 Int8 -> Int8

svec16_3 :: SVec16 Int8 -> Int8

svec16_4 :: SVec16 Int8 -> Int8

svec16_5 :: SVec16 Int8 -> Int8

svec16_6 :: SVec16 Int8 -> Int8

svec16_7 :: SVec16 Int8 -> Int8

svec16_8 :: SVec16 Int8 -> Int8

svec16_9 :: SVec16 Int8 -> Int8

svec16_10 :: SVec16 Int8 -> Int8

svec16_11 :: SVec16 Int8 -> Int8

svec16_12 :: SVec16 Int8 -> Int8

svec16_13 :: SVec16 Int8 -> Int8

svec16_14 :: SVec16 Int8 -> Int8

svec16_15 :: SVec16 Int8 -> Int8

svec16_16 :: SVec16 Int8 -> Int8