Math.Matrix.Simple

Description

This module provides matrices with indices. Matrices are constructed using smatrix and svec operations, example:

let m = smatrix (*) (svec3 1 2 3) (svec4 5 4 3 2) :: (ThreeD :&: FourD) Int
== 5   4 3 2
   10  8 6 4
   15 12 9 6

indexing occurs with the (#) operator, for example, m # (1,0) == 10.

The number after svec is the number of elements in the dimension along row or column. Indices start from zero. The binary operation is applied to each pair of elements to construct the matrix element.

Matrices can also be constructed using Matrix constructor, with indices as input, e.g.:

identity :: (ThreeD :&: ThreeD) Integer

identity = Matrix $ \i j -> if i == j then 1 else 0

Note that index types are specifically restricted to be small types, since the Universe type class operations are often used to enumerate all elements of the type. This is enforced in show operations, which only show ten elements per dimension. For convenience, pre-defined index types, OneD, TwoD, ThreeD, FourD, FiveD are defined, for up to 5 dimensional matrices.

Documentation

type (:&:) row col elem = (row -> elem) :-> (col -> elem)

type Matrix3S row col dep elem = (row -> dep -> elem) :-> (col :&: dep) elem

type Matrix4S row col dep tim elem = (row -> (dep :&: tim) elem) :-> Matrix3S col dep tim elem

type MatrixA arr row col elem = arr row elem :-> arr col elem

type SMatrix1 elem = (OneD :&: OneD) elem

type SMatrix2 elem = (TwoD :&: TwoD) elem

type SMatrix3 elem = (ThreeD :&: ThreeD) elem

type SMatrix4 elem = (FourD :&: FourD) elem

type InfMatrix elem = (Integer :&: Integer) elem

type RatioMatrix elem = (Rational :&: Rational) elem

mapDimensions :: (Universe a, Universe b, Integral a, ConjugateSymmetric c, Num c) => ((->) a :~> f) -> ((->) b :~> g) -> (a :&: b) c -> (f :*: g) c

simpleMatrix :: (Integral a, Universe b, Universe a, ConjugateSymmetric c, Num c, Functor m, Functor n) => (a :&: b) c -> m a -> n b -> (m :*: n) c

index_delta :: (Eq ind, Num a) => ind -> a -> (ind -> a) -> ind -> a

partial_derivate_ind :: (Integral ind, Universe ind, Eq ind, ConjugateSymmetric a, Closed a, Infinitesimal Stream a) => ind -> Dual (ind -> a) -> Dual (ind -> a)

partial_derivate_list :: (Universe a, Eq a, Integral a, ConjugateSymmetric b, Closed b, Infinitesimal Stream b) => [Dual (a -> b) -> Dual (a -> b)]

divergence_index :: (ConjugateSymmetric b, Closed b, Infinitesimal Stream b, Integral ind, Eq ind, Universe a, Universe ind) => ((ind -> b) :-> (a -> b)) -> Dual (ind -> b)

grad_index :: (Scalar (ind -> a) ~ a, Integral ind, Universe ind, ConjugateSymmetric a, Closed a, Infinitesimal Stream a, Eq ind) => Dual (ind -> a) -> (ind -> a) :-> (ind -> a)

laplace_index :: (ConjugateSymmetric a, Closed a, Infinitesimal Stream a, Integral ind, Eq ind, Universe ind) => Dual (ind -> a) -> Dual (ind -> a)

sumS :: (Num b, Universe a) => (a -> b) -> b

productS :: (Num b, Universe a) => (a -> b) -> b

cov_index :: (Integral a, Universe a, Num (Scalar a), ConjugateSymmetric (Scalar a)) => a -> Dual (a -> Scalar a)

kronecker_delta :: (Eq a, Num b) => a -> a -> b

https://en.wikipedia.org/wiki/Kronecker_delta

(%***%) :: (Integral row, Universe row, Universe col, Universe mid, SupportsMatrixMultiplication ((->) row) ((->) col) ((->) mid) a) => (row :&: mid) a -> (mid :&: col) a -> (row :&: col) a

class CoFactorDimension dim where

Methods

cofactorDim :: (Num a, ConjugateSymmetric a) => dim -> (dim :&: dim) a -> a

leviCivita2S :: Num a => ((->) TwoD :*: (->) TwoD) a

https://en.wikipedia.org/wiki/Levi-Civita_symbol Note indices are from 0..(n-1).

leviCivita3S :: (((->) ThreeD :: Type -> Type) :*: (((->) ThreeD :: Type -> Type) :*: ((->) ThreeD :: Type -> Type))) Integer

https://en.wikipedia.org/wiki/Levi-Civita_symbol Note indices are from 0..(n-1).

leviCivita4S :: (((->) FourD :: Type -> Type) :*: (((->) FourD :: Type -> Type) :*: (((->) FourD :: Type -> Type) :*: ((->) FourD :: Type -> Type)))) Integer

https://en.wikipedia.org/wiki/Levi-Civita_symbol Note indices are from 0..(n-1).

leviCivitaS :: (Num b, Enum a, Universe a) => (a -> b) -> b

https://en.wikipedia.org/wiki/Levi-Civita_symbol

leviCivita :: Num a => [a] -> a

https://en.wikipedia.org/wiki/Levi-Civita_symbol Note indices are from 0..(n-1).

crossProductS :: (Integral a, Num b, Universe a) => (a -> b) -> (a -> b) -> a -> b

https://en.wikipedia.org/wiki/Cross_product

crossProduct :: (Integral a, Monoid a) => [a] -> [a] -> [a]

https://en.wikipedia.org/wiki/Cross_product

smatrix :: (Universe col, Integral row, Universe row, ConjugateSymmetric c, Num c) => (a -> b -> c) -> (row -> a) -> (col -> b) -> (row :&: col) c

swapRows :: (Num a, ConjugateSymmetric a, Eq row, Integral row, Universe row, Universe col) => row -> row -> (row :&: col) a -> (row :&: col) a

mulRow :: (Eq row, ConjugateSymmetric a, Num a, Integral row, Universe row, Universe col) => a -> row -> (row :&: col) a -> (row :&: col) a

swapIndex :: Eq a => a -> a -> a -> a

smatrixA :: (Arrow arr, Diagonalizable (arr row) c, Linearizable LinearMap (:*:) (arr row) (arr col) c, LinearTransform (arr row) (arr col) c) => arr (a, b) c -> arr row a -> arr col b -> MatrixA arr row col c

smatrix3 :: (a -> b -> c -> d) -> (row -> a) -> (col -> b) -> (depth -> c) -> ((->) row :*: ((->) col :*: (->) depth)) d

function_smatrix :: (Universe a, Universe b, Integral a, ConjugateSymmetric c, Num c) => (a -> b -> c) -> (a :&: b) c

smatrix_inverse_image :: (ConjugateSymmetric a, Num a, Integral row, Integral row', Universe col', Universe row', Universe row, Universe col) => ((row, col) -> (row', col')) -> (row' :&: col') a -> (row :&: col) a

inverse_image_smatrix :: (Integral row', Integral row, Universe row, ConjugateSymmetric a, Num a, Universe row', Universe col', Universe col) => (row -> row', col -> col') -> (row' :&: col') a -> (row :&: col) a

sapply :: (Integral (a -> b), ConjugateSymmetric b, Num b, Eq a, Universe a, Universe b) => ((a -> b) :&: a) b

(#) :: (ConjugateSymmetric a, Integral row, Num a, Universe row, Universe col) => (row :&: col) a -> (row, col) -> a

indexA :: (ArrowApply arr, Diagonalizable (arr row) a, Linearizable LinearMap (:*:) (arr row) (arr col) a, LinearTransform (arr row) (arr col) a) => MatrixA arr row col a -> arr (row, col) a

sdeterminant2 :: (ConjugateSymmetric a, Num a) => (TwoD :&: TwoD) a -> a

cofactor4_11 :: (ConjugateSymmetric a, Num a) => (FourD :&: FourD) a -> a

cofactor4_21 :: (ConjugateSymmetric a, Num a) => (FourD :&: FourD) a -> a

cofactor4_31 :: (ConjugateSymmetric a, Num a) => (FourD :&: FourD) a -> a

cofactor4_41 :: (ConjugateSymmetric a, Num a) => (FourD :&: FourD) a -> a

cofactor3_11 :: (ConjugateSymmetric a, Num a) => (ThreeD :&: ThreeD) a -> a

cofactor3_21 :: (ConjugateSymmetric a, Num a) => (ThreeD :&: ThreeD) a -> a

cofactor3_31 :: (ConjugateSymmetric a, Num a) => (ThreeD :&: ThreeD) a -> a

cofactor2_11 :: (ConjugateSymmetric a, Num a) => (TwoD :&: TwoD) a -> a

cofactor2_21 :: (ConjugateSymmetric a, Num a) => (TwoD :&: TwoD) a -> a

cofactorDim2 :: (ConjugateSymmetric a, Num a) => TwoD -> (TwoD :&: TwoD) a -> a

cofactorDim3 :: (ConjugateSymmetric a, Num a) => ThreeD -> (ThreeD :&: ThreeD) a -> a

cofactorDim4 :: (ConjugateSymmetric a, Num a) => FourD -> (FourD :&: FourD) a -> a

sdeterminant :: (ConjugateSymmetric a, Num a, Integral dim, Universe dim, CoFactorDimension dim) => (dim :&: dim) a -> a

sdeterminant3 :: (Num a, ConjugateSymmetric a) => (ThreeD :&: ThreeD) a -> a

sdeterminant4 :: (Num a, ConjugateSymmetric a) => (FourD :&: FourD) a -> a

svec_inf :: Stream a -> Integer -> a

srotate_x :: (Floating a, ConjugateSymmetric a) => a -> SMatrix3 a

Orphan instances

(Universe col, Num a) => LinearTransform Vector4 ((->) col :: Type -> Type) a

Instance details

Methods

(<*>>) :: (col -> a) -> (Vector4 :*: (->) col) a -> Vector4 a

(<<*>) :: (Vector4 :*: (->) col) a -> Vector4 a -> col -> a

(Num a, Universe col) => LinearTransform Vector3 ((->) col :: Type -> Type) a

Instance details

Methods

(<*>>) :: (col -> a) -> (Vector3 :*: (->) col) a -> Vector3 a

(<<*>) :: (Vector3 :*: (->) col) a -> Vector3 a -> col -> a

(Num a, Universe col) => LinearTransform Vector2 ((->) col :: Type -> Type) a

Instance details

Methods

(<*>>) :: (col -> a) -> (Vector2 :*: (->) col) a -> Vector2 a

(<<*>) :: (Vector2 :*: (->) col) a -> Vector2 a -> col -> a

(Num a, Universe col) => LinearTransform Vector1 ((->) col :: Type -> Type) a

Instance details

Methods

(<*>>) :: (col -> a) -> (Vector1 :*: (->) col) a -> Vector1 a

(<<*>) :: (Vector1 :*: (->) col) a -> Vector1 a -> col -> a

Num a => DecomposableVectorSpace (Vector4 a) ((->) FourD :: Type -> Type)

Instance details

Methods

decompose :: (Scalar (Vector4 a) -> res) -> Vector4 a -> FourD -> res

project :: Vector4 a -> FourD -> Scalar (Vector4 a)

Num a => DecomposableVectorSpace (Vector3 a) ((->) ThreeD :: Type -> Type)

Instance details

Methods

decompose :: (Scalar (Vector3 a) -> res) -> Vector3 a -> ThreeD -> res

project :: Vector3 a -> ThreeD -> Scalar (Vector3 a)

Num a => DecomposableVectorSpace (Vector2 a) ((->) TwoD :: Type -> Type)

Instance details

Methods

decompose :: (Scalar (Vector2 a) -> res) -> Vector2 a -> TwoD -> res

project :: Vector2 a -> TwoD -> Scalar (Vector2 a)

Num a => DecomposableVectorSpace (Vector1 a) ((->) OneD :: Type -> Type)

Instance details

Methods

decompose :: (Scalar (Vector1 a) -> res) -> Vector1 a -> OneD -> res

project :: Vector1 a -> OneD -> Scalar (Vector1 a)

Num a => DecomposableVectorSpace (Stream a) ((->) Integer :: Type -> Type)

Instance details

Methods

decompose :: (Scalar (Stream a) -> res) -> Stream a -> Integer -> res

project :: Stream a -> Integer -> Scalar (Stream a)

Num a => Num (FourD -> a)

Instance details

Methods

(+) :: (FourD -> a) -> (FourD -> a) -> FourD -> a

(-) :: (FourD -> a) -> (FourD -> a) -> FourD -> a

(*) :: (FourD -> a) -> (FourD -> a) -> FourD -> a

negate :: (FourD -> a) -> FourD -> a

abs :: (FourD -> a) -> FourD -> a

signum :: (FourD -> a) -> FourD -> a

fromInteger :: Integer -> FourD -> a

Num a => Num (ThreeD -> a)

Instance details

Methods

(+) :: (ThreeD -> a) -> (ThreeD -> a) -> ThreeD -> a

(-) :: (ThreeD -> a) -> (ThreeD -> a) -> ThreeD -> a

(*) :: (ThreeD -> a) -> (ThreeD -> a) -> ThreeD -> a

negate :: (ThreeD -> a) -> ThreeD -> a

abs :: (ThreeD -> a) -> ThreeD -> a

signum :: (ThreeD -> a) -> ThreeD -> a

fromInteger :: Integer -> ThreeD -> a

Num a => Num (TwoD -> a)

Instance details

Methods

(+) :: (TwoD -> a) -> (TwoD -> a) -> TwoD -> a

(-) :: (TwoD -> a) -> (TwoD -> a) -> TwoD -> a

(*) :: (TwoD -> a) -> (TwoD -> a) -> TwoD -> a

negate :: (TwoD -> a) -> TwoD -> a

abs :: (TwoD -> a) -> TwoD -> a

signum :: (TwoD -> a) -> TwoD -> a

fromInteger :: Integer -> TwoD -> a

Num a => Num (OneD -> a)

Instance details

Methods

(+) :: (OneD -> a) -> (OneD -> a) -> OneD -> a

(-) :: (OneD -> a) -> (OneD -> a) -> OneD -> a

(*) :: (OneD -> a) -> (OneD -> a) -> OneD -> a

negate :: (OneD -> a) -> OneD -> a

abs :: (OneD -> a) -> OneD -> a

signum :: (OneD -> a) -> OneD -> a

fromInteger :: Integer -> OneD -> a

(Floating b, Universe a) => NormedSpace (a -> b)

Instance details

Methods

norm :: (a -> b) -> Scalar (a -> b)

(Integral a, Universe a, Num b, ConjugateSymmetric b) => Dualizable (a -> b) Dual

Instance details

Methods

covector :: ((a -> b) -> Scalar (a -> b)) -> Dual (a -> b)

bracket :: Dual (a -> b) -> (a -> b) -> Scalar (a -> b)

(b ~ Scalar a, Scalar (a -> b) ~ b, Integral a, VectorSpace b, ConjugateSymmetric b, Closed b, Infinitesimal Stream b, Eq a, Universe a) => VectorDerivative (a -> b) Dual LinearMap

Instance details

Methods

divergence :: LinearMap (a -> b) (a -> b) -> Dual (a -> b)

grad :: Dual (a -> b) -> LinearMap (a -> b) (a -> b)

directional_derivative :: (a -> b) -> Dual (a -> b) -> Dual (a -> b)

laplace :: Dual (a -> b) -> Dual (a -> b)

(Integral col, Floating elem, Scalar (col -> elem) ~ elem, VectorSpace elem, Fractional (Scalar elem), Integral row, Universe row, Universe col, ConjugateSymmetric elem) => Num ((row :&: col) elem)

Instance details

Methods

(+) :: (row :&: col) elem -> (row :&: col) elem -> (row :&: col) elem

(-) :: (row :&: col) elem -> (row :&: col) elem -> (row :&: col) elem

(*) :: (row :&: col) elem -> (row :&: col) elem -> (row :&: col) elem

negate :: (row :&: col) elem -> (row :&: col) elem

abs :: (row :&: col) elem -> (row :&: col) elem

signum :: (row :&: col) elem -> (row :&: col) elem

fromInteger :: Integer -> (row :&: col) elem

Universe a => PpShowVerticalF ((->) a :: Type -> Type)

Instance details

Methods

ppf_vertical :: PpShow a0 => (a -> a0) -> Doc

Universe a => PpShowF ((->) a :: Type -> Type)

Instance details

Methods

ppf :: PpShow a0 => (a -> a0) -> Doc

Comonad ((->) SixD :: Type -> Type)

Instance details

Methods

extract :: (SixD -> a) -> a

duplicate :: (SixD -> a) -> SixD -> (SixD -> a)

extend :: ((SixD -> a) -> b) -> (SixD -> a) -> SixD -> b

(=>>) :: (SixD -> a) -> ((SixD -> a) -> b) -> SixD -> b

(.>>) :: (SixD -> a) -> b -> SixD -> b

Comonad ((->) FiveD :: Type -> Type)

Instance details

Methods

extract :: (FiveD -> a) -> a

duplicate :: (FiveD -> a) -> FiveD -> (FiveD -> a)

extend :: ((FiveD -> a) -> b) -> (FiveD -> a) -> FiveD -> b

(=>>) :: (FiveD -> a) -> ((FiveD -> a) -> b) -> FiveD -> b

(.>>) :: (FiveD -> a) -> b -> FiveD -> b

Comonad ((->) FourD :: Type -> Type)

Instance details

Methods

extract :: (FourD -> a) -> a

duplicate :: (FourD -> a) -> FourD -> (FourD -> a)

extend :: ((FourD -> a) -> b) -> (FourD -> a) -> FourD -> b

(=>>) :: (FourD -> a) -> ((FourD -> a) -> b) -> FourD -> b

(.>>) :: (FourD -> a) -> b -> FourD -> b

Comonad ((->) ThreeD :: Type -> Type)

Instance details

Methods

extract :: (ThreeD -> a) -> a

duplicate :: (ThreeD -> a) -> ThreeD -> (ThreeD -> a)

extend :: ((ThreeD -> a) -> b) -> (ThreeD -> a) -> ThreeD -> b

(=>>) :: (ThreeD -> a) -> ((ThreeD -> a) -> b) -> ThreeD -> b

(.>>) :: (ThreeD -> a) -> b -> ThreeD -> b

Comonad ((->) TwoD :: Type -> Type)

Instance details

Methods

extract :: (TwoD -> a) -> a

duplicate :: (TwoD -> a) -> TwoD -> (TwoD -> a)

extend :: ((TwoD -> a) -> b) -> (TwoD -> a) -> TwoD -> b

(=>>) :: (TwoD -> a) -> ((TwoD -> a) -> b) -> TwoD -> b

(.>>) :: (TwoD -> a) -> b -> TwoD -> b

(Integral row, Integral col, Floating a, Universe row, Universe col, ConjugateSymmetric a) => NormedSpace ((((->) row :: Type -> Type) :*: ((->) col :: Type -> Type)) a)

Instance details

Methods

norm :: ((->) row :*: (->) col) a -> Scalar (((->) row :*: (->) col) a)

(VectorSpace a, Integral col, Integral row, Universe row, ConjugateSymmetric a, Universe col, Floating a, Integral row, Scalar (col -> a) ~ a) => NormedSpace ((row :&: col) a)

Instance details

Methods

norm :: (row :&: col) a -> Scalar ((row :&: col) a)

(Integral row, Universe row, SupportsMatrixMultiplication ((->) row :: Type -> Type) ((->) row :: Type -> Type) ((->) row :: Type -> Type) a) => LieAlgebra ((row :&: row) a)

Instance details

Methods

(%<>%) :: (row :&: row) a -> (row :&: row) a -> (row :&: row) a

(Integral col, Integral row, Universe row, Universe col, Num a, ConjugateSymmetric a) => InnerProductSpace ((((->) row :: Type -> Type) :*: ((->) col :: Type -> Type)) a)
https://en.wikipedia.org/wiki/Frobenius_inner_product

Instance details

Methods

(%.) :: ((->) row :*: (->) col) a -> ((->) row :*: (->) col) a -> Scalar (((->) row :*: (->) col) a)

Num a => VectorSpace ((((->) row :: Type -> Type) :*: Vector4) a)

Instance details

Associated Types

type Scalar (((->) row :*: Vector4) a) :: Type

Methods

vzero :: ((->) row :*: Vector4) a

vnegate :: ((->) row :*: Vector4) a -> ((->) row :*: Vector4) a

(%+) :: ((->) row :*: Vector4) a -> ((->) row :*: Vector4) a -> ((->) row :*: Vector4) a

(%*) :: Scalar (((->) row :*: Vector4) a) -> ((->) row :*: Vector4) a -> ((->) row :*: Vector4) a

Num a => VectorSpace ((((->) row :: Type -> Type) :*: Vector3) a)

Instance details

Associated Types

type Scalar (((->) row :*: Vector3) a) :: Type

Methods

vzero :: ((->) row :*: Vector3) a

vnegate :: ((->) row :*: Vector3) a -> ((->) row :*: Vector3) a

(%+) :: ((->) row :*: Vector3) a -> ((->) row :*: Vector3) a -> ((->) row :*: Vector3) a

(%*) :: Scalar (((->) row :*: Vector3) a) -> ((->) row :*: Vector3) a -> ((->) row :*: Vector3) a

Num a => VectorSpace ((((->) row :: Type -> Type) :*: Vector2) a)

Instance details

Associated Types

type Scalar (((->) row :*: Vector2) a) :: Type

Methods

vzero :: ((->) row :*: Vector2) a

vnegate :: ((->) row :*: Vector2) a -> ((->) row :*: Vector2) a

(%+) :: ((->) row :*: Vector2) a -> ((->) row :*: Vector2) a -> ((->) row :*: Vector2) a

(%*) :: Scalar (((->) row :*: Vector2) a) -> ((->) row :*: Vector2) a -> ((->) row :*: Vector2) a

Num a => VectorSpace ((((->) row :: Type -> Type) :*: Vector1) a)

Instance details

Associated Types

type Scalar (((->) row :*: Vector1) a) :: Type

Methods

vzero :: ((->) row :*: Vector1) a

vnegate :: ((->) row :*: Vector1) a -> ((->) row :*: Vector1) a

(%+) :: ((->) row :*: Vector1) a -> ((->) row :*: Vector1) a -> ((->) row :*: Vector1) a

(%*) :: Scalar (((->) row :*: Vector1) a) -> ((->) row :*: Vector1) a -> ((->) row :*: Vector1) a

Num a => VectorSpace ((((->) row :: Type -> Type) :*: ((->) col :: Type -> Type)) a)

Instance details

Associated Types

type Scalar (((->) row :*: (->) col) a) :: Type

Methods

vzero :: ((->) row :*: (->) col) a

vnegate :: ((->) row :*: (->) col) a -> ((->) row :*: (->) col) a

(%+) :: ((->) row :*: (->) col) a -> ((->) row :*: (->) col) a -> ((->) row :*: (->) col) a

(%*) :: Scalar (((->) row :*: (->) col) a) -> ((->) row :*: (->) col) a -> ((->) row :*: (->) col) a

Num a => VectorSpace ((Vector4 :*: ((->) col :: Type -> Type)) a)

Instance details

Associated Types

type Scalar ((Vector4 :*: (->) col) a) :: Type

Methods

vzero :: (Vector4 :*: (->) col) a

vnegate :: (Vector4 :*: (->) col) a -> (Vector4 :*: (->) col) a

(%+) :: (Vector4 :*: (->) col) a -> (Vector4 :*: (->) col) a -> (Vector4 :*: (->) col) a

(%*) :: Scalar ((Vector4 :*: (->) col) a) -> (Vector4 :*: (->) col) a -> (Vector4 :*: (->) col) a

Num a => VectorSpace ((Vector3 :*: ((->) col :: Type -> Type)) a)

Instance details

Associated Types

type Scalar ((Vector3 :*: (->) col) a) :: Type

Methods

vzero :: (Vector3 :*: (->) col) a

vnegate :: (Vector3 :*: (->) col) a -> (Vector3 :*: (->) col) a

(%+) :: (Vector3 :*: (->) col) a -> (Vector3 :*: (->) col) a -> (Vector3 :*: (->) col) a

(%*) :: Scalar ((Vector3 :*: (->) col) a) -> (Vector3 :*: (->) col) a -> (Vector3 :*: (->) col) a

Num a => VectorSpace ((Vector2 :*: ((->) col :: Type -> Type)) a)

Instance details

Associated Types

type Scalar ((Vector2 :*: (->) col) a) :: Type

Methods

vzero :: (Vector2 :*: (->) col) a

vnegate :: (Vector2 :*: (->) col) a -> (Vector2 :*: (->) col) a

(%+) :: (Vector2 :*: (->) col) a -> (Vector2 :*: (->) col) a -> (Vector2 :*: (->) col) a

(%*) :: Scalar ((Vector2 :*: (->) col) a) -> (Vector2 :*: (->) col) a -> (Vector2 :*: (->) col) a

Num a => VectorSpace ((Vector1 :*: ((->) col :: Type -> Type)) a)

Instance details

Associated Types

type Scalar ((Vector1 :*: (->) col) a) :: Type

Methods

vzero :: (Vector1 :*: (->) col) a

vnegate :: (Vector1 :*: (->) col) a -> (Vector1 :*: (->) col) a

(%+) :: (Vector1 :*: (->) col) a -> (Vector1 :*: (->) col) a -> (Vector1 :*: (->) col) a

(%*) :: Scalar ((Vector1 :*: (->) col) a) -> (Vector1 :*: (->) col) a -> (Vector1 :*: (->) col) a

(Eq dim, Num dim, Integral dim, Universe dim, CoFactorDimension dim, ConjugateSymmetric a, Num a) => Traceable ((->) dim :: Type -> Type) a

Instance details

Methods

trace_impl :: ((->) dim :*: (->) dim) a -> a

determinant_impl :: ((->) dim :*: (->) dim) a -> a

(Num a, Eq dim, Integral dim) => Diagonalizable ((->) dim :: Type -> Type) a

Instance details

Methods

vector_dimension :: (dim -> a) -> dim -> Integer

identity_impl :: (dim -> Integer) -> ((->) dim :*: (->) dim) a

identity :: ((->) dim :*: (->) dim) a

diagonal_impl :: ((->) dim :*: (->) dim) a -> dim -> a

diagonal_matrix_impl :: (dim -> a) -> ((->) dim :*: (->) dim) a

(Num a, Integral row) => Indexable ((->) row :: Type -> Type) a

Instance details

Methods

diagonal_projections :: row -> Index ((->) row) a

basis_vector :: Index ((->) row) a -> row -> a

index_project :: Index ((->) row) a -> (row -> a) -> a

indexable_indices :: row -> a

(Universe row, Num a) => LinearTransform ((->) row :: Type -> Type) Vector4 a

Instance details

Methods

(<*>>) :: Vector4 a -> ((->) row :*: Vector4) a -> row -> a

(<<*>) :: ((->) row :*: Vector4) a -> (row -> a) -> Vector4 a

(Universe row, Num a) => LinearTransform ((->) row :: Type -> Type) Vector3 a

Instance details

Methods

(<*>>) :: Vector3 a -> ((->) row :*: Vector3) a -> row -> a

(<<*>) :: ((->) row :*: Vector3) a -> (row -> a) -> Vector3 a

(Universe row, Num a) => LinearTransform ((->) row :: Type -> Type) Vector2 a

Instance details

Methods

(<*>>) :: Vector2 a -> ((->) row :*: Vector2) a -> row -> a

(<<*>) :: ((->) row :*: Vector2) a -> (row -> a) -> Vector2 a

(Universe row, Num a) => LinearTransform ((->) row :: Type -> Type) Vector1 a

Instance details

Methods

(<*>>) :: Vector1 a -> ((->) row :*: Vector1) a -> row -> a

(<<*>) :: ((->) row :*: Vector1) a -> (row -> a) -> Vector1 a

(Num a, Universe row, Universe col) => LinearTransform ((->) row :: Type -> Type) ((->) col :: Type -> Type) a

Instance details

Methods

(<*>>) :: (col -> a) -> ((->) row :*: (->) col) a -> row -> a

(<<*>) :: ((->) row :*: (->) col) a -> (row -> a) -> col -> a