Math.Matrix.Vector3

Documentation

data Vector3 s

Three element vector

Constructors

Vector3

Fields

xcoord3 :: !s

ycoord3 :: !s

zcoord3 :: !s

type ComplexVector3 a = (Vector3 :*: Complex) a

dim3 :: Vector3 Integer

i3 :: Num a => Vector3 a

j3 :: Num a => Vector3 a

k3 :: Num a => Vector3 a

type Matrix3 a = (Vector3 :*: Vector3) a

vec3 :: (a, a, a) -> Vector3 a

finite3 :: Vector3 a -> Vec 3 a

constant3 :: a -> Vector3 a

vec3Unit :: a -> Scalar (Vector3 a)

vec3Counit :: Vector3 (Scalar (Vector3 a)) -> Vector3 a

destructvec3 :: Vector3 a -> (a, a, a)

setx3 :: s -> Vector3 s -> Vector3 s

sety3 :: s -> Vector3 s -> Vector3 s

setz3 :: s -> Vector3 s -> Vector3 s

setEndo :: Vector3 s -> Vector3 (Endo (Vector3 s))

setVectorAction :: Vector3 s -> Vector3 (Vector3 s -> Vector3 s)

updateRow3 :: g a -> Vector3 ((Vector3 :*: g) a -> (Vector3 :*: g) a)

update_column3 v is a vector of all update operations that replace one row with v

updateColumn3 :: Applicative f => f a -> Vector3 ((f :*: Vector3) a -> (f :*: Vector3) a)

update_column3 v is a vector of all update operations that replace one column with v example use: update_column3 (Vector3 3 4 5) ycoord3 identity3 == [[1,3,0],[0,4,0],[0,5,1]].

removex3 :: Vector3 a -> Vector2 a

removey3 :: Vector3 a -> Vector2 a

removez3 :: Vector3 a -> Vector2 a

remove3 :: Vector3 (Vector3 a -> Vector2 a)

determinancett3 :: Num a => Determinance Vector3 Vector3 Vector2 Vector2 Vector2 a a a

remove_index3 :: Matrix3 a -> Matrix3 (Matrix2 a)

cofactor3 :: Num a => Matrix3 a -> Matrix3 a

adjucate3 :: Num a => Matrix3 a -> Matrix3 a

https://en.wikipedia.org/wiki/Adjugate_matrix

inverse3 :: Fractional a => Matrix3 a -> Matrix3 a

https://en.wikipedia.org/wiki/Invertible_matrix

set3 :: Vector3 a -> Vector3 a -> (Vector3 :*: Vector3) a

swap3 :: (Vector3 :*: Vector3) (Vector3 a -> Vector3 a)

http://en.wikipedia.org/wiki/Elementary_matrix#Operations

mul3 :: VectorSpace s => Scalar s -> Vector3 (Vector3 s -> Vector3 s)

add3Matrix :: VectorSpace a => Scalar a -> ((Vector3 :*: Vector3) :*: Vector3) a -> ((Vector3 :*: Vector3) :*: Vector3) a

add3 :: VectorSpace s => Scalar s -> (Vector3 :*: Vector3) (Vector3 s -> Vector3 s)

swapxy3 :: Vector3 a -> Vector3 a

swapxz3 :: Vector3 a -> Vector3 a

swapyz3 :: Vector3 a -> Vector3 a

mul3Matrix :: (Num a, VectorSpace a) => Scalar a -> Vector3 ((Vector3 :*: Vector3) a)

signs3 :: Integral a => Matrix3 a

dx3_endo :: Num a => a -> Endo (Vector3 a)

dy3_endo :: Num a => a -> Endo (Vector3 a)

dz3_endo :: Num a => a -> Endo (Vector3 a)

dx3 :: Num a => a -> Vector3 a -> Vector3 a

dy3 :: Num a => a -> Vector3 a -> Vector3 a

dz3 :: Num a => a -> Vector3 a -> Vector3 a

op_dot :: Num b => Vector3 (a -> b) -> a -> b

applicative_matrix :: Num b => Vector3 (b -> b) -> (Vector3 :*: Vector3) b

sum_coordinates3 :: Num a => Vector3 a -> a

cube_norm3 :: Floating a => Vector3 a -> a

cube root of a sum of cubes.

nth_norm3 :: Floating a => a -> Vector3 a -> a

nth root of a sum of nth powers

index3 :: Vector3 (Vector3 a -> a)

mapBasis :: Numeric a => (Vector3 a -> Vector3 a) -> Vector3 a -> Vector3 a

functionMatrix3 :: Numeric a => (Vector3 a -> Vector3 a) -> Matrix3 a

splitx :: Vector3 s -> (Vector2 s, Vector1 s)

splity :: Vector3 s -> (Vector2 s, Vector1 s)

splitz :: Vector3 s -> (Vector2 s, Vector1 s)

trace3 :: Num a => Matrix3 a -> a

vector3 :: [a] -> Vector3 a

cross_product :: Num a => Vector3 a -> Vector3 a -> Vector3 a

https://en.wikipedia.org/wiki/Cross_product WARNING: a cross product produces an axial vector. https://en.wikipedia.org/wiki/Pseudovector

cross_product3 :: Num a => (Vector2 :*: Vector3) a -> Vector3 a

cross_product3 computes the three dimensional vector that is orthogonal to both rows of the 2x3 vector.

xid :: Num a => Vector3 a

yid :: Num a => Vector3 a

zid :: Num a => Vector3 a

identity3 :: Num a => Matrix3 a

matrixPower3 :: (Ord a, ConjugateSymmetric a, Numeric a) => Matrix3 a -> Integer -> Matrix3 a

vectorIndices3 :: Integral a => Vector3 a

matrixIndices3 :: Integral a => (Vector3 :*: Vector3) (a, a)

matrixMultiply3 :: (Ord a, ConjugateSymmetric a, Num a) => Matrix3 a -> Matrix3 a -> Matrix3 a

dot3 :: (Num a, ConjugateSymmetric a) => Vector3 a -> Vector3 a -> a

leftMultiply3Gen :: (Functor f, Num a, ConjugateSymmetric a) => Vector3 a -> (f :*: Vector3) a -> f a

rightMultiply3Gen :: (VectorSpace (f a), ConjugateSymmetric a, Scalar (f a) ~ a) => (Vector3 :*: f) a -> Vector3 a -> f a

leftMultiply3 :: (Num a, ConjugateSymmetric a) => Vector3 a -> Matrix3 a -> Vector3 a

rightMultiply3 :: (Num a, ConjugateSymmetric a) => Matrix3 a -> Vector3 a -> Vector3 a

transpose3 :: Num a => Matrix3 a -> Matrix3 a

transpose3Impl :: Matrix3 a -> Matrix3 a

outer3 :: (a -> b -> c) -> Vector3 a -> Vector3 b -> Matrix3 c

zipWithV3 :: (a -> b -> c) -> Vector3 a -> Vector3 b -> Vector3 c

rotate_x :: Floating a => a -> Matrix3 a

rotate_y :: Floating a => a -> Matrix3 a

rotate_z :: Floating a => a -> Matrix3 a

rotate_coordinates3 :: Vector3 a -> Vector3 a

zero_codiagonal3 :: Num a => Codiagonal Vector3 a

diagonal_matrix3 :: Vector3 a -> (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a

matrix3 :: Vector3 a -> Codiagonal Vector3 a -> (Vector3 :*: Vector3) a

codiag3 :: Matrix3 a -> Codiagonal Vector3 a

codiagonal3 :: Matrix3 a -> Vector2 (Vector2 a, Vector2 a)

diagonal3 :: Matrix3 a -> Vector3 a

diagonal_projections3 :: Vector3 (Index Vector3 a)

diagonal3x :: Num a => Matrix3 a -> Vector3 a

diagonal3y :: Num a => Matrix3 a -> Vector3 a

diagonal3z :: Num a => Matrix3 a -> Vector3 a

determinant3 :: Num a => Matrix3 a -> a

algorithm from http://en.wikipedia.org/wiki/Determinant

determinant3_ :: Num a => Matrix3 a -> a

gram_schmidt3 :: (Fractional (Scalar a), Num a, InnerProductSpace a, VectorSpace a) => Vector3 a -> Vector3 a

eigenvalue3 :: (Floating a, Ord a, ConjugateSymmetric a) => (Vector3 :*: Vector3) a -> Vector3 a

https://en.wikipedia.org/wiki/Eigenvalue_algorithm based on algorithm by Smith, Oliver K.: Eigenvalues of symmetric 3x3 matrix, 1961, Communications of the ACM., doi:10.1145/355578.366316 The algorithm works reasonably when the matrix is real and symmetric.

levi_civita3 :: ((Vector3 :*: Vector3) :*: Vector3) Int

https://en.wikipedia.org/wiki/Levi-Civita_symbol

cycle3 :: StreamBuilder str => Vector3 a -> str a

Orphan instances

AppendableVector Vector1 Vector2

Instance details

Associated Types

type Vector1 :+: Vector2 :: Type -> Type

Methods

(||>>) :: Vector1 a -> Vector2 a -> (Vector1 :+: Vector2) a

AppendableVector Vector2 Vector1

Instance details

Associated Types

type Vector2 :+: Vector1 :: Type -> Type

Methods

(||>>) :: Vector2 a -> Vector1 a -> (Vector2 :+: Vector1) a

SplittableVector Vector1 Vector2

Instance details

Methods

vsplit :: (Vector1 :+: Vector2) a -> (Vector1 a, Vector2 a)

SplittableVector Vector2 Vector1

Instance details

Methods

vsplit :: (Vector2 :+: Vector1) a -> (Vector2 a, Vector1 a)