Math.Matrix.Vector4

Documentation

data Vector4 s

Constructors

Vector4

Fields

tcoord4 :: !s

xcoord4 :: !s

ycoord4 :: !s

zcoord4 :: !s

x4 :: Num a => Vector4 a

y4 :: Num a => Vector4 a

z4 :: Num a => Vector4 a

t4 :: Num a => Vector4 a

dx4 :: forall {str :: Type -> Type} {a}. (Infinitesimal str a, Num (Closure str a)) => Vector4 (Closure str a)

dy4 :: forall {str :: Type -> Type} {a}. (Infinitesimal str a, Num (Closure str a)) => Vector4 (Closure str a)

dz4 :: forall {str :: Type -> Type} {a}. (Infinitesimal str a, Num (Closure str a)) => Vector4 (Closure str a)

dt4 :: forall {str :: Type -> Type} {a}. (Infinitesimal str a, Num (Closure str a)) => Vector4 (Closure str a)

setx4 :: s -> Vector4 s -> Vector4 s

sety4 :: s -> Vector4 s -> Vector4 s

setz4 :: s -> Vector4 s -> Vector4 s

sett4 :: s -> Vector4 s -> Vector4 s

updateRow4 :: g a -> Vector4 ((Vector4 :*: g) a -> (Vector4 :*: g) a)

updateColumn4 :: Applicative f => f a -> Vector4 ((f :*: Vector4) a -> (f :*: Vector4) a)

type Matrix4 a = (Vector4 :*: Vector4) a

codiag4 :: (Vector4 :*: Vector4) a -> Codiagonal Vector4 a

matrix4 :: Vector4 a -> Codiagonal Vector4 a -> (Vector4 :*: Vector4) a

diagonalProjections4 :: Num a => Vector4 (Index Vector4 a)

cycle4 :: Vector4 a -> Stream a

vector4 :: [a] -> Vector4 a

unzipV4 :: Vector4 (a, b) -> (Vector4 a, Vector4 b)

constant4 :: a -> Vector4 a

dim4 :: Vector4 Integer

sumCoordinates4 :: Num a => Vector4 a -> a

leftMultiply4Gen :: (Functor f, Num a, ConjugateSymmetric a) => Vector4 a -> (f :*: Vector4) a -> f a

rightMultiply4Gen :: (VectorSpace (f a), ConjugateSymmetric a, Scalar (f a) ~ a) => (Vector4 :*: f) a -> Vector4 a -> f a

matrixMultiply4 :: (Num a, ConjugateSymmetric a) => Matrix4 a -> Matrix4 a -> Matrix4 a

dot4 :: (Num a, ConjugateSymmetric a) => Vector4 a -> Vector4 a -> a

versor :: (Floating a, ConjugateSymmetric a) => Vector4 a -> Vector4 a

index4 :: Int -> Vector4 a -> a

diagonal4 :: (Vector4 :*: Vector4) a -> Vector4 a

dt_4 :: Num s => s -> Vector4 s -> Vector4 s

dx_4 :: Num s => s -> Vector4 s -> Vector4 s

dy_4 :: Num s => s -> Vector4 s -> Vector4 s

dz_4 :: Num s => s -> Vector4 s -> Vector4 s

transpose4 :: Num a => Matrix4 a -> Matrix4 a

transpose4Impl :: Matrix4 a -> Matrix4 a

identity4 :: Num a => Matrix4 a

diagonal4x :: Matrix4 a -> Vector4 a

zero_codiagonal4 :: Num a => Codiagonal Vector4 a

diagonal_matrix4 :: Num a => Vector4 a -> (Vector4 :*: Vector4) a

mat4 :: Vector4 a -> Codiagonal Vector4 a -> (Vector4 :*: Vector4) a

trace4 :: Num a => Matrix4 a -> a

removet4 :: Vector4 a -> Vector3 a

removex4 :: Vector4 a -> Vector3 a

removey4 :: Vector4 a -> Vector3 a

removez4 :: Vector4 a -> Vector3 a

removes4 :: Vector4 (Vector4 a -> Vector3 a)

project4 :: Vector4 (Vector4 a -> a)

removeIndex4 :: Matrix4 a -> Matrix4 (Matrix3 a)

cofactor4 :: Num a => Matrix4 a -> Matrix4 a

adjucate4 :: Num a => Matrix4 a -> Matrix4 a

signs4 :: Integral a => Matrix4 a

vectorIndices4 :: Integral a => Vector4 a

matrixIndices4 :: Integral a => (Vector4 :*: Vector4) (a, a)

inverse4 :: Fractional a => Matrix4 a -> Matrix4 a

https://en.wikipedia.org/wiki/Invertible_matrix

cross4 :: Num a => Vector4 a -> Vector4 a -> Vector4 a -> Vector4 a

Generalization of cross product to four dimensions This is computed using formal determinant representation of cross product expanded to four dimensions https://en.wikipedia.org/wiki/Cross_product This computes vector that is linearly independent of all three vectors given in four dimensional space.

cross_product4 :: Num a => (Vector3 :*: Vector4) a -> Vector4 a

determinant4 :: Num a => Matrix4 a -> a

determinancett4 :: Num a => Determinance Vector4 Vector4 Vector3 Vector3 Vector3 a a a

coordDeterminance4 :: Num a => (forall c. Vector4 c -> c) -> (forall d. Vector4 d -> d) -> Determinance Vector4 Vector4 Vector3 Vector3 Vector3 a a a

Orphan instances

AppendableVector Vector1 Vector3

Instance details

Associated Types

type Vector1 :+: Vector3 :: Type -> Type

Methods

(||>>) :: Vector1 a -> Vector3 a -> (Vector1 :+: Vector3) a

AppendableVector Vector2 Vector2

Instance details

Associated Types

type Vector2 :+: Vector2 :: Type -> Type

Methods

(||>>) :: Vector2 a -> Vector2 a -> (Vector2 :+: Vector2) a

AppendableVector Vector3 Vector1

Instance details

Associated Types

type Vector3 :+: Vector1 :: Type -> Type

Methods

(||>>) :: Vector3 a -> Vector1 a -> (Vector3 :+: Vector1) a

SplittableVector Vector1 Vector3

Instance details

Methods

vsplit :: (Vector1 :+: Vector3) a -> (Vector1 a, Vector3 a)

SplittableVector Vector2 Vector2

Instance details

Methods

vsplit :: (Vector2 :+: Vector2) a -> (Vector2 a, Vector2 a)

SplittableVector Vector3 Vector1

Instance details

Methods

vsplit :: (Vector3 :+: Vector1) a -> (Vector3 a, Vector1 a)