Math.Number.Complex

Documentation

complex_sin :: Floating a => Complex a -> Complex a

https://en.wikipedia.org/wiki/Trigonometric_functions

complex_pi_ :: Floating a => Complex a

complex_sqrt :: RealFloat a => Complex a -> Complex a

compared to Prelude instance of Floating for Complex, this implementation doesn't attempt to compare scalars for equality so complex_sqrt (0 :+ 0) == (0 :+ NaN).

complex_cos :: Floating a => Complex a -> Complex a

https://en.wikipedia.org/wiki/Trigonometric_functions

complex_mul :: Num a => Complex a -> Complex a -> Complex a

Multiplication of complex numbers. Doesn't require RealFloat.

complex_exp :: Floating a => Complex a -> Complex a

using euler's formula https://en.wikipedia.org/wiki/Euler%27s_formula

complex_integral :: (Foldable t, Functor t) => (Rational -> t (Complex R)) -> (Complex R -> Complex R) -> Complex R

fourier :: (Foldable t, Functor t) => (Rational -> t (Complex R)) -> (Complex R -> Complex R) -> Complex R -> Complex R

type ComplexStream a = (Stream :*: Complex) a

Orphan instances

Show a => Show (ComplexStream a)

Instance details

Methods

showsPrec :: Int -> ComplexStream a -> ShowS

show :: ComplexStream a -> String

showList :: [ComplexStream a] -> ShowS

RealFloat a => Floating ((Stream :*: Complex) a)

Instance details

Methods

pi :: (Stream :*: Complex) a

exp :: (Stream :*: Complex) a -> (Stream :*: Complex) a

log :: (Stream :*: Complex) a -> (Stream :*: Complex) a

sqrt :: (Stream :*: Complex) a -> (Stream :*: Complex) a

(**) :: (Stream :*: Complex) a -> (Stream :*: Complex) a -> (Stream :*: Complex) a

logBase :: (Stream :*: Complex) a -> (Stream :*: Complex) a -> (Stream :*: Complex) a

sin :: (Stream :*: Complex) a -> (Stream :*: Complex) a

cos :: (Stream :*: Complex) a -> (Stream :*: Complex) a

tan :: (Stream :*: Complex) a -> (Stream :*: Complex) a

asin :: (Stream :*: Complex) a -> (Stream :*: Complex) a

acos :: (Stream :*: Complex) a -> (Stream :*: Complex) a

atan :: (Stream :*: Complex) a -> (Stream :*: Complex) a

sinh :: (Stream :*: Complex) a -> (Stream :*: Complex) a

cosh :: (Stream :*: Complex) a -> (Stream :*: Complex) a

tanh :: (Stream :*: Complex) a -> (Stream :*: Complex) a

asinh :: (Stream :*: Complex) a -> (Stream :*: Complex) a

acosh :: (Stream :*: Complex) a -> (Stream :*: Complex) a

atanh :: (Stream :*: Complex) a -> (Stream :*: Complex) a

log1p :: (Stream :*: Complex) a -> (Stream :*: Complex) a

expm1 :: (Stream :*: Complex) a -> (Stream :*: Complex) a

log1pexp :: (Stream :*: Complex) a -> (Stream :*: Complex) a

log1mexp :: (Stream :*: Complex) a -> (Stream :*: Complex) a

RealFloat a => Fractional ((Stream :*: Complex) a)

Instance details

Methods

(/) :: (Stream :*: Complex) a -> (Stream :*: Complex) a -> (Stream :*: Complex) a

recip :: (Stream :*: Complex) a -> (Stream :*: Complex) a

fromRational :: Rational -> (Stream :*: Complex) a

RealFloat a => Num ((Stream :*: Complex) a)

Instance details

Methods

(+) :: (Stream :*: Complex) a -> (Stream :*: Complex) a -> (Stream :*: Complex) a

(-) :: (Stream :*: Complex) a -> (Stream :*: Complex) a -> (Stream :*: Complex) a

(*) :: (Stream :*: Complex) a -> (Stream :*: Complex) a -> (Stream :*: Complex) a

negate :: (Stream :*: Complex) a -> (Stream :*: Complex) a

abs :: (Stream :*: Complex) a -> (Stream :*: Complex) a

signum :: (Stream :*: Complex) a -> (Stream :*: Complex) a

fromInteger :: Integer -> (Stream :*: Complex) a