Math.Number.Interface

Documentation

nthroot :: Floating a => a -> a -> a

logroot :: Floating a => a -> a -> a

class Ord q => DedekindCut q r where

Dedekind cut. Notice typically in constructive reals, it's possible to implement comparison between rational and real by utilizing denominator of the rational as describing desired accuracy but comparison between two reals is undecidable. Note, naming is chosen so the percentage mark is at side that can always be compared.

Minimal complete definition

(%<), (<%)

Methods

(%<) :: q -> r -> Bool

(<%) :: r -> q -> Bool

(%>) :: q -> r -> Bool

(>%) :: r -> q -> Bool

class Approximations a where

Methods

floating_approximations :: a -> Stream Double

rational_approximations :: a -> Stream Rational

class Infinitary a where

Methods

infinite :: a

class Infinitary a => PotentiallyInfinite a where

Methods

is_infinite :: a -> Bool

class Numerics a where

Methods

newtons_method :: (a -> a) -> a -> a

class RationalRoots a where

Methods

rational_power :: a -> Rational -> a

class Num r => DifferentiallyClosed r where

Methods

derivate :: (r -> r) -> r -> r

integral :: (r, r) -> (r -> r) -> r

class DifferentialOperator t where

Methods

partial :: DifferentiallyClosed a => (t a -> a) -> t a -> t a

partial1_2 :: DifferentiallyClosed a => (a -> b -> a) -> a -> b -> a

partial2_2 :: DifferentiallyClosed a => (b -> a -> a) -> b -> a -> a

partial1_3 :: DifferentiallyClosed a => (a -> b -> c -> a) -> a -> b -> c -> a

partial2_3 :: DifferentiallyClosed a => (b -> a -> c -> a) -> b -> a -> c -> a

partial3_3 :: DifferentiallyClosed a => (b -> c -> a -> a) -> b -> c -> a -> a

partial1_4 :: DifferentiallyClosed a => (a -> b -> c -> d -> a) -> a -> b -> c -> d -> a

partial2_4 :: DifferentiallyClosed a => (b -> a -> c -> d -> a) -> b -> a -> c -> d -> a

partial3_4 :: DifferentiallyClosed a => (b -> c -> a -> d -> a) -> b -> c -> a -> d -> a

partial4_4 :: DifferentiallyClosed a => (b -> c -> d -> a -> a) -> b -> c -> d -> a -> a

class Show r => ShowPrecision r where

Methods

show_at_precision :: r -> Integer -> String

differential :: DifferentiallyClosed r => (r -> r) -> r -> Endo r

https://en.wikipedia.org/wiki/Differential_form

derivate_commutator :: DifferentiallyClosed r => (r -> r) -> (r -> r) -> r -> r

computes \(f'(x)*g(x) - f(x)*g'(x)\)

derivate_anticommutator :: DifferentiallyClosed r => (r -> r) -> (r -> r) -> r -> r

computes \(f'(x)*g(x) + f(x)*g'(x)\). Notice this is product rule.

derivates :: DifferentiallyClosed r => (r -> r) -> Stream (r -> r)

taylor :: (Fractional a, DifferentiallyClosed a) => (a -> a) -> a -> (Stream :*: Stream) a

https://en.wikipedia.org/wiki/Taylor_series

atan2_generic :: Floating a => a -> a -> a

Atan2

cot :: Floating a => a -> a

https://en.wikipedia.org/wiki/Trigonometric_functions

sec :: Floating a => a -> a

https://en.wikipedia.org/wiki/Trigonometric_functions

csc :: Floating a => a -> a

https://en.wikipedia.org/wiki/Trigonometric_functions

arsinh :: Floating a => a -> a

https://en.wikipedia.org/wiki/Hyperbolic_function

arcosh :: Floating a => a -> a

https://en.wikipedia.org/wiki/Hyperbolic_function

artanh :: Floating a => a -> a

https://en.wikipedia.org/wiki/Hyperbolic_function

arcoth :: Floating a => a -> a

https://en.wikipedia.org/wiki/Hyperbolic_function

arsech :: Floating a => a -> a

https://en.wikipedia.org/wiki/Hyperbolic_function

arcsch :: Floating a => a -> a

https://en.wikipedia.org/wiki/Hyperbolic_function

coth :: Floating a => a -> a

https://en.wikipedia.org/wiki/Hyperbolic_function

sech :: Floating a => a -> a

https://en.wikipedia.org/wiki/Hyperbolic_function

csch :: Floating a => a -> a

https://en.wikipedia.org/wiki/Hyperbolic_function

line_integral :: DifferentiallyClosed r => (r -> r) -> (r -> r) -> (r, r) -> r

Line integral

precisionCompare :: (Ord a, Num a) => a -> a -> a -> Bool

compare to a certain precision, appropriate for floating point style numeric types. first argument is precision.