Math.Number.NumericExpression

Documentation

data ctx :- a where

Constructors

ELam :: (Var a -> (ctx, a) :- b) -> ctx :- (a -> b)

EApp :: (ctx :- (a -> b)) -> (ctx, a) :- b

ECompose :: (b :- c) -> (a :- b) -> a :- c

EVar :: Var a -> ctx :- a

EPair :: (ctx :- a) -> (ctx :- b) -> ctx :- (a, b)

EFst :: (ctx, a) :- ctx

ESnd :: (ctx, a) :- a

ECase :: (a :- c) -> (b :- c) -> Either a b :- c

EIn1 :: a :- Either a b

EIn2 :: b :- Either a b

ETerminal :: ctx :- ()

EInitial :: Contradiction :- b

EVec :: VectorSpaceExpr v a -> () :- v a

ENum :: NumExpr v a -> () :- a

data Contradiction

identity_expr :: ctx :- (b -> b)

pair_identity_expr :: (a, b) :- (a, b)

either_identity_expr :: Either a b :- Either a b

pair_expr :: ctx :- (a -> b -> (a, b))

var_intro :: (Var a -> (ctx, a) :- b) -> (ctx, a) :- b

data Var a

Constructors

Var

Fields

var_name :: !String

var_debruijn_index :: !Integer

Val !a

varIso :: IsoA Var a b -> Var a :==: Var b

class Expression e ctx a where

Methods

eval :: ctx a -> e a -> a

equal_in_context :: (Eq a, Expression f ctx a) => ctx a -> f a -> f a -> Bool

evalVar :: [a] -> Var a -> a

withCtx :: Functor f => [a] -> (f :*: Var) a -> f a

eVar :: Applicative f => String -> Integer -> (f :*: Var) a

expr :: Functor f => f a -> (f :*: Var) a

data VectorCtx v a

Constructors

VectorCtx

Fields

vectorSpaceCtx :: Map String (VectorSpaceExpr v a)

numCtx :: [a]

data VectorSpaceExpr v a

Constructors

VZero

VNegate (VectorSpaceExpr v a)

VPlus (VectorSpaceExpr v a) (VectorSpaceExpr v a)

VScalar (NumExpr v a) (VectorSpaceExpr v a)

VVectorVar (Var (VectorSpaceExpr v a))

VPrim (v a)

reduceVectorSpaceExpr :: VectorSpaceExpr v a -> VectorSpaceExpr v a

vectorSpaceIso :: FunctorArrow v (->) (->) => IsoA (VectorSpaceExpr v) a b -> VectorSpaceExpr v a :==: VectorSpaceExpr v b

evalVectorSpaceExpr :: (VectorSpace (v a), IdVisitor a, Expression Var (Map String) (VectorSpaceExpr v a), Scalar (v a) ~ a) => VectorCtx v a -> VectorSpaceExpr v a -> v a

data NumExpr v a

Basically same as methods in Num type class from prelude, but as data.

Constructors

Plus !(NumExpr v a) !(NumExpr v a)

Product !(NumExpr v a) !(NumExpr v a)

Subtract !(NumExpr v a) !(NumExpr v a)

Negate !(NumExpr v a)

Abs !(NumExpr v a)

Signum !(NumExpr v a)

NumPrim !(Var a)

FromInteger !Integer

InnerProduct !(VectorSpaceExpr v a) !(VectorSpaceExpr v a)

numIso :: FunctorArrow v (->) (->) => IsoA (NumExpr v) a b -> NumExpr v a :==: NumExpr v b

plus_assoc_iso :: NumExpr v a :==: NumExpr v a

product_assoc_iso :: NumExpr v a :==: NumExpr v a

data FracExpr v a

basically same as Fractional class in Prelude.

Constructors

PrimFracExpr !(NumExpr v a)

Divide !(FracExpr v a) !(FracExpr v a)

Recip !(FracExpr v a)

FromRational !Rational

data FloatExpr v a

basically same as Floating class in Prelude, but as data.

Constructors

PrimPi

Exp !(FloatExpr v a)

Log !(FloatExpr v a)

Sqrt !(FloatExpr v a)

Power !(FloatExpr v a) !(FloatExpr v a)

Sin !(FloatExpr v a)

Cos !(FloatExpr v a)

Tan !(FloatExpr v a)

Asin !(FloatExpr v a)

Acos !(FloatExpr v a)

Atan !(FloatExpr v a)

Sinh !(FloatExpr v a)

Cosh !(FloatExpr v a)

Tanh !(FloatExpr v a)

Asinh !(FloatExpr v a)

Acosh !(FloatExpr v a)

Atanh !(FloatExpr v a)

PrimFloatExpr !(FracExpr v (FloatExpr v a))

pp_unary :: PpShow a => Text -> a -> Doc

class VectorShow v where

Methods

show_vector :: Show a => v a -> String

liftNumFrac1 :: forall {k} {v :: Type -> Type} {a} {v :: Type -> Type} {a :: k}. (NumExpr v a -> NumExpr v (FloatExpr v a)) -> a -> FloatExpr v a

liftNumFrac2 :: forall {k} {v :: Type -> Type} {a} {v :: Type -> Type} {a} {v :: Type -> Type} {a :: k}. (NumExpr v a -> NumExpr v a -> NumExpr v (FloatExpr v a)) -> a -> a -> FloatExpr v a

data EnumExpr a

Constructors

SuccExpr !(EnumExpr a)

PredExpr !(EnumExpr a)

ToEnumExpr !(Var a)

runEnumExpr :: Enum a => [a] -> EnumExpr a -> a

runNumExpr :: (Num a, IdVisitor a) => [a] -> NumExpr v a -> a

runFracExpr :: (Fractional a, IdVisitor a) => [a] -> FracExpr v a -> a

varFold :: [a] -> Fold (Var a) a

vecSpaceFold :: (Visitor (v a), VectorSpace b) => Fold (Var b) b -> Fold (NumExpr v a) (Scalar b) -> Fold (v a) b -> Fold (VectorSpaceExpr v a) b

numExprFold :: Num a => Fold (Var a) a -> Fold (NumExpr v a) a

fracExprFold :: Fractional b => Fold (NumExpr v a) b -> Fold (FracExpr v a) b

enumExprFold :: Enum b => Fold (Var a) b -> Fold (EnumExpr a) b