Math.Number.R

Documentation

newtype R

This real representation takes epsilon as input as in epsilon-delta proof.

Constructors

Limit (Endo Rational)

lim :: Stream R -> R

real :: (Rational -> Rational) -> R

approximate :: R -> Rational -> Rational

approximate takes a real \(r\), and a rational \(q\) and produces a rational \(q_2\) such that \(|r - q_2| \le q\)

\(r_1 (r_2 q) = (r_1 r_2) q\), where \(r_1,r_2 \in {\mathbf{R}}\), \(q \in {\mathbf{Q}}\).

that is, \[{\mathbf{approximate}}(r_1) \circ {\mathbf{approximate}}(r_2) = {\mathbf{approximate}}(r_1 r_2)\] that is, \[{\mathbf{approximate}}(r_1,{\mathbf{approximate}}(r_2,q)) = {\mathbf{approximate}}(r_1 r_2, q)\] that is \[{\mathbf{approximate}}(r_1,q_3)=q_2 \Leftrightarrow |r_1 - q_2| \le q_3\] \[{\mathbf{approximate}}(r_2,q) =q_3 \Leftrightarrow |r_2 - q_3| \le q\] \[{\mathbf{approximate}}(r_1 r_2, q)=q_2 \Leftrightarrow |r_1 r_2 - q_2| \le q\] \[{\mathbf{approximate}}(r_1 r_2, q)={\mathbf{approximate}}(r_1,q_3)={\mathbf{approximate}}(r_1,{\mathbf{approximate}}(r_2,q))\] \[\Leftrightarrow |r_1 - q_2| \le q_3 \land |r_2 - q_3| \le q\]

rationalApproximations :: Approximations str a => a -> str Rational

average :: R -> R -> R

floatingApproximations :: Approximations str a => a -> str Double

epsilon :: R

infiniteR :: R

liftR :: (Rational -> Rational) -> R -> R

this propagates accuracy without changing it.

derivateGenericStream :: Fractional a => (a -> a) -> Stream a -> Stream a

liftR2 :: (Rational -> Rational -> Rational) -> R -> R -> R

this propagates accuracy without changing it

inverseImage :: (Rational -> Rational) -> R -> R

inverseImage transforms accuracy computation of the real.

liftWithAccuracy :: (Rational -> Rational) -> (Rational -> Rational) -> R -> R

The first argument is the function lifted. The second argument describes change in accuracy by the function.

approximateAs :: R -> R -> R

approxCompose will use the second real as the level of approximation to compute the first.

limitCompose :: (Rational -> R) -> R -> R

inverseImageEndo :: Endo Rational -> Endo R

differentialLiftR :: (Rational -> Rational) -> R -> R

This lifts a rational function to a real function. This computes accuracy using the formula \(dx = df / f'(x)\). The accuracy behaves badly when \(f'(x) = 0\) due to divide-by-zero.

derivateRational :: Rational -> (Rational -> Rational) -> Rational -> Rational

derivate for rational functions. The first argument is epsilon.

derivateR :: (R -> R) -> R -> R

computes derivate. \[{{df}\over{dt}} = \lim_{\epsilon\rightarrow 0}{{f(t+\epsilon)-f(t)}\over{\epsilon}}\]

newtonsMethodR :: (R -> R) -> R -> R

newton's method for finding root of function. \[x_{i+1} = x_i - {{f(x_i)}\over{f'(x_i)}}\]

sqrtR :: R -> R

square root of a real number \(x\), computed with newton's method as root of \(t \mapsto t^2 - x\).

logByNewton :: R -> R

logarithm of a real number \(x\), computed with newton's method as root of \(t \mapsto e^t - x\).

expR :: R -> R

piR :: R

Using Simon Plouffe's BPP digit extraction algorithm for computing pi. See https://plouffe.fr/Simon%20Plouffe.htm See Pi for details. \[\pi = \sum_{k=0}^{\infty}{16^{-k}}({{4}\over{8k+1}}-{{2}\over{8k+4}}-{1\over{8k+5}}-{1\over{8k+6}})\]

logR :: R -> R

integralR :: (R, R) -> (R -> R) -> R

foldableIntegral :: (Foldable t, Functor t) => (Rational -> t b) -> (b -> R) -> R

foldableSimpleIntegral :: (Num a, Foldable t, Functor t) => (a -> t b) -> (b -> a) -> a -> a

integralAccuracy :: Fractional t => (t -> R) -> R -> R -> Rational -> [Rational]

asinR :: R -> R

https://en.wikipedia.org/wiki/Inverse_trigonometric_functions

acosR :: R -> R

https://en.wikipedia.org/wiki/Inverse_trigonometric_functions

atanR :: R -> R

https://en.wikipedia.org/wiki/Inverse_trigonometric_functions

sinBySeries :: R -> R

trigonometric functions Computed using taylor series expansion of sin function. \[\sin(x) = \sum_{k=0}^{\infty}{{{(-1)^k}\over{(2k+1)!}}{x^{2k+1}}}\]

inverseOfR :: (R -> R) -> R -> R

cosBySeries :: R -> R

trigonometric functions Computed using taylor series expansion of cos function. \[\cos(x) = \sum_{k=0}^{\infty}{{{(-1)^k}\over{(2k)!}}{x^{2k}}}\]

gammaR :: R -> R

lubR :: [R] -> Maybe R

glbR :: [R] -> Maybe R

computableR :: (Rational -> Rational -> Rational) -> R

For \(a,b \in {\mathbb Q}\) and \(0 \in (a,b) \) the invocation \(f(a)(b)\) of \(f\) in \(computable\_r(f)\) must produce approximation \(q\) of another real \(r\) as rational number, such that \( r - q \in (a,b) \).

unsafeRealToRational :: R -> Rational

this realToRational operation sets precision \(\epsilon\) to 0 and produces corresponding rational number. Many algorithms for reals will go to infinite loop, if this is used, or produce divide-by-zero. In particular, this is expected for irrationals. if rational number is produced, it's exact. However, computations on rational numbers, where real number was created using fromRational, will produce that rational. It also allows operations lifted from rationals to work, producing the result you'd obtain from rational numbers.

floorR :: Integral b => R -> b

ceilingR :: Integral b => R -> b

roundR :: Integral b => R -> b

truncateR :: Integral b => R -> b

properFractionR :: Integral b => R -> (b, R)

approximatelyLessThanOrEqualR :: Rational -> R -> R -> Bool

approximatelyLessThanR :: Rational -> R -> R -> Bool

approximatelyGreaterThanOrEqualR :: Rational -> R -> R -> Bool

approximatelyGreaterThanR :: Rational -> R -> R -> Bool

approximatelyEqualToR :: Rational -> R -> R -> Bool

maxR :: R -> R -> R

minR :: R -> R -> R

realToFloat :: R -> Float

realToDouble :: R -> Double

data Limited a where

Constructors

LimitedRationalsAbove :: (Rational -> Rational) -> NonEmpty Rational -> R -> Limited Rational

LimitedRationalsBelow :: (Rational -> Rational) -> NonEmpty Rational -> R -> Limited Rational

LimitedRationalsAboveBelow :: (Rational -> Rational) -> NonEmpty Rational -> NonEmpty Rational -> R -> R -> Limited Rational

supremum :: Limited Rational -> Maybe R

infinum :: Limited Rational -> Maybe R

boundedApproximations :: Limited Rational -> Rational -> Rational

upperBounds :: Limited Rational -> [Rational]

lowerBounds :: Limited Rational -> [Rational]

piByIntegration :: R

integralAccuracySimple :: R -> R -> Rational -> [Rational]

asinBySeries :: R -> R

https://proofwiki.org/wiki/Power_Series_Expansion_for_real_Arcsine_Function Computed by MacLaurin series expansion of arc sine: \( \arcsin x = \sum^{\infty}_{n=0}{{(2n)! x^{2n+1}}\over{2^{2n}(n!)^2(2n+1)}} \)

acosBySeries :: R -> R

integrateRationalR :: (R, R) -> (Rational -> Rational -> Rational) -> R

integrateRealR :: (R, R) -> (R -> R -> R) -> R

atanBySeries :: R -> R

https://en.wikipedia.org/wiki/Arctangent_series Computed using Gregory's series: \( \arctan x = \sum^{\infty}_{n=0}{{(-1)^kx^{2k+1}}\over{2k+1}} \)

errorFunctionR :: R -> R

Orphan instances

StreamBuilder ((->) Rational)

Instance details

Methods

pre :: a -> (Rational -> a) -> Rational -> a

StreamObserver ((->) Rational)

Instance details

Methods

shead :: (Rational -> a) -> a

stail :: (Rational -> a) -> Rational -> a