Math.Number.Real

Documentation

newtype R

Problem with this representation: real number ranges cannot be represented. given two rational numbers \(r_1 < r_2\), it should be there are infinitely many real numbers between \(r_1\) and \(r_2\), so support density of irrationals. https://en.wikipedia.org/wiki/Dense_order

In constructive mathematics, given two real numbers \(r_1' < r_2'\), there should exist a rational number r, such that \(r_1' < r\) and \(r < r_2'\). HOWEVER, these comparisons are a type error for Eq class. In particular this cannot be used to _define_ how to compare two real numbers.

what is meant is if \(r_1' < r_2'\) then there exists a rational r such that \(r_1' < {\mathbf{fromRational}}(r)\) and \({\mathbf{fromRational}}(r) < r_2'\)

In here properties of \({\mathbf{fromRational}}\) are important. Note that comparisons invoke the real comparison, so at most is a recursive definition of real comparison in terms of itself. But it doesn't include termination condition.

It should be so that _if_ two real numbers \(r_1\) and \(r_2\) are not the same real number, then \(r_1 < r_2\) makes sense. But comparing whether two real numbers are the same is undecidable (algorithm for this comparison will not halt if they are same). Would this mean that if \(r_1 < r_2\) is computable, wouldn't \(\neg(r_1 < r_2 \lor r_2 < r_1)\) be computable equality of reals; This is computable in the limit , but not computable? https://en.wikipedia.org/wiki/Computation_in_the_limit

So we are forced to consider only comparison between a rational and a real, but not between two reals. This is encapsulated in the Math.Tools.Interface.DedekindCut type class.

The solution is that when comparing a rational with a real, the rational's denominator determines the accuracy used. This additional information makes comparison between a rational and a real decidable, but doesn't help when comparing two reals.

Constructors

Limit (Stream Rational)

data Modulus

Constructors

Modulus

Fields

modulus_value :: R

modulus :: Rational -> Integer

d :: Infinitesimal Stream a => (a -> a) -> a -> Closure Stream a

d x = partialDerivate x

liftReal :: (Rational -> Rational) -> R -> R

convergenceRatioTest :: (Fractional a, Ord a, Closed a) => Stream a -> Bool

averageConvergenceRatio :: (MetricSpace a, Fractional (Scalar a)) => Stream a -> Stream (Scalar a)

listAverage :: Fractional a => [a] -> a

cauchy :: MetricSpace a => Stream a -> Stream [Scalar a]

cseqDifferenceMatrix :: (Functor m, Functor n, Diagonalizable m a, LinearTransform m n a, Num a) => m a -> n a -> (m :*: n) a

cauchySequenceEquivalence :: (ConjugateSymmetric a, Closed a, Num a) => Stream a -> Stream a -> Stream a

See Suppes: Axiomatic Set Theory. Note that we would like to check that \[\forall \epsilon > 0. \exists n. \forall k > n. |s_k - t_k| < \epsilon \]. However, this is undecidable in particular with respect to what to choose for n. We can nonetheless compute the stream of distances.

cseqEquivalenceMatrix :: (Num a, ConjugateSymmetric a, Closed a, Infinitesimal Stream eps) => (eps -> Integer) -> Stream a -> Stream a -> (Stream :*: Stream) a

cseqEquivalenceMatrix computes equivalence of cauchy sequences. However, the modulus argument is undecidable. But if you produce a modulus, this chooses the part of the cauchy sequence differences that should converge according to the convergence ratio of the epsilon from the infinitesimal instance. Thus if you can prove that the matrix elements converge to zero when elements of increasing index are chosen, you would have proven equality of the two cauchy sequences. This could be possible in special cases, even if it's undecidable in general. Note that it suffices to consider the convergence of the first column, because you could move any part to the first column by choosing reasonable values of the modulus. The other differences are included so computations and analysis of the situation can be easier to perform.

cseqConvergence :: (ConjugateSymmetric a, Closed a, Num a) => Stream a -> Stream a -> Stream [a]

This version of cauchy sequence equality convergence check produces a stream of lists, where each list contains those elements which are close to each other in terms of choice of indices, e.g. \[C(s,t)_i = [ \lvert s_j - t_k \rvert | j \leftarrow naturals, k \leftarrow naturals, i = j + k]\] thus the complexity of the stream element lists of the result should increase in the same way than the complexity of the input stream elements. The lists in the result stream have monotonically increasing length. thus to compute results for particular level of approximation, it should suffice to choose a list from the result stream, and perform the equivalence check using the values in the list. To do this, it would be necessary to be able to determine how close to zero the elements in the list should be. the middle element (if any) of the list is probably the most interesting, because that would mean \(j == k\), therefore the computation result would be \(\lvert s_j - t_j \rvert\), but due to different rate of convergence of the input sequences, it can happen that the elementwise differences are not really representative of the stream convergence behaviour at the diagonal. however, analyzing these numbers should produce useful results about convergence of the input stream element differences. e.g choosing minimum and maximum element from each list determines best and worst approximation in the same depth of approximations. for the equality of the cauchy sequences to hold, the best approximation should stay close to zero and better so when further elements of the stream are obtained.

cseqEquivalence_list :: (Closed a, ConjugateSymmetric a, Num a, Infinitesimal Stream eps) => (eps -> Integer) -> Stream a -> Stream a -> Stream [a]

converges :: MetricSpace a => Stream a -> a -> Stream [Scalar a]

increaseAccuracyBy :: Limiting Stream a => Integer -> Closure Stream a -> Closure Stream a

partialDerive :: (Closed a, Infinitesimal Stream a, Fractional a) => (a -> a -> a) -> (a -> a) -> a -> a

\[ {\rm{partial\_derive}}(\Delta, f, x) = \lim_{\epsilon \rightarrow 0}{{f(\Delta_{\epsilon}(x)) - f(x)}\over{\Delta_{\epsilon}(x) - x}} \]

partialDerivateClosure :: (Limiting str a, Fractional a) => (a -> b -> b) -> (b -> a) -> b -> Endo (Closure str a)

partialDerivate :: (Closed eps, Fractional eps) => (eps -> a -> a) -> (a -> eps) -> a -> eps

\[ {\rm{partial\_derivate}}(\Delta, f, x) = \lim_{\epsilon \rightarrow 0}{{f(\Delta_{\epsilon}(x)) - f(x)}\over{\Delta_{\epsilon}(x) - x}} \]

partialDerivateEndo :: (Limiting str a, Fractional a) => (a -> Endo b) -> (b -> a) -> b -> Endo (Closure str a)

complexDerivate :: (RealFloat r, Closed r) => (Complex r -> Complex r) -> Complex r -> Complex r

https://en.wikipedia.org/wiki/Wirtinger_derivatives

average :: R -> R -> R

completenessOfReals :: Closure Stream R :==: R

An isomorphism \(Cl(R) \cong R\).

liftRClosure :: (R -> R) -> Closure Stream R -> Closure Stream R

limitReal :: Stream R -> R

liftReal :: (Rational -> Rational) -> R -> R

series :: (Integer -> R) -> R

supremumGen :: (Fractional a, Ord a, Limiting Stream a) => [Closure Stream a] -> Closure Stream a

Least upper bound property | note this is really undecidable.

negate_limit :: (Limiting str a, Num a) => Closure str a -> Closure str a

infimumGen :: (Fractional a, Ord a, Limiting Stream a) => [Closure Stream a] -> Closure Stream a

infinity :: R

infinityGen :: (Fractional a, Limiting Stream a) => Closure Stream a

negative_infinityGen :: (Fractional a, Limiting Stream a) => Closure Stream a

limitBelow :: R -> (R -> R) -> Closure Stream R

Squeeze theorem

limitAbove :: R -> (R -> R) -> Closure Stream R

Squeeze_theorem

limitBoth :: R -> (R -> R) -> Closure Stream R

Squeeze_theorem

negative_infinity :: R

epsilonLinear :: R

increaseAccuracy :: R -> R

lessthanPrecision :: R -> R -> Integer -> Bool

limitingDistance :: (Num a, Limiting str a) => Closure str a -> Closure str a -> Closure str a

distanceMatrix :: R -> R -> (Stream :*: Stream) Rational

atPrecisionRational :: R -> Rational -> Rational

compute rational approximation more precise than the given rational the precision is expressed as rational number close to zero, the result will be within range \([r-p,r+p]\), in the sense that in the sequence of rational approximations \(r = (r_i | i \in {\mathbf N})\) we have \(\lvert r_{i+1} - r_i \rvert \leq p\). Note that we do _not_ attempt to prove that all successive approximations have this same property, we need the differences to be monotonically decreasing for this to represent correct precision.

atPrecision :: R -> Integer -> R

here precision is expresses as integer power of 10, e.g. \(1 == 10^{-1}\), \(2 == 10^{-2}\), \(3 == 10^{-3}\) and so on. so how many digits after the decimal point are required.

precisionRational :: R -> Integer -> Rational

rational precision expressed as integer power of 10

realDerivate :: (R -> R) -> R -> R

realExp :: R -> R

derivateClosed :: Infinitesimal str a => (a -> a) -> a -> Closure str a

vectorDerivate :: (Infinitesimal str (Scalar a), VectorSpace a, Limiting str a) => (Scalar a -> a) -> Scalar a -> Closure str a

pseudoDerivate :: (Fractional r, Limiting str r, Infinitesimal str a) => (a -> r) -> a -> Closure str r

https://en.wikipedia.org/wiki/Differential_calculus

derivatesAt :: DifferentiallyClosed a => (a -> a) -> Stream a -> Stream a

i'th element of derivatesAt(f,s) is \(D^i[f](s_i)\)

derivateAround :: Infinitesimal str a => (a -> a) -> a -> Closure str a

derivate_around doesn't require f to be defined at x, but requires limits from both sides of x to exist [it never evaluates f at x]. \[ \lim_{\epsilon \rightarrow 0} {{f(x+\epsilon)-f(x-\epsilon)}\over{2\epsilon}} \]

partialDerivate1_2 :: Infinitesimal str a => (a -> b -> a) -> a -> b -> Closure str a

\[\lim_{\epsilon\rightarrow 0}{{f(a+\epsilon,b)-f(a,b)}\over{\epsilon}}\]

partialDerivate2_2 :: Infinitesimal str a => (b -> a -> a) -> b -> a -> Closure str a

\[\lim_{\epsilon\rightarrow 0}{{f(a,b+\epsilon)-f(a,b)}\over{\epsilon}}\]

partialDerivate1_3 :: Infinitesimal str a => (a -> b -> c -> a) -> a -> b -> c -> Closure str a

\[\lim_{\epsilon\rightarrow 0}{{f(a+\epsilon,b,c)-f(a,b,c)}\over{\epsilon}}\]

partialDerivate2_3 :: Infinitesimal str a => (b -> a -> c -> a) -> b -> a -> c -> Closure str a

\[\lim_{\epsilon\rightarrow 0}{{f(a,b+\epsilon,c)-f(a,b,c)}\over{\epsilon}}\]

partialDerivate3_3 :: Infinitesimal str a => (b -> c -> a -> a) -> b -> c -> a -> Closure str a

\[\lim_{\epsilon\rightarrow 0}{{f(a,b,c+\epsilon)-f(a,b,c)}\over{\epsilon}}\]

agm :: (Floating a, Limiting str a) => a -> a -> Closure str a

Arithmetic-geometric mean

expo :: R -> R

Exponential function using limit definition \(\exp(x) = \lim_{n\rightarrow\infty}(1+{{x}\over{n}})^n\).

newtonsMethod :: Numerics a => (a -> a) -> a -> a

eigenvalue :: (Closed a, LinearTraceable LinearMap m a, Linearizable LinearMap (:*:) m m a, Applicative m, a ~ Scalar (m a), VectorSpace (m a), LinearTransform m m a) => (m a :-> m a) -> Closure Stream a

integrate_vector :: (Enum (Scalar a), VectorSpace a, Limiting str a, Limiting str (Scalar a)) => (Scalar a -> a) -> Closure str (Scalar a) -> (Scalar a, Scalar a) -> Closure str a

integrate :: (Enum a, Num a, Limiting str a) => (a -> a) -> Closure str a -> (a, a) -> Closure str a

https://en.wikipedia.org/wiki/Methods_of_contour_integration

integralReal :: (R, R) -> (R -> R) -> R

integralRational :: (R, R) -> (Rational -> Rational) -> R

integrateCurve :: (Enum a, Num a, Num r, Limiting str a, Limiting str r) => (a, a) -> (a -> r) -> (r -> r) -> Closure str a -> Closure str r

for integrateCurve (xa,ya) curve f dx, Requires: \(\lim_{x \rightarrow 0}curve(x) = 0 \land \lim_{i\rightarrow \infty} dx_i = 0\)

gamma_via_integral :: R -> Closure Stream R

Integral This doesn't converge well.

approximatelyEqual :: R -> R -> Bool

approximatelyEq :: R -> R -> Bool

constructively_less :: Rational -> R -> R -> Closure Stream Bool

equality :: Closure Stream R -> Closure Stream R -> Closure Stream Bool

signumReal :: R -> Int -> Int

showAtPrecisionReal :: R -> Integer -> String

properFractionReal :: Num a => R -> (a, R)

enumFromThenToReal :: Enum e => (e -> R) -> e -> e -> e -> [R]

This attempts to work around the density problem by computing range | in some type which does have enum instance, | then using the given function to map to reals.

fromThenTo :: Rational -> Rational -> Rational -> [R]

fromTo :: Rational -> Rational -> [R]

enumFromToReal :: Enum e => (e -> R) -> e -> e -> [R]

This attempts to work around the density problem by computing range | in some type which does have enum instance, | then using the given function to map to reals.

realEnumFromThenTo :: R -> R -> R -> Stream [Rational]

strEnumFromThenTo :: Stream Rational -> Stream Rational -> Stream Rational -> Stream [Rational]

toReal :: Rational -> R

logarithm :: R -> R

Natural logarithm log(1+z) = z - z^22 + z^33 - ...

eulerConstant :: R

napiersConstant :: R

logarithmByNewton :: R -> Closure Stream R

invertByNewton :: Closed a => (a -> a) -> a -> Closure Stream a

expByPowers :: (Fractional r, Limiting Stream r) => r -> Closure Stream r

expByPowersReal :: R -> R

sumGeneratingFunction :: Stream Rational -> R -> R

riemann_zeta :: Floating a => a -> Stream a

https://en.wikipedia.org/wiki/Riemann_zeta_function

riemann_zetaReal :: Integral a => a -> R

https://en.wikipedia.org/wiki/Riemann_zeta_function

riemann_zetaComplex :: RealFloat a => Complex a -> Stream (Complex a)

https://en.wikipedia.org/wiki/Riemann_zeta_function

encodeInBase :: Integer -> Integer -> Int -> R

fromFloat :: RealFloat a => a -> R

goldenRatio :: R

https://en.wikipedia.org/wiki/Golden_ratio

slowGoldenRatio :: R

logistic :: Floating a => a -> a

https://en.wikipedia.org/wiki/Logistic_function

floatingApproximations :: Approximations str a => a -> str Double

logarithmByPower :: (Floating a, Limiting Stream a) => a -> Closure Stream a

tetrate :: (Eq b, Num b, Floating a) => b -> a -> a

Tetration

tetrate2 :: Floating a => a -> a

logNewton :: (Floating a, Closed a) => a -> a

expBySeries2 :: R -> R

expBySeries :: R -> R

sinBySeries :: R -> R

cosBySeries :: R -> R

approximateSumsModulus :: Modulus -> Modulus -> Modulus

withInverseModulus :: (Fractional r, Real u) => (R -> R) -> (r -> u) -> Modulus -> Modulus

partialDerivateLinear :: (LinearTransform f f a, VectorSpace (f a), Closed a, Closed (f a), Fractional a, Scalar (f a) ~ a, Linearizable LinearMap (:*:) f f a, Diagonalizable f a) => (a -> f a :-> f a) -> (f a :-> f a) -> f a :-> f a

Orphan instances

MetricSpace Integer

Instance details

Methods

distance :: Integer -> Integer -> Scalar Integer

MetricSpace Int

Instance details

Methods

distance :: Int -> Int -> Scalar Int

Infinitesimal Stream Rational

Instance details

Methods

epsilonClosure :: Closure Stream Rational

infiniteClosure :: Closure Stream Rational

minusInfiniteClosure :: Closure Stream Rational

nanClosure :: Closure Stream Rational

epsilonStream :: Stream Rational

infiniteStream :: Stream Rational

minusInfiniteStream :: Stream Rational

MetricSpace (Ratio Integer)

Instance details

Methods

distance :: Ratio Integer -> Ratio Integer -> Scalar (Ratio Integer)