Math.Number.Stream

Description

This module implements lazy infinite streams, stream generating functions and lots of operations on the lazy streams. some examples for use of this module:

fib = 1 `div` (1 - z - z*z)

pascal_triangle = Matrix $ 1 `div` (1 - z - z2*z)

take 4 fib == [1,1,2,3]

alternating_bits = cycle [0,1]

Documentation

data Stream a

Data structure of infinite lazy streams.

Constructors

Pre infixr 5

Fields

shead :: a

stail_strict :: Stream a

stail :: Stream a -> Stream a

matrix_powers :: (SquareMatrix g a, InnerProductSpace (g a), Scalar (g a) ~ a) => (g :*: g) a -> Stream ((g :*: g) a)

matrix_exponential_stream :: (Fractional (Scalar ((g :*: g) a)), Num ((g :*: g) a), SquareMatrix g a, InnerProductSpace (g a), VectorSpace ((g :*: g) a), Scalar (g a) ~ a) => (g :*: g) a -> Stream ((g :*: g) a)

matrix_exponential :: (Fractional (Scalar ((g :*: g) a)), Num ((g :*: g) a), SquareMatrix g a, InnerProductSpace (g a), VectorSpace ((g :*: g) a), Scalar (g a) ~ a) => (g :*: g) a -> (g :*: g) a

https://en.wikipedia.org/wiki/Matrix_exponential

class Limiting a where

Idea: limit produces an exact value from a monotone sequence of approximations

https://ncatlab.org/nlab/show/closure+operator

The Closure represents a sequentially closed set of elements

Minimal complete definition

limit, approximations

Associated Types

data Closure a

Methods

limit :: Stream a -> Closure a

approximations :: Closure a -> Stream a

class Limiting a => Closed a where

Minimal complete definition

accumulation_point

Methods

accumulation_point :: Closure a -> a

limiting_iso :: (BiArrow arr, Limiting a) => arr (Stream a) (Closure a)

vec_stream :: Stream a -> Integer -> a

cmap :: (Limiting a, Limiting b) => (a -> b) -> Closure a -> Closure b

mapping operation for closures. Note we cannot make this a Functor instance due to constraint in the type.

cliftA2 :: (Limiting a, Limiting b, Limiting c) => (a -> b -> c) -> Closure a -> Closure b -> Closure c

close :: Limiting a => a -> Closure a

operation for producing a closure out of single element.

closure_limit :: Limiting a => Stream (Closure a) -> Closure a

(.>>=.) :: (Limiting a, Limiting b) => Closure a -> (a -> Closure b) -> Closure b

Monadic bind operation for closures, again cannot be made instance of Monad.

eigenvector_by_power_iteration :: (Fractional (Scalar (m a)), NormedSpace (m a), LinearTransform m m a) => (m :*: m) a -> m a -> Stream (m a)

https://en.wikipedia.org/wiki/Power_iteration

stream_distance_count :: Eq a => Stream a -> Stream a -> Integer

stream_distance :: Eq a => Stream a -> Stream a -> Rational

stream_identity_matrix :: Integral a => (Stream :*: Stream) a

This implementation of infinite identity matrix requires Integral constraint. it is not as efficient as identity, but is implemented using generating functions.

stream_diagonal :: (Stream :*: Stream) a -> Stream a

stream_diagonal is the "depth-first" traversal over two-dimensional streams, where the choice always goes to diagonal elements.

stream_diagonal (matrix f x y) == liftA2 f x y

ssum :: (Closed a, Num a) => Stream a -> a

sum of a stream elements.

increasing_substream :: Ord a => Stream a -> Stream a

sub_member :: Ord a => Stream a -> a -> Bool

bind_matrix :: (Stream :*: Stream) a -> (a -> Stream a -> Stream a -> Stream b) -> Stream b

bind diagonal element and horizontal and vertical strip

stream_as_diagonal :: Num a => Stream a -> (Stream :*: Stream) a

stream matrix with a specified diagonal, other elements are zero.

add_row :: f a -> (Stream :*: f) a -> (Stream :*: f) a

adding a row to a matrix.

remove_row :: (Stream :*: f) a -> (Stream :*: f) a

add_column :: Applicative f => f a -> (f :*: Stream) a -> (f :*: Stream) a

adding a column to a matrix.

remove_column :: Functor f => (f :*: Stream) a -> (f :*: Stream) a

join3 :: (Stream :*: Stream) a -> Stream (a, a, a)

tails :: Stream a -> (Stream :*: Stream) a

skippers :: Stream a -> (Stream :*: Stream) a

adjusted_sequence_product :: Fractional b => Stream b -> Stream b -> Stream b

average_seq :: Fractional a => Seq a -> a

sum_seq :: Num a => Seq a -> a

stream_powers :: Num a => Stream a -> (Stream :*: Stream) a

compose :: (Num a, Eq a) => Stream a -> Stream a -> Stream a

https://en.wikipedia.org/wiki/Formal_power_series

Note that identity of composition is z.

matrix_convolution :: (Functor g, Transposable h f, InnerProductSpace (h a), VectorSpace ((g :*: f) (Scalar (h a)))) => Stream ((g :*: h) a) -> Stream ((h :*: f) a) -> Stream ((g :*: f) (Scalar (h a)))

matrix_convolution_product :: (Fractional a, ConjugateSymmetric a) => (Stream :*: Stream) a -> (Stream :*: Stream) a -> (Stream :*: Stream) a

distance_product :: (Floating a, ConjugateSymmetric a) => Stream a -> Stream a -> Stream a

distance_product a b computes l_k = sqrt(sum[a_i*conj(b_j) | i + j = k]).

and_convolution :: Stream Bool -> Stream Bool -> Stream Bool

min_convolution :: (Num a, Ord a) => Stream a -> Stream a -> Stream a

and_stream :: Stream Bool -> Stream Bool -> Stream Bool

or_stream :: Stream Bool -> Stream Bool -> Stream Bool

fromNum :: Num a => a -> Stream a

cycleQueueBy :: (a -> b -> b) -> Queue a -> b

cycleQueue :: Queue a -> Stream a

convolve_with :: ([a] -> b) -> (c -> d -> a) -> Stream c -> Stream d -> Stream b

convolve_with_seq :: (Seq a -> b) -> (c -> d -> a) -> Stream c -> Stream d -> Stream b

data StreamIndex a

Constructors

StreamIndex

Fields

sindex_position :: Integer

sindex_element :: a

streamfold :: (a -> b -> b) -> Fold (Stream a) b

fold over streams

fold_codiagonals :: ([a] -> b -> b) -> (Stream :*: Stream) a -> b

streamfold2 :: (a -> c -> b) -> (a -> b -> c) -> Fold (Stream a) b

streamfold_lst :: [a -> b -> b] -> Fold (Stream a) b

streamfold_stream :: Stream (a -> b -> b) -> Fold (Stream a) b

rotate_prefix :: Integer -> Stream a -> Stream a

threeNplusOne :: Integer -> Integer

toNatSeq :: Stream a -> Rec Maybe -> a

fromIntegerSeq :: (Integer -> a) -> Stream a

sumSeq :: (Num b, Num a, Ord a) => (a -> b) -> a -> Stream b

sumSeq f i == Sum[n=i..Infinity](f(n))

boolValue :: Num a => Bool -> a

fromBoolSeq :: Num a => (Integer -> Bool) -> Stream a

streamIndexFold :: Fold (Stream a) (Integer -> a)

streamindex :: Integer -> Stream a -> a

streamindex2 :: (Integer, Integer) -> (Stream :*: Stream) a -> a

map_indices :: (Integer -> Integer) -> Stream a -> Stream a

constant :: a -> Stream a

stream consisting of the same element repeated.

limit_fold :: (a -> a -> a) -> Fold (Stream a) (Integer -> a)

limit_fold_gen :: (a -> b -> b) -> (a -> b) -> Fold (Stream a) (Integer -> b)

iterate_stream :: (a -> a) -> a -> Stream a

stream_differences :: Num a => Stream a -> Stream a

stream_quotients :: Fractional a => Stream a -> Stream a

fixpoint :: Limiting a => (a -> a) -> a -> Closure a

matrix_inverse :: (VectorSpace ((h :*: f) a), Transposable h f, Transposable f h, InnerProductSpace (h a), InnerProductSpace (f a), Num a, Scalar (h a) ~# a, Scalar (f a) ~# a) => (f :*: h) a -> (h :*: f) a -> Stream ((h :*: f) a)

stream_min :: Ord a => Fold (Stream a) (Integer -> a)

stream_max :: Ord a => Fold (Stream a) (Integer -> a)

liftStream2 :: (a -> a -> a) -> Stream a -> Stream a

min_stream :: Ord a => Stream a -> Stream a

max_stream :: Ord a => Stream a -> Stream a

stream_product :: Num a => Fold (Stream a) (Integer -> a)

stream_sum :: Num a => Fold (Stream a) (Integer -> a)

product_stream :: Num a => Stream a -> Stream a

product_stream produce from a stream [a0,a1,a2,...] a stream [a0,a0*a1,a0*a1*a2,a0*a1*a2*a3,...]

sum_stream :: Num a => Stream a -> Stream a

sum_stream will produce from a stream [a0,a1,a2,...] a stream [a0,a0+a1,a0+a1+a2,...].

sum_stream_integral :: Integral a => Stream a -> Stream a

sum_stream_fractional :: Fractional a => Stream a -> Stream a

cobind :: (a -> b) -> (Stream :*: Stream) a -> Stream b

if f in cobind is a fixed-point free function, then this will find a stream that is not an element of any of the streams.

indexed_lookup :: Stream Integer -> Stream a -> Stream a

universal :: (Stream :*: Stream) Bool -> Stream Bool

existential :: (Stream :*: Stream) Bool -> Stream Bool

topleft :: (Stream :*: Stream) a -> a

stream_codiagonal :: (Stream :*: Stream) a -> Stream (Stream a, Stream a)

Everything but a diagonal

zero_codiagonal :: Num a => Codiagonal Stream a

stream_diagonal_matrix :: Num a => Stream a -> (Stream :*: Stream) a

stream_matrix :: Stream a -> Stream (Stream a, Stream a) -> (Stream :*: Stream) a

stream_matrix creates matrix from a diagonal and a codiagonal

prefix_stream_matrix :: Stream a -> (Stream a, Stream a) -> (Stream :*: Stream) a -> (Stream :*: Stream) a

pseudo_codiagonal :: (Stream :*: Stream) a -> Codiagonal Stream a

pseudo_matrix :: Stream a -> Stream (Stream a, Stream a) -> (Stream :*: Stream) a

map_matrix :: (a -> b) -> ((Stream a, Stream a) -> (Stream b, Stream b)) -> (Stream :*: Stream) a -> (Stream :*: Stream) b

map_diagonal :: (a -> a) -> (Stream :*: Stream) a -> (Stream :*: Stream) a

joinWith :: (a -> b -> c) -> Stream a -> Stream b -> Stream c

zipList :: ([a] -> b) -> [Stream a] -> Stream b

sequence :: [Stream a] -> Stream [a]

unzipList :: Stream [a] -> [Stream a]

zipStream :: Functor f => (f b -> a) -> f (Stream b) -> Stream a

sequenceStream :: Functor f => f (Stream a) -> Stream (f a)

apply3 :: Stream (a -> b -> c) -> Stream a -> Stream b -> Stream c

applyA :: Arrow arr => Stream (arr a b) -> arr (Stream a) (Stream b)

dematrix :: (Stream :*: Stream) a -> (a, (Stream a, Stream a), (Stream :*: Stream) a)

codiagonals_seq :: (Stream :*: Stream) a -> Stream (Seq a)

codiagonals :: (Stream :*: Stream) a -> Stream [a]

The codiagonals function will divide a two-dimensional stream of streams (where elements are e_(i,j) into a stream of lists

l_k where l_k = [e_(i,j) | i <- naturals, j <- naturals, i+j == k]

The indices in the streams are closely related to "Cantor pairing function" (which is a bijection)

f :: (N,N) -> N
f(x,y) = (x+y)(x+y+1)/2 + x

Note that the list at each element of the resulting stream is of finite (but increasing) size. The order in each list is such that the diagonal elements are at the center of (the every second) list.

It is possible to think of this as "breadth-first" traversal over two-dimensional streams.

 d   x xr0 xr1  xr2 ..
 y   d' x' xr0' xr1' ..
 yr0 y'
 yr1 yr0'
 yr2 yr1'
 ..

There is symmetry between primed and unprimed (e.g. d and d') towards the diagonal.

The resulting lists would be: [d], [x,y], [xr0,d',yr0], [xr1,x',y',yr1],...

data Transform2D a b

Constructors

Transform2D

Fields

transform_diagonal :: a -> b

transform_commutator :: a -> a -> b

transform_symmetry :: a -> b -> a -> b

codiagonalsWith :: Transform2D a b -> (Stream :*: Stream) a -> Stream b

This operation allows customizing the behaviour of the codiagonals transformation.

element2D :: a -> a -> Transform2D () a

functor2D :: (FunctorArrow f arr, Category arr) => (arr a a -> arr a a -> arr (f a) (f a)) -> Transform2D (arr a a) (arr (f a) (f a))

category2D :: Category cat => Transform2D (cat a a) (cat a a)

arrow2D :: (BiArrow arr, Arrow arr) => Transform2D (a -> a) (arr a a)

median2D :: MedianAlgebra a => Transform2D a a

commutatorSum2D :: LieAlgebra a => Transform2D a a

transpose2D :: Transform2D a b -> Transform2D a b

compose2D :: Transform2D b c -> Transform2D a b -> Transform2D a c

monoid2D :: Monoid a => (a -> a -> a -> a) -> Transform2D a a

productSum2D :: Num a => Transform2D a a

set2D :: Ord a => Transform2D a (Set a)

project2D :: (a -> (a, a) -> a -> (a, a)) -> Transform2D a (a, a)

flattenList2D :: Transform2D [a] [a]

list2D :: Transform2D a [a]

seq2D :: Transform2D a (Seq a)

uncodiagonals :: Stream [a] -> (Stream :*: Stream) a

for uncodiagonals, the i'th list element of the input stream must have length i.

uncodiagonals ([a] `Pre` [b,c] `Pre` [d,e,f] ...) == Pre (a `Pre` c `Pre` f ...) $ Pre (b `Pre` e `Pre` ...) $ Pre (d `Pre` ... ) $ ...

https://en.wikipedia.org/wiki/Formal_power_series

takes :: Stream Integer -> Stream a -> Stream [a]

split_dimension :: Stream a -> (Stream :*: Stream) a

split_planar :: Stream a -> (Stream :*: Stream) a

codiagonalsIso :: BiArrow arr => arr ((Stream :*: Stream) a) (Stream [a])

lower_triangle :: (Stream :*: Stream) a -> (Stream :*: Stream) a

lower_triangle takes lower half of a two-dimensional stream split at diagonal.

upper_triangle :: (Stream :*: Stream) a -> (Stream :*: Stream) a

upper_triangle takes upper half of a two-dimensional stream split at diagonal.

pairing_matrix :: Fractional a => (Stream :*: Stream) a

codiagonals3 :: (Stream :*: (Stream :*: Stream)) a -> Stream [[a]]

concatenate :: Stream [a] -> Stream a

concatenate a stream of lists to a stream of elements

diagonalize :: Stream [a] -> Stream a

skip_even :: Stream a -> Stream a

nice_tail :: [a] -> [a]

diagonal_differences :: (a -> b) -> (Stream :*: Stream) a -> Stream b

find_equal :: Eq a => Stream a -> Stream a -> a

fixedpoint :: Eq a => (a -> a) -> (Stream :*: Stream) a -> a

sum_join :: Num a => (Stream :*: Stream) a -> Stream a

sum_bind :: Num b => Stream a -> (a -> Stream b) -> Stream b

cojoin :: (Stream :*: Stream) a -> Stream a

cojoin is "better" version of join for streams. cojoin does not satisfy monad laws (Right identity and associativity do not hold.). It eventually produces every element of the original two-dimensional stream. It first produces all elements e_(i,j) where i+j=k and k increasing.

(|*|) :: Stream a -> Stream b -> Stream (a, b)

naturals |*| naturals == (0,0),   [note sum=0]
                         (0,1),(1,0),  [note sum=1]
                         (0,2),(2,0),(1,1),  [note sum=2]
                         (0,3),(3,0),(1,2),(2,1),  [note sum=3]
                         (0,4),(4,0),(1,3),(3,1),(2,2),  [note sum=4]
                          ...

(>>!=) :: Stream a -> (a -> Stream b) -> Stream b

For a better "bind" operation, there exists (>>!=), which produces the following result (note duplicates):

naturals >>!= \x -> naturals >>!= \y -> return (x,y)
== (0,0),
  (0,0),(1,0),
     (0,1),(1,0),(2,0),
  (0,0),(1,1),(2,0),(3,0),
    (0,1),(1,0),(2,1),(3,0),(4,0),
      (0,2),(1,1),(2,0),(3,1),(4,0),(5,0),
  (0,0),(1,2),(2,1),(3,0),(4,1),(5,0),(6,0),
    (0,1),(1,0),(2,2),(3,1),(4,0),(5,1),(6,0),(7,0),
      (0,2),(1,1),(2,0),(3,2),(4,1),(5,0),(6,1),(7,0),(8,0),
       (0,3),(1,2),(2,1),(3,0),(4,2),(5,1),(6,0),(7,1),(8,0),(9,0),
  ...

(>!=) :: Stream a -> (a -> Stream b) -> Stream [b]

The (>!=) is like binding, but the results are packaged to a finite list. Again note duplicates.

naturals >!= \x -> naturals >!= \y -> return (x,y)
== [[(0,0)]
  ],
  [[(0,0),(0,1)],
   [(1,0)]
  ],
  [[(0,0),(0,1),(0,2)],
   [(1,0),(1,1)],
   [(2,0)],
  ],
  [[(0,0),(0,1),(0,2),(0,3)],
 [(1,0),(1,1),(1,2)],
 [(2,0),(2,1)],
 [(3,0)]
],
...

integers :: Num a => Stream a

integers_stream :: Stream Integer

A stream of integers. This is specialized version of naturals for integers

naturals :: Num a => Stream a

A stream of increasing numbers starting from 0

naturals_starting_from :: (Num a, Ord a) => a -> Stream a

A stream of increasing numbers starting from a given number.

nonzero_naturals :: Num a => Stream a

A stream of increasing numbers starting from 1

integral_nonzero_naturals :: Integral a => Stream a

integral_naturals :: Integral a => Stream a

z :: Num a => Stream a

This is the variable used in generating functions. Note that in multidimensional streams, this is the variable associated with the largest dimension.

z2 :: Num a => Stream (Stream a)

The variable associated with second largest dimension:

z2_matrix :: Num a => (Stream :*: Stream) a

z3 :: Num a => Stream (Stream (Stream a))

variable associated with third largest dimension

z3_matrix :: Num a => ((Stream :*: Stream) :*: Stream) a

log_stream :: Num a => (Stream :*: Stream) a

stream_logarithm :: Fractional a => Stream a -> Stream a

stream_log :: Integral a => Stream a -> Stream a

approximate_sums :: Fractional a => Stream a -> Stream a -> Stream a

For a stream of terms a=[a0,a1,a2,...], and x=[x0,x1,x2,...], 'approximate_sums a x' computes \n -> Sum[i=0..n]{a_i*x_i^i}, where n is the index to the result stream.

exponential_stream :: Fractional a => Stream a

exponential :: (Eq a, Fractional a) => Stream a -> Stream a

log_generating_function :: Fractional a => Stream a

cos_stream :: Fractional a => Stream a -> Stream a

https://en.wikipedia.org/wiki/Trigonometric_functions

sin_stream :: Fractional a => Stream a -> Stream a

https://en.wikipedia.org/wiki/Trigonometric_functions

sin_generating_function :: Fractional a => Stream a

https://en.wikipedia.org/wiki/Trigonometric_functions

cos_generating_function :: Fractional a => Stream a

https://en.wikipedia.org/wiki/Trigonometric_functions

factorial :: Num a => Stream a

reciprocal :: Fractional a => Stream a -> Stream a

both reciprocal and inversion are the inverse of a Cauchy product. see http://en.wikipedia.org/wiki/Formal_power_series

reciprocal s * s == unit_product

inversion :: Integral a => Stream a -> Stream a

see http://en.wikipedia.org/wiki/Formal_power_series

inversion s * s == unit_product

quotient_invert :: Integral a => Stream a -> Stream a

unit_product :: Num a => Stream a

unit_product is the unit element of (*) for streams.

zero :: Num a => Stream a

zero is the unit element of (+) for streams

stream_if :: Stream Bool -> Stream a -> Stream a -> Stream a

either :: (a -> c) -> (b -> c) -> Stream (Either a b) -> Stream c

split_either :: Stream (Either a b) -> (Stream a, Stream b)

split_either_bool :: Stream (Either a b) -> (Stream Bool, Stream a, Stream b)

join_either :: Stream Bool -> Stream a -> Stream b -> Stream (Either a b)

subtract :: Num a => Stream a -> Stream a -> Stream a

prefix :: [a] -> Stream a -> Stream a

add a list as a prefix to a stream

fromList :: Num a => [a] -> Stream a

The cycle operation is better than fromList, if the list is infinite, because it will ensure the result is infinite (and no Num constraint is needed).

toList :: Stream a -> [a]

filter :: (a -> Bool) -> Stream a -> Stream a

filter f s for streams does not terminate, if the input stream never contains any elements e for which f e == True.

remove :: Eq a => [a] -> Stream a -> Stream a

interleave_count :: Integer -> Stream a -> Stream a -> Stream a

interleave_count is like interleave except that i elements of each stream are taken from each stream at each stream.

uninterleave_index :: Integer -> Stream a -> Stream a

interleave :: Stream a -> Stream a -> Stream a

interleave two streams such that even indexed elements of result are from first input stream and odd indexed elements of result are from second stream.

interleave_either :: Stream a -> Stream b -> Stream (Either a b)

version of interleave that produces either instances

uninterleave :: Stream a -> (Stream a, Stream a)

split a stream, elements with even index go to first result stream, rest to second.

interleave3 :: Stream a -> Stream a -> Stream a -> Stream a

three stream interleave

uninterleave3 :: Stream a -> (Stream a, Stream a, Stream a)

three stream uninterleave

uninterleave_indices :: Integer -> Integer -> Stream a -> Stream a

interleave_lst :: [Stream a] -> Stream a

interleave a non-empty list of streams.

uninterleave_lst :: Integer -> Stream a -> [Stream a]

uninterleave to an indicated number of streams.

interleave_queue :: (Queue :*: Stream) a -> Stream a

interleave a queue of streams.

uninterleave_queue :: Integer -> Stream a -> (Queue :*: Stream) a

uninterleave to a queue of streams

square :: Num a => [[a]] -> (Stream :*: Stream) a

drop :: Integer -> Stream a -> Stream a

drop a specified number of elements from beginning of a stream.

take :: Integer -> Stream a -> [a]

take a specified number of elements from the beginning of the stream.

splitAt :: Integer -> Stream a -> ([a], Stream a)

split a stream from index.

take2d :: (Integer, Integer) -> (Stream :*: Stream) a -> ([] :*: []) a

drop2d :: (Integer, Integer) -> (Stream :*: Stream) a -> (Stream :*: Stream) a

countWhile :: (a -> Bool) -> Stream a -> Integer

count how many elements from a beginning of a stream satisfy a predicate.

takeWhile :: (a -> Bool) -> Stream a -> [a]

take elements from a stream while predicate is true.

dropWhile :: (a -> Bool) -> Stream a -> Stream a

drop elements from a stream while predicate is true.

span :: (a -> Bool) -> Stream a -> ([a], Stream a)

cycle :: [a] -> Stream a

The cycle operation is better than fromList for converting a list to stream, if the list is infinite, because it will ensure the result is infinite.

cycle_queue :: Queue a -> Stream a

use elements of a queue to produce an infinite stream.

stirling_numbers :: Num a => (Stream :*: Stream) a

stirling numbers are the numbers obtained from z * (z+1) * ... * (z+n)

negative_stirling_numbers :: Num a => (Stream :*: Stream) a

natural_prefixes :: Num a => Stream [a]

toSquare :: (Stream :*: Stream) a -> [[a]]

print_square :: PpShow a => (Stream :*: Stream) a -> Doc

power :: Fractional a => Integer -> Stream a

stream of powers of a given integer for fractional items

power_integral :: Integral a => Integer -> Stream a

stream of powers of a given integer for integral items.

index_powers :: Num a => Stream a -> Stream a

substitute :: Num a => Stream a -> Stream a -> Stream a

fourier_ :: RealFloat a => Complex a -> Stream (Complex a) -> Stream (Complex a)

exp_generating_function :: Stream Double

pre_subst :: Num a => Stream a -> Stream a -> Stream [a]

subst :: Num a => Stream a -> Stream a -> Stream a

multiply :: Num a => a -> Stream a -> Stream a

powers :: Num a => Stream a -> (Stream :*: Stream) a

input stream elements are raised to n'th power.

cauchy_powers :: Num a => a -> Stream a

input element is raised to successive increasing powers (first element of the stream is n^1)

binomial_coefficient :: Integral a => Integer -> Integer -> a

The elements of pascal triangle are binomial coefficients.

http://en.wikipedia.org/wiki/Binomial_coefficient

generating_sqrt :: Integral a => Stream (Ratio a) -> Stream (Ratio a)

catalan_numbers :: Integral a => Stream a

http://en.wikipedia.org/wiki/Catalan_number

catalan_numbers_floating :: (Eq a, Floating a) => Stream a

http://en.wikipedia.org/wiki/Catalan_number of floating elements.

pascal_triangle :: Integral a => (Stream :*: Stream) a

pascal triangle, elements are binomial coefficients. https://en.wikipedia.org/wiki/Pascal%27s_triangle this expands "down" (leaving zero elements)

pascal_triangle_diag :: Integral a => (Stream :*: Stream) a

pascal triangle which expands towards the diagonal

binomial_coefficients :: Integral a => Stream [a]

exp_approx :: Fractional a => a -> Stream a

polynomial_powers :: Integral a => a -> a -> Stream a

computes (a + b)^n for all n, using binomial coefficients. Hint: use z in one of the argument. pascal_triangle == Matrix $ polynomial_powers 1 z Note: stream_powers can sometimes be a more efficient alternative.

fib :: Integral a => Stream a

a stream of fibonacci numbers, each element is a sum of two previous elements. 1,1,2,3,5,8,13,21,34,55,...

triangular_numbers :: Integral a => Stream a

Triangular numbers. https://en.wikipedia.org/wiki/Generating_function 1,3,6,10,15,21,28,36,...

squares :: Integral a => Stream a

alternating_signs :: Fractional a => Stream a

1.0,-1.0,1.0,-1.0,...

alternating_bits :: Fractional a => Stream a

1.0,0.0,1.0,0.0,1.0,0.0,...

alternating_possibly_negative_bits :: Fractional a => Stream a

odd_integers :: Integral a => Stream a

stream of odd integers

odd_integers_ :: Integral a => Stream a

integers_divisible_by :: Integer -> Stream Integer

This is an ideal of the set of integers, usually denoted nZ, e.g. set of integers divisible by n.

positive_residue_class_modulo :: Integer -> Integer -> Stream Integer

negative_residue_class_modulo :: Integer -> Integer -> Stream Integer

residue_class_modulo :: Integer -> Integer -> Stream Integer

This function produces the equivalence class of m mod n. The numbers produced by the stream are considered as equivalent by the equivalence class.

peirces_triangle_func :: (Num a, Ord a) => a -> a -> a

peirces_triangle is the number of set partitions. From Knuth: The art of Computer Programming, Volume 4 "Generating all combinations and partitions", page 64.

peirces_triangle :: (Num a, Ord a) => (Stream :*: Stream) a

peirces_triangle is the number of set partitions. From Knuth: The art of Computer Programming, Volume 4 "Generating all combinations and partitions", page 64.

peirces_triangle_list :: (Num a, Ord a) => Stream [a]

falling_factorial_powers_diag :: (Stream :*: Stream) Integer

https://en.wikipedia.org/wiki/Injective_function?wprof=sfla1 falling factorial powers specify number of injective functions from domain with specified number of elements to codomain with specified number of elements. https://en.wikipedia.org/wiki/Falling_and_rising_factorials?wprof=sfla1

Be careful about indices here, could cause off-by-one error!

falling_factorial_powers :: (Eq a, Num a) => (Stream :*: Stream) a

rising_factorial_powers :: (Eq a, Num a) => (Stream :*: Stream) a

bell_numbers :: (Num a, Ord a) => Stream a

bell numbers https://en.wikipedia.org/wiki/Bell_number. also see Knuth: The Art of Computer Programming, Volume 4.

integer_partitions_with_parts :: Integral a => Integer -> Stream a

square_root :: Integral a => a -> Stream a

newton's method of computing square root approximations:

fractional_square_root :: (Fractional a, Eq a) => a -> Stream a

detect_cycle :: Eq a => Stream a -> [a]

This is an implementation of Floyd's cycle-finding algorithm http://en.wikipedia.org/wiki/Cycle_detection.

integer_partitions :: Integral a => (Stream :*: Stream) a

primes :: Integral t => Stream t

Stream of prime numbers generated using a sieve.

https://wiki.haskell.org/Prime_numbers

derivative :: Num a => Stream a -> Stream a

derivative calculates the derivative of the generating function.

subtract_ordered :: Ord a => Stream a -> Stream a -> Stream a

join_ordered :: Ord a => Stream a -> Stream a -> Stream a

monotonic_membership :: Ord a => Stream a -> Stream a -> Stream Bool

for two monotone streams, compares whether all elements of the first stream are also in the second stream. produces True for every element of the first stream, if that element is part of the second stream. produces error if the streams are not monotone.

sqrt_stream :: Floating a => Stream a -> Stream a

Square root algorithm is from Simpson: Power Series, http://caps.gsfc.nasa.gov/simpson/ref/series.pdf

stream_epsilon :: Fractional a => (Stream :*: Stream) a

another_epsilon :: Fractional a => (Stream :*: Stream) a

stream_derivate :: (Fractional a, Limiting a) => (Stream a -> Stream a) -> Stream a -> Closure (Stream a)

derivate a stream function at a particular point.

stream_integral :: (Limiting b, Fractional b, Enum b) => (Stream b -> Stream b) -> (Stream b, Stream b) -> Closure (Stream b)

stream_pi :: (Limiting a, Fractional a, Enum a, Eq a, Floating a) => Closure (Stream a)

https://en.wikipedia.org/wiki/Pi

complex_pi :: RealFloat a => Stream (Complex a)

twin_primes :: Integral a => Stream a

http://en.wikipedia.org/wiki/Twin_prime

gcd_primes_diff :: Stream Integer

http://en.wikipedia.org/wiki/Formula_for_primes

gcd_primes :: Stream Integer

gcd_prime_gen :: Integral p => p -> p