Math.Number.StreamInterface

Documentation

class (Monad str, StreamBuilder str, StreamObserver str) => Limiting str a where

Idea: limit produces an exact value from a monotone sequence of approximations

https://ncatlab.org/nlab/show/closure+operator

The Closure represents a sequentially closed set of elements

Associated Types

data Closure str a

Methods

limit :: str a -> Closure str a

approximations :: Closure str a -> str a

class (Fractional a, Limiting str a) => Infinitesimal str a where

Minimal complete definition

Nothing

Methods

epsilon_closure :: Closure str a

epsilon_stream :: str a

class Applicative str => StreamBuilder str where

Methods

pre :: a -> str a -> str a infixr 5

class Applicative str => StreamObserver str where

Methods

shead :: str a -> a

stail :: str a -> str a

data Stream a

Data structure of infinite lazy streams.

Constructors

Pre infixr 5

Fields

shead_impl :: a

stail_lazy :: Stream a

class Infinitesimal Stream a => Closed a where

Methods

accumulation_point :: Closure Stream a -> a

type DerivateConstraint str a = (Monad str, Num a)

partial_derivate :: (Closed eps, Fractional eps) => (eps -> a -> a) -> (a -> eps) -> a -> eps

partial_derive :: (Closed a, Infinitesimal Stream a, Fractional a) => (a -> a -> a) -> (a -> a) -> a -> a

derivate_around :: Infinitesimal str a => (a -> a) -> a -> Closure str a

derivate_around doesn't require f to be defined at x, but requires limits from both sides of x to exist [it never evaluates f at x]. \[ \lim_{\epsilon \rightarrow 0} {{f(x+\epsilon)-f(x-\epsilon)}\over{2\epsilon}} \]

vector_derivate :: (Infinitesimal str (Scalar a), VectorSpace a, Limiting str a) => (Scalar a -> a) -> Scalar a -> Closure str a

derivate_closed :: Infinitesimal str a => (a -> a) -> a -> Closure str a

newtons_method_real :: (Infinitesimal Stream a, Closed a) => (a -> a) -> a -> Closure Stream a

iterate_stream :: StreamBuilder str => (a -> a) -> a -> str a

pseudo_derivate :: (Fractional r, Limiting str r, Infinitesimal str a) => (a -> r) -> a -> Closure str r

https://en.wikipedia.org/wiki/Differential_calculus

partial_derivate1_2 :: Infinitesimal str a => (a -> b -> a) -> a -> b -> Closure str a

\[\lim_{\epsilon\rightarrow 0}{{f(a+\epsilon,b)-f(a,b)}\over{\epsilon}}\]

partial_derivate2_2 :: Infinitesimal str a => (b -> a -> a) -> b -> a -> Closure str a

\[\lim_{\epsilon\rightarrow 0}{{f(a,b+\epsilon)-f(a,b)}\over{\epsilon}}\]

partial_derivate1_3 :: Infinitesimal str a => (a -> b -> c -> a) -> a -> b -> c -> Closure str a

\[\lim_{\epsilon\rightarrow 0}{{f(a+\epsilon,b,c)-f(a,b,c)}\over{\epsilon}}\]

partial_derivate2_3 :: Infinitesimal str a => (b -> a -> c -> a) -> b -> a -> c -> Closure str a

\[\lim_{\epsilon\rightarrow 0}{{f(a,b+\epsilon,c)-f(a,b,c)}\over{\epsilon}}\]

partial_derivate3_3 :: Infinitesimal str a => (b -> c -> a -> a) -> b -> c -> a -> Closure str a

\[\lim_{\epsilon\rightarrow 0}{{f(a,b,c+\epsilon)-f(a,b,c)}\over{\epsilon}}\]

constant :: a -> Stream a

stream consisting of the same element repeated.

stream_matrix_impl :: Stream a -> Stream (Stream a, Stream a) -> (Stream :*: Stream) a

stream_diagonal_impl :: (Stream :*: Stream) a -> Stream a

stream_codiagonal_impl :: (Stream :*: Stream) a -> Stream (Stream a, Stream a)