Math.Number.TypeUnits

Documentation

data DDimension

Constructors

DDimension

Fields

dlength_power :: SNat

dweight_power :: SNat

dtime_power :: SNat

dcurrent_power :: SNat

dtemperature_power :: SNat

dluminosity_power :: SNat

dsubstance_power :: SNat

dcircular_power :: SNat

type DDimensionless = DDimension Ze Ze Ze Ze Ze Ze Ze Ze

type DLength = DDimension SOne Ze Ze Ze Ze Ze Ze Ze

type DWeight = DDimension Ze SOne Ze Ze Ze Ze Ze Ze

type DTime = DDimension Ze Ze SOne Ze Ze Ze Ze Ze

type DCurrent = DDimension Ze Ze Ze SOne Ze Ze Ze Ze

type DTemperature = DDimension Ze Ze Ze Ze SOne Ze Ze Ze

type DLuminosity = DDimension Ze Ze Ze Ze Ze SOne Ze Ze

type DSubstance = DDimension Ze Ze Ze Ze Ze Ze SOne Ze

type DAngle = DDimension Ze Ze Ze Ze Ze Ze Ze SOne

type DAngularSpeed = DDimension Ze Ze (SNegate SOne) Ze Ze Ze Ze SOne

type DAngularAcceleration = DDimension Ze Ze (SNegate STwo) Ze Ze Ze Ze SOne

type DSolidAngle = DDimension Ze Ze Ze Ze Ze Ze Ze STwo

type DSquareLength = DDimension STwo Ze Ze Ze Ze Ze Ze Ze

type DCubicLength = DDimension SThree Ze Ze Ze Ze Ze Ze Ze

type DVelocity = DDimension SOne Ze (SNegate SOne) Ze Ze Ze Ze Ze

type DAcceleration = DDimension SOne Ze (SNegate STwo) Ze Ze Ze Ze Ze

type DMass = DDimension Ze SOne Ze Ze Ze Ze Ze Ze

type DFrequency = DDimension Ze Ze (SNegate SOne) Ze Ze Ze Ze Ze

type DForce = DDimension SOne SOne (SNegate STwo) Ze Ze Ze Ze Ze

type DInductance = DDimension STwo SOne (SNegate STwo) (SNegate STwo) Ze Ze Ze Ze

type DTorque = DDimension STwo SOne (SNegate STwo) Ze Ze Ze Ze (SNegate SOne)

https://en.wikipedia.org/wiki/Torque

type DPressure = DDimension (SNegate SOne) SOne (SNegate STwo) Ze Ze Ze Ze Ze

type DEnergy = DDimension STwo SOne (SNegate STwo) Ze Ze Ze Ze Ze

type DPower = DDimension STwo SOne (SNegate SThree) Ze Ze Ze Ze Ze

type DCharge = DDimension Ze Ze SOne SOne Ze Ze Ze Ze

type DVoltage = DDimension STwo SOne (SNegate SThree) (SNegate SOne) Ze Ze Ze Ze

type DCapacitance = DDimension (SNegate STwo) (SNegate SOne) SFour STwo Ze Ze Ze Ze

type DResistance = DDimension STwo SOne (SNegate SThree) (SNegate STwo) Ze Ze Ze Ze

type DConductance = DDimension (SNegate STwo) (SNegate SOne) SThree STwo Ze Ze Ze Ze

type DFlux = DDimension STwo SOne (SNegate STwo) (SNegate SOne) Ze Ze Ze Ze

type DFluxDensity = DDimension Ze SOne (SNegate STwo) (SNegate SOne) Ze Ze Ze Ze

type DLuminousEnergy = DDimension Ze Ze SOne Ze Ze SOne Ze STwo

https://en.wikipedia.org/wiki/Luminous_flux

type DLuminousIntensity = DDimension Ze Ze Ze Ze Ze SOne Ze Ze

type DLuminousFlux = DDimension Ze Ze Ze Ze Ze SOne Ze STwo

type DLuminousExitance = DDimension (SNegate STwo) Ze Ze Ze Ze SOne Ze STwo

type DIlluminance = DDimension (SNegate STwo) Ze Ze Ze Ze SOne Ze STwo

type DLuminance = DDimension (SNegate STwo) Ze Ze Ze Ze SOne Ze Ze

type DRadioactivity = DDimension Ze (SNegate SOne) Ze Ze Ze Ze Ze Ze

type DAbsorbedDose = DDimension STwo (SNegate STwo) Ze Ze Ze Ze Ze Ze

type DEquivalentDose = DDimension (SNegate STwo) (SNegate SOne) SThree STwo Ze Ze Ze Ze

type DCatalyticActivity = DDimension Ze (SNegate SOne) Ze Ze Ze Ze SOne Ze

type family DimensionPlus (a :: DDimension) (b :: DDimension) :: DDimension

type family DimensionMinus (a :: DDimension) (b :: DDimension) :: DDimension

unit_multiply :: (Scalar (DUnit (DimensionPlus u u')) ~ Double, Scalar (DUnit u) ~ Double, Scalar (DUnit u') ~ Double, Unit (DUnit u), Unit (DUnit u'), Unit (DUnit (DimensionPlus u u'))) => DUnit u -> DUnit u' -> DUnit (DimensionPlus u u')

unit_divide :: (Scalar (DUnit (DimensionMinus u u')) ~ Double, Scalar (DUnit u) ~ Double, Scalar (DUnit u') ~ Double, Unit (DUnit u), Unit (DUnit u'), Unit (DUnit (DimensionMinus u u'))) => DUnit u -> DUnit u' -> DUnit (DimensionMinus u u')

half_circle_angle :: DUnit DAngle

full_circle_angle :: DUnit DAngle

data family DUnit (u :: DDimension)