Math.Tools.CoFunctor

Documentation

type CoFunctor p = Contravariant p

For exposition of classical logic used here, see "Lawvere,Rosebrugh: Sets for mathematics".

(x |>> f) |>> g == x |>> (f . g)
inverse_image g (inverse_image f x) == inverse_image (f . g) x

Notice, CoFunctor is nowadays in base, so please use Data.Functor.Contravariant.Contravariant

inverseImage :: CoFunctor p => (a -> b) -> p b -> p a

(|>>) :: CoFunctor p => p b -> (a -> b) -> p a

class Representable p where

Associated Types

type Representation p :: * -> *

Methods

represent :: Representation p a -> p a

(!>) :: p a -> (Representation p a -> p a) -> p a

class Propositional p where

Methods

runProposition :: p a -> a -> p ()

class Propositional p => BinaryLogic p where

Methods

toBool :: p () -> Bool

class CoFunctor p => Relational p where

Methods

relation :: (a -> p b) -> p (a, b)

unrelation :: p (a, b) -> a -> p b

class CoFunctor p => Existential p where

Methods

directImage :: (Universe a, Eq b) => (a -> b) -> p a -> p b

universalImage :: (Universe a, Eq b) => (a -> b) -> p a -> p b

class CoFunctor p => Difference p minus | p -> minus where

Methods

difference :: (p a -> p b) -> p (b `minus` a)

undifference :: p (b `minus` a) -> p a -> p b

class (Eq c, CoFunctor p) => HasEqualizers p c where

Minimal complete definition

equalizer

Methods

equalizer :: (a -> c) -> (a -> c) -> p a

pullback :: (a -> c) -> (b -> c) -> p (a, b)

class Functor p => Topology p where

Methods

arbitraryUnion :: [p a] -> p a

finiteIntersect :: [p a] -> p a

class HasTerminalObject t where

Methods

terminalArrow :: a -> t

class CoFunctor p => HasVariables p t where

Methods

variable :: String -> p t

class CoFunctor p => PropositionalLogic p where

Methods

(-&-) :: p a -> p a -> p a

(-|-) :: p a -> p a -> p a

invert :: p a -> p a

class PropositionalLogic p => ModalLogic p where

https://en.wikipedia.org/wiki/Modal_logic

Minimal complete definition

Nothing

Methods

eventually :: p a -> p a

always :: p a -> p a

class CoFunctor p => Continuous p where

Methods

expand :: p (Either a b) -> (p a, p b)

class CoFunctor p => SubtractiveLogic p where

Methods

(-\-) :: p a -> p a -> p a

class CoFunctor p => ImplicativeLogic p where

Methods

(-=>-) :: p a -> p a -> p a

class CoFunctor p => BooleanLogic p where

Methods

true :: p a

class CoFunctor p => UniversalLogic p where

Methods

singleton :: Eq a => a -> p a

allMembers :: Universe a => p a

class CoFunctor p => Universal p where

Methods

exist :: Universe b => p (a, b) -> p a

univ :: Universe b => p (a, b) -> p a

member :: Propositional p => a -> p a -> p ()

isIn :: BinaryLogic p => a -> p a -> Bool

isNotIn :: BinaryLogic p => a -> p a -> Bool

terminalPart :: Propositional p => p () -> p ()

propositionalImplication :: PropositionalLogic p => p a -> p a -> p a

relativeComplement :: PropositionalLogic p => p a -> p a -> p a

intersectList :: (BooleanLogic p, PropositionalLogic p) => [p a] -> p a

unionList :: (BooleanLogic p, PropositionalLogic p) => [p a] -> p a

characteristic :: BooleanLogic p => (a -> p ()) -> p a

falseProp :: (BooleanLogic p, PropositionalLogic p) => p a

(-<=>-) :: (ImplicativeLogic p, PropositionalLogic p) => p a -> p a -> p a

symmetricDifference :: (PropositionalLogic p, SubtractiveLogic p) => p a -> p a -> p a

boundary :: (PropositionalLogic p, BooleanLogic p, SubtractiveLogic p) => p a -> p a

currySubst :: CoFunctor p => ((a, b) -> c) -> p (b -> c) -> p a

uncurrySubst :: CoFunctor p => (a -> b -> c) -> p c -> p (a, b)

dualize :: Difference p minus => (a -> b) -> p (a `minus` b)

coin1 :: CoFunctor p => p a -> p (a, b)

coin2 :: CoFunctor p => p b -> p (a, b)

cofst :: CoFunctor p => p (Either a b) -> p a

cosnd :: CoFunctor p => p (Either a b) -> p b

cocase :: CoFunctor p => (Either a b -> c) -> p c -> (p a, p b)

copair :: (Propositional p, BooleanLogic p) => p a -> p b -> p (Either a b)

pairSubst :: CoFunctor p => (a -> b) -> (a -> c) -> p (b, c) -> p a

graphSubst :: CoFunctor p => (a -> b) -> p (a, b) -> p a

data Assertion a

Constructors

PrimAssert

Fields

assertedCondition :: a -> Bool

NotAssert

Fields

assertedNegations :: Assertion a

AndAssert

Fields

assertedConjunctions :: [Assertion a]

OrAssert

Fields

assertedDisjunctions :: [Assertion a]