Math.Tools.Map

Documentation

allMaps :: Ord k => [k] -> [a] -> [Map k a]

allTotalMaps :: Ord k => [k] -> [a] -> [Map k a]

lookupMapM :: (MonadFail m, Show i, Ord i) => Map i a -> i -> m a

unzip :: Map i (a, b) -> (Map i a, Map i b)

unzip3 :: Map i (a, b, c) -> (Map i a, Map i b, Map i c)

foldM :: Monad m => (a -> b -> m b) -> m b -> Map i a -> m b

foldKeyM :: Monad m => (i -> a -> b -> m b) -> m b -> Map i a -> m b

mapM :: (Monad m, Ord i) => (a -> m b) -> Map i a -> m (Map i b)

sequence_ :: (Monad m, Ord i) => Map i (m a) -> m ()

sequence_map :: (Monad m, Ord i) => Map i (m a) -> m (Map i a)

sequenceByKeyM :: (Monad m, Ord i) => Map i (i -> m a) -> m (Map i a)

mapByKeyM :: (Monad m, Ord i) => (i -> a -> m b) -> Map i a -> m (Map i b)

joinMap :: Map String (Map String a) -> Map String a

splitMapByKey :: (Ord s, Ord m) => Map (m, s) a -> Map m (Map s a)

partitionMap :: Ord k => Map k a -> (a -> Either b c) -> (Map k b, Map k c)

intersectZipMapA :: (Ord i, ArrowChoice arr) => arr (Map i a, Map i b) (Map i (a, b))

intersectMapA :: (Ord i, ArrowChoice arr) => arr (a, b) c -> arr (Map i a, Map i b) (Map i c)

zipWithMapM :: (MonadFail m, Ord i) => (a -> b -> m c) -> Map i a -> Map i b -> m (Map i c, Map i a, Map i b)

intersectMapKeyA :: (Ord i, ArrowChoice arr) => arr (i, (a, b)) c -> arr (Map i a, Map i b) (Map i c)

mapMapWithKey :: (ArrowChoice arr, Eq i) => arr (i, e) (i, f) -> arr (Map i e) (Map i f)

unionMapA :: (Ord i, ArrowChoice arr) => arr (a, a) a -> arr (Map i a, Map i a) (Map i a)

sequenceMap :: (Ord i, Monad m) => Map i (m b) -> m (Map i b)

sequenceMapA :: (Arrow arr, Ord i) => Map i (arr a b) -> arr a (Map i b)

sequenceMapWithKeyA :: (Arrow arr, Ord i) => Map i (arr (a, i) b) -> arr a (Map i b)

foldMapA :: ArrowApply arr => arr (a, b) b -> arr (Map i a, b) b

foldMapKeyA :: ArrowApply arr => arr (i, a, b) b -> arr (Map i a, b) b

composeMapA :: (Arrow arr, Ord i, Ord a) => arr (Map i a, Map a b) (Map i b)

count_frequency :: Ord b => (a -> b) -> [a] -> Map b Integer

benford :: Show a => a -> Char

Benford's law: https://en.wikipedia.org/wiki/Benford%27s_law

count_benford :: Show a => [a] -> Map Char Integer

Orphan instances

(Ord i, ArrowChoice arr) => FunctorArrow (Map i :: Type -> Type) (arr :: Type -> Type -> Type) (arr :: Type -> Type -> Type)

Instance details

Methods

amap :: forall (c :: k) (d :: k). arr c d -> arr (Map i c) (Map i d)

MonadFail (Map String)

Instance details

Methods

fail :: String -> Map String a

Applicative (Map String)

Instance details

Methods

pure :: a -> Map String a

(<*>) :: Map String (a -> b) -> Map String a -> Map String b

liftA2 :: (a -> b -> c) -> Map String a -> Map String b -> Map String c

(*>) :: Map String a -> Map String b -> Map String b

(<*) :: Map String a -> Map String b -> Map String a

Monad (Map String)

Instance details

Methods

(>>=) :: Map String a -> (a -> Map String b) -> Map String b

(>>) :: Map String a -> Map String b -> Map String b

return :: a -> Map String a

(Ord a, Eq b, Universe a, Universe b) => Universe (Map a b)

Instance details

Methods

allElements :: [Map a b]