Math.Tools.Queue

Documentation

data Queue a

a queue is really a list with a missing element which is the focus of the current operations. This idea comes from zipper pattern. http://www.haskell.org/haskellwiki/Zipper

Constructors

MakeQueue

Fields

queue_newest_first_prefix :: ![a]

queue_oldest_first_postfix :: ![a]

newQueueM :: IO (IORef (Queue a))

enqueueM :: IORef (Queue a) -> a -> IO ()

dequeueM :: IORef (Queue a) -> IO a

split_list :: [a] -> Int -> (a, Queue a)

join_list :: a -> Queue a -> [a]

isEmpty :: Queue a -> Bool

join :: Queue (Queue a) -> Queue a

join works in round-robin fashion. Note if all inner queues are empty, this goes to infinite loop trying to find elements from empty queues.

join' :: Queue (Queue a) -> Queue a

Alternative implementation of join, works in reverse compared to join.

fold :: Arrow arr => (a -> arr b b) -> Queue a -> arr b b

interleave_queue :: Queue a -> Queue a -> Queue a

singleton :: a -> Queue a

enqueue :: a -> Queue a -> Queue a

enqueue_back :: a -> Queue a -> Queue a

dequeue :: Monad m => Queue a -> m (a, Queue a)

dequeue_back :: Monad m => Queue a -> m (Queue a, a)

dequeue_rotate :: Monad m => Queue a -> m (a, Queue a)

dequeue_rotate_back :: Monad m => Queue a -> m (Queue a, a)

rotate :: Queue a -> Queue a

rotate_back :: Queue a -> Queue a

reduce :: Queue a -> Queue a

unreduce :: Queue a -> Queue a

fromList :: [a] -> Queue a

toList :: Queue a -> [a]

unzip :: Queue (a, b) -> (Queue a, Queue b)

zip :: Queue a -> Queue b -> Queue (a, b)

partitionEither :: Queue (Either a b) -> (Queue a, Queue b)

append :: Queue a -> Queue a -> Queue a

dequeue_current :: Monad m => Queue a -> m (a, Queue a)

replace_current :: a -> Queue a -> Queue a

forward :: a -> Queue a -> Queue a

backward :: a -> Queue a -> Queue a

reverse :: Queue a -> Queue a

sequence :: Monad m => Queue (m a) -> m (Queue a)

mapM :: Monad m => (a -> m b) -> Queue a -> m (Queue b)