Math.Tools.Visitor

Documentation

class Visitor v where

loosely based on GOF: Design Patterns visitor pattern, and also on Haskell foldr function and common variations of it.

Associated Types

data Fold v :: * -> *

Methods

visit :: Fold v a -> v -> a

class Builder v where

Associated Types

data Unfold v :: * -> *

Methods

build :: Unfold v a -> a -> v

class Unfoldable v where

http://blog.jle.im/entry/fixed-length-vector-types-in-haskell-2015.html | we are allowing return of the state from the operation. | and are using state transformer monad for easier writing | of instance declarations

Methods

unfoldF :: Monad m => StateT s m a -> StateT s m (v a)

unfoldVectorM :: (Monad m, Unfoldable v) => (s -> (a, s)) -> s -> m (v a, s)

unfoldVector :: Unfoldable v => (s -> (a, s)) -> s -> (v a, s)

class FunctorBuilder f where

Methods

mapBuilder :: Builder v => Unfold v a -> Unfold (f v) (f a)

class Builder v => ConstantBuilder v where

Methods

newBuilder :: v -> Unfold v a

data VArrow a b where

Constructors

VArrowCompose :: VArrow b c -> VArrow a b -> VArrow a c

VArrowId :: (Visitor a, Builder a) => Fold a b -> Unfold a b -> VArrow a a

varrow :: VArrow a b -> a -> b

transform_visitor :: (Visitor v, Builder v) => Fold v b -> Unfold v a -> a -> b

class Builder v => Unbuilder v b where

Methods

unbuild :: (Unfold v b -> b -> b) -> v -> Unfold v b

class Visitor v => IdVisitor v where

Methods

idFold :: Fold v v

class Visitor v => ComposableVisitor v where

Minimal complete definition

embed

Methods

embed :: (v -> b) -> Fold v b

(+.+) :: Visitor b => Fold b c -> Fold v b -> Fold v c

class Builder v => IdBuilder v where

Methods

idUnfold :: Unfold v v

fix2 :: (a -> b -> b) -> a -> b

fix3 :: (a -> b -> c -> c) -> a -> b -> c

bind2 :: (a -> b -> b) -> a -> a -> b

bind3 :: (a -> b -> b) -> a -> a -> a -> b

bindn :: (a -> b -> b) -> [a] -> b

mappingBuilder :: Ord i => Unfold (i -> v) (Map i v)

predicate :: Ord i => i -> Unfold Bool (Set i)

var :: Ord i => ((Map i v -> v) -> t) -> i -> t

boolvar :: Ord i => i -> Unfold Bool (Map i Bool)

intvar :: Ord i => i -> Unfold Int (Map i Int)

doublevar :: Ord i => i -> Unfold Double (Map i Double)

integervar :: Ord i => i -> Unfold Integer (Map i Integer)

listfold :: a -> (v -> a -> a) -> Fold [v] a

ziplistfold :: Fold [v] ([t] -> [(v, t)])