Math.Graph.Algorithms

Documentation

dfsM :: (Monad m, Ord a, GraphMonoid mon) => a -> InGraphM mon a m [a]

Compute depth first search on a graph, starting from a given root vertex.

dfsIntM :: (Monad m, Ord a, GraphMonoid mon) => a -> StateT (Set a) (InGraphM mon a m) [a]

bfsM :: (Monad m, Ord n, GraphMonoid mon) => n -> InGraphM mon n m (Map n n)

Breadth first search: bfsM produces a map from child to parent nodes in a bfs tree.

bfsIntM :: (Monad m, GraphMonoid mon, Ord n) => StateT (Queue n, Map n Integer) (InGraphM mon n m) (Map n n)

spanningTreeM :: (Monad m, GraphMonoid mon, Ord n) => n -> InGraphM mon n m (Tree n)

Spanning tree rooted at a given graph vertex.

spanningForestM :: (Monad m, GraphMonoid mon, Ord n) => [n] -> InGraphM mon n m (Forest n)

Spanning forest rooted at the given list of vertices.

spanningTreeIntM :: (Monad m, GraphMonoid mon, Ord n) => n -> StateT (Set n) (InGraphM mon n m) (Tree n)

spanningForestIntM :: (Monad m, GraphMonoid mon, Ord n) => [n] -> StateT (Set n) (InGraphM mon n m) (Forest n)