Math.Graph.Reversible

Description

Graph representation as a set with action of a monoid. See Lawvere,Rosebrugh: Sets for Mathematics for details.

This module implements graphs using a representation as a set with an action of a monoid.

For directed graph, the monoid Three is used, which represents operations of id,source and target

For undirected graph, the monoid Four is used, which represents operations of id,source, target and inverse.

See Math.Graph.InGraphMonad for how to read graphs in monadic context

See Math.Graph.InGraphA for how to read graphs in arrow context

Documentation

data Graph m a

Graph representation containing a set of elements. an element of a graph is either an edge or a vertex.

the action_endomorphism allows obtaining source vertex, target vertex and inverse edge of an element See GraphMonoid for details.

Note that the monoid data structure contains only names of the monoid elements - the action_endomorphism is then used to convert it to actual operation.

Constructors

Graph

Fields

elements :: !(Set a)

action_endomorphism :: m -> Endo a

action :: Graph m a -> a -> m -> a

action_rep :: Graph m a -> Action (Endo a) m

inverseImageG :: (m -> n) -> Graph n a -> Graph m a

emptyG :: Graph m b

vertexG :: a -> Graph m a

verticesFromSetG :: Set a -> Graph m a

verticesG :: Ord a => [a] -> Graph m a

edgeG :: Ord a => a -> a -> a -> Graph Three a

scalarVectorG :: VectorSpace v => Set v -> Graph (Scalar v) v

https://en.wikipedia.org/wiki/Module_(mathematics)

listG :: [a] -> Graph a [a]

plusVectorG :: VectorSpace v => Graph v v

integerG :: Num a => Graph a a

vectorG :: LinearTransform v v a => Set (v a) -> Graph ((v :*: v) a) (v a)

appG :: Set a -> Graph (a -> a) a

binopG :: Set b -> (a -> b -> b) -> Graph a b

outerG :: (Ord a, Ord b) => Graph m a -> Graph n b -> Graph (m, n) (a, b)

discriminatedUnionG :: (Ord a, Ord b) => Graph m a -> Graph m b -> Graph m (Either a b)

intersectionG :: Ord a => Graph m a -> Graph m a -> Graph m a

complementG :: ReversibleGraphMonoid m => Graph m a -> Graph m a

unionG :: (Ord a, Show a) => Graph m a -> Graph m a -> Graph m a

completeG :: Ord a => [a] -> (a -> a -> a) -> Graph Three a

In a complete graph, every pair of vertices (x,y) has a single link.

completeG \["1","2","3"\] \(++\) == [1, 2, 3] ; [11 = 1 -> 1, 12 = 1 -> 2, 13 = 1 -> 3, 21 = 2 -> 1,
                                             22 = 2 -> 2, 23 = 2 -> 3, 31 = 3 -> 1, 32 = 3 -> 2, 33 = 3 -> 3]

Note though that edge names produced from pair of vertices by the function given must be unique, no overlaps are allowed, so it is required that for call 'completeG vertices edge'

forall v,w : vertices . (v != v' || w != w') => edge v w != edge v' w'

edgesFromMapG :: Ord a => Map a (a, a) -> Graph Three a

edgesG :: Ord a => [(a, (a, a))] -> Graph Three a

reversibleToDirectedG :: Graph Four a -> Graph Three a

converts from undirected graph to directed graph

inverseGraphG :: Group m => Graph m a -> Graph m a

inverts all edges

mapG :: (Ord a, Ord b) => (a :==: b) -> Graph m a :==: Graph m b

mapG uses an isomorphism of graph elements to permute a graph.

reversibleEdgeG :: Ord a => a -> a -> a -> a -> Graph Four a

reversibleCompleteG :: Ord a => [a] -> (a -> a -> (a, a)) -> Graph Four a

edges argument has inputs from and to and should return names of forward and backward edges.

reversibleEdgesG :: Ord a => [((a, a), (a, a))] -> Graph Four a

The input data contains ((edge,reverse-edge),(startNode,endNode))

subobjectClassifierGraphG :: Graph Four String