Math.Matrix.Covector

Description

License: LGPL

For background material, see K. Chandrasekhara Rao: Functional Analysis.

See also: http://theoreticalminimum.com/courses/advanced-quantum-mechanics/2013/fall/lecture-1

Documentation

data Dual v

Data type for dual vectors.

Constructors

Covector

Fields

bracket :: v -> Scalar v

(*><) :: Dual v -> v -> Scalar v

class VectorSpace v => VectorDerivative v where

https://en.wikipedia.org/wiki/Curl_(mathematics) https://en.wikipedia.org/wiki/Laplace_operator https://en.wikipedia.org/wiki/Divergence https://en.wikipedia.org/wiki/Gradient

Minimal complete definition

divergence, grad

Methods

divergence :: (v -> v) -> Dual v

grad :: Dual v -> v -> v

laplace :: Dual v -> Dual v

class VectorDerivative v => VectorCrossProduct v where

Minimal complete definition

curl

Methods

curl :: (v -> v) -> v -> v

(∇) :: VectorDerivative v => Dual v -> v -> v

unicode support for gradient. Use with parenthesis.

(∇·) :: VectorDerivative v => (v -> v) -> Dual v

unicode support for divergence. Use with parenthesis.

(∇×) :: VectorCrossProduct v => (v -> v) -> v -> v

unicode support for curl. Use with parenthesis.

(∇·∇) :: VectorDerivative v => Dual v -> Dual v

unicode support for laplace. Use with parenthesis.

differential_commutator :: (VectorDerivative v, Scalar v ~ v) => Dual v -> Dual v -> v -> v

differential_commutator_endo :: (VectorDerivative v, Scalar v ~ v) => Dual v -> Dual v -> Endo v

curl_endo :: VectorCrossProduct v => Endo v -> Endo v

grad_endo :: VectorDerivative v => Dual v -> Endo v

divergence_endo :: VectorDerivative v => Endo v -> Dual v

class VectorSpace v => FiniteDimensional v where

https://en.wikipedia.org/wiki/Dual_space#Injection_into_the_double-dual

Minimal complete definition

finite

Methods

finite :: (Dual :*: Dual) v -> v

directional_derivative :: (VectorDerivative v, InnerProductSpace v) => v -> Dual v -> Dual v

dconst :: Scalar v -> Dual v

outer_vector :: (VectorSpace v, Scalar w ~ Scalar v) => Dual w -> v -> w -> v

conjugate_transpose :: (Transposable m n, ConjugateSymmetric a) => (m :*: n) a -> (n :*: m) a

http://en.wikipedia.org/wiki/Conjugate_transpose

bravector :: InnerProductSpace v => v -> Dual v

partial application of dot product on second argument (- %. v).

ketvector :: InnerProductSpace v => v -> Dual v

partial application of dot product on first argument: (v %. -).

kronecker :: (Eq i, Functor n, Functor m, Num c) => m i -> n i -> (m :*: n) c

bilinear :: (v -> v -> Scalar v) -> v -> Dual v

natural_double_dual :: v -> (Dual :*: Dual) v

https://en.wikipedia.org/wiki/Dual_space#Injection_into_the_double-dual

covector_kernel :: (Num (Scalar v), Eq (Scalar v)) => Dual v -> v -> Bool

adjoint :: Scalar a ~ Scalar b => (a -> b) -> Dual b -> Dual a

would like to make adjoint as instance of CoFunctor, but the constraint prevents. Instead we declare VectorSpaceMap.

data VectorSpaceMap v w where

Constructors

VectorSpaceMap :: Scalar v ~ Scalar w => (v -> w) -> VectorSpaceMap v w

scalar_map :: (Scalar a -> Scalar a) -> Dual a -> Dual a

operator_map :: ((v -> Scalar v) -> w -> Scalar w) -> Dual v -> Dual w

class (Num (Scalar v), VectorSpace v) => NumSpace v

vector space with scalars in Num class

class (Fractional (Scalar v), NumSpace v) => FractionalSpace v

vector space with fractional scalars