Math.Matrix.Covector

Description

License: LGPL

For background material, see K. Chandrasekhara Rao: Functional Analysis.

See also: http://theoreticalminimum.com/courses/advanced-quantum-mechanics/2013/fall/lecture-1

Documentation

data Dual v

Data type for dual vectors.

Constructors

Covector

Fields

bracketMap :: v -> Scalar v

contract :: (Dual :*: v) a -> (I :*: v) a -> Scalar (v a)

bracket :: Dual v -> v -> Scalar v

covector :: (v -> Scalar v) -> Dual v

(*><) :: Dual v -> v -> Scalar v

class VectorSpace v => VectorDerivative v where

https://en.wikipedia.org/wiki/Laplace_operator https://en.wikipedia.org/wiki/Divergence https://en.wikipedia.org/wiki/Gradient

Minimal complete definition

divergence, grad

Methods

divergence :: LinearMap v v -> Dual v

grad :: Dual v -> LinearMap v v

laplace :: Dual v -> Dual v

class VectorDerivative v => VectorCrossProduct v where

https://en.wikipedia.org/wiki/Curl_(mathematics)

Methods

curl :: LinearMap v v -> LinearMap v v

class VectorLaplacian v where

https://en.wikipedia.org/wiki/Vector_Laplacian

Minimal complete definition

Nothing

Methods

vector_laplace :: LinearMap v v -> LinearMap v v

vector_laplace :: VectorCrossProduct v => LinearMap v v -> LinearMap v v

class Functor f => ProjectionDual f a where

Methods

projection_dual :: f (Dual (f a))

(∇) :: VectorDerivative v => Dual v -> LinearMap v v

unicode support for gradient. Use with parenthesis.

(∇·) :: VectorDerivative v => LinearMap v v -> Dual v

unicode support for divergence. Use with parenthesis.

(·∇) :: (VectorDerivative v, InnerProductSpace v) => v -> Dual v -> Dual v

unicode support for directional derivative. Binary operator.

(∇×) :: VectorCrossProduct v => LinearMap v v -> LinearMap v v

unicode support for curl. Use with parenthesis.

(∇·∇) :: VectorDerivative v => Dual v -> Dual v

unicode support for laplace. Use with parenthesis.

differential_commutator :: (VectorDerivative v, Scalar v ~ v) => Dual v -> Dual v -> LinearMap v v

differential_commutator_endo :: (VectorDerivative v, Scalar v ~ v) => Dual v -> Dual v -> Endo v

curl_endo :: VectorCrossProduct v => Endo v -> Endo v

grad_endo :: VectorDerivative v => Dual v -> Endo v

divergence_endo :: VectorDerivative v => Endo v -> Dual v

laplace_endo :: VectorDerivative v => Endo (Dual v)

class VectorSpace v => FiniteDimensional v where

https://en.wikipedia.org/wiki/Dual_space#Injection_into_the_double-dual

Methods

finite :: (Dual :*: Dual) v -> v

directional_derivative :: (VectorDerivative v, InnerProductSpace v) => v -> Dual v -> Dual v

directional_derivative_endo :: (VectorDerivative v, InnerProductSpace v) => v -> Endo (Dual v)

dconst :: Scalar v -> Dual v

outer_vector :: (VectorSpace v, Scalar w ~ Scalar v) => Dual w -> v -> w -> v

conjugate_transpose :: (Transposable m n, ConjugateSymmetric a) => (m :*: n) a -> (n :*: m) a

http://en.wikipedia.org/wiki/Conjugate_transpose

(⊹) :: (Transposable m n, ConjugateSymmetric a) => (m :*: n) a -> (n :*: m) a

unicode alias (unicode HERMITIAN CONJUGATE MATRIX character)

is_hermitian :: (Eq a, ConjugateSymmetric a, Transposable m m, Applicative m, Foldable m) => (m :*: m) a -> Bool

https://en.wikipedia.org/wiki/Conjugate_transpose

is_skew_hermitian :: (Num a, Eq a, ConjugateSymmetric a, Applicative m, Transposable m m, Foldable m) => (m :*: m) a -> Bool

https://en.wikipedia.org/wiki/Conjugate_transpose

is_normal :: (Num a, Eq (Scalar (m a)), ConjugateSymmetric a, Foldable m, Applicative m, Transposable m m, InnerProductSpace (m a)) => (m :*: m) a -> Bool

https://en.wikipedia.org/wiki/Conjugate_transpose

bravector :: InnerProductSpace v => v -> Dual v

partial application of dot product on second argument (- %. v).

ketvector :: InnerProductSpace v => v -> Dual v

partial application of dot product on first argument: (v %. -).

kronecker :: (Eq i, Functor n, Functor m, Num c) => m i -> n i -> (m :*: n) c

dual_tensor :: (Scalar a -> Scalar b -> Scalar (a, b)) -> Dual a -> Dual b -> Dual (a, b)

bilinear :: (v -> v -> Scalar v) -> v -> Dual v

kronecker_dual :: (Eq a, Num (Scalar a)) => a -> Dual a

natural_double_dual :: v -> (Dual :*: Dual) v

https://en.wikipedia.org/wiki/Dual_space#Injection_into_the_double-dual

covector_kernel :: (Num (Scalar v), Eq (Scalar v)) => Dual v -> v -> Bool

adjoint :: LinearMap a b -> LinearMap (Dual b) (Dual a)

would like to make adjoint as instance of CoFunctor, but the constraint prevents. Instead we declare LinearMap.

adjoint_map :: Scalar b ~ Scalar a => (Scalar b -> Scalar a) -> LinearMap a b -> Dual b -> Dual a

matrix_arrow :: LinearTransform f g a => (f :*: g) a -> LinearMap (f a) (g a)

scalar_map :: LinearMap (Scalar a) (Scalar a) -> Dual a -> Dual a

operator_map :: ((v -> Scalar v) -> w -> Scalar w) -> Dual v -> Dual w

class (Num (Scalar v), VectorSpace v) => NumSpace v

vector space with scalars in Num class

class (Fractional (Scalar v), NumSpace v) => FractionalSpace v

vector space with fractional scalars

(-!<) :: Scalar w ~ Scalar v => LinearMap v w -> v -> w

data LinearMap v w where

Constructors

LinearMap :: Scalar w ~ Scalar v => (v -> w) -> LinearMap v w

dual_derivative :: (VectorDerivative v, VectorDerivative w, InnerProductSpace v, InnerProductSpace w) => (Dual v -> Dual w -> c) -> v -> w -> Dual v -> Dual w -> c

This computes \(DDM(\otimes,{\bar{x}},{\bar{y}}) = [({\bar{x}} \cdot \nabla) \times ({\bar{y}} \cdot \nabla)]\) where \(\times\) is the outer product for linear operators derived from product of dual spaces \(\otimes\). Notice \({\bar{x}} \cdot \nabla\) is normally called the directional derivative, so this is really a two dimensional version of directional derivative with some possibility to choose how two dual vectors are combined.

dual_differential_outer_product :: (VectorDerivative v, InnerProductSpace v) => v -> v -> Dual v -> Dual v -> Dual v

dual_differential_dot_product :: (VectorDerivative v, InnerProductSpace v, StandardBasis v) => v -> v -> Dual v -> Dual v -> Scalar v

norm_covector :: NormedSpace v => Dual v