Math.Matrix.Interface

Description

These should match standard definitions of vector spaces. Used for reference: K. Chandrasekhara Rao: Functional Analysis. also see Warner: Modern algebra.

Documentation

data (f :*: g) a infixl 5

The primary data type for matrices. Note that indices are represented in the functors, If you want to use numeric indices, use Simple.

Constructors

Matrix

Fields

cells :: f (g a)

class Num (Scalar v) => VectorSpace v where

https://en.wikipedia.org/wiki/Matrix_%28mathematics%29

Associated Types

type Scalar v

Methods

vzero :: v

vnegate :: v -> v

(%+) :: v -> v -> v infixl 6

(%*) :: Scalar v -> v -> v infixl 7

class (VectorSpace v, Scalar v ~ Complex a) => ComplexVectorSpace v a

class InnerProductSpace m where

Methods

(%.) :: m -> m -> Scalar m

class VectorSpace m => LieAlgebra m where

https://en.wikipedia.org/wiki/Lie_algebra

Methods

(%<>%) :: m -> m -> m

class VectorSpace m => NormedSpace m where

Methods

norm :: m -> Scalar m

class CompleteSpace m

class Num m => ConjugateSymmetric m where

Methods

conj :: m -> m

class Linear f where

Methods

linear :: f (Scalar (f a)) -> f a

unlinear :: f a -> f (Scalar (f a))

class LinearTransform m n a where

Methods

(<*>>)

Arguments

:: n a

-> (m :*: n) a

-> m a
vector times matrix

(<<*>)

Arguments

:: (m :*: n) a

-> m a

-> n a
matrix times vector

class (Scalar (f a) ~ a) => VectorLike f a

class (Functor m, Functor n) => Transposable m n where

Methods

transpose :: (m :*: n) a -> (n :*: m) a

type Index m a = m a -> a

class Applicative m => Indexable m where

Methods

diagonal_projections :: m (Index m a)

indexable_indices :: Integral a => m a

class Indexable m => Summable m where

Methods

sum_coordinates :: Num a => m a -> a

class Transposable m m => SquareMatrix m a where

https://en.wikipedia.org/wiki/Square_matrix

Methods

identity :: (m :*: m) a

diagonal :: (m :*: m) a -> m a

diagonal_matrix :: m a -> (m :*: m) a

class SquareMatrix m a => FiniteSquareMatrix m a where

Methods

determinant :: (m :*: m) a -> a

trace :: (m :*: m) a -> a

class FiniteSquareMatrix m a => InvertibleMatrix m a where

http://en.wikipedia.org/wiki/Adjugate

Methods

cofactor :: (m :*: m) a -> (m :*: m) a

adjucate :: (m :*: m) a -> (m :*: m) a

inverse :: (m :*: m) a -> (m :*: m) a

class Functor m => EigenDecomposable m a where

Methods

eigenvalues :: (m :*: m) a -> m a

class EigenDecomposable m a => EigenVectorable m a where

Methods

eigenvectors :: (m :*: m) a -> (m :*: m) a

class (Applicative m, Applicative n) => AppendableVector m n where

Associated Types

type m :+: n :: * -> *

Methods

(|>) :: m a -> n a -> (m :+: n) a

class AppendableVector m n => SplittableVector m n where

Methods

vsplit :: (m :+: n) a -> (m a, n a)

class Conditional a where

Iverson bracket: http://en.wikipedia.org/wiki/Iverson_bracket

Methods

fromBoolean :: Bool -> a

class StandardBasis m where

Methods

unit_vectors :: [m]

class VectorSpace v => CoordinateSpace v where

Associated Types

type Coordinate v

Methods

index :: Coordinate v -> v -> Scalar v

listVector :: [Scalar v] -> v

dimension_size :: v -> Int

coordinates :: v -> [Coordinate v]

matrix :: (Functor m, Functor n) => (a -> b -> c) -> m a -> n b -> (m :*: n) c

matrixMatrix :: (Functor m, Functor n, Functor m', Functor n') => (a -> b -> c) -> (m :*: m') a -> (n :*: n') b -> ((m :*: n) :*: (m' :*: n')) c

invert_matrix :: (Eq a, Num a, InvertibleMatrix m a) => (m :*: m) a -> Maybe ((m :*: m) a)

isInvertibleMatrix :: (Eq a, Num a, FiniteSquareMatrix m a) => (m :*: m) a -> Bool

data Basis m

Constructors

Basis [m]

(%-) :: VectorSpace v => v -> v -> v infixl 6

(%/) :: (Fractional (Scalar v), VectorSpace v) => v -> Scalar v -> v

angle :: (InnerProductSpace m, Floating (Scalar m)) => m -> m -> Scalar m

https://en.wikipedia.org/wiki/Angle

i_vec :: StandardBasis v => v

j_vec :: StandardBasis v => v

k_vec :: StandardBasis v => v

l_vec :: StandardBasis v => v

innerproductspace_norm :: (Floating (Scalar m), InnerProductSpace m) => m -> Scalar m

vsum :: (Foldable t, VectorSpace a) => t a -> a

fold_rows :: Functor m => (n a -> b) -> (m :*: n) a -> m b

fold_columns :: (Functor n, Transposable m n) => (m a -> b) -> (m :*: n) a -> n b

index_unit :: StandardBasis v => Int -> v

projection :: (Fractional (Scalar v), VectorSpace v, InnerProductSpace v) => v -> v -> v

normalize :: (Fractional (Scalar a), NormedSpace a) => a -> a

vector_length :: (Floating (Scalar m), InnerProductSpace m) => m -> Scalar m

divide :: (Fractional (Scalar v), VectorSpace v) => v -> Scalar v -> v

vaverage :: (Num v, Fractional (Scalar v), VectorSpace v) => [v] -> v

toScalarList :: (StandardBasis m, InnerProductSpace m) => m -> [Scalar m]

class PpShowF f => PpShowVerticalF f where

Methods

ppf_vertical :: PpShow a => f a -> Doc

fromScalarList :: (VectorSpace a, StandardBasis a) => [Scalar a] -> a

toListCS :: CoordinateSpace v => v -> [Scalar v]

toListCS2 :: (CoordinateSpace m, CoordinateSpace (Scalar m)) => m -> [[Scalar (Scalar m)]]

coordinates2 :: (CoordinateSpace m, CoordinateSpace (Scalar m)) => m -> [[(Coordinate m, Coordinate (Scalar m))]]

basis_of :: StandardBasis m => Basis m

listMatrix :: (CoordinateSpace n, CoordinateSpace (Scalar n)) => [[Scalar (Scalar n)]] -> n

(%*%) :: (Functor g, Transposable h f, InnerProductSpace (h a)) => (g :*: h) a -> (h :*: f) a -> (g :*: f) (Scalar (h a))

generalized implementation of matrix multiplication see http://en.wikipedia.org/wiki/Matrix_multiplication

(%**%) :: (Functor g, Transposable h f, InnerProductSpace (h a), Scalar (h a) ~ a) => (g :*: h) a -> (h :*: f) a -> (g :*: f) a

In this version, we must assume Scalar (h a) ~ a constraint, but the result type is nicer.

(%^%) :: (Functor h, SquareMatrix h b, InnerProductSpace (h b), b ~ Scalar (h b)) => (h :*: h) b -> Integer -> (h :*: h) b

(|><|) :: (Functor m, Functor n, ConjugateSymmetric a) => m a -> n a -> (m :*: n) a

identityCS :: (CoordinateSpace m, CoordinateSpace (Scalar m), Num (Scalar (Scalar m))) => (Int, Int) -> m

(%.%) :: (Num (Scalar m), CoordinateSpace m) => m -> m -> Scalar m infixl 7

basis_coordinates :: InnerProductSpace v => Basis v -> v -> [Scalar v]

coordinateSpaceFunctionMatrix :: (CoordinateSpace m, StandardBasis v) => (v -> Scalar m) -> m

functionMatrix :: (Functor f, SquareMatrix f a) => (f a -> g b) -> (f :*: g) b

This is the linearity condition:

hilbertSpace :: (Num m, NormedSpace m) => m -> m -> Scalar m

lie_adjoint :: LieAlgebra v => v -> Endo v

is_on_unit_circle :: (NormedSpace v, Eq (Scalar v)) => v -> Bool

https://en.wikipedia.org/wiki/Norm_(mathematics)

is_inside_unit_circle :: (NormedSpace v, Ord (Scalar v)) => v -> Bool

https://en.wikipedia.org/wiki/Norm_(mathematics)