Math.Matrix.Interface

Description

These should match standard definitions of vector spaces. Used for reference: K. Chandrasekhara Rao: Functional Analysis. also see Warner: Modern algebra.

Documentation

data (f :*: g) a infixl 5

The primary data type for matrices. Note that indices are represented in the functors, If you want to use numeric indices, use Simple.

Constructors

Matrix

Fields

cells :: f (g a)

matrix :: (Functor m, Functor n) => (a -> b -> c) -> m a -> n b -> (m :*: n) c

This method of matrix construction is especially nice.

applicativeMatrix :: (Applicative f, Functor m, Functor n) => f (a -> b -> c) -> (m :*: f) a -> (n :*: f) b -> (m :*: n) (f c)

(>*<) :: (Applicative f, Functor m, Functor n) => f (a -> b -> c) -> ((m :*: f) a, (n :*: f) b) -> (m :*: n) (f c)

class Num (Scalar v) => VectorSpace v where

https://en.wikipedia.org/wiki/Matrix_%28mathematics%29

Associated Types

type Scalar v

Methods

vzero :: v

vnegate :: v -> v

(%+) :: v -> v -> v infixl 6

(%*) :: Scalar v -> v -> v infix 7

type VectorSpaceOver v a = (VectorSpace v, Scalar v ~ a)

type PrimitiveSpace v = (v ~ Scalar v, VectorSpace v)

type ComplexVectorSpace v a = VectorSpaceOver v (Complex a)

type Linear a b = (VectorSpace a, VectorSpace b, Scalar a ~ Scalar b)

type LinearInnerProductSpace a b = (Linear a b, InnerProductSpace a, InnerProductSpace b)

class VectorSpace v => BilinearVectorSpace v where

Methods

biLin :: v -> v -> Scalar v

class InnerProductSpace m where

https://en.wikipedia.org/wiki/Dot_product

Methods

(%.) :: m -> m -> Scalar m infix 7

class VectorSpace m => LieAlgebra m where

https://en.wikipedia.org/wiki/Lie_algebra

Methods

(%<>%) :: m -> m -> m

class VectorSpace s => MetricSpace s where

Methods

distance :: s -> s -> Scalar s

compose_matrix_with :: (Transposable g n, Functor m) => (f a -> g b -> c) -> (m :*: f) a -> (g :*: n) b -> (m :*: n) c

lie_compose :: (LieAlgebra (g a), Transposable g n, Functor m) => (m :*: g) a -> (g :*: n) a -> (m :*: n) (g a)

class VectorSpace m => NormedSpace m where

Methods

norm :: m -> Scalar m

matrix_norm :: (Functor f, NormedSpace (g a), NormedSpace (f (Scalar (g a)))) => (f :*: g) a -> Scalar (f (Scalar (g a)))

This computes norm of each row, then computes the norm of the resulting column vector.

class CompleteSpace m

class ConjugateSymmetric m where

Methods

conj :: m -> m

class Scalar (m a) ~ Scalar (n a) => LinearTransform m n a where

Methods

(<*>>)

Arguments

:: n a

-> (m :*: n) a

-> m a
vector times matrix

(<<*>)

Arguments

:: (m :*: n) a

-> m a

-> n a
matrix times vector

class (Functor m, Functor n) => Transposable m n where

Methods

transpose :: (m :*: n) a -> (n :*: m) a

type Index m a = m a -> a

class Applicative m => Indexable m where

Methods

diagonal_projections :: m (Index m a)

indexable_indices :: Integral a => m a

class Indexable m => Summable m where

Methods

sum_coordinates :: Num a => m a -> a

class Transposable m m => SquareMatrix m a where

https://en.wikipedia.org/wiki/Square_matrix

Methods

identity :: (m :*: m) a

diagonal :: (m :*: m) a -> m a

diagonal_matrix :: m a -> (m :*: m) a

class (Functor m, Functor n) => ProjectionSpace (m :: * -> *) (n :: * -> *) where

Associated Types

data (m \\\ n) a

Methods

project_first :: m a -> n a

project_second :: m a -> (m \\\ n) a

join_vector :: n a -> (m \\\ n) a -> m a

class CodiagonalMatrix m a where

CodiagonalMatrix represents a matrix that can be split along the diagonal. The Codiagonal type represents a matrix without its diagonal. The ProjectionVector type represents a vector down from first element of diagonal when the diagonal is removed. This vector often has less elements than the original vector. Similarly for vector right from the first element of diagonal.

Associated Types

data Codiagonal m a

type m \\ a

Methods

codiagonal :: (m :*: m) a -> Codiagonal m a

(|\|) :: m a -> Codiagonal m a -> (m :*: m) a

down_project :: Codiagonal m a -> m \\ a

right_project :: Codiagonal m a -> m \\ a

class SquareMatrix m a => FiniteSquareMatrix m a where

Methods

determinant :: (m :*: m) a -> a

trace :: (m :*: m) a -> a

class FiniteSquareMatrix m a => InvertibleMatrix m a where

http://en.wikipedia.org/wiki/Adjugate \(cofactor(A) = |A|(A^{-1})^T\) https://en.wikipedia.org/wiki/Cross_product

Methods

cofactor :: (m :*: m) a -> (m :*: m) a

adjucate :: (m :*: m) a -> (m :*: m) a

inverse :: (m :*: m) a -> (m :*: m) a

is_unitary :: (InvertibleMatrix m a, Eq ((m :*: m) a), ConjugateSymmetric ((m :*: m) a)) => (m :*: m) a -> Bool

class Functor m => EigenDecomposable m a where

Methods

eigenvalues :: (m :*: m) a -> m a

class EigenDecomposable m a => EigenVectorable m a where

Methods

eigenvectors :: (m :*: m) a -> (m :*: m) a

class (Applicative m, Applicative n) => AppendableVector m n where

Associated Types

type m :+: n :: * -> *

Methods

(|>) :: m a -> n a -> (m :+: n) a

class AppendableVector m n => SplittableVector m n where

Methods

vsplit :: (m :+: n) a -> (m a, n a)

class Conditional a where

Iverson bracket: http://en.wikipedia.org/wiki/Iverson_bracket

Methods

fromBoolean :: Bool -> a

class StandardBasis m where

Methods

unit_vectors :: [m]

class VectorSpace v => CoordinateSpace v where

Associated Types

type Coordinate v

Methods

index :: Coordinate v -> v -> Scalar v

listVector :: [Scalar v] -> v

dimension_size :: v -> Int

coordinates :: v -> [Coordinate v]

matrixM :: (Traversable f, Traversable g, Monad m) => (a -> b -> m c) -> f a -> g b -> m ((f :*: g) c)

matrixMatrix :: (Functor m, Functor n, Functor m', Functor n') => (a -> b -> c) -> (m :*: m') a -> (n :*: n') b -> ((m :*: n) :*: (m' :*: n')) c

invert_matrix :: (Eq a, Num a, InvertibleMatrix m a) => (m :*: m) a -> Maybe ((m :*: m) a)

is_invertible_matrix :: (Eq a, Num a, FiniteSquareMatrix m a) => (m :*: m) a -> Bool

data Basis m

Constructors

Basis [m]

(%-) :: VectorSpace v => v -> v -> v infixl 6

(%/) :: (Fractional (Scalar v), VectorSpace v) => v -> Scalar v -> v

angle :: (InnerProductSpace m, Floating (Scalar m)) => m -> m -> Scalar m

https://en.wikipedia.org/wiki/Angle

cross_product_matrix :: (SquareMatrix m a, LieAlgebra (m a)) => m a -> (m :*: m) a

https://en.wikipedia.org/wiki/Cross_product This is a skew-symmetric matrix whose application to a vector is same as cross product of a with the vector. cross_product_matrix v <* w == v %<>% w.

i_vec :: StandardBasis v => v

j_vec :: StandardBasis v => v

k_vec :: StandardBasis v => v

l_vec :: StandardBasis v => v

innerproductspace_norm :: (Floating (Scalar m), InnerProductSpace m) => m -> Scalar m

vsum :: (Foldable t, VectorSpace a) => t a -> a

fold_rows :: Functor m => (n a -> b) -> (m :*: n) a -> m b

fold_columns :: (Functor n, Transposable m n) => (m a -> b) -> (m :*: n) a -> n b

index_unit :: StandardBasis v => Int -> v

projection :: (Fractional (Scalar v), VectorSpace v, InnerProductSpace v) => v -> v -> v

normalize :: (Fractional (Scalar a), NormedSpace a) => a -> a

vector_length :: (Floating (Scalar m), InnerProductSpace m) => m -> Scalar m

divide :: (Fractional (Scalar v), VectorSpace v) => v -> Scalar v -> v

vaverage :: (Num v, Fractional (Scalar v), VectorSpace v) => [v] -> v

toScalarList :: (StandardBasis m, InnerProductSpace m) => m -> [Scalar m]

fromScalarList :: (VectorSpace a, StandardBasis a) => [Scalar a] -> a

toListCS :: CoordinateSpace v => v -> [Scalar v]

toListCS2 :: (CoordinateSpace m, CoordinateSpace (Scalar m)) => m -> [[Scalar (Scalar m)]]

coordinates2 :: (CoordinateSpace m, CoordinateSpace (Scalar m)) => m -> [[(Coordinate m, Coordinate (Scalar m))]]

basis_of :: StandardBasis m => Basis m

listMatrix :: (CoordinateSpace n, CoordinateSpace (Scalar n)) => [[Scalar (Scalar n)]] -> n

(%*%) :: (Functor g, Transposable h f, InnerProductSpace (h a)) => (g :*: h) a -> (h :*: f) a -> (g :*: f) (Scalar (h a))

generalized implementation of matrix multiplication see http://en.wikipedia.org/wiki/Matrix_multiplication

(%**%) :: (Functor g, Transposable h f, InnerProductSpace (h a), VectorSpaceOver (h a) a) => (g :*: h) a -> (h :*: f) a -> (g :*: f) a

In this version, we must assume VectorSpaceOver (h a) a constraint, but the result type is nicer.

(%^%) :: (Functor h, SquareMatrix h b, InnerProductSpace (h b), VectorSpaceOver (h b) b) => (h :*: h) b -> Integer -> (h :*: h) b

(|><|) :: (Functor m, Functor n, ConjugateSymmetric a, Num a) => m a -> n a -> (m :*: n) a

identityCS :: (CoordinateSpace m, CoordinateSpace (Scalar m), Num (Scalar (Scalar m))) => (Int, Int) -> m

(%.%) :: (Num (Scalar m), CoordinateSpace m) => m -> m -> Scalar m infixl 7

basis_coordinates :: InnerProductSpace v => Basis v -> v -> [Scalar v]

coordinateSpaceFunctionMatrix :: (CoordinateSpace m, StandardBasis v) => (v -> Scalar m) -> m

functionMatrix :: (Functor f, SquareMatrix f a) => (f a -> g b) -> (f :*: g) b

This is the linearity condition:

convex_combination :: (VectorSpace v, Fractional (Scalar v), Foldable t, Applicative t) => t (Scalar v) -> t v -> v

https://en.wikipedia.org/wiki/Convex_combination This computes \[f([a_0,a_1,...,a_n], [{\mathbf b}_0,{\mathbf b}_1,...,{\mathbf b}_n]) = {{\sum_{j=0}^n{a_j{\mathbf b}_j}} \over \sum_{i=0}^n{a_i}}\]

hilbertSpace :: (Num m, NormedSpace m) => m -> m -> Scalar m

lie_adjoint :: LieAlgebra v => v -> Endo v

is_on_unit_circle :: (NormedSpace v, Eq (Scalar v)) => v -> Bool

https://en.wikipedia.org/wiki/Norm_(mathematics)

is_inside_unit_circle :: (NormedSpace v, Ord (Scalar v)) => v -> Bool

https://en.wikipedia.org/wiki/Norm_(mathematics)

Orphan instances

PpShowVerticalF Complex

Instance details

Methods

ppf_vertical :: PpShow a => Complex a -> Doc

Num a => Num (Endo a)

Instance details

Methods

(+) :: Endo a -> Endo a -> Endo a

(-) :: Endo a -> Endo a -> Endo a

(*) :: Endo a -> Endo a -> Endo a

negate :: Endo a -> Endo a

abs :: Endo a -> Endo a

signum :: Endo a -> Endo a

fromInteger :: Integer -> Endo a