Math.Matrix.Matrix

Documentation

type family Transpose a

conjIso :: ConjugateSymmetric a => a :==: a

transposeIso :: (Transposable m n, Transposable n m) => (m :*: n) a :==: (n :*: m) a

vnegateIso :: VectorSpace v => v :==: v

(!$!) :: (Functor f, Functor g) => f (a -> b) -> g a -> (f :*: g) b

matrixBind :: Functor f => f a -> (a -> g b) -> (f :*: g) b

matrix2d :: (Functor f, Functor g) => (a -> b -> c) -> f a -> g b -> (f :*: g) c

this is same as Matrix.Interface.matrix

matrix3d :: (Functor f, Functor g, Functor h) => (a -> b -> c -> d) -> f a -> g b -> h c -> (f :*: (g :*: h)) d

matrix4d :: (Functor f, Functor g, Functor h, Functor i) => (a -> b -> c -> d -> e) -> f a -> g b -> h c -> i d -> (f :*: (g :*: (h :*: i))) e

constantMatrix :: (Applicative m, Applicative n) => a -> (m :*: n) a

needs type annotation

applyCol :: Applicative f => (f :*: (->) a) b -> f a -> f b

applyRow :: ((->) a :*: g) b -> a -> g b

matrixA :: (ArrowApply arr, FunctorArrow f arr, FunctorArrow g arr) => arr (a, b) c -> arr (f a, g b) ((f :*: g) c)

modular_exponentiation :: (Integral a, SquareMatrix f a, Functor f, InnerProductSpace (f a), a ~ Scalar (f a)) => (f :*: f) a -> Integer -> a -> (f :*: f) a

https://en.wikipedia.org/wiki/Modular_exponentiation modular_exponentiation m b c == m^b (mod c)

isSymmetric :: (Transposable m m, Applicative m, Foldable m, Eq a) => (m :*: m) a -> Bool

apply_vector :: SquareMatrix m b => m (a -> b) -> m a -> m b

apply_columns :: (Functor n, Functor m, Applicative f) => (m :*: f) (a -> b) -> (n :*: f) a -> (m :*: n) (f b)

matrixAssoc :: Functor f => ((f :*: g) :*: h) a -> (f :*: (g :*: h)) a

matrixUnassoc :: Functor f => (f :*: (g :*: h)) a -> ((f :*: g) :*: h) a

matrixLeftId :: (I :*: f) a -> f a

matrixRightId :: Functor f => (f :*: I) a -> f a

(<!>) :: (Functor f, Functor g) => (g :*: f) a -> (g c -> b, f a -> c) -> b

(!*!) :: (Functor f, Functor g, Num a) => g a -> f a -> (g :*: f) a

(<!!>) :: (Ix i, Ix j) => (Array i :*: Array j) a -> (i, j) -> a

matrix_commutator :: (Num a, Applicative h, Transposable h h, InnerProductSpace (h a), VectorSpaceOver (h a) a) => (h :*: h) a -> (h :*: h) a -> (h :*: h) a

matrix_anticommutator :: (Num a, Applicative h, Transposable h h, InnerProductSpace (h a), VectorSpaceOver (h a) a) => (h :*: h) a -> (h :*: h) a -> (h :*: h) a

normalized_anticommutator :: (Fractional a, Applicative h, Transposable h h, InnerProductSpace (h a), VectorSpaceOver (h a) a) => (h :*: h) a -> (h :*: h) a -> (h :*: h) a

commute :: (Fractional a, Applicative f, Transposable f f, InnerProductSpace (f a), VectorSpaceOver (f a) a) => Complex ((f :*: f) a) -> Complex ((f :*: f) a) -> (f :*: f) (Complex a)

bind_diagonal :: Monad m => (m :*: m) a -> (a -> m b) -> m b

pairMatrix :: (Functor f, Functor g) => f a -> g b -> (f :*: g) (a, b)

(!+!) :: (Functor f, Functor g, Num a) => g a -> f a -> (g :*: f) a

cofunctor_inverse_image :: (Functor f, CoFunctor g) => (a -> b) -> (g :*: f) b -> (g :*: f) a

cofunctor_map :: (CoFunctor f, CoFunctor g) => (a -> b) -> (g :*: f) a -> (g :*: f) b

in_transpose :: (Monad g, Indexable f) => (f :*: g) a -> (a -> (g :*: f) b) -> (f :*: g) b

fmap_split :: (SplittableVector m n, AppendableVector m' n') => (m a -> m' a') -> (n a -> n' a') -> (m :+: n) a -> (m' :+: n') a'

split_matrix :: (SplittableVector m n, SplittableVector f f', Functor (f :+: f'), Functor (m :+: n)) => ((f :+: f') :*: (m :+: n)) a -> (((f :*: m) a, (f :*: n) a), ((f' :*: m) a, (f' :*: n) a))

join_matrix :: (AppendableVector m' n', AppendableVector m n) => (m :*: m') a -> (m :*: n') a -> (n :*: m') a -> (n :*: n') a -> ((m :+: n) :*: (m' :+: n')) a

matrix_iso :: f (g a) :==: (f :*: g) a

Orphan instances

Ix i => PpShowVerticalF (Array i)

Instance details

Methods

ppf_vertical :: PpShow a => Array i a -> Doc

Ix i => PpShowF (Array i)

Instance details

Methods

ppf :: PpShow a => Array i a -> Doc

(Monad f, Monad g, Indexable f) => Monad (f :*: g)

Instance details

Methods

(>>=) :: (f :*: g) a -> (a -> (f :*: g) b) -> (f :*: g) b

(>>) :: (f :*: g) a -> (f :*: g) b -> (f :*: g) b

return :: a -> (f :*: g) a

fail :: String -> (f :*: g) a

(Functor f, Functor g) => Functor (g :*: f)

Instance details

Methods

fmap :: (a -> b) -> (g :*: f) a -> (g :*: f) b

(<$) :: a -> (g :*: f) b -> (g :*: f) a

(Applicative f, Applicative g) => Applicative (f :*: g)

Instance details

Methods

pure :: a -> (f :*: g) a

(<*>) :: (f :*: g) (a -> b) -> (f :*: g) a -> (f :*: g) b

liftA2 :: (a -> b -> c) -> (f :*: g) a -> (f :*: g) b -> (f :*: g) c

(*>) :: (f :*: g) a -> (f :*: g) b -> (f :*: g) b

(<*) :: (f :*: g) a -> (f :*: g) b -> (f :*: g) a

(Foldable f, Foldable g) => Foldable (f :*: g)

Instance details

Methods

fold :: Monoid m => (f :*: g) m -> m

foldMap :: Monoid m => (a -> m) -> (f :*: g) a -> m

foldr :: (a -> b -> b) -> b -> (f :*: g) a -> b

foldr' :: (a -> b -> b) -> b -> (f :*: g) a -> b

foldl :: (b -> a -> b) -> b -> (f :*: g) a -> b

foldl' :: (b -> a -> b) -> b -> (f :*: g) a -> b

foldr1 :: (a -> a -> a) -> (f :*: g) a -> a

foldl1 :: (a -> a -> a) -> (f :*: g) a -> a

toList :: (f :*: g) a -> [a]

null :: (f :*: g) a -> Bool

length :: (f :*: g) a -> Int

elem :: Eq a => a -> (f :*: g) a -> Bool

maximum :: Ord a => (f :*: g) a -> a

minimum :: Ord a => (f :*: g) a -> a

sum :: Num a => (f :*: g) a -> a

product :: Num a => (f :*: g) a -> a

(Traversable f, Traversable g) => Traversable (f :*: g)

Instance details

Methods

traverse :: Applicative f0 => (a -> f0 b) -> (f :*: g) a -> f0 ((f :*: g) b)

sequenceA :: Applicative f0 => (f :*: g) (f0 a) -> f0 ((f :*: g) a)

mapM :: Monad m => (a -> m b) -> (f :*: g) a -> m ((f :*: g) b)

sequence :: Monad m => (f :*: g) (m a) -> m ((f :*: g) a)

(MonadZip g, MonadZip f, Indexable g) => MonadZip (g :*: f)

Instance details

Methods

mzip :: (g :*: f) a -> (g :*: f) b -> (g :*: f) (a, b)

mzipWith :: (a -> b -> c) -> (g :*: f) a -> (g :*: f) b -> (g :*: f) c

munzip :: (g :*: f) (a, b) -> ((g :*: f) a, (g :*: f) b)

(Alternative f, Alternative g) => Alternative (g :*: f)

Instance details

Methods

empty :: (g :*: f) a

(<|>) :: (g :*: f) a -> (g :*: f) a -> (g :*: f) a

some :: (g :*: f) a -> (g :*: f) [a]

many :: (g :*: f) a -> (g :*: f) [a]

(MonadPlus g, Indexable g, MonadPlus f) => MonadPlus (g :*: f)

Instance details

Methods

mzero :: (g :*: f) a

mplus :: (g :*: f) a -> (g :*: f) a -> (g :*: f) a

(Comonad f, Comonad g, Transposable g f, Transposable f g) => Comonad (f :*: g)
See video by Bartosz Milewski ("Category theory II 7.2: Comonads categorically and examples")

Instance details

Methods

extract :: (f :*: g) a -> a

duplicate :: (f :*: g) a -> (f :*: g) ((f :*: g) a)

extend :: ((f :*: g) a -> b) -> (f :*: g) a -> (f :*: g) b

(=>>) :: (f :*: g) a -> ((f :*: g) a -> b) -> (f :*: g) b

(.>>) :: (f :*: g) a -> b -> (f :*: g) b

(CoFunctor f, Functor g) => CoFunctor (g :*: f)

Instance details

Methods

inverse_image :: (a -> b) -> (g :*: f) b -> (g :*: f) a

(|>>) :: (g :*: f) b -> (a -> b) -> (g :*: f) a

(Eq a, Applicative f, Applicative g, Foldable f, Foldable g) => Eq ((f :*: g) a)

Instance details

Methods

(==) :: (f :*: g) a -> (f :*: g) a -> Bool

(/=) :: (f :*: g) a -> (f :*: g) a -> Bool

(Eq a, Applicative f, Applicative g, Foldable f, Foldable g, Ord (f (g a))) => Ord ((f :*: g) a)

Instance details

Methods

compare :: (f :*: g) a -> (f :*: g) a -> Ordering

(<) :: (f :*: g) a -> (f :*: g) a -> Bool

(<=) :: (f :*: g) a -> (f :*: g) a -> Bool

(>) :: (f :*: g) a -> (f :*: g) a -> Bool

(>=) :: (f :*: g) a -> (f :*: g) a -> Bool

max :: (f :*: g) a -> (f :*: g) a -> (f :*: g) a

min :: (f :*: g) a -> (f :*: g) a -> (f :*: g) a

(Functor f, Functor g) => Builder ((g :*: f) a)

Instance details

Associated Types

data Unfold ((g :*: f) a) a :: Type

Methods

build :: Unfold ((g :*: f) a) a0 -> a0 -> (g :*: f) a

Functor g => Visitor ((g :*: f) a)

Instance details

Associated Types

data Fold ((g :*: f) a) a :: Type

Methods

visit :: Fold ((g :*: f) a) a0 -> (g :*: f) a -> a0