Math.Matrix.SIMD

Description

This modules provides SIMD optimized versions of vectors. Relies heavily on GHC SIMD support. SIMD works only when using LLVM and GHC 8.x The only combinations of vectors supported are 4 * Int32, 8 * Int16 and 16 * Int8, 4*Float, 2*Double Conversions to and from optimized types and Vector4 Int32, (Vector2 :*: Vector2) Int32, (Vector4 Int16, Vector4 Int16) Vector4 Float, (Vector2 :*: Vector2) Float, and (Vector4 :*: Vector4) Int8 can be performed as appropriate for each type.

Documentation

data SIMDVec (len :: Nat) a

eqSVec4 :: SIMDVec 4 Int32 -> SIMDVec 4 Int32 -> Bool

eqSVec8 :: SIMDVec 8 Int16 -> SIMDVec 8 Int16 -> Bool

eqSVec16 :: SIMDVec 16 Int8 -> SIMDVec 16 Int8 -> Bool

to2x2Matrix :: SIMDVec 4 Int32 -> (Vector2 :*: Vector2) Int32

from2x2Matrix :: (Vector2 :*: Vector2) Int32 -> SIMDVec 4 Int32

svec16_1 :: SIMDVec 16 Int8 -> Int8

svec16_2 :: SIMDVec 16 Int8 -> Int8

svec16_3 :: SIMDVec 16 Int8 -> Int8

svec16_4 :: SIMDVec 16 Int8 -> Int8

svec16_5 :: SIMDVec 16 Int8 -> Int8

svec16_6 :: SIMDVec 16 Int8 -> Int8

svec16_7 :: SIMDVec 16 Int8 -> Int8

svec16_8 :: SIMDVec 16 Int8 -> Int8

svec16_9 :: SIMDVec 16 Int8 -> Int8

svec16_10 :: SIMDVec 16 Int8 -> Int8

svec16_11 :: SIMDVec 16 Int8 -> Int8

svec16_12 :: SIMDVec 16 Int8 -> Int8

svec16_13 :: SIMDVec 16 Int8 -> Int8

svec16_14 :: SIMDVec 16 Int8 -> Int8

svec16_15 :: SIMDVec 16 Int8 -> Int8

svec16_16 :: SIMDVec 16 Int8 -> Int8

mapSVec16 :: (Int8 -> Int8) -> SIMDVec 16 Int8 -> SIMDVec 16 Int8

mapSVec8 :: (Int16 -> Int16) -> SIMDVec 8 Int16 -> SIMDVec 8 Int16

mapSVec4 :: (Int32 -> Int32) -> SIMDVec 4 Int32 -> SIMDVec 4 Int32

mapFVec4 :: (Float -> Float) -> SIMDVec 4 Float -> SIMDVec 4 Float

zipSVec4 :: (Int32 -> Int32 -> Int32) -> SIMDVec 4 Int32 -> SIMDVec 4 Int32 -> SIMDVec 4 Int32

zipFVec4 :: (Float -> Float -> Float) -> SIMDVec 4 Float -> SIMDVec 4 Float -> SIMDVec 4 Float

zipSVec8 :: (Int16 -> Int16 -> Int16) -> SIMDVec 8 Int16 -> SIMDVec 8 Int16 -> SIMDVec 8 Int16

zipSVec16 :: (Int8 -> Int8 -> Int8) -> SIMDVec 16 Int8 -> SIMDVec 16 Int8 -> SIMDVec 16 Int8

constantFVec4 :: Float -> SIMDVec 4 Float

constantSVec4 :: Int32 -> SIMDVec 4 Int32

constantSVec8 :: Int16 -> SIMDVec 8 Int16

constantSVec16 :: Int8 -> SIMDVec 16 Int8

svec8_1 :: SIMDVec 8 Int16 -> Int16

svec8_2 :: SIMDVec 8 Int16 -> Int16

svec8_3 :: SIMDVec 8 Int16 -> Int16

svec8_4 :: SIMDVec 8 Int16 -> Int16

svec8_5 :: SIMDVec 8 Int16 -> Int16

svec8_6 :: SIMDVec 8 Int16 -> Int16

svec8_7 :: SIMDVec 8 Int16 -> Int16

svec8_8 :: SIMDVec 8 Int16 -> Int16

makeFVec4 :: Float -> Float -> Float -> Float -> SIMDVec 4 Float

makeSVec4 :: Int32 -> Int32 -> Int32 -> Int32 -> SIMDVec 4 Int32

makeSVec8 :: Int16 -> Int16 -> Int16 -> Int16 -> Int16 -> Int16 -> Int16 -> Int16 -> SIMDVec 8 Int16

makeSVec16 :: Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> Int8 -> SIMDVec 16 Int8

svec4_x :: SIMDVec 4 Int32 -> Int32

svec4_y :: SIMDVec 4 Int32 -> Int32

svec4_z :: SIMDVec 4 Int32 -> Int32

svec4_t :: SIMDVec 4 Int32 -> Int32

fvec4_x :: SIMDVec 4 Float -> Float

fvec4_y :: SIMDVec 4 Float -> Float

fvec4_z :: SIMDVec 4 Float -> Float

fvec4_t :: SIMDVec 4 Float -> Float

fromInt8Vector16 :: SIMDVec 16 Int8 -> (Vector4 :*: Vector4) Int8

toInt8Vector16 :: (Vector4 :*: Vector4) Int8 -> SIMDVec 16 Int8

fromDoubleVector2 :: Vector2 Double -> SIMDVec 2 Double

toDoubleVector2 :: SIMDVec 2 Double -> Vector2 Double

fromFloatVector4 :: Vector4 Float -> SIMDVec 4 Float

toFloatVector4 :: SIMDVec 4 Float -> Vector4 Float

from2x2FloatVector4 :: (Vector2 :*: Vector2) Float -> SIMDVec 4 Float

to2x2FloatVector4 :: SIMDVec 4 Float -> (Vector2 :*: Vector2) Float

fromInt32Vector4 :: Vector4 Int32 -> SIMDVec 4 Int32

fromInt16Vector8 :: Vector4 Int16 -> Vector4 Int16 -> SIMDVec 8 Int16

toInt16Vector8 :: SIMDVec 8 Int16 -> (Vector4 Int16, Vector4 Int16)

toInt32Vector4 :: SIMDVec 4 Int32 -> Vector4 Int32

transposeFloatVector4 :: SIMDVec 4 Float -> SIMDVec 4 Float

class VectorSpace a => Optimal a where

Associated Types

type Optimized a

Methods

toO :: a -> Optimized a

fromO :: Optimized a -> a

zipO :: (Scalar a -> Scalar a -> Scalar a) -> Optimized a -> Optimized a -> Optimized a

mapO :: (Scalar a -> Scalar a) -> Optimized a -> Optimized a

sumCoordinatesO :: Optimized a -> Scalar a

constantO :: Scalar a -> Optimized a

fast_multiply4_float :: Matrix4 Float -> Vector4 Float -> Vector4 Float

fast_multiply2_double :: Scalar (SIMDVec 2 Double) ~ Double => Matrix2 Double -> Vector2 Double -> Vector2 Double

Orphan instances

Scalar (SIMDVec 2 Double) ~ Double => LinearTransform Vector2 Vector2 Double

Instance details

Methods

(<*>>) :: Vector2 Double -> (Vector2 :*: Vector2) Double -> Vector2 Double

(<<*>) :: (Vector2 :*: Vector2) Double -> Vector2 Double -> Vector2 Double

LinearTransform Vector3 Vector3 Float

Instance details

Methods

(<*>>) :: Vector3 Float -> (Vector3 :*: Vector3) Float -> Vector3 Float

(<<*>) :: (Vector3 :*: Vector3) Float -> Vector3 Float -> Vector3 Float

LinearTransform Vector4 Vector4 Float

Instance details

Methods

(<*>>) :: Vector4 Float -> (Vector4 :*: Vector4) Float -> Vector4 Float

(<<*>) :: (Vector4 :*: Vector4) Float -> Vector4 Float -> Vector4 Float

Scalar (SIMDVec 2 Double) ~ Double => InnerProductSpace (Vector2 Double)

Instance details

Methods

(%.) :: Vector2 Double -> Vector2 Double -> Scalar (Vector2 Double)

InnerProductSpace (Vector3 Float)

Instance details

Methods

(%.) :: Vector3 Float -> Vector3 Float -> Scalar (Vector3 Float)

InnerProductSpace (Vector4 Float)

Instance details

Methods

(%.) :: Vector4 Float -> Vector4 Float -> Scalar (Vector4 Float)