Math.Matrix.Simple

Description

This module provides matrices with indices. Matrices are constructed using smatrix and svec operations, example:

smatrix (*) (svec3 1 2 3) (svec4 5 4 3 2) :: (ThreeD :&: FourD) Int
== 5   4 3 2
   10  8 6 4
   15 12 9 6

The number after svec is the number of elements in the dimension along row or column. The binary operation is applied to each pair of elements to construct the matrix element.

Matrices can also be constructed using Matrix constructor, with indices as input, e.g.:

identity :: (ThreeD :&: ThreeD) Integer

identity = Matrix $ \i j -> if i == j then 1 else 0

Note that index types are specifically restricted to be small types, since the Universe type class operations enumerate all elements of the type. This is enforced in show operations, which only show ten elements per dimension.

Documentation

type (:&:) row col elem = ((->) row :*: (->) col) elem

type Matrix3S row col dep elem = ((->) row :*: ((->) col :*: (->) dep)) elem

type Matrix4S row col dep tim elem = ((->) row :*: ((->) col :*: ((->) dep :*: (->) tim))) elem

type MatrixA arr row col elem = (arr row :*: arr col) elem

type SMatrix1 elem = (OneD :&: OneD) elem

type SMatrix2 elem = (TwoD :&: TwoD) elem

type SMatrix3 elem = (ThreeD :&: ThreeD) elem

type SMatrix4 elem = (FourD :&: FourD) elem

type InfMatrix elem = (Integer :&: Integer) elem

mapDimensions :: ((->) a :~> f) -> ((->) b :~> g) -> (a :&: b) c -> (f :*: g) c

simpleMatrix :: (Functor m, Functor n) => (a :&: b) c -> m a -> n b -> (m :*: n) c

dimensionFlip :: (Functor m, Functor n) => (a :&: b) (c -> d) -> m a -> n c -> (m :*: n) (b -> d)

sumS :: (Num b, Universe a) => (a -> b) -> b

productS :: (Num b, Universe a) => (a -> b) -> b

kronecker_delta :: (Eq a, Num b) => a -> a -> b

https://en.wikipedia.org/wiki/Kronecker_delta

(%***%) :: (Universe mid, ConjugateSymmetric a) => (row :&: mid) a -> (mid :&: col) a -> (row :&: col) a

class CoFactorDimension dim where

Methods

cofactorDim :: Num a => dim -> (dim :&: dim) a -> a

leviCivita2S :: Num a => (TwoD :&: TwoD) a

https://en.wikipedia.org/wiki/Levi-Civita_symbol Note indices are from 0..(n-1).

leviCivita3S :: Num a => Matrix3S ThreeD ThreeD ThreeD a

https://en.wikipedia.org/wiki/Levi-Civita_symbol Note indices are from 0..(n-1).

leviCivita4S :: Num a => Matrix4S FourD FourD FourD FourD a

https://en.wikipedia.org/wiki/Levi-Civita_symbol Note indices are from 0..(n-1).

leviCivitaS :: (Num b, Enum a, Universe a) => (a -> b) -> b

https://en.wikipedia.org/wiki/Levi-Civita_symbol

leviCivita :: Num a => [a] -> a

https://en.wikipedia.org/wiki/Levi-Civita_symbol Note indices are from 0..(n-1).

crossProductS :: (Integral a, Num b, Universe a) => (a -> b) -> (a -> b) -> a -> b

https://en.wikipedia.org/wiki/Cross_product

crossProduct :: Integral a => [a] -> [a] -> [a]

https://en.wikipedia.org/wiki/Cross_product

smatrix :: (a -> b -> c) -> (row -> a) -> (col -> b) -> (row :&: col) c

swapRows :: Eq row => row -> row -> (row :&: col) a -> (row :&: col) a

mulRow :: (Eq row, Num a) => a -> row -> (row :&: col) a -> (row :&: col) a

swapIndex :: Eq a => a -> a -> a -> a

smatrixA :: Arrow arr => arr (a, b) c -> arr row a -> arr col b -> MatrixA arr row col c

smatrix3 :: (a -> b -> c -> d) -> (row -> a) -> (col -> b) -> (depth -> c) -> ((->) row :*: ((->) col :*: (->) depth)) d

function_smatrix :: (a -> b -> c) -> (a :&: b) c

smatrix_inverse_image :: ((row, col) -> (row', col')) -> (row' :&: col') a -> (row :&: col) a

inverse_image_smatrix :: (row -> row', col -> col') -> (row' :&: col') a -> (row :&: col) a

index_function_matrix :: (row -> b -> c) -> (col -> a -> b) -> (row :&: col) (a -> c)

sapply :: ((a -> b) :&: a) b

(#) :: (row :&: col) a -> (row, col) -> a

index3S :: Matrix3S row col dep elem -> (row, col, dep) -> elem

index4S :: Matrix4S row col dep tim elem -> (row, col, dep, tim) -> elem

indexA :: ArrowApply arr => MatrixA arr row col a -> arr (row, col) a

sdeterminant2 :: Num a => (TwoD :&: TwoD) a -> a

cofactor4_11 :: Num a => (FourD :&: FourD) a -> a

cofactor4_21 :: Num a => (FourD :&: FourD) a -> a

cofactor4_31 :: Num a => (FourD :&: FourD) a -> a

cofactor4_41 :: Num a => (FourD :&: FourD) a -> a

cofactor3_11 :: Num a => (ThreeD :&: ThreeD) a -> a

cofactor3_21 :: Num a => (ThreeD :&: ThreeD) a -> a

cofactor3_31 :: Num a => (ThreeD :&: ThreeD) a -> a

cofactor2_11 :: Num a => (TwoD :&: TwoD) a -> a

cofactor2_21 :: Num a => (TwoD :&: TwoD) a -> a

cofactorDim2 :: Num a => TwoD -> (TwoD :&: TwoD) a -> a

cofactorDim3 :: Num a => ThreeD -> (ThreeD :&: ThreeD) a -> a

cofactorDim4 :: Num a => FourD -> (FourD :&: FourD) a -> a

sdeterminant :: (Num a, Num dim, Universe dim, CoFactorDimension dim) => (dim :&: dim) a -> a

determinant3S :: Num a => (ThreeD :&: ThreeD) a -> a

sdeterminant3 :: Num a => (ThreeD :&: ThreeD) a -> a

sdeterminant4 :: Num a => (FourD :&: FourD) a -> a

remove41 :: ThreeD -> FourD

remove42 :: ThreeD -> FourD

remove43 :: ThreeD -> FourD

remove44 :: ThreeD -> FourD

remove31 :: TwoD -> ThreeD

remove32 :: TwoD -> ThreeD

remove33 :: TwoD -> ThreeD

remove21 :: OneD -> TwoD

remove22 :: OneD -> TwoD

svec1 :: a -> OneD -> a

svec2 :: a -> a -> TwoD -> a

svec3 :: a -> a -> a -> ThreeD -> a

svec4 :: a -> a -> a -> a -> FourD -> a

svec5 :: a -> a -> a -> a -> a -> FiveD -> a

data OneD

Constructors

OneD0

data TwoD

Constructors

TwoD0

TwoD1

data ThreeD

Constructors

ThreeD0

ThreeD1

ThreeD2

data FourD

Constructors

FourD0

FourD1

FourD2

FourD3

data FiveD

Constructors

FiveD0

FiveD1

FiveD2

FiveD3

FiveD4

toInt5 :: FiveD -> Int

toInt4 :: FourD -> Int

toInt3 :: ThreeD -> Int

toInt2 :: TwoD -> Int

toInt1 :: OneD -> Int

fromInt5 :: Int -> FiveD

fromInt4 :: Int -> FourD

fromInt3 :: Int -> ThreeD

fromInt2 :: Int -> TwoD

fromInt1 :: Int -> OneD

srotate_x :: Floating a => a -> SMatrix3 a