Math.Matrix.Vector3

Documentation

type ComplexVector3 a = (Vector3 :*: Complex) a

dim3 :: Vector3 Integer

i3 :: Num a => Vector3 a

j3 :: Num a => Vector3 a

k3 :: Num a => Vector3 a

type Matrix3 a = (Vector3 :*: Vector3) a

vec3 :: (a, a, a) -> Vector3 a

finite3 :: Vector3 a -> Vec 3 a

constant3 :: a -> Vector3 a

vec3Unit :: a -> Scalar (Vector3 a)

vec3Counit :: Vector3 (Scalar (Vector3 a)) -> Vector3 a

destructvec3 :: Vector3 a -> (a, a, a)

setx3 :: s -> Vector3 s -> Vector3 s

sety3 :: s -> Vector3 s -> Vector3 s

setz3 :: s -> Vector3 s -> Vector3 s

set_endo :: Vector3 s -> Vector3 (Endo (Vector3 s))

set_vector_action :: Vector3 s -> Vector3 (Vector3 s -> Vector3 s)

update_row3 :: g a -> Vector3 ((Vector3 :*: g) a -> (Vector3 :*: g) a)

update_column3 v is a vector of all update operations that replace one row with v

update_column3 :: Applicative f => f a -> Vector3 ((f :*: Vector3) a -> (f :*: Vector3) a)

update_column3 v is a vector of all update operations that replace one column with v example use: update_column3 (Vector3 3 4 5) ycoord3 identity3 == [[1,3,0],[0,4,0],[0,5,1]].

removex3 :: Vector3 a -> Vector2 a

removey3 :: Vector3 a -> Vector2 a

removez3 :: Vector3 a -> Vector2 a

remove3 :: Vector3 (Vector3 a -> Vector2 a)

remove_index3 :: Matrix3 a -> Matrix3 (Matrix2 a)

cofactor3 :: Num a => Matrix3 a -> Matrix3 a

adjucate3 :: Num a => Matrix3 a -> Matrix3 a

https://en.wikipedia.org/wiki/Adjugate_matrix

inverse3 :: Fractional a => Matrix3 a -> Matrix3 a

https://en.wikipedia.org/wiki/Invertible_matrix

set3 :: Vector3 a -> Vector3 a -> (Vector3 :*: Vector3) a

swap3 :: (Vector3 :*: Vector3) (Vector3 a -> Vector3 a)

http://en.wikipedia.org/wiki/Elementary_matrix#Operations

mul3 :: VectorSpace s => Scalar s -> Vector3 (Vector3 s -> Vector3 s)

add3Matrix :: VectorSpace a => Scalar a -> ((Vector3 :*: Vector3) :*: Vector3) a -> ((Vector3 :*: Vector3) :*: Vector3) a

add3 :: VectorSpace s => Scalar s -> (Vector3 :*: Vector3) (Vector3 s -> Vector3 s)

swapxy3 :: Vector3 a -> Vector3 a

swapxz3 :: Vector3 a -> Vector3 a

swapyz3 :: Vector3 a -> Vector3 a

mul3Matrix :: (Num a, VectorSpace a) => Scalar a -> Vector3 ((Vector3 :*: Vector3) a)

signs3 :: Integral a => Matrix3 a

dx3_endo :: Num a => a -> Endo (Vector3 a)

dy3_endo :: Num a => a -> Endo (Vector3 a)

dz3_endo :: Num a => a -> Endo (Vector3 a)

dx3 :: Num a => a -> Vector3 a -> Vector3 a

dy3 :: Num a => a -> Vector3 a -> Vector3 a

dz3 :: Num a => a -> Vector3 a -> Vector3 a

op_dot :: Num b => Vector3 (a -> b) -> a -> b

sum_coordinates3 :: Num a => Vector3 a -> a

cube_norm3 :: Floating a => Vector3 a -> a

cube root of a sum of cubes.

nth_norm3 :: Floating a => a -> Vector3 a -> a

nth root of a sum of nth powers

index3 :: Vector3 (Vector3 a -> a)

mapBasis :: Num a => (Vector3 a -> Vector3 a) -> Vector3 a -> Vector3 a

functionMatrix3 :: Num a => (Vector3 a -> Vector3 a) -> Matrix3 a

splitx :: Vector3 s -> (Vector2 s, Vector1 s)

splity :: Vector3 s -> (Vector2 s, Vector1 s)

splitz :: Vector3 s -> (Vector2 s, Vector1 s)

trace3 :: Num a => Matrix3 a -> a

vector3 :: [a] -> Vector3 a

cross_product :: Num a => Vector3 a -> Vector3 a -> Vector3 a

https://en.wikipedia.org/wiki/Cross_product WARNING: a cross product produces an axial vector. https://en.wikipedia.org/wiki/Pseudovector

cross_product3 :: Num a => (Vector2 :*: Vector3) a -> Vector3 a

cross_product3 computes the three dimensional vector that is orthogonal to both rows of the 2x3 vector.

xid :: Num a => Vector3 a

yid :: Num a => Vector3 a

zid :: Num a => Vector3 a

identity3 :: Num a => Matrix3 a

matrix_power3 :: (ConjugateSymmetric a, Num a) => Matrix3 a -> Integer -> Matrix3 a

vector_indices3 :: Integral a => Vector3 a

matrix_indices3 :: Integral a => (Vector3 :*: Vector3) (a, a)

matrix_multiply3 :: (ConjugateSymmetric a, Num a) => Matrix3 a -> Matrix3 a -> Matrix3 a

dot3 :: (Num a, ConjugateSymmetric a) => Vector3 a -> Vector3 a -> a

left_multiply3_gen :: (Functor f, Num a, ConjugateSymmetric a) => Vector3 a -> (f :*: Vector3) a -> f a

right_multiply3_gen :: (VectorSpace (f a), ConjugateSymmetric a, Scalar (f a) ~ a) => (Vector3 :*: f) a -> Vector3 a -> f a

left_multiply3 :: (Num a, ConjugateSymmetric a) => Vector3 a -> Matrix3 a -> Vector3 a

right_multiply3 :: (Num a, ConjugateSymmetric a) => Matrix3 a -> Vector3 a -> Vector3 a

transpose3 :: Num a => Matrix3 a -> Matrix3 a

transpose3_impl :: Matrix3 a -> Matrix3 a

outer3 :: (a -> b -> c) -> Vector3 a -> Vector3 b -> Matrix3 c

zipWithV3 :: (a -> b -> c) -> Vector3 a -> Vector3 b -> Vector3 c

rotate_x :: Floating a => a -> Matrix3 a

rotate_y :: Floating a => a -> Matrix3 a

rotate_z :: Floating a => a -> Matrix3 a

rotate_coordinates3 :: Vector3 a -> Vector3 a

zero_codiagonal3 :: Num a => Codiagonal Vector3 a

diagonal_matrix3 :: Num a => Vector3 a -> (Vector3 :*: Vector3) a

matrix3 :: Vector3 a -> Codiagonal Vector3 a -> (Vector3 :*: Vector3) a

codiag3 :: Matrix3 a -> Codiagonal Vector3 a

codiagonal3 :: Matrix3 a -> Vector2 (Vector2 a, Vector2 a)

diagonal3 :: Matrix3 a -> Vector3 a

diagonal_projections3 :: Num a => Vector3 (Index Vector3 a)

diagonal3x :: Num a => Matrix3 a -> Vector3 a

diagonal3y :: Num a => Matrix3 a -> Vector3 a

diagonal3z :: Num a => Matrix3 a -> Vector3 a

determinant3 :: Num a => Matrix3 a -> a

algorithm from http://en.wikipedia.org/wiki/Determinant

determinant3_ :: Num a => Matrix3 a -> a

gram_schmidt3 :: (Fractional (Scalar a), Num a, InnerProductSpace a, VectorSpace a) => Vector3 a -> Vector3 a

eigenvalue3 :: (Floating a, Ord a, ConjugateSymmetric a) => (Vector3 :*: Vector3) a -> Vector3 a

https://en.wikipedia.org/wiki/Eigenvalue_algorithm based on algorithm by Smith, Oliver K.: Eigenvalues of symmetric 3x3 matrix, 1961, Communications of the ACM., doi:10.1145/355578.366316 The algorithm works reasonably when the matrix is real and symmetric.

levi_civita3 :: ((Vector3 :*: Vector3) :*: Vector3) Int

https://en.wikipedia.org/wiki/Levi-Civita_symbol

cycle3 :: StreamBuilder str => Vector3 a -> str a

Orphan instances

Monad Vector3

Instance details

Methods

(>>=) :: Vector3 a -> (a -> Vector3 b) -> Vector3 b

(>>) :: Vector3 a -> Vector3 b -> Vector3 b

return :: a -> Vector3 a

fail :: String -> Vector3 a

Functor Vector3

Instance details

Methods

fmap :: (a -> b) -> Vector3 a -> Vector3 b

(<$) :: a -> Vector3 b -> Vector3 a

Applicative Vector3

Instance details

Methods

pure :: a -> Vector3 a

(<*>) :: Vector3 (a -> b) -> Vector3 a -> Vector3 b

liftA2 :: (a -> b -> c) -> Vector3 a -> Vector3 b -> Vector3 c

(*>) :: Vector3 a -> Vector3 b -> Vector3 b

(<*) :: Vector3 a -> Vector3 b -> Vector3 a

Foldable Vector3

Instance details

Methods

fold :: Monoid m => Vector3 m -> m

foldMap :: Monoid m => (a -> m) -> Vector3 a -> m

foldr :: (a -> b -> b) -> b -> Vector3 a -> b

foldr' :: (a -> b -> b) -> b -> Vector3 a -> b

foldl :: (b -> a -> b) -> b -> Vector3 a -> b

foldl' :: (b -> a -> b) -> b -> Vector3 a -> b

foldr1 :: (a -> a -> a) -> Vector3 a -> a

foldl1 :: (a -> a -> a) -> Vector3 a -> a

toList :: Vector3 a -> [a]

null :: Vector3 a -> Bool

length :: Vector3 a -> Int

elem :: Eq a => a -> Vector3 a -> Bool

maximum :: Ord a => Vector3 a -> a

minimum :: Ord a => Vector3 a -> a

sum :: Num a => Vector3 a -> a

product :: Num a => Vector3 a -> a

Traversable Vector3

Instance details

Methods

traverse :: Applicative f => (a -> f b) -> Vector3 a -> f (Vector3 b)

sequenceA :: Applicative f => Vector3 (f a) -> f (Vector3 a)

mapM :: Monad m => (a -> m b) -> Vector3 a -> m (Vector3 b)

sequence :: Monad m => Vector3 (m a) -> m (Vector3 a)

MonadZip Vector3

Instance details

Methods

mzip :: Vector3 a -> Vector3 b -> Vector3 (a, b)

mzipWith :: (a -> b -> c) -> Vector3 a -> Vector3 b -> Vector3 c

munzip :: Vector3 (a, b) -> (Vector3 a, Vector3 b)

PpShowVerticalF Vector3

Instance details

Methods

ppf_vertical :: PpShow a => Vector3 a -> Doc

PpShowF Vector3

Instance details

Methods

ppf :: PpShow a => Vector3 a -> Doc

FiniteComonad Vector3

Instance details

Methods

inverseRotate :: Vector3 a -> Vector3 a

CircularComonad Vector3

Instance details

Methods

rotate :: Vector3 a -> Vector3 a

Comonad Vector3

Instance details

Methods

extract :: Vector3 a -> a

duplicate :: Vector3 a -> Vector3 (Vector3 a)

extend :: (Vector3 a -> b) -> Vector3 a -> Vector3 b

(=>>) :: Vector3 a -> (Vector3 a -> b) -> Vector3 b

(.>>) :: Vector3 a -> b -> Vector3 b

Coapplicative Vector3

Instance details

Methods

coeval :: Vector3 a -> a

colambda :: (Vector3 a -> Vector3 b) -> Vector3 (a -> b)

Unfoldable Vector3

Instance details

Methods

unfoldF :: Monad m => StateT s m a -> StateT s m (Vector3 a)

SplittableVector Vector2 Vector1

Instance details

Methods

vsplit :: (Vector2 :+: Vector1) a -> (Vector2 a, Vector1 a)

SplittableVector Vector1 Vector2

Instance details

Methods

vsplit :: (Vector1 :+: Vector2) a -> (Vector1 a, Vector2 a)

AppendableVector Vector2 Vector1

Instance details

Associated Types

type Vector2 :+: Vector1 :: Type -> Type

Methods

(|>) :: Vector2 a -> Vector1 a -> (Vector2 :+: Vector1) a

AppendableVector Vector1 Vector2

Instance details

Associated Types

type Vector1 :+: Vector2 :: Type -> Type

Methods

(|>) :: Vector1 a -> Vector2 a -> (Vector1 :+: Vector2) a

(Floating a, Ord a, ConjugateSymmetric a) => EigenDecomposable Vector3 a

Instance details

Methods

eigenvalues :: (Vector3 :*: Vector3) a -> Vector3 a

Fractional a => Invertible Vector3 a

Instance details

Methods

cofactor_impl :: (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a

adjucate_impl :: (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a

inverse_impl :: (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a

Num a => Traceable Vector3 a

Instance details

Methods

trace_impl :: (Vector3 :*: Vector3) a -> a

determinant_impl :: (Vector3 :*: Vector3) a -> a

vector_dimension :: Vector3 a -> a

CodiagonalMatrix Vector3 a

Instance details

Associated Types

data Codiagonal Vector3 a :: Type

type Vector3 \\ a :: Type

Methods

codiagonal_impl :: (Vector3 :*: Vector3) a -> Codiagonal Vector3 a

(|\|) :: Vector3 a -> Codiagonal Vector3 a -> (Vector3 :*: Vector3) a

down_project :: Codiagonal Vector3 a -> Vector3 \\ a

right_project :: Codiagonal Vector3 a -> Vector3 \\ a

ProjectionSpace Vector3 Vector2

Instance details

Associated Types

data (Vector3 \\\ Vector2) a :: Type

Methods

project_first :: Vector3 a -> Vector2 a

project_second :: Vector3 a -> (Vector3 \\\ Vector2) a

join_vector :: Vector2 a -> (Vector3 \\\ Vector2) a -> Vector3 a

ProjectionSpace Vector3 Vector1

Instance details

Associated Types

data (Vector3 \\\ Vector1) a :: Type

Methods

project_first :: Vector3 a -> Vector1 a

project_second :: Vector3 a -> (Vector3 \\\ Vector1) a

join_vector :: Vector1 a -> (Vector3 \\\ Vector1) a -> Vector3 a

Num a => Diagonalizable Vector3 a

Instance details

Methods

identity_impl :: Vector3 Integer -> (Vector3 :*: Vector3) a

identity :: (Vector3 :*: Vector3) a

diagonal_impl :: (Vector3 :*: Vector3) a -> Vector3 a

diagonal_matrix_impl :: Vector3 a -> (Vector3 :*: Vector3) a

Num a => Indexable Vector3 a

Instance details

Methods

diagonal_projections :: Vector3 (Index Vector3 a)

basis_vector :: Index Vector3 a -> Vector3 a

index_project :: Index Vector3 a -> Vector3 a -> a

indexable_indices :: Vector3 a

Transposable Vector3 Vector3 a

Instance details

Methods

transpose_impl :: (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a

Transposable Vector3 Vector2 a

Instance details

Methods

transpose_impl :: (Vector3 :*: Vector2) a -> (Vector2 :*: Vector3) a

Transposable Vector3 Vector1 a

Instance details

Methods

transpose_impl :: (Vector3 :*: Vector1) a -> (Vector1 :*: Vector3) a

Transposable Vector2 Vector3 a

Instance details

Methods

transpose_impl :: (Vector2 :*: Vector3) a -> (Vector3 :*: Vector2) a

Transposable Vector1 Vector3 a

Instance details

Methods

transpose_impl :: (Vector1 :*: Vector3) a -> (Vector3 :*: Vector1) a

Transposable Stream Vector3 a

Instance details

Methods

transpose_impl :: (Stream :*: Vector3) a -> (Vector3 :*: Stream) a

(Num a, ConjugateSymmetric a) => LinearTransform Vector3 Vector3 a

Instance details

Methods

(<*>>) :: Vector3 a -> (Vector3 :*: Vector3) a -> Vector3 a

(<<*>) :: (Vector3 :*: Vector3) a -> Vector3 a -> Vector3 a

(Num a, ConjugateSymmetric a) => LinearTransform Vector3 Vector2 a

Instance details

Methods

(<*>>) :: Vector2 a -> (Vector3 :*: Vector2) a -> Vector3 a

(<<*>) :: (Vector3 :*: Vector2) a -> Vector3 a -> Vector2 a

(Num a, ConjugateSymmetric a) => LinearTransform Vector3 Vector1 a

Instance details

Methods

(<*>>) :: Vector1 a -> (Vector3 :*: Vector1) a -> Vector3 a

(<<*>) :: (Vector3 :*: Vector1) a -> Vector3 a -> Vector1 a

(Num a, ConjugateSymmetric a) => LinearTransform Vector2 Vector3 a

Instance details

Methods

(<*>>) :: Vector3 a -> (Vector2 :*: Vector3) a -> Vector2 a

(<<*>) :: (Vector2 :*: Vector3) a -> Vector2 a -> Vector3 a

(Num a, ConjugateSymmetric a) => LinearTransform Vector1 Vector3 a

Instance details

Methods

(<*>>) :: Vector3 a -> (Vector1 :*: Vector3) a -> Vector1 a

(<<*>) :: (Vector1 :*: Vector3) a -> Vector1 a -> Vector3 a

(Monad str, Limiting str a) => Limiting str (Vector3 a)

Instance details

Associated Types

data Closure str (Vector3 a) :: Type

Methods

limit :: str (Vector3 a) -> Closure str (Vector3 a)

approximations :: Closure str (Vector3 a) -> str (Vector3 a)

Transposable Vector3 ((->) row :: Type -> Type) a

Instance details

Methods

transpose_impl :: (Vector3 :*: (->) row) a -> ((->) row :*: Vector3) a

Eq a => Eq (Vector3 a)

Instance details

Methods

(==) :: Vector3 a -> Vector3 a -> Bool

(/=) :: Vector3 a -> Vector3 a -> Bool

Data a => Data (Vector3 a)

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> Vector3 a -> c (Vector3 a)

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c (Vector3 a)

toConstr :: Vector3 a -> Constr

dataTypeOf :: Vector3 a -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c (Vector3 a))

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c (Vector3 a))

gmapT :: (forall b. Data b => b -> b) -> Vector3 a -> Vector3 a

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> Vector3 a -> r

gmapQr :: (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> Vector3 a -> r

gmapQ :: (forall d. Data d => d -> u) -> Vector3 a -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> Vector3 a -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> Vector3 a -> m (Vector3 a)

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> Vector3 a -> m (Vector3 a)

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> Vector3 a -> m (Vector3 a)

Functor (Codiagonal Vector3)

Instance details

Methods

fmap :: (a -> b) -> Codiagonal Vector3 a -> Codiagonal Vector3 b

(<$) :: a -> Codiagonal Vector3 b -> Codiagonal Vector3 a

Num a => Num (Vector3 a)

Instance details

Methods

(+) :: Vector3 a -> Vector3 a -> Vector3 a

(-) :: Vector3 a -> Vector3 a -> Vector3 a

(*) :: Vector3 a -> Vector3 a -> Vector3 a

negate :: Vector3 a -> Vector3 a

abs :: Vector3 a -> Vector3 a

signum :: Vector3 a -> Vector3 a

fromInteger :: Integer -> Vector3 a

Ord a => Ord (Vector3 a)

Instance details

Methods

compare :: Vector3 a -> Vector3 a -> Ordering

(<) :: Vector3 a -> Vector3 a -> Bool

(<=) :: Vector3 a -> Vector3 a -> Bool

(>) :: Vector3 a -> Vector3 a -> Bool

(>=) :: Vector3 a -> Vector3 a -> Bool

max :: Vector3 a -> Vector3 a -> Vector3 a

min :: Vector3 a -> Vector3 a -> Vector3 a

Read a => Read (Vector3 a)

Instance details

Methods

readsPrec :: Int -> ReadS (Vector3 a)

readList :: ReadS [Vector3 a]

readPrec :: ReadPrec (Vector3 a)

readListPrec :: ReadPrec [Vector3 a]

Show a => Show (Vector3 a)

Instance details

Methods

showsPrec :: Int -> Vector3 a -> ShowS

show :: Vector3 a -> String

showList :: [Vector3 a] -> ShowS

Applicative (Codiagonal Vector3)

Instance details

Methods

pure :: a -> Codiagonal Vector3 a

(<*>) :: Codiagonal Vector3 (a -> b) -> Codiagonal Vector3 a -> Codiagonal Vector3 b

liftA2 :: (a -> b -> c) -> Codiagonal Vector3 a -> Codiagonal Vector3 b -> Codiagonal Vector3 c

(*>) :: Codiagonal Vector3 a -> Codiagonal Vector3 b -> Codiagonal Vector3 b

(<*) :: Codiagonal Vector3 a -> Codiagonal Vector3 b -> Codiagonal Vector3 a

Generic a => Generic (Vector3 a)

Instance details

Associated Types

type Rep (Vector3 a) :: Type -> Type

Methods

from :: Vector3 a -> Rep (Vector3 a) x

to :: Rep (Vector3 a) x -> Vector3 a

Num s => Semigroup (Vector3 s)

Instance details

Methods

(<>) :: Vector3 s -> Vector3 s -> Vector3 s

sconcat :: NonEmpty (Vector3 s) -> Vector3 s

stimes :: Integral b => b -> Vector3 s -> Vector3 s

Num s => Monoid (Vector3 s)

Instance details

Methods

mempty :: Vector3 s

mappend :: Vector3 s -> Vector3 s -> Vector3 s

mconcat :: [Vector3 s] -> Vector3 s

Binary s => Binary (Vector3 s)

Instance details

Methods

put :: Vector3 s -> Put

get :: Get (Vector3 s)

putList :: [Vector3 s] -> Put

PpShow a => PpShow (Vector3 a)

Instance details

Methods

pp :: Vector3 a -> Doc

Universe a => Universe (Vector3 a)

Instance details

Methods

all_elements :: [Vector3 a]

Visitor (Vector3 a)

Instance details

Associated Types

data Fold (Vector3 a) a :: Type

Methods

visit :: Fold (Vector3 a) a0 -> Vector3 a -> a0

Num a => CoordinateSpace (Vector3 a)

Instance details

Associated Types

type Coordinate (Vector3 a) :: Type

Methods

index :: Coordinate (Vector3 a) -> Vector3 a -> Scalar (Vector3 a)

listVector :: [Scalar (Vector3 a)] -> Vector3 a

dimension_size :: Vector3 a -> Int

coordinates :: Vector3 a -> [Coordinate (Vector3 a)]

Num a => StandardBasis (Vector3 a)

Instance details

Methods

unit_vectors :: [Vector3 a]

ConjugateSymmetric a => ConjugateSymmetric (Vector3 a)

Instance details

Methods

conj :: Vector3 a -> Vector3 a

(Floating a, ConjugateSymmetric a) => NormedSpace (Vector3 a)

Instance details

Methods

norm :: Vector3 a -> Scalar (Vector3 a)

Num a => LieAlgebra (Vector3 a)
https://en.wikipedia.org/wiki/Lie_algebra

Instance details

Methods

(%<>%) :: Vector3 a -> Vector3 a -> Vector3 a

(ConjugateSymmetric a, Num a) => InnerProductSpace (Vector3 a)

Instance details

Methods

(%.) :: Vector3 a -> Vector3 a -> Scalar (Vector3 a)

Num a => VectorSpace (Vector3 a)

Instance details

Associated Types

type Scalar (Vector3 a) :: Type

Methods

vzero :: Vector3 a

vnegate :: Vector3 a -> Vector3 a

(%+) :: Vector3 a -> Vector3 a -> Vector3 a

(%*) :: Scalar (Vector3 a) -> Vector3 a -> Vector3 a

MedianAlgebra s => MedianAlgebra (Vector3 s)

Instance details

Methods

med :: Vector3 s -> Vector3 s -> Vector3 s -> Vector3 s

ShowPrecision s => ShowPrecision (Vector3 s)

Instance details

Methods

show_at_precision :: Vector3 s -> Integer -> String

Num s => Group (Vector3 s)

Instance details

Methods

ginvert :: Vector3 s -> Vector3 s

Functor (Vector3 \\\ Vector2)

Instance details

Methods

fmap :: (a -> b) -> (Vector3 \\\ Vector2) a -> (Vector3 \\\ Vector2) b

(<$) :: a -> (Vector3 \\\ Vector2) b -> (Vector3 \\\ Vector2) a

Functor (Vector3 \\\ Vector1)

Instance details

Methods

fmap :: (a -> b) -> (Vector3 \\\ Vector1) a -> (Vector3 \\\ Vector1) b

(<$) :: a -> (Vector3 \\\ Vector1) b -> (Vector3 \\\ Vector1) a

Show a => Show (Codiagonal Vector3 a)

Instance details

Methods

showsPrec :: Int -> Codiagonal Vector3 a -> ShowS

show :: Codiagonal Vector3 a -> String

showList :: [Codiagonal Vector3 a] -> ShowS

(Show a, Monad str, Limiting str a) => Show (Closure str (Vector3 a))

Instance details

Methods

showsPrec :: Int -> Closure str (Vector3 a) -> ShowS

show :: Closure str (Vector3 a) -> String

showList :: [Closure str (Vector3 a)] -> ShowS

Applicative (Vector3 \\\ Vector2)

Instance details

Methods

pure :: a -> (Vector3 \\\ Vector2) a

(<*>) :: (Vector3 \\\ Vector2) (a -> b) -> (Vector3 \\\ Vector2) a -> (Vector3 \\\ Vector2) b

liftA2 :: (a -> b -> c) -> (Vector3 \\\ Vector2) a -> (Vector3 \\\ Vector2) b -> (Vector3 \\\ Vector2) c

(*>) :: (Vector3 \\\ Vector2) a -> (Vector3 \\\ Vector2) b -> (Vector3 \\\ Vector2) b

(<*) :: (Vector3 \\\ Vector2) a -> (Vector3 \\\ Vector2) b -> (Vector3 \\\ Vector2) a

Applicative (Vector3 \\\ Vector1)

Instance details

Methods

pure :: a -> (Vector3 \\\ Vector1) a

(<*>) :: (Vector3 \\\ Vector1) (a -> b) -> (Vector3 \\\ Vector1) a -> (Vector3 \\\ Vector1) b

liftA2 :: (a -> b -> c) -> (Vector3 \\\ Vector1) a -> (Vector3 \\\ Vector1) b -> (Vector3 \\\ Vector1) c

(*>) :: (Vector3 \\\ Vector1) a -> (Vector3 \\\ Vector1) b -> (Vector3 \\\ Vector1) b

(<*) :: (Vector3 \\\ Vector1) a -> (Vector3 \\\ Vector1) b -> (Vector3 \\\ Vector1) a

(Monad str, PpShow a, Limiting str a) => PpShow (Closure str (Vector3 a))

Instance details

Methods

pp :: Closure str (Vector3 a) -> Doc

ProjectionSpace (Vector3 \\\ Vector1) Vector1

Instance details

Associated Types

data ((Vector3 \\\ Vector1) \\\ Vector1) a :: Type

Methods

project_first :: (Vector3 \\\ Vector1) a -> Vector1 a

project_second :: (Vector3 \\\ Vector1) a -> ((Vector3 \\\ Vector1) \\\ Vector1) a

join_vector :: Vector1 a -> ((Vector3 \\\ Vector1) \\\ Vector1) a -> (Vector3 \\\ Vector1) a

(Fractional a, ConjugateSymmetric a) => Fractional ((Vector3 :*: Vector3) a)

Instance details

Methods

(/) :: (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a

recip :: (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a

fromRational :: Rational -> (Vector3 :*: Vector3) a

(Num a, ConjugateSymmetric a) => Num ((Vector3 :*: Vector3) a)

Instance details

Methods

(+) :: (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a

(-) :: (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a

(*) :: (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a

negate :: (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a

abs :: (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a

signum :: (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a

fromInteger :: Integer -> (Vector3 :*: Vector3) a

Show a => Show ((Vector3 \\\ Vector2) a)

Instance details

Methods

showsPrec :: Int -> (Vector3 \\\ Vector2) a -> ShowS

show :: (Vector3 \\\ Vector2) a -> String

showList :: [(Vector3 \\\ Vector2) a] -> ShowS

Show a => Show ((Vector3 \\\ Vector1) a)

Instance details

Methods

showsPrec :: Int -> (Vector3 \\\ Vector1) a -> ShowS

show :: (Vector3 \\\ Vector1) a -> String

showList :: [(Vector3 \\\ Vector1) a] -> ShowS

Show (f a) => Show ((Vector3 :*: f) a)

Instance details

Methods

showsPrec :: Int -> (Vector3 :*: f) a -> ShowS

show :: (Vector3 :*: f) a -> String

showList :: [(Vector3 :*: f) a] -> ShowS

(ConjugateSymmetric a, Num a) => Semigroup ((Vector3 :*: Vector3) a)
see "Lawvere,Rosebrugh: Sets for mathematics", pg. 167.

Instance details

Methods

(<>) :: (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a

sconcat :: NonEmpty ((Vector3 :*: Vector3) a) -> (Vector3 :*: Vector3) a

stimes :: Integral b => b -> (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a

(Num a, ConjugateSymmetric a) => Monoid ((Vector3 :*: Vector3) a)
see "Lawvere,Rosebrugh: Sets for mathematics", pg. 167.

Instance details

Methods

mempty :: (Vector3 :*: Vector3) a

mappend :: (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a

mconcat :: [(Vector3 :*: Vector3) a] -> (Vector3 :*: Vector3) a

(ConjugateSymmetric a, Num a) => ConjugateSymmetric ((Vector3 :*: Vector3) a)
https://en.wikipedia.org/wiki/Conjugate_transpose

Instance details

Methods

conj :: (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a

(Floating a, ConjugateSymmetric a) => NormedSpace ((Vector3 :*: Vector3) a)

Instance details

Methods

norm :: (Vector3 :*: Vector3) a -> Scalar ((Vector3 :*: Vector3) a)

(Floating a, ConjugateSymmetric a) => InnerProductSpace ((Vector3 :*: Vector3) a)

Instance details

Methods

(%.) :: (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a -> Scalar ((Vector3 :*: Vector3) a)

Num a => VectorSpace ((Vector3 :*: Vector3) a)
3 x 3 matrices:

Instance details

Associated Types

type Scalar ((Vector3 :*: Vector3) a) :: Type

Methods

vzero :: (Vector3 :*: Vector3) a

vnegate :: (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a

(%+) :: (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a

(%*) :: Scalar ((Vector3 :*: Vector3) a) -> (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a

Num a => VectorSpace ((Vector3 :*: Vector2) a)
3 x 2 matrices

Instance details

Associated Types

type Scalar ((Vector3 :*: Vector2) a) :: Type

Methods

vzero :: (Vector3 :*: Vector2) a

vnegate :: (Vector3 :*: Vector2) a -> (Vector3 :*: Vector2) a

(%+) :: (Vector3 :*: Vector2) a -> (Vector3 :*: Vector2) a -> (Vector3 :*: Vector2) a

(%*) :: Scalar ((Vector3 :*: Vector2) a) -> (Vector3 :*: Vector2) a -> (Vector3 :*: Vector2) a

Num a => VectorSpace ((Vector3 :*: Vector1) a)
3 x 1 matrices:

Instance details

Associated Types

type Scalar ((Vector3 :*: Vector1) a) :: Type

Methods

vzero :: (Vector3 :*: Vector1) a

vnegate :: (Vector3 :*: Vector1) a -> (Vector3 :*: Vector1) a

(%+) :: (Vector3 :*: Vector1) a -> (Vector3 :*: Vector1) a -> (Vector3 :*: Vector1) a

(%*) :: Scalar ((Vector3 :*: Vector1) a) -> (Vector3 :*: Vector1) a -> (Vector3 :*: Vector1) a

Num a => VectorSpace ((Vector2 :*: Vector3) a)
2 x 3 matrices:

Instance details

Associated Types

type Scalar ((Vector2 :*: Vector3) a) :: Type

Methods

vzero :: (Vector2 :*: Vector3) a

vnegate :: (Vector2 :*: Vector3) a -> (Vector2 :*: Vector3) a

(%+) :: (Vector2 :*: Vector3) a -> (Vector2 :*: Vector3) a -> (Vector2 :*: Vector3) a

(%*) :: Scalar ((Vector2 :*: Vector3) a) -> (Vector2 :*: Vector3) a -> (Vector2 :*: Vector3) a

Num a => VectorSpace ((Vector1 :*: Vector3) a)
1 x 3 matrices:

Instance details

Associated Types

type Scalar ((Vector1 :*: Vector3) a) :: Type

Methods

vzero :: (Vector1 :*: Vector3) a

vnegate :: (Vector1 :*: Vector3) a -> (Vector1 :*: Vector3) a

(%+) :: (Vector1 :*: Vector3) a -> (Vector1 :*: Vector3) a -> (Vector1 :*: Vector3) a

(%*) :: Scalar ((Vector1 :*: Vector3) a) -> (Vector1 :*: Vector3) a -> (Vector1 :*: Vector3) a

(Fractional a, ConjugateSymmetric a) => Group ((Vector3 :*: Vector3) a)

Instance details

Methods

ginvert :: (Vector3 :*: Vector3) a -> (Vector3 :*: Vector3) a

Transposable ((->) row :: Type -> Type) Vector3 a

Instance details

Methods

transpose_impl :: ((->) row :*: Vector3) a -> (Vector3 :*: (->) row) a