Math.Matrix.Vector3

Documentation

data Vector3 s

Three element vector

Constructors

Vector3

Fields

xcoord3 :: s

ycoord3 :: s

zcoord3 :: s

del_vector3 :: Infinitesimal a => (Vector3 a -> a) -> Vector3 a -> Vector3 (Closure a)

this computes partial derivates of the scalar-valued 3D vector field along each variable simultaneously. \[\nabla f({\mathbb{v}}) = \frac{\partial \mathbb{f}}{\partial x}{\mathbb{i}} + \frac{\partial \mathbb{f}}{\partial y}{\mathbb{j}} + \frac{\partial \mathbb{f}}{\partial z}{\mathbb{k}}\]

del_partial3 :: DifferentiallyClosed a => (Vector3 a -> a) -> Vector3 a -> Vector3 a

partial_derivate_vector3 :: (Infinitesimal (Scalar a), VectorSpace a, Num a, Limiting a) => (Vector3 (Scalar a) -> a) -> Vector3 (Scalar a) -> Closure (Vector3 a)

cov3 :: a ~ Scalar a => Vector3 (Dual (Vector3 a))

type ComplexVector3 a = (Vector3 :*: Complex) a

i3 :: Num a => Vector3 a

j3 :: Num a => Vector3 a

k3 :: Num a => Vector3 a

type Matrix3 a = (Vector3 :*: Vector3) a

vec3 :: (a, a, a) -> Vector3 a

finite3 :: Vector3 a -> Vec 3 a

vector3_to_vec3 :: Vector3 a -> ThreeD -> a

vec3_iso_obj :: Vector3 ThreeD

vec3_to_vector3 :: (ThreeD -> a) -> Vector3 a

vector3_natiso :: (->) ThreeD :<~>: Vector3

constant3 :: a -> Vector3 a

vec3Unit :: a -> Scalar (Vector3 a)

vec3Counit :: Vector3 (Scalar (Vector3 a)) -> Vector3 a

destructvec3 :: Vector3 a -> (a, a, a)

setx3 :: s -> Vector3 s -> Vector3 s

sety3 :: s -> Vector3 s -> Vector3 s

setz3 :: s -> Vector3 s -> Vector3 s

set_endo :: Vector3 s -> Vector3 (Endo (Vector3 s))

set_vector_action :: Vector3 s -> Vector3 (Vector3 s -> Vector3 s)

removex3 :: Vector3 a -> Vector2 a

removey3 :: Vector3 a -> Vector2 a

removez3 :: Vector3 a -> Vector2 a

remove3 :: Vector3 (Vector3 a -> Vector2 a)

remove_index3 :: Matrix3 a -> Matrix3 (Matrix2 a)

cofactor3 :: Num a => Matrix3 a -> Matrix3 a

adjucate3 :: Num a => Matrix3 a -> Matrix3 a

https://en.wikipedia.org/wiki/Adjugate_matrix

inverse3 :: Fractional a => Matrix3 a -> Matrix3 a

https://en.wikipedia.org/wiki/Invertible_matrix

set3 :: Vector3 a -> Vector3 a -> (Vector3 :*: Vector3) a

swap3 :: (Vector3 :*: Vector3) (Vector3 a -> Vector3 a)

http://en.wikipedia.org/wiki/Elementary_matrix#Operations

mul3 :: VectorSpace s => Scalar s -> Vector3 (Vector3 s -> Vector3 s)

add3Matrix :: VectorSpace a => Scalar a -> ((Vector3 :*: Vector3) :*: Vector3) a -> ((Vector3 :*: Vector3) :*: Vector3) a

add3 :: VectorSpace s => Scalar s -> (Vector3 :*: Vector3) (Vector3 s -> Vector3 s)

swapxy3 :: Vector3 a -> Vector3 a

swapxz3 :: Vector3 a -> Vector3 a

swapyz3 :: Vector3 a -> Vector3 a

mul3Matrix :: (Num a, VectorSpace a) => Scalar a -> Vector3 ((Vector3 :*: Vector3) a)

signs3 :: Integral a => Matrix3 a

vector_epsilon :: Infinitesimal a => Stream (Vector3 a)

pderive3 :: (Closed a, Infinitesimal a) => Vector3 (a -> a) -> Vector3 a -> Vector3 a

partial derivate a function defined for each coordinate along each dimension of three-dimensional vector.

vector_field_derivate :: (Closed a, Infinitesimal a) => (Vector3 a -> a) -> Vector3 a -> Vector3 a

dx3_endo :: Num a => a -> Endo (Vector3 a)

dy3_endo :: Num a => a -> Endo (Vector3 a)

dz3_endo :: Num a => a -> Endo (Vector3 a)

dx3 :: Num a => a -> Vector3 a -> Vector3 a

dy3 :: Num a => a -> Vector3 a -> Vector3 a

dz3 :: Num a => a -> Vector3 a -> Vector3 a

partial_derivate3x :: (Infinitesimal a, Closed a) => Dual (Vector3 a) -> Dual (Vector3 a)

partial_derivate3y :: (Infinitesimal a, Closed a) => Dual (Vector3 a) -> Dual (Vector3 a)

partial_derivate3z :: (Infinitesimal a, Closed a) => Dual (Vector3 a) -> Dual (Vector3 a)

del3 :: (Infinitesimal v, Closed v) => Vector3 (Dual (Vector3 v) -> Dual (Vector3 v))

\[\nabla_3\], three-dimensional partial derivate. Use Applicative.* for applying it.

del3_squared :: (v ~ Scalar v, Infinitesimal v, Closed v) => (Vector3 :*: Vector3) (Dual (Vector3 v) -> Dual (Vector3 v))

op_dot :: Num b => Vector3 (a -> b) -> a -> b

hessian3 :: (v ~ Scalar v, Infinitesimal v, Closed v) => Dual (Vector3 v) -> (Vector3 :*: Vector3) (Dual (Vector3 v))

vector_laplace3 :: VectorDerivative v => LinearMap v (Vector3 (Scalar v)) -> LinearMap v (Vector3 (Scalar v))

divergence3 :: (Infinitesimal a, Closed a) => LinearMap (Vector3 a) (Vector3 a) -> Dual (Vector3 a)

https://en.wikipedia.org/wiki/Divergence

linear_dual_3x :: LinearMap v (Vector3 (Scalar v)) -> Dual v

linear_dual_3y :: LinearMap v (Vector3 (Scalar v)) -> Dual v

linear_dual_3z :: LinearMap v (Vector3 (Scalar v)) -> Dual v

cycle3 :: Vector3 a -> Stream a

grad3 :: (Infinitesimal s, Closed s) => Dual (Vector3 s) -> LinearMap (Vector3 s) (Vector3 s)

https://en.wikipedia.org/wiki/Gradient

laplace3 :: (Infinitesimal a, Closed a) => Dual (Vector3 a) -> Dual (Vector3 a)

https://en.wikipedia.org/wiki/Laplace_operator

curl3 :: (Infinitesimal a, Closed a) => LinearMap (Vector3 a) (Vector3 a) -> LinearMap (Vector3 a) (Vector3 a)

https://en.wikipedia.org/wiki/Curl_(mathematics)

sum_coordinates3 :: Num a => Vector3 a -> a

cube_norm3 :: Floating a => Vector3 a -> a

cube root of a sum of cubes.

nth_norm3 :: Floating a => a -> Vector3 a -> a

nth root of a sum of nth powers

index3 :: Vector3 (Vector3 a -> a)

mapBasis :: Num a => (Vector3 a -> Vector3 a) -> Vector3 a -> Vector3 a

functionMatrix3 :: Num a => (Vector3 a -> Vector3 a) -> Matrix3 a

splitx :: Vector3 s -> (Vector2 s, Vector1 s)

splity :: Vector3 s -> (Vector2 s, Vector1 s)

splitz :: Vector3 s -> (Vector2 s, Vector1 s)

trace3 :: Num a => Matrix3 a -> a

vector3 :: [a] -> Vector3 a

cross_product :: Num a => Vector3 a -> Vector3 a -> Vector3 a

https://en.wikipedia.org/wiki/Cross_product WARNING: a cross product produces an axial vector. https://en.wikipedia.org/wiki/Pseudovector

cross_product3 :: Num a => (Vector2 :*: Vector3) a -> Vector3 a

cross_product3 computes the three dimensional vector that is orthogonal to both rows of the 2x3 vector.

matrix_exponential3 :: (Closed b, ConjugateSymmetric b, Fractional b) => Matrix3 b -> Matrix3 b

xid :: Num a => Vector3 a

yid :: Num a => Vector3 a

zid :: Num a => Vector3 a

identity3 :: Num a => Matrix3 a

matrix_power3 :: (ConjugateSymmetric a, Num a) => Matrix3 a -> Integer -> Matrix3 a

vector_indices3 :: Integral a => Vector3 a

matrix_indices3 :: Integral a => (Vector3 :*: Vector3) (a, a)

matrix_multiply3 :: (ConjugateSymmetric a, Num a) => Matrix3 a -> Matrix3 a -> Matrix3 a

dot3 :: Num a => Vector3 a -> Vector3 a -> a

dot3 doesn't work on complex numbers

left_multiply3 :: Num a => Vector3 a -> Matrix3 a -> Vector3 a

right_multiply3 :: Num a => Matrix3 a -> Vector3 a -> Vector3 a

transpose3 :: Matrix3 a -> Matrix3 a

outer3 :: (a -> b -> c) -> Vector3 a -> Vector3 b -> Matrix3 c

zipWithV3 :: (a -> b -> c) -> Vector3 a -> Vector3 b -> Vector3 c

rotate_x :: Floating a => a -> Matrix3 a

rotate_y :: Floating a => a -> Matrix3 a

rotate_z :: Floating a => a -> Matrix3 a

rotate_coordinates3 :: Vector3 a -> Vector3 a

zero_codiagonal3 :: Num a => Codiagonal Vector3 a

diagonal_matrix3 :: Num a => Vector3 a -> (Vector3 :*: Vector3) a

matrix3 :: Vector3 a -> Codiagonal Vector3 a -> (Vector3 :*: Vector3) a

codiag3 :: Matrix3 a -> Codiagonal Vector3 a

codiagonal3 :: Matrix3 a -> Vector2 (Vector2 a, Vector2 a)

diagonal3 :: Matrix3 a -> Vector3 a

diagonal_projections3 :: Vector3 (Vector3 a -> a)

diagonal3x :: Matrix3 a -> Vector3 a

diagonal3y :: Matrix3 a -> Vector3 a

diagonal3z :: Matrix3 a -> Vector3 a

determinant3 :: Num a => Matrix3 a -> a

algorithm from http://en.wikipedia.org/wiki/Determinant

determinant3_ :: Num a => Matrix3 a -> a

gram_schmidt3 :: (Fractional (Scalar a), Num a, InnerProductSpace a, VectorSpace a) => Vector3 a -> Vector3 a

eigenvalue3 :: (Floating a, Ord a, ConjugateSymmetric a) => (Vector3 :*: Vector3) a -> Vector3 a

https://en.wikipedia.org/wiki/Eigenvalue_algorithm based on algorithm by Smith, Oliver K.: Eigenvalues of symmetric 3x3 matrix, 1961, Communications of the ACM., doi:10.1145/355578.366316 The algorithm works reasonably when the matrix is real and symmetric.

levi_civita3 :: ((Vector3 :*: Vector3) :*: Vector3) Int

https://en.wikipedia.org/wiki/Levi-Civita_symbol

Orphan instances

SplittableVector Vector1 Vector2

Instance details

Methods

vsplit :: (Vector1 :+: Vector2) a -> (Vector1 a, Vector2 a)

SplittableVector Vector2 Vector1

Instance details

Methods

vsplit :: (Vector2 :+: Vector1) a -> (Vector2 a, Vector1 a)

AppendableVector Vector1 Vector2

Instance details

Associated Types

type Vector1 :+: Vector2 :: Type -> Type

Methods

(|>) :: Vector1 a -> Vector2 a -> (Vector1 :+: Vector2) a

AppendableVector Vector2 Vector1

Instance details

Associated Types

type Vector2 :+: Vector1 :: Type -> Type

Methods

(|>) :: Vector2 a -> Vector1 a -> (Vector2 :+: Vector1) a