Math.Matrix.Vector4

Documentation

x4 :: Num a => Vector4 a

y4 :: Num a => Vector4 a

z4 :: Num a => Vector4 a

t4 :: Num a => Vector4 a

dx4 :: forall {str :: Type -> Type} {a}. (Infinitesimal str a, Num (Closure str a)) => Vector4 (Closure str a)

dy4 :: forall {str :: Type -> Type} {a}. (Infinitesimal str a, Num (Closure str a)) => Vector4 (Closure str a)

dz4 :: forall {str :: Type -> Type} {a}. (Infinitesimal str a, Num (Closure str a)) => Vector4 (Closure str a)

dt4 :: forall {str :: Type -> Type} {a}. (Infinitesimal str a, Num (Closure str a)) => Vector4 (Closure str a)

setx4 :: s -> Vector4 s -> Vector4 s

sety4 :: s -> Vector4 s -> Vector4 s

setz4 :: s -> Vector4 s -> Vector4 s

sett4 :: s -> Vector4 s -> Vector4 s

update_row4 :: g a -> Vector4 ((Vector4 :*: g) a -> (Vector4 :*: g) a)

update_column4 :: Applicative f => f a -> Vector4 ((f :*: Vector4) a -> (f :*: Vector4) a)

type Matrix4 a = (Vector4 :*: Vector4) a

codiag4 :: (Vector4 :*: Vector4) a -> Codiagonal Vector4 a

matrix4 :: Vector4 a -> Codiagonal Vector4 a -> (Vector4 :*: Vector4) a

diagonal_projections4 :: Num a => Vector4 (Index Vector4 a)

cycle4 :: Vector4 a -> Stream a

vector4 :: [a] -> Vector4 a

unzipV4 :: Vector4 (a, b) -> (Vector4 a, Vector4 b)

constant4 :: a -> Vector4 a

dim4 :: Vector4 Integer

sum_coordinates4 :: Num a => Vector4 a -> a

left_multiply4_gen :: (Functor f, Num a, ConjugateSymmetric a) => Vector4 a -> (f :*: Vector4) a -> f a

right_multiply4_gen :: (VectorSpace (f a), ConjugateSymmetric a, Scalar (f a) ~ a) => (Vector4 :*: f) a -> Vector4 a -> f a

matrix_multiply4 :: (Num a, ConjugateSymmetric a) => Matrix4 a -> Matrix4 a -> Matrix4 a

dot4 :: (Num a, ConjugateSymmetric a) => Vector4 a -> Vector4 a -> a

versor :: (Floating a, ConjugateSymmetric a) => Vector4 a -> Vector4 a

index4 :: Int -> Vector4 a -> a

diagonal4 :: (Vector4 :*: Vector4) a -> Vector4 a

dt_4 :: Num s => s -> Vector4 s -> Vector4 s

dx_4 :: Num s => s -> Vector4 s -> Vector4 s

dy_4 :: Num s => s -> Vector4 s -> Vector4 s

dz_4 :: Num s => s -> Vector4 s -> Vector4 s

transpose4 :: Num a => Matrix4 a -> Matrix4 a

transpose4_impl :: Matrix4 a -> Matrix4 a

identity4 :: Num a => Matrix4 a

diagonal4x :: Matrix4 a -> Vector4 a

zero_codiagonal4 :: Num a => Codiagonal Vector4 a

diagonal_matrix4 :: Num a => Vector4 a -> (Vector4 :*: Vector4) a

mat4 :: Vector4 a -> Codiagonal Vector4 a -> (Vector4 :*: Vector4) a

trace4 :: Num a => Matrix4 a -> a

removet4 :: Vector4 a -> Vector3 a

removex4 :: Vector4 a -> Vector3 a

removey4 :: Vector4 a -> Vector3 a

removez4 :: Vector4 a -> Vector3 a

removes4 :: Vector4 (Vector4 a -> Vector3 a)

remove_index4 :: Matrix4 a -> Matrix4 (Matrix3 a)

cofactor4 :: Num a => Matrix4 a -> Matrix4 a

adjucate4 :: Num a => Matrix4 a -> Matrix4 a

signs4 :: Integral a => Matrix4 a

vector_indices4 :: Integral a => Vector4 a

matrix_indices4 :: Integral a => (Vector4 :*: Vector4) (a, a)

inverse4 :: Fractional a => Matrix4 a -> Matrix4 a

https://en.wikipedia.org/wiki/Invertible_matrix

cross4 :: Num a => Vector4 a -> Vector4 a -> Vector4 a -> Vector4 a

Generalization of cross product to four dimensions This is computed using formal determinant representation of cross product expanded to four dimensions https://en.wikipedia.org/wiki/Cross_product This computes vector that is linearly independent of all three vectors given in four dimensional space.

cross_product4 :: Num a => (Vector3 :*: Vector4) a -> Vector4 a

determinant4 :: Num a => Matrix4 a -> a

Orphan instances

MonadZip Vector4

Instance details

Methods

mzip :: Vector4 a -> Vector4 b -> Vector4 (a, b)

mzipWith :: (a -> b -> c) -> Vector4 a -> Vector4 b -> Vector4 c

munzip :: Vector4 (a, b) -> (Vector4 a, Vector4 b)

Foldable Vector4

Instance details

Methods

fold :: Monoid m => Vector4 m -> m

foldMap :: Monoid m => (a -> m) -> Vector4 a -> m

foldMap' :: Monoid m => (a -> m) -> Vector4 a -> m

foldr :: (a -> b -> b) -> b -> Vector4 a -> b

foldr' :: (a -> b -> b) -> b -> Vector4 a -> b

foldl :: (b -> a -> b) -> b -> Vector4 a -> b

foldl' :: (b -> a -> b) -> b -> Vector4 a -> b

foldr1 :: (a -> a -> a) -> Vector4 a -> a

foldl1 :: (a -> a -> a) -> Vector4 a -> a

toList :: Vector4 a -> [a]

null :: Vector4 a -> Bool

length :: Vector4 a -> Int

elem :: Eq a => a -> Vector4 a -> Bool

maximum :: Ord a => Vector4 a -> a

minimum :: Ord a => Vector4 a -> a

sum :: Num a => Vector4 a -> a

product :: Num a => Vector4 a -> a

Traversable Vector4

Instance details

Methods

traverse :: Applicative f => (a -> f b) -> Vector4 a -> f (Vector4 b)

sequenceA :: Applicative f => Vector4 (f a) -> f (Vector4 a)

mapM :: Monad m => (a -> m b) -> Vector4 a -> m (Vector4 b)

sequence :: Monad m => Vector4 (m a) -> m (Vector4 a)

Applicative Vector4

Instance details

Methods

pure :: a -> Vector4 a

(<*>) :: Vector4 (a -> b) -> Vector4 a -> Vector4 b

liftA2 :: (a -> b -> c) -> Vector4 a -> Vector4 b -> Vector4 c

(*>) :: Vector4 a -> Vector4 b -> Vector4 b

(<*) :: Vector4 a -> Vector4 b -> Vector4 a

Functor Vector4

Instance details

Methods

fmap :: (a -> b) -> Vector4 a -> Vector4 b

(<$) :: a -> Vector4 b -> Vector4 a

Monad Vector4

Instance details

Methods

(>>=) :: Vector4 a -> (a -> Vector4 b) -> Vector4 b

(>>) :: Vector4 a -> Vector4 b -> Vector4 b

return :: a -> Vector4 a

UpdateableMatrixDimension Vector4

Instance details

Methods

write_row :: Applicative h => h a -> Vector4 ((Vector4 :*: h) a -> (Vector4 :*: h) a)

write_column :: Applicative h => h a -> Vector4 ((h :*: Vector4) a -> (h :*: Vector4) a)

Comonad Vector4

Instance details

Methods

extract :: Vector4 a -> a

duplicate :: Vector4 a -> Vector4 (Vector4 a)

extend :: (Vector4 a -> b) -> Vector4 a -> Vector4 b

(=>>) :: Vector4 a -> (Vector4 a -> b) -> Vector4 b

(.>>) :: Vector4 a -> b -> Vector4 b

PpShowF Vector4

Instance details

Methods

ppf :: PpShow a => Vector4 a -> Doc

PpShowVerticalF Vector4

Instance details

Methods

ppf_vertical :: PpShow a => Vector4 a -> Doc

Unfoldable Vector4

Instance details

Methods

unfoldF :: forall (m :: Type -> Type) s a. Monad m => StateT s m a -> StateT s m (Vector4 a)

AppendableVector Vector1 Vector3

Instance details

Associated Types

type Vector1 :+: Vector3 :: Type -> Type

Methods

(||>>) :: Vector1 a -> Vector3 a -> (Vector1 :+: Vector3) a

AppendableVector Vector2 Vector2

Instance details

Associated Types

type Vector2 :+: Vector2 :: Type -> Type

Methods

(||>>) :: Vector2 a -> Vector2 a -> (Vector2 :+: Vector2) a

AppendableVector Vector3 Vector1

Instance details

Associated Types

type Vector3 :+: Vector1 :: Type -> Type

Methods

(||>>) :: Vector3 a -> Vector1 a -> (Vector3 :+: Vector1) a

StreamBuilder str => AppendableVector Vector4 str

Instance details

Associated Types

type Vector4 :+: str :: Type -> Type

Methods

(||>>) :: Vector4 a -> str a -> (Vector4 :+: str) a

CodiagonalMatrix Vector4 a

Instance details

Associated Types

data Codiagonal Vector4 a

type Vector4 \\ a

Methods

codiagonal_impl :: (Vector4 :*: Vector4) a -> Codiagonal Vector4 a

(|\|) :: Vector4 a -> Codiagonal Vector4 a -> (Vector4 :*: Vector4) a

down_project :: Codiagonal Vector4 a -> Vector4 \\ a

right_project :: Codiagonal Vector4 a -> Vector4 \\ a

Num a => Diagonalizable Vector4 a

Instance details

Methods

identity_impl :: Vector4 Integer -> (Vector4 :*: Vector4) a

identity :: (Vector4 :*: Vector4) a

diagonal_impl :: (Vector4 :*: Vector4) a -> Vector4 a

diagonal_matrix_impl :: Vector4 a -> (Vector4 :*: Vector4) a

Num a => Indexable Vector4 a

Instance details

Methods

diagonal_projections :: Vector4 (Index Vector4 a)

basis_vector :: Index Vector4 a -> Vector4 a

index_project :: Index Vector4 a -> Vector4 a -> a

indexable_indices :: Vector4 a

Fractional a => Invertible Vector4 a

Instance details

Methods

cofactor_impl :: (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a

adjucate_impl :: (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a

inverse_impl :: (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a

SplittableVector Vector1 Vector3

Instance details

Methods

vsplit :: (Vector1 :+: Vector3) a -> (Vector1 a, Vector3 a)

SplittableVector Vector2 Vector2

Instance details

Methods

vsplit :: (Vector2 :+: Vector2) a -> (Vector2 a, Vector2 a)

SplittableVector Vector3 Vector1

Instance details

Methods

vsplit :: (Vector3 :+: Vector1) a -> (Vector3 a, Vector1 a)

Num a => Traceable Vector4 a

Instance details

Methods

trace_impl :: (Vector4 :*: Vector4) a -> a

determinant_impl :: (Vector4 :*: Vector4) a -> a

vector_dimension :: Vector4 a -> a

(Num a, ConjugateSymmetric a) => LinearTransform Vector1 Vector4 a

Instance details

Methods

(<*>>) :: Vector4 a -> (Vector1 :*: Vector4) a -> Vector1 a

(<<*>) :: (Vector1 :*: Vector4) a -> Vector1 a -> Vector4 a

(Num a, ConjugateSymmetric a) => LinearTransform Vector2 Vector4 a

Instance details

Methods

(<*>>) :: Vector4 a -> (Vector2 :*: Vector4) a -> Vector2 a

(<<*>) :: (Vector2 :*: Vector4) a -> Vector2 a -> Vector4 a

(Num a, ConjugateSymmetric a) => LinearTransform Vector3 Vector4 a

Instance details

Methods

(<*>>) :: Vector4 a -> (Vector3 :*: Vector4) a -> Vector3 a

(<<*>) :: (Vector3 :*: Vector4) a -> Vector3 a -> Vector4 a

(Num a, ConjugateSymmetric a) => LinearTransform Vector4 Vector1 a

Instance details

Methods

(<*>>) :: Vector1 a -> (Vector4 :*: Vector1) a -> Vector4 a

(<<*>) :: (Vector4 :*: Vector1) a -> Vector4 a -> Vector1 a

(Num a, ConjugateSymmetric a) => LinearTransform Vector4 Vector2 a

Instance details

Methods

(<*>>) :: Vector2 a -> (Vector4 :*: Vector2) a -> Vector4 a

(<<*>) :: (Vector4 :*: Vector2) a -> Vector4 a -> Vector2 a

(Num a, ConjugateSymmetric a) => LinearTransform Vector4 Vector3 a

Instance details

Methods

(<*>>) :: Vector3 a -> (Vector4 :*: Vector3) a -> Vector4 a

(<<*>) :: (Vector4 :*: Vector3) a -> Vector4 a -> Vector3 a

(Num a, ConjugateSymmetric a) => LinearTransform Vector4 Vector4 a

Instance details

Methods

(<*>>) :: Vector4 a -> (Vector4 :*: Vector4) a -> Vector4 a

(<<*>) :: (Vector4 :*: Vector4) a -> Vector4 a -> Vector4 a

Transposable Vector1 Vector4 a
1 x 4 matrices:

Instance details

Methods

transpose_impl :: (Vector1 :*: Vector4) a -> (Vector4 :*: Vector1) a

Transposable Vector2 Vector4 a
2 x 4 matrices:

Instance details

Methods

transpose_impl :: (Vector2 :*: Vector4) a -> (Vector4 :*: Vector2) a

Transposable Vector3 Vector4 a
3 x 4 matrices

Instance details

Methods

transpose_impl :: (Vector3 :*: Vector4) a -> (Vector4 :*: Vector3) a

Transposable Vector4 Vector1 a
4 x 1 matrices:

Instance details

Methods

transpose_impl :: (Vector4 :*: Vector1) a -> (Vector1 :*: Vector4) a

Transposable Vector4 Vector2 a
4 x 2 matrices

Instance details

Methods

transpose_impl :: (Vector4 :*: Vector2) a -> (Vector2 :*: Vector4) a

Transposable Vector4 Vector3 a
4 x 3 matrices:

Instance details

Methods

transpose_impl :: (Vector4 :*: Vector3) a -> (Vector3 :*: Vector4) a

Transposable Vector4 Vector4 a
4 x 4 matrices

Instance details

Methods

transpose_impl :: (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a

Transposable Stream Vector4 a

Instance details

Methods

transpose_impl :: (Stream :*: Vector4) a -> (Vector4 :*: Stream) a

(Monad str, Limiting str a) => Limiting str (Vector4 a)

Instance details

Associated Types

data Closure str (Vector4 a)

Methods

limit :: str (Vector4 a) -> Closure str (Vector4 a)

approximations :: Closure str (Vector4 a) -> str (Vector4 a)

Transposable Vector4 ((->) row) a

Instance details

Methods

transpose_impl :: (Vector4 :*: (->) row) a -> ((->) row :*: Vector4) a

Data a => Data (Vector4 a)

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> Vector4 a -> c (Vector4 a)

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c (Vector4 a)

toConstr :: Vector4 a -> Constr

dataTypeOf :: Vector4 a -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c (Vector4 a))

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c (Vector4 a))

gmapT :: (forall b. Data b => b -> b) -> Vector4 a -> Vector4 a

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> Vector4 a -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> Vector4 a -> r

gmapQ :: (forall d. Data d => d -> u) -> Vector4 a -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> Vector4 a -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> Vector4 a -> m (Vector4 a)

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> Vector4 a -> m (Vector4 a)

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> Vector4 a -> m (Vector4 a)

Applicative (Codiagonal Vector4)

Instance details

Methods

pure :: a -> Codiagonal Vector4 a

(<*>) :: Codiagonal Vector4 (a -> b) -> Codiagonal Vector4 a -> Codiagonal Vector4 b

liftA2 :: (a -> b -> c) -> Codiagonal Vector4 a -> Codiagonal Vector4 b -> Codiagonal Vector4 c

(*>) :: Codiagonal Vector4 a -> Codiagonal Vector4 b -> Codiagonal Vector4 b

(<*) :: Codiagonal Vector4 a -> Codiagonal Vector4 b -> Codiagonal Vector4 a

Functor (Codiagonal Vector4)

Instance details

Methods

fmap :: (a -> b) -> Codiagonal Vector4 a -> Codiagonal Vector4 b

(<$) :: a -> Codiagonal Vector4 b -> Codiagonal Vector4 a

Num s => Monoid (Vector4 s)

Instance details

Methods

mempty :: Vector4 s

mappend :: Vector4 s -> Vector4 s -> Vector4 s

mconcat :: [Vector4 s] -> Vector4 s

Num s => Semigroup (Vector4 s)

Instance details

Methods

(<>) :: Vector4 s -> Vector4 s -> Vector4 s

sconcat :: NonEmpty (Vector4 s) -> Vector4 s

stimes :: Integral b => b -> Vector4 s -> Vector4 s

Generic a => Generic (Vector4 a)

Instance details

Associated Types

type Rep (Vector4 a) :: Type -> Type

Methods

from :: Vector4 a -> Rep (Vector4 a) x

to :: Rep (Vector4 a) x -> Vector4 a

Num a => Num (Vector4 a)

Instance details

Methods

(+) :: Vector4 a -> Vector4 a -> Vector4 a

(-) :: Vector4 a -> Vector4 a -> Vector4 a

(*) :: Vector4 a -> Vector4 a -> Vector4 a

negate :: Vector4 a -> Vector4 a

abs :: Vector4 a -> Vector4 a

signum :: Vector4 a -> Vector4 a

fromInteger :: Integer -> Vector4 a

Read a => Read (Vector4 a)

Instance details

Methods

readsPrec :: Int -> ReadS (Vector4 a)

readList :: ReadS [Vector4 a]

readPrec :: ReadPrec (Vector4 a)

readListPrec :: ReadPrec [Vector4 a]

Show a => Show (Vector4 a)

Instance details

Methods

showsPrec :: Int -> Vector4 a -> ShowS

show :: Vector4 a -> String

showList :: [Vector4 a] -> ShowS

Binary s => Binary (Vector4 s)

Instance details

Methods

put :: Vector4 s -> Put

get :: Get (Vector4 s)

putList :: [Vector4 s] -> Put

ConjugateSymmetric a => ConjugateSymmetric (Vector4 a)

Instance details

Methods

conj :: Vector4 a -> Vector4 a

Num a => CoordinateSpace (Vector4 a)

Instance details

Associated Types

type Coordinate (Vector4 a)

Methods

index :: Coordinate (Vector4 a) -> Vector4 a -> Scalar (Vector4 a)

listVector :: [Scalar (Vector4 a)] -> Vector4 a

dimension_size :: Vector4 a -> Int

coordinates :: Vector4 a -> [Coordinate (Vector4 a)]

(Num a, ConjugateSymmetric a) => InnerProductSpace (Vector4 a)

Instance details

Methods

(%.) :: Vector4 a -> Vector4 a -> Scalar (Vector4 a)

(Floating a, ConjugateSymmetric a) => NormedSpace (Vector4 a)

Instance details

Methods

norm :: Vector4 a -> Scalar (Vector4 a)

norm_squared :: Vector4 a -> Scalar (Vector4 a)

Num a => StandardBasis (Vector4 a)

Instance details

Methods

unit_vectors :: [Vector4 a]

Num a => VectorSpace (Vector4 a)

Instance details

Associated Types

type Scalar (Vector4 a)

Methods

vzero :: Vector4 a

vnegate :: Vector4 a -> Vector4 a

(%+) :: Vector4 a -> Vector4 a -> Vector4 a

(%*) :: Scalar (Vector4 a) -> Vector4 a -> Vector4 a

ShowPrecision s => ShowPrecision (Vector4 s)

Instance details

Methods

show_at_precision :: Vector4 s -> Integer -> String

MedianAlgebra s => MedianAlgebra (Vector4 s)

Instance details

Methods

med :: Vector4 s -> Vector4 s -> Vector4 s -> Vector4 s

PpShow a => PpShow (Vector4 a)

Instance details

Methods

pp :: Vector4 a -> Doc

Universe a => Universe (Vector4 a)

Instance details

Methods

all_elements :: [Vector4 a]

IdVisitor (Vector4 a)

Instance details

Methods

idFold :: Fold (Vector4 a) (Vector4 a)

Visitor (Vector4 a)

Instance details

Associated Types

data Fold (Vector4 a) :: Type -> Type

Methods

visit :: Fold (Vector4 a) a0 -> Vector4 a -> a0

Eq a => Eq (Vector4 a)

Instance details

Methods

(==) :: Vector4 a -> Vector4 a -> Bool

(/=) :: Vector4 a -> Vector4 a -> Bool

Ord a => Ord (Vector4 a)

Instance details

Methods

compare :: Vector4 a -> Vector4 a -> Ordering

(<) :: Vector4 a -> Vector4 a -> Bool

(<=) :: Vector4 a -> Vector4 a -> Bool

(>) :: Vector4 a -> Vector4 a -> Bool

(>=) :: Vector4 a -> Vector4 a -> Bool

max :: Vector4 a -> Vector4 a -> Vector4 a

min :: Vector4 a -> Vector4 a -> Vector4 a

Show a => Show (Codiagonal Vector4 a)

Instance details

Methods

showsPrec :: Int -> Codiagonal Vector4 a -> ShowS

show :: Codiagonal Vector4 a -> String

showList :: [Codiagonal Vector4 a] -> ShowS

Transposable ((->) row) Vector4 a

Instance details

Methods

transpose_impl :: ((->) row :*: Vector4) a -> (Vector4 :*: (->) row) a

(Num a, ConjugateSymmetric a) => Monoid ((Vector4 :*: Vector4) a)
see "Lawvere,Rosebrugh: Sets for mathematics", pg. 167.

Instance details

Methods

mempty :: (Vector4 :*: Vector4) a

mappend :: (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a

mconcat :: [(Vector4 :*: Vector4) a] -> (Vector4 :*: Vector4) a

(Num a, ConjugateSymmetric a) => Semigroup ((Vector4 :*: Vector4) a)

Instance details

Methods

(<>) :: (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a

sconcat :: NonEmpty ((Vector4 :*: Vector4) a) -> (Vector4 :*: Vector4) a

stimes :: Integral b => b -> (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a

(Num a, ConjugateSymmetric a) => Num ((Vector4 :*: Vector4) a)

Instance details

Methods

(+) :: (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a

(-) :: (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a

(*) :: (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a

negate :: (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a

abs :: (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a

signum :: (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a

fromInteger :: Integer -> (Vector4 :*: Vector4) a

Show (f a) => Show ((Vector4 :*: f) a)

Instance details

Methods

showsPrec :: Int -> (Vector4 :*: f) a -> ShowS

show :: (Vector4 :*: f) a -> String

showList :: [(Vector4 :*: f) a] -> ShowS

ConjugateSymmetric a => ConjugateSymmetric ((Vector4 :*: Vector4) a)
https://en.wikipedia.org/wiki/Conjugate_transpose

Instance details

Methods

conj :: (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a

(Floating a, ConjugateSymmetric a) => InnerProductSpace ((Vector4 :*: Vector4) a)
4 x 4 matrices

Instance details

Methods

(%.) :: (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a -> Scalar ((Vector4 :*: Vector4) a)

(Floating a, ConjugateSymmetric a) => NormedSpace ((Vector4 :*: Vector4) a)
4 x 4 matrices

Instance details

Methods

norm :: (Vector4 :*: Vector4) a -> Scalar ((Vector4 :*: Vector4) a)

norm_squared :: (Vector4 :*: Vector4) a -> Scalar ((Vector4 :*: Vector4) a)

Num a => VectorSpace ((Vector1 :*: Vector4) a)
1 x 4 matrices:

Instance details

Associated Types

type Scalar ((Vector1 :*: Vector4) a)

Methods

vzero :: (Vector1 :*: Vector4) a

vnegate :: (Vector1 :*: Vector4) a -> (Vector1 :*: Vector4) a

(%+) :: (Vector1 :*: Vector4) a -> (Vector1 :*: Vector4) a -> (Vector1 :*: Vector4) a

(%*) :: Scalar ((Vector1 :*: Vector4) a) -> (Vector1 :*: Vector4) a -> (Vector1 :*: Vector4) a

Num a => VectorSpace ((Vector2 :*: Vector4) a)
2 x 4 matrices

Instance details

Associated Types

type Scalar ((Vector2 :*: Vector4) a)

Methods

vzero :: (Vector2 :*: Vector4) a

vnegate :: (Vector2 :*: Vector4) a -> (Vector2 :*: Vector4) a

(%+) :: (Vector2 :*: Vector4) a -> (Vector2 :*: Vector4) a -> (Vector2 :*: Vector4) a

(%*) :: Scalar ((Vector2 :*: Vector4) a) -> (Vector2 :*: Vector4) a -> (Vector2 :*: Vector4) a

Num a => VectorSpace ((Vector3 :*: Vector4) a)
3 x 4 matrices

Instance details

Associated Types

type Scalar ((Vector3 :*: Vector4) a)

Methods

vzero :: (Vector3 :*: Vector4) a

vnegate :: (Vector3 :*: Vector4) a -> (Vector3 :*: Vector4) a

(%+) :: (Vector3 :*: Vector4) a -> (Vector3 :*: Vector4) a -> (Vector3 :*: Vector4) a

(%*) :: Scalar ((Vector3 :*: Vector4) a) -> (Vector3 :*: Vector4) a -> (Vector3 :*: Vector4) a

Num a => VectorSpace ((Vector4 :*: Vector1) a)
4 x 1 matrices:

Instance details

Associated Types

type Scalar ((Vector4 :*: Vector1) a)

Methods

vzero :: (Vector4 :*: Vector1) a

vnegate :: (Vector4 :*: Vector1) a -> (Vector4 :*: Vector1) a

(%+) :: (Vector4 :*: Vector1) a -> (Vector4 :*: Vector1) a -> (Vector4 :*: Vector1) a

(%*) :: Scalar ((Vector4 :*: Vector1) a) -> (Vector4 :*: Vector1) a -> (Vector4 :*: Vector1) a

Num a => VectorSpace ((Vector4 :*: Vector2) a)
4 x 2 matrices

Instance details

Associated Types

type Scalar ((Vector4 :*: Vector2) a)

Methods

vzero :: (Vector4 :*: Vector2) a

vnegate :: (Vector4 :*: Vector2) a -> (Vector4 :*: Vector2) a

(%+) :: (Vector4 :*: Vector2) a -> (Vector4 :*: Vector2) a -> (Vector4 :*: Vector2) a

(%*) :: Scalar ((Vector4 :*: Vector2) a) -> (Vector4 :*: Vector2) a -> (Vector4 :*: Vector2) a

Num a => VectorSpace ((Vector4 :*: Vector3) a)
4 x 3 matrices:

Instance details

Associated Types

type Scalar ((Vector4 :*: Vector3) a)

Methods

vzero :: (Vector4 :*: Vector3) a

vnegate :: (Vector4 :*: Vector3) a -> (Vector4 :*: Vector3) a

(%+) :: (Vector4 :*: Vector3) a -> (Vector4 :*: Vector3) a -> (Vector4 :*: Vector3) a

(%*) :: Scalar ((Vector4 :*: Vector3) a) -> (Vector4 :*: Vector3) a -> (Vector4 :*: Vector3) a

Num a => VectorSpace ((Vector4 :*: Vector4) a)
4 x 4 matrices

Instance details

Associated Types

type Scalar ((Vector4 :*: Vector4) a)

Methods

vzero :: (Vector4 :*: Vector4) a

vnegate :: (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a

(%+) :: (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a

(%*) :: Scalar ((Vector4 :*: Vector4) a) -> (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a

(Fractional a, ConjugateSymmetric a) => Group ((Vector4 :*: Vector4) a)

Instance details

Methods

ginvert :: (Vector4 :*: Vector4) a -> (Vector4 :*: Vector4) a

PpShow (f a) => PpShow ((Vector4 :*: f) a)

Instance details

Methods

pp :: (Vector4 :*: f) a -> Doc