Math.Matrix.Vector4

Documentation

data Vector4 s

Constructors

Vector4

Fields

tcoord4 :: !s

xcoord4 :: !s

ycoord4 :: !s

zcoord4 :: !s

cov4 :: a ~ Scalar a => Vector4 (Dual (Vector4 a))

x4 :: Num a => Vector4 a

y4 :: Num a => Vector4 a

z4 :: Num a => Vector4 a

t4 :: Num a => Vector4 a

dx4 :: (Infinitesimal a, Num (Closure a)) => Vector4 (Closure a)

dy4 :: (Infinitesimal a, Num (Closure a)) => Vector4 (Closure a)

dz4 :: (Infinitesimal a, Num (Closure a)) => Vector4 (Closure a)

dt4 :: (Infinitesimal a, Num (Closure a)) => Vector4 (Closure a)

type Matrix4 a = (Vector4 :*: Vector4) a

codiag4 :: (Vector4 :*: Vector4) a -> Codiagonal Vector4 a

matrix4 :: Vector4 a -> Codiagonal Vector4 a -> (Vector4 :*: Vector4) a

cycle4 :: Vector4 a -> Stream a

vector4 :: [a] -> Vector4 a

unzipV4 :: Vector4 (a, b) -> (Vector4 a, Vector4 b)

constant4 :: a -> Vector4 a

sum_coordinates4 :: Num a => Vector4 a -> a

left_multiply4 :: Num a => Vector4 a -> Matrix4 a -> Vector4 a

right_multiply4 :: Num a => Matrix4 a -> Vector4 a -> Vector4 a

matrix_multiply4 :: Num a => Matrix4 a -> Matrix4 a -> Matrix4 a

dot4 :: Num a => Vector4 a -> Vector4 a -> a

versor :: (Floating a, ConjugateSymmetric a) => Vector4 a -> Vector4 a

index4 :: Int -> Vector4 a -> a

index4_dual :: a ~ Scalar a => Int -> Dual (Vector4 a)

vector4_to_vec4 :: Vector4 a -> FourD -> a

diagonal4 :: (Vector4 :*: Vector4) a -> Vector4 a

diagonal_projections4 :: a ~ Scalar a => Vector4 (Dual (Vector4 a))

dt_4 :: Num s => s -> Vector4 s -> Vector4 s

dx_4 :: Num a => a -> Vector4 a -> Vector4 a

dy_4 :: Num a => a -> Vector4 a -> Vector4 a

dz_4 :: Num a => a -> Vector4 a -> Vector4 a

partial_derivate4x :: (Infinitesimal a, Closed a) => (Vector4 a -> a) -> Vector4 a -> a

partial_derivate4y :: (Infinitesimal a, Closed a) => (Vector4 a -> a) -> Vector4 a -> a

partial_derivate4z :: (Infinitesimal a, Closed a) => (Vector4 a -> a) -> Vector4 a -> a

partial_derivate4t :: (Infinitesimal a, Closed a) => (Vector4 a -> a) -> Vector4 a -> a

derivate4t_squared :: (Closed a, Infinitesimal a, a ~ Scalar a) => Dual (Vector4 a) -> Dual (Vector4 a)

derivate4t :: (Closed a, Infinitesimal a, a ~ Scalar a) => Dual (Vector4 a) -> Dual (Vector4 a)

derivate4x :: (Closed a, Infinitesimal a, a ~ Scalar a) => Dual (Vector4 a) -> Dual (Vector4 a)

derivate4y :: (Closed a, Infinitesimal a, a ~ Scalar a) => Dual (Vector4 a) -> Dual (Vector4 a)

derivate4z :: (Closed a, Infinitesimal a, a ~ Scalar a) => Dual (Vector4 a) -> Dual (Vector4 a)

del4 :: (Closed a, Infinitesimal a, a ~ Scalar a) => Vector4 (Dual (Vector4 a) -> Dual (Vector4 a))

del4_ :: (Infinitesimal a, Closed a) => Vector4 ((Vector4 a -> a) -> Vector4 a -> a)

hessian4 :: (Infinitesimal v, Closed v, v ~ Scalar v) => Dual (Vector4 v) -> (Vector4 :*: Vector4) (Dual (Vector4 v))

grad4 :: (Infinitesimal a, Closed a) => Dual (Vector4 a) -> LinearMap (Vector4 a) (Vector4 a)

curl4 :: (Infinitesimal a, Closed a) => Vector4 (Dual (Vector4 a)) -> Vector4 a -> (Vector4 :*: Vector4) a

divergence4 :: (Infinitesimal a, Closed a, a ~ Scalar a) => LinearMap (Vector4 a) (Vector4 a) -> Dual (Vector4 a)

laplace4 :: (Infinitesimal a, Closed a, a ~ Scalar a) => Dual (Vector4 a) -> Dual (Vector4 a)

vector_laplace4 :: (VectorDerivative v, Scalar (Scalar v) ~ Scalar v) => LinearMap v (Vector4 (Scalar v)) -> LinearMap v (Vector4 (Scalar v))

transpose4 :: Matrix4 a -> Matrix4 a

identity4 :: Num a => Matrix4 a

diagonal4x :: Matrix4 a -> Vector4 a

zero_codiagonal4 :: Num a => Codiagonal Vector4 a

diagonal_matrix4 :: Num a => Vector4 a -> (Vector4 :*: Vector4) a

mat4 :: Vector4 a -> Codiagonal Vector4 a -> (Vector4 :*: Vector4) a

trace4 :: Num a => Matrix4 a -> a

removet4 :: Vector4 a -> Vector3 a

removex4 :: Vector4 a -> Vector3 a

removey4 :: Vector4 a -> Vector3 a

removez4 :: Vector4 a -> Vector3 a

removes4 :: Vector4 (Vector4 a -> Vector3 a)

remove_index4 :: Matrix4 a -> Matrix4 (Matrix3 a)

cofactor4 :: Num a => Matrix4 a -> Matrix4 a

adjucate4 :: Num a => Matrix4 a -> Matrix4 a

signs4 :: Integral a => Matrix4 a

vector_indices4 :: Integral a => Vector4 a

matrix_indices4 :: Integral a => (Vector4 :*: Vector4) (a, a)

inverse4 :: Fractional a => Matrix4 a -> Matrix4 a

https://en.wikipedia.org/wiki/Invertible_matrix

cross4 :: Num a => Vector4 a -> Vector4 a -> Vector4 a -> Vector4 a

Generalization of cross product to four dimensions This is computed using formal determinant representation of cross product expanded to four dimensions https://en.wikipedia.org/wiki/Cross_product This computes vector that is linearly independent of all three vectors given in four dimensional space.

cross_product4 :: Num a => (Vector3 :*: Vector4) a -> Vector4 a

determinant4 :: Num a => Matrix4 a -> a

Orphan instances

SplittableVector Vector1 Vector3

Instance details

Methods

vsplit :: (Vector1 :+: Vector3) a -> (Vector1 a, Vector3 a)

SplittableVector Vector2 Vector2

Instance details

Methods

vsplit :: (Vector2 :+: Vector2) a -> (Vector2 a, Vector2 a)

SplittableVector Vector3 Vector1

Instance details

Methods

vsplit :: (Vector3 :+: Vector1) a -> (Vector3 a, Vector1 a)

AppendableVector Vector1 Vector3

Instance details

Associated Types

type Vector1 :+: Vector3 :: Type -> Type

Methods

(|>) :: Vector1 a -> Vector3 a -> (Vector1 :+: Vector3) a

AppendableVector Vector2 Vector2

Instance details

Associated Types

type Vector2 :+: Vector2 :: Type -> Type

Methods

(|>) :: Vector2 a -> Vector2 a -> (Vector2 :+: Vector2) a

AppendableVector Vector3 Vector1

Instance details

Associated Types

type Vector3 :+: Vector1 :: Type -> Type

Methods

(|>) :: Vector3 a -> Vector1 a -> (Vector3 :+: Vector1) a