Math.Number.DimensionalAnalysis

Description

This module provides dimensional analysis according to SI system of units. For reference have used https://en.wikipedia.org/wiki/Dimensional_analysis and https://en.wikipedia.org/wiki/International_System_of_Units.

This module supports run-time checked quantities.

In reality this should be according to the International system of units, but I have not used the official standard documents when writing this code. However it is recognized that any major deviation from the standard would be considered a bug in this code.

For official references, see e.g. "ISO80000-1:2009:Quantities and Units" and "NIST Special Publication 330, 2008: The International system of units".

For c++ approach to dimensional analysis, see "Barton&Nackman: Scientific and Engineering C++".

Example uses:

show (3 %* meter) = "3 m"

3 %* meter + 4 %* kilogram == throw (InvalidDimensionsException meter_dimension kilogram_dimension "...")

convert (3 %* meter) (milli meter) == return 3000.0

convert (3 %* meter) kilogram == fail "..."

(3 %* meter) =/ (milli meter) == 3000.0

(3 %* meter) =/ kilogram == error "..."

convert lightyear (kilo meter) == return 9.4607304725808e12

Documentation

data Quantity r

Constructors

As

Fields

value_amount :: r

value_dimension :: Dimension

type Prefix r = Quantity r -> Quantity r

hasDimension :: Quantity r -> Dimension -> Bool

data DimensionException

Constructors

InvalidDimensionsException Dimension Dimension String

mapQuantity :: (a -> b) -> (Dimension -> Dimension) -> Quantity a -> Quantity b

mapQuantity2 :: (a -> b -> c) -> (Dimension -> Dimension -> Dimension) -> Quantity a -> Quantity b -> Quantity c

complexQuantity :: Show r => Complex (Quantity r) -> Quantity (Complex r)

quantityComplex :: Quantity (Complex r) -> Complex (Quantity r)

ratioQuantity :: (Integral r, Show r) => Ratio (Quantity r) -> Quantity (Ratio r)

class VectorSpace u => Unit u where

the Unit class should be defined by any newtype based types that should interact well with the dimensional analysis mechanism.

The fromAmount method must check that compile-time information about dimensions is sufficient to determine dimension of the given input e.g. (fromAmount 3 :: Mass) is ok, but (fromAmount 3 :: Quantity Double) is not.

Minimal complete definition

amount, fromAmount, fromQuantity, unitOf, dimension

Methods

amount :: u -> Scalar u

fromAmount :: Scalar u -> u

fromQuantity :: (Alternative m, Monad m) => Quantity (Scalar u) -> m u

unitOf :: u -> String

dimension :: u -> Dimension

zeroAmount :: (Scalar u -> u) -> Scalar u

conversionFactor :: (Scalar u -> u) -> Scalar u

fromQuantityDef :: (Monad m, Alternative m, Show a) => Dimension -> (a -> b) -> Quantity a -> m b

data Level r

https://en.wikipedia.org/wiki/Level_(logarithmic_quantity)

Constructors

Level

Fields

reference_value :: Quantity r

base_of_logarithm :: r

level :: (Show a, Floating a, Real a) => Quantity a -> Level a -> Quantity a

bit_level :: Num a => Level a

bel_level :: Num a => Level a

ratioToDecibel :: (Fractional a, Show a, Floating a, Real a) => Quantity a -> Quantity a

decibelToRatio :: (Floating a, Real a, Show a) => Quantity a -> Quantity a

bit :: (Floating a, Real a, Show a) => Quantity a

WARNING: native representation is in number of bits. Use code such as the following for conversion: gibi byte `convert` kilo bit == return 8589934.592

byte :: (Floating a, Real a, Show a) => Quantity a

Byte as the number of bits. WARNING: native representation is in number of bits. use code such as the following for conversion: gibi byte `convert` kilo bit == return 8589934.592

fromAlternatives :: (Floating a, Real a, Show a) => a -> Quantity a

toAlternatives :: (Floating a, Ord a, ShowPrecision a, Monad m) => Quantity a -> m a

Note that the number of alternatives grows quickly.

(@@) :: r -> Dimension -> Quantity r

convertTo :: (Monad m, Floating a, ShowPrecision a, Ord a, VectorSpace a) => Quantity a -> (String, Quantity a) -> m String

(=/) :: (Fractional (Scalar a), Unit a, Show a) => a -> a -> Scalar a

convert :: (Scalar u ~ Scalar v, Unit v, Unit u, Show v, Show u, Monad m, Fractional (Scalar u)) => v -> u -> m (Scalar u)

conversions between units. Dimensions have to match.

data Dimension

Constructors

Dimension

Fields

length_power :: Rational

weight_power :: Rational

time_power :: Rational

current_power :: Rational

temperature_power :: Rational

luminosity_power :: Rational

substance_power :: Rational

dimension_basis :: [Dimension]

isDimensionless :: Dimension -> Bool

mapDimension :: (Rational -> Rational) -> Dimension -> Dimension

zipWithDimension :: (Rational -> Rational -> Rational) -> Dimension -> Dimension -> Dimension

(=*) :: VectorSpace v => Scalar v -> v -> v

deca :: (VectorSpace u, Num (Scalar u)) => u -> u

hecto :: (VectorSpace u, Num (Scalar u)) => u -> u

kilo :: (VectorSpace u, Num (Scalar u)) => u -> u

mega :: (VectorSpace u, Num (Scalar u)) => u -> u

giga :: (VectorSpace u, Num (Scalar u)) => u -> u

tera :: (VectorSpace u, Num (Scalar u)) => u -> u

peta :: (VectorSpace u, Num (Scalar u)) => u -> u

exa :: (VectorSpace u, Num (Scalar u)) => u -> u

zetta :: (VectorSpace u, Num (Scalar u)) => u -> u

yotta :: (VectorSpace u, Num (Scalar u)) => u -> u

prefix_value :: Num a => Prefix a -> a

decimal_prefixes :: Num a => [(String, a)]

floating_prefixes :: Floating a => [(String, a)]

binary_prefixes :: Num a => [(String, a)]

kibi :: (VectorSpace u, Num (Scalar u)) => u -> u

https://en.wikipedia.org/wiki/Binary_prefix#kibi

mebi :: (VectorSpace u, Num (Scalar u)) => u -> u

https://en.wikipedia.org/wiki/Binary_prefix#kibi

gibi :: (VectorSpace u, Num (Scalar u)) => u -> u

https://en.wikipedia.org/wiki/Binary_prefix#kibi

tebi :: (VectorSpace u, Num (Scalar u)) => u -> u

https://en.wikipedia.org/wiki/Binary_prefix#kibi

pebi :: (VectorSpace u, Num (Scalar u)) => u -> u

https://en.wikipedia.org/wiki/Binary_prefix#kibi

exbi :: (VectorSpace u, Num (Scalar u)) => u -> u

https://en.wikipedia.org/wiki/Binary_prefix#kibi

zebi :: (VectorSpace u, Num (Scalar u)) => u -> u

https://en.wikipedia.org/wiki/Binary_prefix#kibi

yobi :: (VectorSpace u, Num (Scalar u)) => u -> u

https://en.wikipedia.org/wiki/Binary_prefix#kibi

deci :: (VectorSpace u, Floating (Scalar u)) => u -> u

centi :: (VectorSpace u, Floating (Scalar u)) => u -> u

milli :: (VectorSpace u, Floating (Scalar u)) => u -> u

micro :: (VectorSpace u, Floating (Scalar u)) => u -> u

nano :: (VectorSpace u, Floating (Scalar u)) => u -> u

pico :: (VectorSpace u, Floating (Scalar u)) => u -> u

femto :: (VectorSpace u, Floating (Scalar u)) => u -> u

atto :: (VectorSpace u, Floating (Scalar u)) => u -> u

zepto :: (VectorSpace u, Floating (Scalar u)) => u -> u

yocto :: (VectorSpace u, Floating (Scalar u)) => u -> u

(=+=) :: VectorSpace v => v -> v -> v

(=-=) :: VectorSpace v => v -> v -> v

require_dimensionless :: Show a => String -> (a -> a) -> Quantity a -> Quantity a

readprefix :: ReadPrec (Quantity Double -> Quantity Double)

readunit :: ReadPrec (String, Quantity Double)

invalidDimensions :: (Show b, Show c) => String -> Dimension -> Dimension -> b -> c -> a

invalidDimensionsM :: (Monad m, Show b, Show c) => String -> Dimension -> Dimension -> b -> c -> m a

plusQ :: (Monad m, Num r, Show r) => Quantity r -> Quantity r -> m (Quantity r)

minusQ :: (Monad m, Num r, Show r) => Quantity r -> Quantity r -> m (Quantity r)

show_dimension :: Dimension -> String

full_dimension_table :: [(Dimension, String)]

dimension_table :: [(Dimension, String)]

dimension_product_reader :: ReadPrec [Dimension]

dimension_divider_reader :: ReadPrec [Dimension]

dimension_reader :: ReadPrec Dimension

dimensionless :: Dimension

kelvin_dimension :: Dimension

farad_dimension :: Dimension

ohm_dimension :: Dimension

sievert_dimension :: Dimension

volt_dimension :: Dimension

pascal_dimension :: Dimension

henry_dimension :: Dimension

weber_dimension :: Dimension

joule_dimension :: Dimension

newton_dimension :: Dimension

tesla_dimension :: Dimension

meter_dimension :: Dimension

kilogram_dimension :: Dimension

second_dimension :: Dimension

ampere_dimension :: Dimension

candela_dimension :: Dimension

mol_dimension :: Dimension

radian_dimension :: Dimension

radian_dimension is basically same as dimensionless, using separate name anyway to allow clients to distinguish. No checking for this distinction is implemented.

steradian_dimension :: Dimension

steradian_dimension is basically same as dimensionless, redefining anyway to allow clients to distinguish steradians. No checking for this distinction is implemented.

steradian :: Quantity Double

becquerel_dimension :: Dimension

becquerel :: Quantity Double

gray_dimension :: Dimension

gray :: Quantity Double

lux_dimension :: Dimension

lux :: Quantity Double

lumen_dimension :: Dimension

lumen :: Quantity Double

kelvin_quantity :: Quantity Integer

siemens :: Quantity Double

second :: Floating a => Quantity a

meter :: Floating a => Quantity a

kilogram :: Floating a => Quantity a

gram :: Quantity Double

mole :: Floating a => Quantity a

candela :: Floating a => Quantity a

kelvin :: Floating a => Quantity a

ampere :: Floating a => Quantity a

coulomb :: Floating a => Quantity a

tesla :: Floating a => Quantity a

henry :: Floating a => Quantity a

newton :: Floating a => Quantity a

pascal :: Floating a => Quantity a

joule :: Floating a => Quantity a

weber :: Floating a => Quantity a

volt :: Floating a => Quantity a

sievert :: Floating a => Quantity a

ohm :: Floating a => Quantity a

farad :: Floating a => Quantity a

watt :: Floating a => Quantity a

minute :: Floating a => Quantity a

hour :: Floating a => Quantity a

day :: Floating a => Quantity a

week :: Floating a => Quantity a

year :: Floating a => Quantity a

degree :: Floating a => Quantity a

one degree angle

radian :: Floating a => Quantity a

katal_dimension :: Dimension

katal :: Quantity Double

fromCelsius :: (Show a, Fractional a) => a -> Quantity a

conversion from degrees celcius. WARNING: produces degrees in kelvin, so computations intended in degrees celcius should be done before conversion!

fromFarenheit :: (Fractional a, Show a) => a -> Quantity a

toFarenheit :: (Monad m, Fractional a, VectorSpace a) => Quantity a -> m a

toCelsius :: (Fractional a, Show a, VectorSpace a) => Quantity a -> a

fromRadiansAngle :: Floating a => a -> Quantity a

fromDegreesAngle :: Floating a => a -> Quantity a

conversion from angle. WARNING: produces degrees in radian, so computations intended in degrees should be done before conversion!

hertz_dimension :: Dimension

hertz :: Quantity Double

squaremeter_dimension :: Dimension

squaremeter :: Quantity Double

cubicmeter_dimension :: Dimension

cubicmeter :: Quantity Double

watt_dimension :: Dimension

coulomb_dimension :: Dimension

siemens_dimension :: Dimension

planck_length :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

planck_mass :: Floating a => Quantity a

planck_time :: Floating a => Quantity a

planck_charge :: Floating a => Quantity a

planck_temperature :: Floating a => Quantity a

boltzmann_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

gravitational_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Gravitational_constant

standard_acceleration_of_gravity :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

standard_atmosphere :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

speed_of_light :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Speed_of_light

planck_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Planck_constant

reduced_planck_constant :: Floating a => Quantity a

electronvolt :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Electronvolt

magnetic_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

electric_constant :: (Show a, Floating a) => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

characteristic_impedance_of_vacuum :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

coulombs_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

elementary_charge :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

bohr_magneton :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

conductance_quantum :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

inverse_conductance_quantum :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

josephson_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

magnetic_flux_quantum :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

nuclear_magneton :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

von_klitzing_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

bohr_radius :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

classical_electron_radius :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

electron_mass :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

fermi_coupling_constant :: (Show a, Floating a) => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

fine_structure_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

hartree_energy :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

proton_mass :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

quantum_of_circulation :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

rydberg_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

thomson_cross_section :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

weak_mixing_angle :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

avogadro_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Avogadro_constant

atomic_mass_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

faraday_constant :: Quantity Double

first_radiation_constant :: Quantity Double

loschmidt_constant :: Quantity Double

gas_constant :: Quantity Double

molar_planck_constant :: Quantity Double

second_radiation_constant :: Quantity Double

stefan_boltzmann_constant :: Quantity Double

efimov_factor :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

lightyear :: Num a => Quantity a

https://en.wikipedia.org/wiki/Light-year

astronomicalUnit :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Astronomical_unit

parsec :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Parsec

liter :: Quantity Double

https://en.wikipedia.org/wiki/Litre

litre :: Quantity Double

https://en.wikipedia.org/wiki/Litre

siUnits :: [(String, Quantity Double)]

https://en.wikipedia.org/wiki/Parsec https://en.wikipedia.org/wiki/International_System_of_Units

siPrefixes :: Floating a => [(String, Prefix a)]

https://en.wikipedia.org/wiki/Metric_prefix

siZeros :: [(String, Quantity Double)]

this contains a list of quantities where basis of measurement is not zero in the dimension because the corresponding SI unit has different zero