Math.Number.DimensionalAnalysis

Description

This module provides dimensional analysis according to SI system of units. For reference have used https://en.wikipedia.org/wiki/Dimensional_analysis and https://en.wikipedia.org/wiki/International_System_of_Units.

This module supports run-time checked quantities.

In reality this should be according to the International system of units, but I have not used the official standard documents when writing this code. However it is recognized that any major deviation from the standard would be considered a bug in this code. And definitely the International system of units is basis for this code.

For official references, see e.g. "ISO80000-1:2009:Quantities and Units" and "NIST Special Publication 330, 2008: The International system of units".

For c++ approach to dimensional analysis, see "Barton&Nackman: Scientific and Engineering C++".

Example uses:

show (3 %* meter) = "3 m"

3 %* meter + 4 %* kilogram == throw (InvalidDimensionsException meter_dimension kilogram_dimension "...")

convert (3 %* meter) (milli meter) == return 3000.0

convert (3 %* meter) kilogram == fail "..."

(3 %* meter) =/ (milli meter) == 3000.0

(3 %* meter) =/ kilogram == error "..."

convert lightyear (kilo meter) == return 9.4607304725808e12

3 logarithmic dBV == 1000.0 %* volt

Documentation

data Quantity r

Constructors

As

Fields

value_amount :: r

value_dimension :: !Dimension

dimensionless_amount :: Show r => Quantity r -> r

checked projection

equal_up_to :: (Floating a, Show a, Ord a) => a -> Quantity a -> Quantity a -> Bool

data Dimension

Constructors

Dimension

Fields

length_power :: !Rational

weight_power :: !Rational

time_power :: !Rational

current_power :: !Rational

temperature_power :: !Rational

luminosity_power :: !Rational

substance_power :: !Rational

dimension_multiples :: Dimension -> Stream Dimension

quantity_powers :: (Num a, Show a) => Quantity a -> Stream (Quantity a)

isFractionalDimensional :: Dimension -> Bool

isFractionalDimensional checks if a dimension is fractional dimensional

type Prefix r = Quantity r -> Quantity r

hasDimension :: Quantity r -> Dimension -> Bool

data DimensionException

Constructors

InvalidDimensionsException Dimension Dimension String

mapQuantity :: (a -> b) -> (Dimension -> Dimension) -> Quantity a -> Quantity b

mapQuantity2 :: (a -> b -> c) -> (Dimension -> Dimension -> Dimension) -> Quantity a -> Quantity b -> Quantity c

complexQuantity :: Show r => Complex (Quantity r) -> Quantity (Complex r)

quantityComplex :: Quantity (Complex r) -> Complex (Quantity r)

ratioQuantity :: (Integral r, Show r) => Ratio (Quantity r) -> Quantity (Ratio r)

class VectorSpace u => Unit u where

the Unit class should be defined by any newtype based types that should interact well with the dimensional analysis mechanism.

The fromAmount method must check that compile-time information about dimensions is sufficient to determine dimension of the given input e.g. (fromAmount 3 :: Mass) is ok, but (fromAmount 3 :: Quantity Double) is not.

Minimal complete definition

amount, fromAmount, fromQuantity, unitOf, dimension

Methods

amount :: u -> Scalar u

fromAmount :: Scalar u -> u

fromQuantity :: (Alternative m, Monad m) => Quantity (Scalar u) -> m u

unitOf :: u -> String

dimension :: u -> Dimension

zeroAmount :: (Scalar u -> u) -> Scalar u

conversionFactor :: (Scalar u -> u) -> Scalar u

fromQuantityDef :: (Monad m, Alternative m, Show a) => Dimension -> (a -> b) -> Quantity a -> m b

data Level r

<https://en.wikipedia.org/wiki/Level_(logarithmic_quantity) Level represents a reference to which computations involving logarithmic scales are compared to.

Constructors

Level

Fields

reference_value :: Quantity r

base_of_logarithm :: r

scale :: (Show a, Floating a, Real a) => Quantity a -> Level a -> Quantity a

logarithmic :: Floating a => a -> Level a -> Quantity a

hyperbolic :: RealFloat a => Complex a -> Level (Complex a) -> Quantity (Complex a)

radian_scale :: RealFloat a => Level (Complex a)

This is a way to convert radians to complex numbers e.g. (pi/2) logarithmic radian_scale == 0 :+ 1.

bit_scale :: Num a => Level a

bel_scale :: Num a => Level a

dBV :: Floating a => Level a

logarithmic voltage with respect to base 10 relative to 1V. https://en.wikipedia.org/wiki/Decibel

dBu :: Floating a => Level a

https://en.wikipedia.org/wiki/Decibel logarithmic voltage in base 10 relative to \(\sqrt{0.6}V\).

dB_SPL :: Floating a => Level a

https://en.wikipedia.org/wiki/Decibel logarithmic pressure in base 10 compared to 20 micro pascals.

dB_SIL :: (Floating a, Show a) => Level a

https://en.wikipedia.org/wiki/Decibel

dB_SWL :: Floating a => Level a

length_scale :: Floating a => Level a

for logarithmic lengths in base 10.

frequency_scale :: Floating a => Level a

https://en.wikipedia.org/wiki/Octave logarithmic frequency in base 2 relative to 16.352 Hz. a.k.a. in scientific designation. middle-C is fourth octave, e.g. 4 `logarithmic` octave == 261.626 %* hertz

octave :: Floating a => Level a

https://en.wikipedia.org/wiki/Octave logarithmic frequency in base 2 relative to 16.352 Hz. a.k.a. in scientific designation. middle-C is fourth octave, e.g. 4 `logarithmic` octave == 261.626 %* hertz octave is just a different name for frequency_scale.

moment_magnitude_scale :: (Show a, Floating a) => Level a

https://en.wikipedia.org/wiki/Moment_magnitude_scale

entropy_scale :: Floating a => Level a

https://en.wikipedia.org/wiki/Entropy

hartley_scale :: Num a => Level a

ban_scale :: Num a => Level a

time_scale :: Floating a => Level a

weight_scale :: Floating a => Level a

nines :: Double -> Double

compute how many nines a probability has. nines 0.99999 == 5.0. https://en.wikipedia.org/wiki/9#Probability

fromNines :: Double -> Double

compute probability when given number of nines. https://en.wikipedia.org/wiki/9#Probability

note_pitch :: Int -> Int -> Quantity Double

note frequency from octave and index of note within octave. note_pitch 4 0 is middle-C.

ratioToDecibel :: (Fractional a, Show a, Floating a, Real a) => Quantity a -> Level a -> Quantity a

bit :: (Floating a, Real a, Show a) => Quantity a

WARNING: native representation is in number of bits. Use code such as the following for conversion: gibi byte `convert` kilo bit == return 8589934.592

byte :: (Floating a, Real a, Show a) => Quantity a

Byte as the number of bits. WARNING: native representation is in number of bits. use code such as the following for conversion: gibi byte `convert` kilo bit == return 8589934.592

fromAlternatives :: (Floating a, Real a, Show a) => a -> Quantity a

toAlternatives :: (Floating a, Ord a, ShowPrecision a, Monad m) => Quantity a -> m a

Note that the number of alternatives grows quickly.

(@@) :: r -> Dimension -> Quantity r

convertTo :: (Monad m, Floating a, ShowPrecision a, Ord a, VectorSpace a) => Quantity a -> (String, Quantity a) -> m String

(=/) :: (Fractional (Scalar a), Unit a, Show a) => a -> a -> Scalar a

convert :: (Scalar u ~ Scalar v, Unit v, Unit u, Show v, Show u, Monad m, Fractional (Scalar u)) => v -> u -> m (Scalar u)

conversions between units. Dimensions have to match.

dimension_basis :: [Dimension]

isDimensionless :: Dimension -> Bool

mapDimension :: (Rational -> Rational) -> Dimension -> Dimension

zipWithDimension :: (Rational -> Rational -> Rational) -> Dimension -> Dimension -> Dimension

(=*) :: VectorSpace v => Scalar v -> v -> v

deca :: (VectorSpace u, Num (Scalar u)) => u -> u

hecto :: (VectorSpace u, Num (Scalar u)) => u -> u

kilo :: (VectorSpace u, Num (Scalar u)) => u -> u

mega :: (VectorSpace u, Num (Scalar u)) => u -> u

giga :: (VectorSpace u, Num (Scalar u)) => u -> u

tera :: (VectorSpace u, Num (Scalar u)) => u -> u

peta :: (VectorSpace u, Num (Scalar u)) => u -> u

exa :: (VectorSpace u, Num (Scalar u)) => u -> u

zetta :: (VectorSpace u, Num (Scalar u)) => u -> u

yotta :: (VectorSpace u, Num (Scalar u)) => u -> u

prefix_value :: Num a => Prefix a -> a

decimal_prefixes :: Num a => [(String, a)]

floating_prefixes :: Floating a => [(String, a)]

binary_prefixes :: Num a => [(String, a)]

kibi :: (VectorSpace u, Num (Scalar u)) => u -> u

https://en.wikipedia.org/wiki/Binary_prefix#kibi

mebi :: (VectorSpace u, Num (Scalar u)) => u -> u

https://en.wikipedia.org/wiki/Binary_prefix#kibi

gibi :: (VectorSpace u, Num (Scalar u)) => u -> u

https://en.wikipedia.org/wiki/Binary_prefix#kibi

tebi :: (VectorSpace u, Num (Scalar u)) => u -> u

https://en.wikipedia.org/wiki/Binary_prefix#kibi

pebi :: (VectorSpace u, Num (Scalar u)) => u -> u

https://en.wikipedia.org/wiki/Binary_prefix#kibi

exbi :: (VectorSpace u, Num (Scalar u)) => u -> u

https://en.wikipedia.org/wiki/Binary_prefix#kibi

zebi :: (VectorSpace u, Num (Scalar u)) => u -> u

https://en.wikipedia.org/wiki/Binary_prefix#kibi

yobi :: (VectorSpace u, Num (Scalar u)) => u -> u

https://en.wikipedia.org/wiki/Binary_prefix#kibi

deci :: (VectorSpace u, Floating (Scalar u)) => u -> u

centi :: (VectorSpace u, Floating (Scalar u)) => u -> u

milli :: (VectorSpace u, Floating (Scalar u)) => u -> u

micro :: (VectorSpace u, Floating (Scalar u)) => u -> u

nano :: (VectorSpace u, Floating (Scalar u)) => u -> u

pico :: (VectorSpace u, Floating (Scalar u)) => u -> u

femto :: (VectorSpace u, Floating (Scalar u)) => u -> u

atto :: (VectorSpace u, Floating (Scalar u)) => u -> u

zepto :: (VectorSpace u, Floating (Scalar u)) => u -> u

yocto :: (VectorSpace u, Floating (Scalar u)) => u -> u

(=+=) :: VectorSpace v => v -> v -> v

(=-=) :: VectorSpace v => v -> v -> v

require_dimensionless :: Show a => String -> (a -> a) -> Quantity a -> Quantity a

readprefix :: ReadPrec (Quantity Double -> Quantity Double)

readunit :: ReadPrec (String, Quantity Double)

invalidDimensions :: (Show b, Show c) => String -> Dimension -> Dimension -> b -> c -> a

invalidDimensionsM :: (Monad m, Show b, Show c) => String -> Dimension -> Dimension -> b -> c -> m a

plusQ :: (Monad m, Num r, Show r) => Quantity r -> Quantity r -> m (Quantity r)

minusQ :: (Monad m, Num r, Show r) => Quantity r -> Quantity r -> m (Quantity r)

show_dimension :: Dimension -> String

full_dimension_table :: [(Dimension, String)]

dimension_table :: [(Dimension, String)]

dimension_product_reader :: ReadPrec [Dimension]

dimension_divider_reader :: ReadPrec [Dimension]

dimension_reader :: ReadPrec Dimension

dimensionless :: Dimension

pp_dimension :: Dimension -> Doc

pp_dimension prints dimension in a simple way without interpreting derived dimensions for example, pp_dimension newton_dimension == pp "kg m s^-2"

kelvin_dimension :: Dimension

farad_dimension :: Dimension

ohm_dimension :: Dimension

sievert_dimension :: Dimension

volt_dimension :: Dimension

pascal_dimension :: Dimension

henry_dimension :: Dimension

weber_dimension :: Dimension

joule_dimension :: Dimension

newton_dimension :: Dimension

tesla_dimension :: Dimension

meter_dimension :: Dimension

kilogram_dimension :: Dimension

second_dimension :: Dimension

ampere_dimension :: Dimension

candela_dimension :: Dimension

mol_dimension :: Dimension

radian_dimension :: Dimension

radian_dimension is basically same as dimensionless, using separate name anyway to allow clients to distinguish. No checking for this distinction is implemented.

steradian_dimension :: Dimension

steradian_dimension is basically same as dimensionless, redefining anyway to allow clients to distinguish steradians. No checking for this distinction is implemented.

steradian :: Floating a => Quantity a

becquerel_dimension :: Dimension

becquerel :: Floating a => Quantity a

gray_dimension :: Dimension

gray :: Floating a => Quantity a

lux_dimension :: Dimension

lux :: Floating a => Quantity a

lumen_dimension :: Dimension

lumen :: Floating a => Quantity a

kelvin_quantity :: Floating a => Quantity a

siemens :: Floating a => Quantity a

second :: Floating a => Quantity a

meter :: Floating a => Quantity a

kilogram :: Floating a => Quantity a

gram :: Floating a => Quantity a

mole :: Floating a => Quantity a

candela :: Floating a => Quantity a

kelvin :: Floating a => Quantity a

ampere :: Floating a => Quantity a

coulomb :: Floating a => Quantity a

tesla :: Floating a => Quantity a

henry :: Floating a => Quantity a

newton :: Floating a => Quantity a

pascal :: Floating a => Quantity a

joule :: Floating a => Quantity a

weber :: Floating a => Quantity a

volt :: Floating a => Quantity a

sievert :: Floating a => Quantity a

ohm :: Floating a => Quantity a

farad :: Floating a => Quantity a

watt :: Floating a => Quantity a

minute :: Floating a => Quantity a

hour :: Floating a => Quantity a

day :: Floating a => Quantity a

week :: Floating a => Quantity a

year :: Floating a => Quantity a

degree :: Floating a => Quantity a

one degree angle

radian :: Floating a => Quantity a

katal_dimension :: Dimension

katal :: Floating a => Quantity a

fromCelsius :: (Show a, Fractional a) => a -> Quantity a

conversion from degrees celcius. WARNING: produces degrees in kelvin, so computations intended in degrees celcius should be done before conversion!

fromFahrenheit :: (Fractional a, Show a) => a -> Quantity a

toFahrenheit :: (Monad m, Fractional a, VectorSpace a) => Quantity a -> m a

toCelsius :: (Fractional a, Show a, VectorSpace a) => Quantity a -> a

fromRadiansAngle :: Floating a => a -> Quantity a

fromDegreesAngle :: Floating a => a -> Quantity a

conversion from angle. WARNING: produces degrees in radian, so computations intended in degrees should be done before conversion!

hertz_dimension :: Dimension

hertz :: Floating a => Quantity a

squaremeter_dimension :: Dimension

squaremeter :: Floating a => Quantity a

cubicmeter_dimension :: Dimension

cubicmeter :: Floating a => Quantity a

watt_dimension :: Dimension

coulomb_dimension :: Dimension

siemens_dimension :: Dimension

planck_length :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

planck_mass :: Floating a => Quantity a

planck_time :: Floating a => Quantity a

planck_charge :: Floating a => Quantity a

planck_temperature :: Floating a => Quantity a

boltzmann_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant https://en.wikipedia.org/wiki/2019_redefinition_of_SI_base_units Warning: don't confuse with stefan_boltzmann_constant. Notice value is exact.

gravitational_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Gravitational_constant

standard_acceleration_of_gravity :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

standard_atmosphere :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

speed_of_light :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Speed_of_light Note value is exact.

planck_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Planck_constant https://en.wikipedia.org/wiki/2019_redefinition_of_SI_base_units Note value is exact.

reduced_planck_constant :: Floating a => Quantity a

electronvolt :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Electronvolt

magnetic_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

electric_constant :: (Show a, Floating a) => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

characteristic_impedance_of_vacuum :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

coulombs_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

elementary_charge :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant https://en.wikipedia.org/wiki/2019_redefinition_of_SI_base_units Notice value is exact.

bohr_magneton :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

conductance_quantum :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

inverse_conductance_quantum :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

josephson_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

magnetic_flux_quantum :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

nuclear_magneton :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

von_klitzing_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

bohr_radius :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

classical_electron_radius :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

electron_mass :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

fermi_coupling_constant :: (Show a, Floating a) => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

fine_structure_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

hartree_energy :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

proton_mass :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

quantum_of_circulation :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

rydberg_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

thomson_cross_section :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

ground_state_cesium_hyperfine_transition_frequency :: Floating a => Quantity a

weak_mixing_angle :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

luminous_efficacy_of_monochromatic_540e12_hertz_radiation :: Floating a => Quantity a

https://en.wikipedia.org/wiki/2019_redefinition_of_SI_base_units value is exact.

avogadro_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Avogadro_constant https://en.wikipedia.org/wiki/2019_redefinition_of_SI_base_units Notice value is exact.

atomic_mass_constant :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

faraday_constant :: (Floating a, Show a) => Quantity a

first_radiation_constant :: (Floating a, Show a) => Quantity a

loschmidt_constant :: Floating a => Quantity a

gas_constant :: (Floating a, Show a) => Quantity a

molar_planck_constant :: (Floating a, Show a) => Quantity a

second_radiation_constant :: (Floating a, Show a) => Quantity a

stefan_boltzmann_constant :: (Floating a, Show a) => Quantity a

Warning: don't confuse with boltzmann_constant.

efimov_factor :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Physical_constant

lightyear :: Num a => Quantity a

https://en.wikipedia.org/wiki/Light-year

astronomicalUnit :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Astronomical_unit

parsec :: Floating a => Quantity a

https://en.wikipedia.org/wiki/Parsec

liter :: Quantity Double

https://en.wikipedia.org/wiki/Litre

litre :: Quantity Double

https://en.wikipedia.org/wiki/Litre

siDimensions :: [(String, Dimension)]

siUnits :: [(String, Quantity Double)]

https://en.wikipedia.org/wiki/Parsec https://en.wikipedia.org/wiki/International_System_of_Units

siPrefixes :: Floating a => [(String, Prefix a)]

https://en.wikipedia.org/wiki/Metric_prefix

siZeros :: [(String, Quantity Double)]

this contains a list of quantities where basis of measurement is not zero in the dimension because the corresponding SI unit has different zero