Math.Number.NumericExpression

Documentation

data Var a

Constructors

Var

Fields

var_name :: String

var_debruijn_index :: Integer

Val a

varIso :: IsoA Var a b -> Var a :==: Var b

class Expression e ctx a where

Minimal complete definition

eval

Methods

eval :: ctx a -> e a -> a

equal_in_context :: (Eq a, Expression f ctx a) => ctx a -> f a -> f a -> Bool

evalVar :: [a] -> Var a -> a

withCtx :: Functor f => [a] -> (f :*: Var) a -> f a

eVar :: Applicative f => String -> Integer -> (f :*: Var) a

expr :: Functor f => f a -> (f :*: Var) a

data VectorCtx a

Constructors

VectorCtx

Fields

vectorSpaceCtx :: Map String (VectorSpaceExpr a)

scalarCtx :: Map String (NumExpr a)

numericScalarCtx :: VectorCtx (NumExpr a)

numCtx :: [a]

data VectorSpaceExpr a

Constructors

VZero

VNegate (VectorSpaceExpr a)

VPlus (VectorSpaceExpr a) (VectorSpaceExpr a)

VScalar (NumExpr a) (VectorSpaceExpr a)

VVectorVar (Var (VectorSpaceExpr a))

VPrim a

vectorSpaceIso :: IsoA VectorSpaceExpr a b -> VectorSpaceExpr a :==: VectorSpaceExpr b

evalVectorSpaceExpr :: (VectorSpace v, Num v, IdVisitor v, Expression Var (Map String) (VectorSpaceExpr v), Expression VectorSpaceExpr VectorCtx v, Scalar v ~# v) => VectorCtx v -> VectorSpaceExpr v -> v

data NumExpr a

Basically same as methods in Num type class from prelude, but as data.

Constructors

Plus (NumExpr a) (NumExpr a)

Product (NumExpr a) (NumExpr a)

Subtract (NumExpr a) (NumExpr a)

Negate (NumExpr a)

Abs (NumExpr a)

Signum (NumExpr a)

NumPrim (Var a)

FromInteger Integer

InnerProduct (VectorSpaceExpr a) (VectorSpaceExpr a)

numIso :: IsoA NumExpr a b -> NumExpr a :==: NumExpr b

plus_assoc_iso :: NumExpr a :==: NumExpr a

product_assoc_iso :: NumExpr a :==: NumExpr a

data FracExpr a

basically same as Fractional class in Prelude.

Constructors

PrimFracExpr (NumExpr a)

Divide (FracExpr a) (FracExpr a)

Recip (FracExpr a)

FromRational Rational

data FloatExpr a

basically same as Floating class in Prelude, but as data.

Constructors

PrimPi

Exp (FloatExpr a)

Log (FloatExpr a)

Sqrt (FloatExpr a)

Power (FloatExpr a) (FloatExpr a)

Sin (FloatExpr a)

Cos (FloatExpr a)

Tan (FloatExpr a)

Asin (FloatExpr a)

Acos (FloatExpr a)

Atan (FloatExpr a)

Sinh (FloatExpr a)

Cosh (FloatExpr a)

Tanh (FloatExpr a)

Asinh (FloatExpr a)

Acosh (FloatExpr a)

Atanh (FloatExpr a)

PrimFloatExpr (FracExpr (FloatExpr a))

pp_unary :: (PpShow a1, PpShow a2) => a1 -> a2 -> Doc

liftNumFrac1 :: (NumExpr a1 -> NumExpr (FloatExpr a2)) -> a1 -> FloatExpr a2

liftNumFrac2 :: (NumExpr a1 -> NumExpr a2 -> NumExpr (FloatExpr a3)) -> a1 -> a2 -> FloatExpr a3

data EnumExpr a

Constructors

SuccExpr (EnumExpr a)

PredExpr (EnumExpr a)

ToEnumExpr (Var a)

runEnumExpr :: Enum a => [a] -> EnumExpr a -> a

runNumExpr :: (Num a, IdVisitor a) => [a] -> NumExpr a -> a

runFracExpr :: (Fractional a, IdVisitor a) => [a] -> FracExpr a -> a

varFold :: [a] -> Fold (Var a) a

vecSpaceFold :: (IdVisitor a, VectorSpace b, a ~ Scalar b) => Fold (Var b) b -> Fold (NumExpr a) (Scalar b) -> Fold a b -> Fold (VectorSpaceExpr a) b

numExprFold :: Num a => Fold (Var a) a -> Fold (NumExpr a) a

fracExprFold :: Fractional b => Fold (NumExpr a) b -> Fold (FracExpr a) b

enumExprFold :: Enum b => Fold (Var a) b -> Fold (EnumExpr a) b