Math.Number.R

Documentation

data R

This real representation takes epsilon as input as in epsilon-delta proof.

Constructors

Limit (Endo Rational)

lim :: Stream R -> R

real :: (Rational -> Rational) -> R

approximate :: R -> Rational -> Rational

rational_approximations :: Approximations a => a -> Stream Rational

floating_approximations :: Approximations a => a -> Stream Double

epsilon :: R

infinite_r :: R

liftR :: (Rational -> Rational) -> R -> R

this propagates accuracy without changing it.

liftR2 :: (Rational -> Rational -> Rational) -> R -> R -> R

this propagates accuracy without changing it

inverseImage :: (Rational -> Rational) -> R -> R

inverseImage transforms accuracy computation of the real.

liftWithAccuracy :: (Rational -> Rational) -> (Rational -> Rational) -> R -> R

The first argument is the function lifted. The second argument describes change in accuracy by the function.

approximate_as :: R -> R -> R

approx_compose will use the second real as the level of approximation to compute the first.

limit_compose :: (Rational -> R) -> R -> R

inverseImageEndo :: Endo Rational -> Endo R

differentialLiftR :: (Rational -> Rational) -> R -> R

This lifts a rational function to a real function. This computes accuracy using the formula \(dx = df / f'(x)\). The accuracy behaves badly when \(f'(x) = 0\) due to divide-by-zero.

derivate_rational :: Rational -> (Rational -> Rational) -> Rational -> Rational

derivate for rational functions. The first argument is epsilon.

derivate_r :: (R -> R) -> R -> R

computes derivate. \[{{df}\over{dt}} = \lim_{\epsilon\rightarrow 0}{{f(t+\epsilon)-f(t-\epsilon)}\over{2\epsilon}}\]

newtons_method :: Numerics a => (a -> a) -> a -> a

sqrt_r :: R -> R

log_by_newton :: R -> R

exp_r :: R -> R

pi_r :: R

Using Simon Plouffe's BPP digit extraction algorithm for computing pi. See Pi for details.

log_r :: R -> R

integral_r :: (R, R) -> (R -> R) -> R

foldable_integral :: (Foldable t, Functor t) => (Rational -> t b) -> (b -> R) -> R

foldable_simple_integral :: (Num a, Foldable t, Functor t) => (a -> t b) -> (b -> a) -> a -> a

integral_accuracy :: Fractional t => (t -> R) -> R -> R -> Rational -> [Rational]

asin_r :: R -> R

https://en.wikipedia.org/wiki/Inverse_trigonometric_functions

acos_r :: R -> R

https://en.wikipedia.org/wiki/Inverse_trigonometric_functions

atan_r :: R -> R

https://en.wikipedia.org/wiki/Inverse_trigonometric_functions

sin_by_series :: R -> R

trigonometric functions

cos_by_series :: R -> R

trigonometric functions