Math.Number.Real

Documentation

data R

Problem with this representation: real number ranges cannot be represented. given two rational numbers \(r_1 < r_2\), it should be there are infinitely many real numbers between \(r_1\) and \(r_2\), so support density of irrationals. https://en.wikipedia.org/wiki/Dense_order

In constructive mathematics, given two real numbers \(r_1' < r_2'\), there should exist a rational number r, such that \(r_1' < r\) and \(r < r_2'\). HOWEVER, these comparisons are a type error. In particular this cannot be used to _define_ how to compare real numbers.

what is meant is if \(r_1' < r_2'\) then there exists a rational r such that \(r_1' < {\mathbf{fromRational}}(r)\) and \({\mathbf{fromRational}}(r) < r_2'\)

In here properties of \({\mathbf{fromRational}}\) are important. Note that comparisons invoke the real comparison, so at most is a recursive definition of real comparison in terms of itself. But it doesn't include termination condition.

It should be so that _if_ two real numbers \(r_1\) and \(r_2\) are not the same real number, then \(r_1 < r_2\) makes sense. But comparing whether two real numbers are the same is undecidable (algorithm for this comparison will not halt if they are same). Would this mean that if \(r_1 < r_2\) is computable, wouldn't \(\neg(r_1 < r_2 \lor r_2 < r_1)\) be computable equality of reals; This is computable in the limit , but not computable? https://en.wikipedia.org/wiki/Computation_in_the_limit

Constructors

Limit (Stream Rational)

data Modulus

Constructors

Modulus

Fields

modulus_value :: R

modulus :: Rational -> Integer

d :: Infinitesimal a => (a -> a) -> a -> Closure a

d x = partial_derivate x

liftReal :: (Rational -> Rational) -> R -> R

convergence_ratio_test :: (Fractional a, Ord a, Closed a) => Stream a -> Bool

average_convergence_ratio :: (MetricSpace a, Fractional (Scalar a)) => Stream a -> Stream (Scalar a)

list_average :: Fractional a => [a] -> a

cauchy :: MetricSpace a => Stream a -> Stream [Scalar a]

cseq_difference_matrix :: (Functor m, Functor n, Num a) => m a -> n a -> (m :*: n) a

cauchy_sequence_equivalence :: Num a => Stream a -> Stream a -> Stream a

See Suppes: Axiomatic Set Theory. Note that we would like to check that \[\forall \epsilon > 0. \exists n. \forall k > n. |s_k - t_k| < \epsilon \]. However, this is undecidable in particular with respect to what to choose for n. We can nonetheless compute the stream of distances.

cseq_equivalence_matrix :: (Num a, Infinitesimal eps) => (eps -> Integer) -> Stream a -> Stream a -> (Stream :*: Stream) a

cseq_equivalence_matrix computes equivalence of cauchy sequences. However, the modulus argument is undecidable. But if you produce a modulus, this chooses the part of the cauchy sequence differences that should converge according to the convergence ratio of the epsilon from the infinitesimal instance. Thus if you can prove that the matrix elements converge to zero when elements of increasing index are chosen, you would have proven equality of the two cauchy sequences. This could be possible in special cases, even if it's undecidable in general. Note that it suffices to consider the convergence of the first column, because you could move any part to the first column by choosing reasonable values of the modulus. The other differences are included so computations and analysis of the situation can be easier to perform.

cseq_convergence :: Num a => Stream a -> Stream a -> Stream [a]

This version of cauchy sequence equality convergence check produces a stream of lists, where each list contains those elements which are close to each other in terms of choice of indices, e.g. \[C(s,t)_i = [ \lvert s_j - t_k \rvert | j \leftarrow naturals, k \leftarrow naturals, i = j + k]\] thus the complexity of the stream element lists of the result should increase in the same way than the complexity of the input stream elements. The lists in the result stream have monotonically increasing length. thus to compute results for particular level of approximation, it should suffice to choose a list from the result stream, and perform the equivalence check using the values in the list. To do this, it would be necessary to be able to determine how close to zero the elements in the list should be. the middle element (if any) of the list is probably the most interesting, because that would mean \(j == k\), therefore the computation result would be \(\lvert s_j - t_j \rvert\), but due to different rate of convergence of the input sequences, it can happen that the elementwise differences are not really representative of the stream convergence behaviour at the diagonal. however, analyzing these numbers should produce useful results about convergence of the input stream element differences. e.g choosing minimum and maximum element from each list determines best and worst approximation in the same depth of approximations. for the equality of the cauchy sequences to hold, the best approximation should stay close to zero and better so when further elements of the stream are obtained.

cseq_equivalence_list :: (Num a, Infinitesimal eps) => (eps -> Integer) -> Stream a -> Stream a -> Stream [a]

converges :: MetricSpace a => Stream a -> a -> Stream [Scalar a]

increase_accuracy_by :: Limiting a => Integer -> Closure a -> Closure a

partial_derive :: (Fractional a, Closed a, Infinitesimal eps) => (eps -> a -> a) -> (a -> a) -> a -> a

partial_derivate_closure :: (Limiting a, Fractional a) => (a -> b -> b) -> (b -> a) -> b -> Endo (Closure a)

partial_derivate :: (Closed eps, Infinitesimal eps) => (eps -> a -> a) -> (a -> eps) -> a -> eps

partial_derivate_endo :: (Limiting a, Fractional a) => (a -> Endo b) -> (b -> a) -> b -> Endo (Closure a)

complex_derivate :: (RealFloat r, Closed r, Infinitesimal r) => (Complex r -> Complex r) -> Complex r -> Complex r

https://en.wikipedia.org/wiki/Wirtinger_derivatives

average :: R -> R -> R

completenessOfReals :: Closure R :==: R

An isomorphism \(Cl(R) \cong R\).

liftRClosure :: (R -> R) -> Closure R -> Closure R

limit_real :: Stream R -> R

lift_real :: (Rational -> Rational) -> R -> R

series :: (Integer -> R) -> R

supremum_gen :: (Fractional a, Ord a, Limiting a) => [Closure a] -> Closure a

Least upper bound property | note this is really undecidable.

negate_limit :: (Limiting a, Num a) => Closure a -> Closure a

infimum_gen :: (Fractional a, Ord a, Limiting a) => [Closure a] -> Closure a

infinity :: R

infinity_gen :: (Fractional a, Limiting a) => Closure a

negative_infinity_gen :: (Fractional a, Limiting a) => Closure a

limit_below :: R -> (R -> R) -> Closure R

Squeeze theorem

limit_above :: R -> (R -> R) -> Closure R

Squeeze_theorem

limit_both :: R -> (R -> R) -> Closure R

Squeeze_theorem

negative_infinity :: R

epsilon_linear :: R

increase_accuracy :: R -> R

lessthan_precision :: R -> R -> Integer -> Bool

limiting_distance :: (Num a, Limiting a) => Closure a -> Closure a -> Closure a

distance_matrix :: R -> R -> (Stream :*: Stream) Rational

at_precision_rational :: R -> Rational -> Rational

compute rational approximation more precise than the given rational the precision is expressed as rational number close to zero, the result will be within range \([r-p,r+p]\), in the sense that in the sequence of rational approximations \(r = (r_i | i \in {\mathbf N})\) we have \(\lvert r_{i+1} - r_i \rvert \leq p\). Note that we do _not_ attempt to prove that all successive approximations have this same property, we need the differences to be monotonically decreasing for this to represent correct precision.

at_precision :: R -> Integer -> R

here precision is expresses as integer power of 10, e.g. \(1 == 10^{-1}\), \(2 == 10^{-2}\), \(3 == 10^{-3}\) and so on. so how many digits after the decimal point are required.

precision_rational :: R -> Integer -> Rational

rational precision expressed as integer power of 10

real_derivate :: (R -> R) -> R -> R

real_exp :: R -> R

derivate_closed :: Infinitesimal a => (a -> a) -> a -> Closure a

https://en.wikipedia.org/wiki/Differential_calculus

vector_derivate :: (Infinitesimal (Scalar a), VectorSpace a, Limiting a) => (Scalar a -> a) -> Scalar a -> Closure a

pseudo_derivate :: (Fractional r, Limiting r, Infinitesimal a) => (a -> r) -> a -> Closure r

derivates_at :: DifferentiallyClosed a => (a -> a) -> Stream a -> Stream a

i'th element of derivates_at(f,s) is \(D^i[f](s_i)\)

derivate_around :: Infinitesimal a => (a -> a) -> a -> Closure a

derivate_around doesn't require f to be defined at x, but requires limits from both sides of x to exist [it never evaluates f at x]. \[ \lim_{\epsilon \rightarrow 0} {{f(x+\epsilon)-f(x-\epsilon)}\over{2\epsilon}} \]

partial_derivate1_2 :: Infinitesimal a => (a -> b -> a) -> a -> b -> Closure a

\[\lim_{\epsilon\rightarrow 0}{{f(a+\epsilon,b)-f(a-\epsilon,b)}\over{2\epsilon}}\]

partial_derivate2_2 :: Infinitesimal a => (b -> a -> a) -> b -> a -> Closure a

\[\lim_{\epsilon\rightarrow 0}{{f(a,b+\epsilon)-f(a,b-\epsilon)}\over{2\epsilon}}\]

partial_derivate1_3 :: Infinitesimal a => (a -> b -> c -> a) -> a -> b -> c -> Closure a

\[\lim_{\epsilon\rightarrow 0}{{f(a+\epsilon,b,c)-f(a-\epsilon,b,c)}\over{2\epsilon}}\]

partial_derivate2_3 :: Infinitesimal a => (b -> a -> c -> a) -> b -> a -> c -> Closure a

\[\lim_{\epsilon\rightarrow 0}{{f(a,b+\epsilon,c)-f(a,b-\epsilon,c)}\over{2\epsilon}}\]

partial_derivate3_3 :: Infinitesimal a => (b -> c -> a -> a) -> b -> c -> a -> Closure a

\[\lim_{\epsilon\rightarrow 0}{{f(a,b,c+\epsilon)-f(a,b,c-\epsilon)}\over{2\epsilon}}\]

agm :: (Floating a, Limiting a) => a -> a -> Closure a

Arithmetic-geometric mean

expo :: R -> R

Exponential function using limit definition \(\exp(x) = \lim_{n\rightarrow\infty}(1+{{x}\over{n}})^n\).

newtons_method :: Numerics a => (a -> a) -> a -> a

eigenvalue :: (Infinitesimal a, Closed a, Applicative m, FiniteSquareMatrix m a, VectorSpace ((m :*: m) a), Scalar ((m :*: m) a) ~ a) => (m :*: m) a -> Closure a

integrate_vector :: (Enum (Scalar a), VectorSpace a, Limiting a, Limiting (Scalar a)) => (Scalar a -> a) -> Closure (Scalar a) -> (Scalar a, Scalar a) -> Closure a

integrate :: (Enum a, Num a, Limiting a) => (a -> a) -> Closure a -> (a, a) -> Closure a

https://en.wikipedia.org/wiki/Methods_of_contour_integration

integral_real :: (R, R) -> (R -> R) -> R

integral_rational :: (R, R) -> (Rational -> Rational) -> R

integrate_curve :: (Enum a, Num a, Num r, Limiting a, Limiting r) => (a, a) -> (a -> r) -> (r -> r) -> Closure a -> Closure r

for integrate_curve (xa,ya) curve f dx, Requires: \(\lim_{x \rightarrow 0}curve(x) = 0 \land \lim_{i\rightarrow \infty} dx_i = 0\)

gamma_via_integral :: R -> Closure R

Integral This doesn't converge well.

approximately_equal :: R -> R -> Bool

approximately_eq :: R -> R -> Bool

constructively_less :: Rational -> R -> R -> Closure Bool

equality :: Closure R -> Closure R -> Closure Bool

signum_real :: R -> Int -> Int

show_at_precision_real :: R -> Integer -> String

properFraction_real :: Num a => R -> (a, R)

enumFromThenToReal :: Enum e => (e -> R) -> e -> e -> e -> [R]

This attempts to work around the density problem by computing range | in some type which does have enum instance, | then using the given function to map to reals.

fromThenTo :: Rational -> Rational -> Rational -> [R]

fromTo :: Rational -> Rational -> [R]

enumFromToReal :: Enum e => (e -> R) -> e -> e -> [R]

This attempts to work around the density problem by computing range | in some type which does have enum instance, | then using the given function to map to reals.

realEnumFromThenTo :: R -> R -> R -> Stream [Rational]

strEnumFromThenTo :: Stream Rational -> Stream Rational -> Stream Rational -> Stream [Rational]

toReal :: Rational -> R

logarithm :: R -> R

Natural logarithm log(1+z) = z - z^22 + z^33 - ...

euler_constant :: R

napiers_constant :: R

logarithm_by_newton :: R -> Closure R

invert_by_newton :: (Closed a, Infinitesimal a) => (a -> a) -> a -> Closure a

exp_by_powers :: (Fractional r, Limiting r) => r -> Closure r

exp_by_powers_real :: R -> R

sum_generating_function :: Stream Rational -> R -> R

riemann_zeta :: Floating a => a -> Stream a

https://en.wikipedia.org/wiki/Riemann_zeta_function

riemann_zeta_real :: Integral a => a -> R

https://en.wikipedia.org/wiki/Riemann_zeta_function

riemann_zeta_complex :: RealFloat a => Complex a -> Stream (Complex a)

https://en.wikipedia.org/wiki/Riemann_zeta_function

encodeInBase :: Integer -> Integer -> Int -> R

fromFloat :: RealFloat a => a -> R

golden_ratio :: R

https://en.wikipedia.org/wiki/Golden_ratio

slow_golden_ratio :: R

logistic :: Floating a => a -> a

https://en.wikipedia.org/wiki/Logistic_function

floating_approximations :: Approximations a => a -> Stream Double

logarithm_by_power :: (Floating a, Limiting a) => a -> Closure a

tetrate :: (Eq b, Num b, Floating a) => b -> a -> a

Tetration

tetrate2 :: Floating a => a -> a

log_newton :: (Floating a, Infinitesimal a, Closed a) => a -> a

exp_by_series2 :: R -> R

exp_by_series :: R -> R

sin_by_series :: R -> R

cos_by_series :: R -> R

approximate_sums_modulus :: Modulus -> Modulus -> Modulus

withInverseModulus :: (Fractional r, Real u) => (R -> R) -> (r -> u) -> Modulus -> Modulus

Orphan instances

MetricSpace Int

Instance details

Methods

distance :: Int -> Int -> Scalar Int

MetricSpace Integer

Instance details

Methods

distance :: Integer -> Integer -> Scalar Integer

Infinitesimal Rational

Instance details

Methods

epsilon_closure :: Closure Rational

epsilon_stream :: Stream Rational

Limiting Bool

Instance details

Associated Types

data Closure Bool :: Type

Methods

limit :: Stream Bool -> Closure Bool

approximations :: Closure Bool -> Stream Bool

MetricSpace (Ratio Integer)

Instance details

Methods

distance :: Ratio Integer -> Ratio Integer -> Scalar (Ratio Integer)

(RealFloat a, Infinitesimal a) => Infinitesimal (Complex a)

Instance details

Methods

epsilon_closure :: Closure (Complex a)

epsilon_stream :: Stream (Complex a)

(Fractional a, Limiting a) => Infinitesimal (Stream a)

Instance details

Methods

epsilon_closure :: Closure (Stream a)

epsilon_stream :: Stream (Stream a)