Math.Number.StreamInterface

Documentation

class (Monad str, StreamBuilder str, StreamObserver str) => Limiting str a where

Idea: limit produces an exact value from a monotone sequence of approximations

https://ncatlab.org/nlab/show/closure+operator

The Closure represents a sequentially closed set of elements

Associated Types

data Closure str a

Methods

limit :: str a -> Closure str a

approximations :: Closure str a -> str a

class (Fractional a, Limiting str a) => Infinitesimal str a where

Minimal complete definition

(epsilonClosure | epsilonStream)

Methods

epsilonClosure :: Closure str a

infiniteClosure :: Closure str a

minusInfiniteClosure :: Closure str a

nanClosure :: Closure str a

epsilonStream :: str a

infiniteStream :: str a

minusInfiniteStream :: str a

class Applicative str => StreamBuilder str where

Methods

pre :: a -> str a -> str a infixr 5

class StreamBuilder str => StreamBuilder2 str where

Methods

pre2 :: str a -> (str :*: str) a -> (str :*: str) a -> (str :*: str) a

class Applicative str => StreamObserver str where

Methods

shead :: str a -> a

stail :: str a -> str a

class StreamObserver str => StreamObserver2 str where

Methods

stail2 :: str a -> str a

class StreamObserver str => StreamObserver3 str where

Methods

stail3 :: str a -> str a

data Stream a

Data structure of infinite lazy streams.

Constructors

Pre infixr 5

Fields

sheadImpl :: a

stailLazy :: Stream a

class Infinitesimal Stream a => Closed a where

Methods

accumulationPoint :: Closure Stream a -> a

type DerivateConstraint str a = (Monad str, Num a)

partialDerivate :: (Closed eps, Fractional eps) => (eps -> a -> a) -> (a -> eps) -> a -> eps

\[ {\rm{partial\_derivate}}(\Delta, f, x) = \lim_{\epsilon \rightarrow 0}{{f(\Delta_{\epsilon}(x)) - f(x)}\over{\Delta_{\epsilon}(x) - x}} \]

partialDerive :: (Closed a, Infinitesimal Stream a, Fractional a) => (a -> a -> a) -> (a -> a) -> a -> a

\[ {\rm{partial\_derive}}(\Delta, f, x) = \lim_{\epsilon \rightarrow 0}{{f(\Delta_{\epsilon}(x)) - f(x)}\over{\Delta_{\epsilon}(x) - x}} \]

derivateAround :: Infinitesimal str a => (a -> a) -> a -> Closure str a

derivate_around doesn't require f to be defined at x, but requires limits from both sides of x to exist [it never evaluates f at x]. \[ \lim_{\epsilon \rightarrow 0} {{f(x+\epsilon)-f(x-\epsilon)}\over{2\epsilon}} \]

vectorDerivate :: (Infinitesimal str (Scalar a), VectorSpace a, Limiting str a) => (Scalar a -> a) -> Scalar a -> Closure str a

derivateClosed :: Infinitesimal str a => (a -> a) -> a -> Closure str a

newtonsMethodReal :: (Infinitesimal Stream a, Closed a) => (a -> a) -> a -> Closure Stream a

iterateStream :: StreamBuilder str => (a -> a) -> a -> str a

pseudoDerivate :: (Fractional r, Limiting str r, Infinitesimal str a) => (a -> r) -> a -> Closure str r

https://en.wikipedia.org/wiki/Differential_calculus

partialDerivate1_2 :: Infinitesimal str a => (a -> b -> a) -> a -> b -> Closure str a

\[\lim_{\epsilon\rightarrow 0}{{f(a+\epsilon,b)-f(a,b)}\over{\epsilon}}\]

partialDerivate2_2 :: Infinitesimal str a => (b -> a -> a) -> b -> a -> Closure str a

\[\lim_{\epsilon\rightarrow 0}{{f(a,b+\epsilon)-f(a,b)}\over{\epsilon}}\]

partialDerivate1_3 :: Infinitesimal str a => (a -> b -> c -> a) -> a -> b -> c -> Closure str a

\[\lim_{\epsilon\rightarrow 0}{{f(a+\epsilon,b,c)-f(a,b,c)}\over{\epsilon}}\]

partialDerivate2_3 :: Infinitesimal str a => (b -> a -> c -> a) -> b -> a -> c -> Closure str a

\[\lim_{\epsilon\rightarrow 0}{{f(a,b+\epsilon,c)-f(a,b,c)}\over{\epsilon}}\]

partialDerivate3_3 :: Infinitesimal str a => (b -> c -> a -> a) -> b -> c -> a -> Closure str a

\[\lim_{\epsilon\rightarrow 0}{{f(a,b,c+\epsilon)-f(a,b,c)}\over{\epsilon}}\]

constant :: a -> Stream a

stream consisting of the same element repeated.

streamMatrixImpl :: Stream a -> Stream (Stream a, Stream a) -> (Stream :*: Stream) a

streamDiagonalImpl :: (Stream :*: Stream) a -> Stream a

streamCodiagonalImpl :: (Stream :*: Stream) a -> Stream (Stream a, Stream a)