Math.Number.Units

Description

This module contains auxiliary definitions related to dimensional analysis. This is based on the SI system of units. This module supports compile-time checking for dimensional units. This is accomplished by a set of newtype wrappers of Double.

It is not necessarily a good idea to make the checking in compile-time. In particular, this can prevent operations to be available, e.g. Num class is not supported for compile-time checked quantities. Num because it has wrong type for the multiplication and division. Instead of Num, VectorSpace operations and operations defined in this module should be used.

However, these types can distinguish quantities even in cases where the dimensions match, e.g. radians and steradians. Explicit conversions are nonetheless available when dimensions match.

Read operations are pretty strict in that they require exactly the syntax "double unit", where unit is the result of unitOf (fromAmount 0 :: T). Example: (read "3 m^2" :: SquareLength) This can sometimes be unintuitive and is sensitive to the algorithm used to print units. For example: unitOf (fromAmount 0 :: Mass :/ Length) == "m^-1 kg". If in doubt, it's possible to use the read algorithm handling all units from DimensionalAnalysis module.

However, it's very much likely that compile-time checked newtypes are faster at run-time than using the run-time checked quantities. This is probably the primary reason for wanting this module instead of run-time checked version.

Also it's not possible to cast input data directly without run-time check to a compile-time checked quantity.

See "Barton&Nackman: Scientific and Engineering C++" for C++ approach to dimensional analysis.

https://en.wikipedia.org/wiki/International_System_of_Units

Documentation

type UnitName u = Scalar u -> u

Notice that fromAmount from Unit class has type (Unit u) => UnitName u. However usually a concrete constructor of one of the newtypes is used as value.

(*%%) :: Scalar u ~ Scalar u' => UnitName u -> UnitName u' -> UnitName (u :* u')

(/%%) :: Scalar u ~ Scalar u' => UnitName u -> UnitName u' -> UnitName (u :/ u')

(*%) :: (LiteralUnit u, LiteralUnit w, Num (Scalar u), Scalar u ~ Scalar w) => u -> w -> u :* w

heterogeneous product respecting units It may be necessary to use quantity operation to convert to run-time representation after using this, because the result type accumulates new parts with each use without considering what the total dimension of the result is. This can produce excessively complicated types, because the type represents the structure of the expression rather than structure of the result.

(/%) :: (Fractional (Scalar u), LiteralUnit u, LiteralUnit w, Scalar u ~ Scalar w) => u -> w -> u :/ w

heterogeneous division respecting units. These simply accumulate type information about the units. It may be necessary to use quantity operation to convert to run-time representation after using this, because the result type accumulates new parts with each use without considering what the total dimension of the result is. This can produce excessively complicated types, because the type represents the structure of the expression rather than the structure of the result.

unitNameToDimension :: (Num (Scalar u), Unit u) => UnitName u -> Dimension

mapAmount :: LiteralUnit a => (Scalar a -> Scalar a) -> a -> a

mapAmount2 :: LiteralUnit a => (Scalar a -> Scalar a -> Scalar a) -> a -> a -> a

asUnit :: (MonadFail m, Show (Scalar u), Show u, LiteralUnit u, Fractional (Scalar u)) => Quantity (Scalar u) -> UnitName u -> m u

Conversion from "Quantity a" to a unit-specific type. The second argument is the constructor for the newtype specific to the unit.

Example: (3 %* meter) asUnit Meters == return (Meters 3) 0.3 asUnit Radians == return (Radians 0.3)

If the dimensions don't match, this raises an exception.

quantity :: LiteralUnit u => u -> Quantity (Scalar u)

Converts a compile-time checked dimensional unit to run-time checked version This often has the effect of reducing the complexity in types.

data a :/ b

Constructors

QDivide

Fields

qdivideAmount :: !(Scalar a)

qdivideDividendUnit :: UnitName a

qdivideDivisorUnit :: UnitName b

data a :* b

Constructors

QProduct

Fields

qproductAmount :: !(Scalar a)

qproduct_firstUnit :: UnitName a

qproduct_secondUnit :: UnitName b

newtype Dimensionless a

Constructors

Dimensionless

Fields

dimensionlessValue :: a

newtype Information a

Constructors

Bits

Fields

numberOfBits :: a

newtype SoundLevel a

Constructors

SoundAmplitude

Fields

soundAmplitude :: a

newtype Angle a

Constructors

Radians

Fields

radians :: a

class Functor angle => DimensionalFloating angle where

Minimal complete definition

fractionOfRotation, dimSin, dimCos, dimTan, dimAsin, dimAcos, dimAtan, dimLog, dimExp, dimFromPolar, dimToPolar

Methods

halfRotation :: angle Double

fullRotation :: angle Double

fractionOfRotation :: Rational -> angle Double

dimSin :: Floating a => angle a -> a

dimCos :: Floating a => angle a -> a

dimTan :: Floating a => angle a -> a

dimAsin :: Floating a => a -> angle a

dimAcos :: Floating a => a -> angle a

dimAtan :: Floating a => a -> angle a

dimLog :: RealFloat a => Complex a -> angle a

dimExp :: RealFloat a => angle a -> Complex a

dimFromPolar :: RealFloat a => a -> angle a -> Complex a

dimToPolar :: RealFloat a => Complex a -> (a, angle a)

piAngle :: Angle Double

fraction_of_circle :: Rational -> Angle Double

asinAngle :: Floating a => a -> Angle a

acosAngle :: Floating a => a -> Angle a

atanAngle :: Floating a => a -> Angle a

sinAngle :: Floating a => Angle a -> a

cosAngle :: Floating a => Angle a -> a

tanAngle :: Floating a => Angle a -> a

logAngle :: RealFloat a => Complex a -> Angle a

expAngle :: Floating a => Angle a -> Complex a

fromPolar :: (Unit u, Floating (Scalar u)) => u -> Angle (Scalar u) -> Complex u

toPolar :: (LiteralUnit u, RealFloat (Scalar u)) => Complex u -> (u, Angle (Scalar u))

newtype DegreesAngle a

Constructors

Degrees

Fields

degrees :: a

degreesAngle :: Floating a => DegreesAngle a -> Angle a

angleDegrees :: Floating a => Angle a -> DegreesAngle a

isoDegreesToAngle :: Floating a => DegreesAngle a :==: Angle a

newtype SolidAngle a

https://en.wikipedia.org/wiki/Steradian

Constructors

Steradians

Fields

steradians :: a

showUnit :: (Unit u, Show (Scalar u)) => u -> String

readUnit :: (Unit u, Read (Scalar u)) => UnitName u -> ReadPrec u

newtype Percentage a

Constructors

Percentages

Fields

percentages :: a

newtype Acceleration a

Constructors

MetersPerSquareSecond

Fields

metersPerSquareSecond :: a

newtype Velocity a

Constructors

MetersPerSecond

Fields

metersPerSecond :: a

newtype SquareLength a

Constructors

SquareMeters

Fields

squaremeters :: a

newtype CubicLength a

Constructors

CubicMeters

Fields

cubicmeters :: a

sqrt_length :: Floating a => SquareLength a -> Length a

square_length :: Num a => Length a -> SquareLength a

times_length :: Num a => Length a -> Length a -> SquareLength a

newtype Length a

Constructors

Meters

Fields

meters :: a

newtype Mass a

Constructors

Kilograms

Fields

kilograms :: a

newtype Time a

Constructors

Seconds

Fields

seconds :: a

newtype Current a

Constructors

Amperes

Fields

amperes :: a

newtype DegreesFahrenheit a

https://en.wikipedia.org/wiki/Fahrenheit

Constructors

DegreesFahrenheit

Fields

fahrenheits :: a

newtype DegreesRankine a

https://en.wikipedia.org/wiki/Rankine_scale

Constructors

DegreesRankine

Fields

rankines :: a

newtype DegreesTemperature a

Constructors

DegreesCelcius

Fields

celciuses :: a

newtype Temperature a

Constructors

DegreesKelvin

Fields

kelvins :: a

isoCelciusToKelvin :: Fractional a => DegreesTemperature a :==: Temperature a

celciusToKelvin :: Fractional a => DegreesTemperature a -> Temperature a

kelvinToCelcius :: Fractional a => Temperature a -> DegreesTemperature a

newtype Substance a

Constructors

Moles

Fields

moles :: a

newtype Intensity a

Constructors

Candelas

Fields

candelas :: a

newtype Frequency a

Constructors

Hertzes

Fields

hertzes :: a

newtype Force a

Constructors

Newtons

Fields

newtons :: a

newtype Torque a

Constructors

NewtonMeters

Fields

newtonmeters :: a

newtype Pressure a

Constructors

Pascals

Fields

pascals :: a

newtype Energy a

Constructors

Joules

Fields

joules :: a

newtype Power a

Constructors

Watts

Fields

watts :: a

newtype Charge a

Constructors

Coulombs

Fields

coulombs :: a

newtype Voltage a

Constructors

Volts

Fields

volts :: a

newtype Capacitance a

Constructors

Farads

Fields

farads :: a

newtype Resistance a

Constructors

Ohms

Fields

ohms :: a

newtype Conductance a

Constructors

Siemenses

Fields

siemenses :: a

newtype Flux a

Constructors

Webers

Fields

webers :: a

newtype FluxDensity a

Constructors

Teslas

Fields

teslas :: a

newtype Inductance a

Constructors

Henrys

Fields

henrys :: a

newtype LuminousFlux a

Constructors

Lumens

Fields

lumens :: a

newtype Illuminance a

Constructors

Luxes

Fields

luxes :: a

newtype Radioactivity a

Constructors

Becquerels

Fields

becquerels :: a

newtype AbsorbedDose a

Constructors

Grays

Fields

grays :: a

newtype EquivalentDose a

Constructors

Sieverts

Fields

sieverts :: a

newtype CatalyticActivity a

Constructors

Katals

Fields

katals :: a

Orphan instances

LiteralUnit R

Instance details

Methods

fromAmount :: Scalar R -> R

zeroAmount :: (Scalar R -> R) -> Scalar R

conversionFactor :: (Scalar R -> R) -> Scalar R

LiteralUnit Double

Instance details

Methods

fromAmount :: Scalar Double -> Double

zeroAmount :: (Scalar Double -> Double) -> Scalar Double

conversionFactor :: (Scalar Double -> Double) -> Scalar Double

LiteralUnit Float

Instance details

Methods

fromAmount :: Scalar Float -> Float

zeroAmount :: (Scalar Float -> Float) -> Scalar Float

conversionFactor :: (Scalar Float -> Float) -> Scalar Float

Unit R

Instance details

Methods

amount :: R -> Scalar R

fromQuantity :: (Alternative m, MonadFail m) => Quantity (Scalar R) -> m R

unitOf :: R -> String

dimension :: R -> Dimension

Unit Double

Instance details

Methods

amount :: Double -> Scalar Double

fromQuantity :: (Alternative m, MonadFail m) => Quantity (Scalar Double) -> m Double

unitOf :: Double -> String

dimension :: Double -> Dimension

Unit Float

Instance details

Methods

amount :: Float -> Scalar Float

fromQuantity :: (Alternative m, MonadFail m) => Quantity (Scalar Float) -> m Float

unitOf :: Float -> String

dimension :: Float -> Dimension