Math.Number.Units

Description

This module contains auxiliary definitions related to dimensional analysis. This is based on the SI system of units. This module supports compile-time checking for dimensional units.

It is not necessarily a good idea to make the checking in compile-time. In particular, this can prevent operations to be available, e.g. Num and Read classes are not supported for compile-time checked quantities. Num because it has wrong type for the multiplication and division. Instead of Num, VectorSpace operations and operations defined in this module should be used.

However, these types can distinguish quantities even in cases where the dimensions match, e.g. radians and steradians. Explicit conversions are nonetheless available when dimensions match.

However, it's very much likely that compile-time checked newtypes are faster at run-time than using the run-time checked quantities. This is probably the primary reason for wanting this module instead of run-time checked version.

Also it's not possible to cast input data directly without run-time check to a compile-time checked quantity.

Notice: Read instances exists only for run-time checked quantities.

See "Barton&Nackman: Scientific and Engineering C++" for C++ approach to dimensional analysis.

https://en.wikipedia.org/wiki/International_System_of_Units

Documentation

type UnitName u = Scalar u -> u

Notice that fromAmount from Unit class has type (Unit u) => UnitName u. However usually a concrete constructor of one of the newtypes is used as value.

(*%%) :: Scalar u ~ Scalar u' => UnitName u -> UnitName u' -> UnitName (u :* u')

(/%%) :: Scalar u ~ Scalar u' => UnitName u -> UnitName u' -> UnitName (u :/ u')

(*%) :: (Unit u, Unit w, Num (Scalar u), Scalar u ~ Scalar w) => u -> w -> u :* w

heterogeneous product respecting units It may be necessary to use quantity operation to convert to run-time representation after using this, because the result type accumulates new parts with each use without considering what the total dimension of the result is. This can produce excessively complicated types, because the type represents the structure of the expression rather than structure of the result.

(/%) :: (Fractional (Scalar u), Unit u, Unit w, Scalar u ~ Scalar w) => u -> w -> u :/ w

heterogeneous division respecting units. These simply accumulate type information about the units. It may be necessary to use quantity operation to convert to run-time representation after using this, because the result type accumulates new parts with each use without considering what the total dimension of the result is. This can produce excessively complicated types, because the type represents the structure of the expression rather than the structure of the result.

unitNameToDimension :: (Num (Scalar u), Unit u) => UnitName u -> Dimension

mapAmount :: Unit a => (Scalar a -> Scalar a) -> a -> a

mapAmount2 :: Unit a => (Scalar a -> Scalar a -> Scalar a) -> a -> a -> a

asUnit :: (Monad m, Show (Scalar u), Show u, Unit u, Fractional (Scalar u)) => Quantity (Scalar u) -> UnitName u -> m u

Conversion from "Quantity a" to a unit-specific type. The second argument is the constructor for the newtype specific to the unit.

Example: (3 %* meter) asUnit Meters == return (Meters 3) 0.3 asUnit Radians == return (Radians 0.3)

If the dimensions don't match, this raises an exception.

quantity :: Unit u => u -> Quantity (Scalar u)

Converts a compile-time checked dimensional unit to run-time checked version This often has the effect of reducing the complexity in types.

data a :/ b

Constructors

QDivide

Fields

qdivide_amount :: !(Scalar a)

qdivide_dividend_unit :: UnitName a

qdivide_divisor_unit :: UnitName b

data a :* b

Constructors

QProduct

Fields

qproduct_amount :: !(Scalar a)

qproduct_first_unit :: UnitName a

qproduct_second_unit :: UnitName b

newtype Dimensionless

Constructors

Dimensionless

Fields

dimensionless_value :: Double

newtype Information

Constructors

Bits

Fields

number_of_bits :: Double

newtype SoundLevel

Constructors

SoundAmplitude

Fields

sound_amplitude :: Double

newtype Angle

Constructors

Radians

Fields

radians :: Double

pi_angle :: Angle

asin_angle :: Double -> Angle

acos_angle :: Double -> Angle

atan_angle :: Double -> Angle

sin_angle :: Angle -> Double

cos_angle :: Angle -> Double

tan_angle :: Angle -> Double

log_angle :: Complex Double -> Angle

exp_angle :: Angle -> Complex Double

fromPolar :: Length -> Angle -> Complex Length

toPolar :: Complex Length -> (Length, Angle)

newtype DegreesAngle

Constructors

Degrees

Fields

degrees :: Double

degreesAngle :: DegreesAngle -> Angle

angleDegrees :: Angle -> DegreesAngle

newtype SolidAngle

https://en.wikipedia.org/wiki/Steradian

Constructors

Steradians

Fields

steradians :: Double

show_unit :: (Unit u, Show (Scalar u)) => u -> String

newtype Percentage

Constructors

Percentages

Fields

percentages :: Double

newtype Acceleration

Constructors

MetersPerSquareSecond

Fields

metersPerSquareSecond :: Double

newtype Velocity

Constructors

MetersPerSecond

Fields

metersPerSecond :: Double

newtype SquareLength

Constructors

SquareMeters

Fields

squaremeters :: Double

sqrt_length :: SquareLength -> Length

square_length :: Length -> SquareLength

times_length :: Length -> Length -> SquareLength

newtype Length

Constructors

Meters

Fields

meters :: Double

newtype Mass

Constructors

Kilograms

Fields

kilograms :: Double

newtype Time

Constructors

Seconds

Fields

seconds :: Double

newtype Current

Constructors

Amperes

Fields

amperes :: Double

newtype DegreesFahrenheit

https://en.wikipedia.org/wiki/Fahrenheit

Constructors

DegreesFahrenheit

Fields

fahrenheits :: Double

newtype DegreesTemperature

Constructors

DegreesCelcius

Fields

celciuses :: Double

newtype Temperature

Constructors

DegreesKelvin

Fields

kelvins :: Double

newtype Substance

Constructors

Moles

Fields

moles :: Double

newtype Intensity

Constructors

Candelas

Fields

candelas :: Double

newtype Frequency

Constructors

Hertz

Fields

hertzs :: Double

newtype Force

Constructors

Newtons

Fields

newtons :: Double

newtype Torque

Constructors

NewtonMeters

Fields

newtonmeters :: Double

newtype Pressure

Constructors

Pascals

Fields

pascals :: Double

newtype Energy

Constructors

Joules

Fields

joules :: Double

newtype Power

Constructors

Watts

Fields

watts :: Double

newtype Charge

Constructors

Coulombs

Fields

coulombs :: Double

newtype Voltage

Constructors

Volts

Fields

volts :: Double

newtype Capacitance

Constructors

Farads

Fields

farads :: Double

newtype Resistance

Constructors

Ohms

Fields

ohms :: Double

newtype Conductance

Constructors

Siemenses

Fields

siemenses :: Double

newtype Flux

Constructors

Webers

Fields

webers :: Double

newtype FluxDensity

Constructors

Teslas

Fields

teslas :: Double

newtype Inductance

Constructors

Henrys

Fields

henrys :: Double

newtype LuminousFlux

Constructors

Lumens

Fields

lumens :: Double

newtype Illuminance

Constructors

Luxes

Fields

luxes :: Double

newtype Radioactivity

Constructors

Becquerels

Fields

becquerels :: Double

newtype AbsorbedDose

Constructors

Grays

Fields

grays :: Double

newtype EquivalentDose

Constructors

Sieverts

Fields

sieverts :: Double

newtype CatalyticActivity

Constructors

Katals

Fields

katals :: Double

Orphan instances

Unit Double

Instance details

Methods

amount :: Double -> Scalar Double

fromAmount :: Scalar Double -> Double

fromQuantity :: (Alternative m, Monad m) => Quantity (Scalar Double) -> m Double

unitOf :: Double -> String

dimension :: Double -> Dimension

zeroAmount :: (Scalar Double -> Double) -> Scalar Double

conversionFactor :: (Scalar Double -> Double) -> Scalar Double

Unit Float

Instance details

Methods

amount :: Float -> Scalar Float

fromAmount :: Scalar Float -> Float

fromQuantity :: (Alternative m, Monad m) => Quantity (Scalar Float) -> m Float

unitOf :: Float -> String

dimension :: Float -> Dimension

zeroAmount :: (Scalar Float -> Float) -> Scalar Float

conversionFactor :: (Scalar Float -> Float) -> Scalar Float

Unit R

Instance details

Methods

amount :: R -> Scalar R

fromAmount :: Scalar R -> R

fromQuantity :: (Alternative m, Monad m) => Quantity (Scalar R) -> m R

unitOf :: R -> String

dimension :: R -> Dimension

zeroAmount :: (Scalar R -> R) -> Scalar R

conversionFactor :: (Scalar R -> R) -> Scalar R