Math.Tools

Documentation

class (Category ar, Category ar') => ArrowTransformation ar ar' where

Methods

mapA :: ar a b -> ar' a b

class (Category arr, Monoid m) => MonoidArrow arr m a where

Methods

monoidA :: m -> arr a a

class Category arr => MonoidalCategory (arr :: k -> k -> *) where

https://en.wikipedia.org/wiki/Monoidal_category

Associated Types

type Prod (arr :: k -> k -> *) (a :: k) (b :: k) :: k

type MUnit (arr :: k -> k -> *) :: k

Methods

(-*-) :: arr a b -> arr a' b' -> arr (Prod arr a a') (Prod arr b b')

monoidal_assoc :: arr (Prod arr (Prod arr a b) c) (Prod arr a (Prod arr b c))

monoidal_deassoc :: arr (Prod arr a (Prod arr b c)) (Prod arr (Prod arr a b) c)

leftunitor :: arr (Prod arr (MUnit arr) a) a

unleftunitor :: arr a (Prod arr (MUnit arr) a)

rightunitor :: arr (Prod arr a (MUnit arr)) a

unrightunitor :: arr a (Prod arr a (MUnit arr))

class MonoidalCategory arr => BraidedCategory (arr :: k -> k -> *) where

Methods

braiding :: arr (Prod arr a b) (Prod arr b a)

class Category c => OverCategory c y where

Methods

overmap :: c a b -> c b y -> c a y

class (Category arr, Category arr') => OppositeCategoryAction arr arr' where

Methods

categoryA :: arr a' a -> arr' x a -> arr' x a'

class Category arr => FailureArrow arr err where

Methods

failA :: arr err b

catchA :: arr a b -> arr err b -> arr a b

class Category arr => Groupoid arr where

Methods

invertA :: arr a b -> arr b a

class Category arr => RelativeMonad arr j m where

https://ncatlab.org/nlab/show/relative+monad

Methods

relativeUnitA :: arr (j x) (m x)

extA :: arr (j x) (m y) -> arr (m x) (m y)

relativeBind :: (RelativeMonad arr j m, ArrowApply arr) => arr (m x, arr (j x) (m y)) (m y)

relativeBind2 :: (RelativeMonad arr j m, RelativeMonad arr k m, ArrowApply arr) => arr (j a, k b) (m c) -> arr (m a, m b) (m c)

class (Category arr, Category arr') => OpArrow p arr arr' where

Methods

inverse_imageA :: arr a b -> arr' (p b) (p a)

class Category arr => CoArrow arr where

Methods

coarr :: (b -> a) -> arr a b

coleft :: arr a b -> arr (Either a d) (Either b d)

coright :: arr a b -> arr (Either d a) (Either d b)

class Category arr => BiArrow arr where

Methods

(<->) :: (a -> b) -> (b -> a) -> arr a b

class Category arr => LinearBiArrow arr m where

Methods

(<==>) :: (a %m -> b) -> (b %m -> a) -> arr a b

class (Category arr, Category m) => MonoidCategory arr m f | f -> m arr where

Methods

monoidArr :: m a a -> arr (f a) (f a)

class (BiArrow arr, FunctorArrow f arr arr) => ApplicativeBiArrow arr f where

Minimal complete definition

bipure, biliftA2

Methods

bipure :: arr a (f a)

bimap :: arr (f (arr a b), f a) (f b)

biliftA2 :: arr a (arr b c) -> arr (f a, f b) (f c)

class Category arr => CPSArrow arr where

Methods

callcc :: (forall bot. arr c bot -> arr b c) -> arr b c

class (Category arr, Category arr') => FunctorArrow f arr arr' | f arr' -> arr, f arr -> arr' where

Arrow version of Functor from Haskell Prelude

Methods

amap :: arr c d -> arr' (f c) (f d)

class (FunctorArrow f arr arr, Category arr) => ApplicativeArrow f arr where

Arrow version of Applicative from Haskell Prelude

Methods

apure :: arr a (f a)

aapply :: f (arr a b) -> arr (f a) (f b)

alift2 :: (ApplicativeArrow f arr, ArrowApply arr) => f (arr a (arr b c)) -> arr (f a, f b) (f c)

class (Category arr, Category arr') => CoFunctorArrow f arr arr' where

Methods

inverseImageArr :: arr c d -> arr' (f d) (f c)

(|>) :: CoFunctorArrow f (->) (->) => f d -> (c -> d) -> f c

class (Category arr, Category arr') => AdjunctionArrow arr arr' f g | f -> g, g -> f, arr -> arr', arr' -> arr where

Minimal complete definition

ladA, radA

Methods

ladA :: arr (f a) b -> arr' a (g b)

radA :: arr' a (g b) -> arr (f a) b

unitA :: arr' a (g (f a))

counitA :: arr (f (g b)) b

outerA :: (Arrow arr2, FunctorArrow f arr arr', FunctorArrow g arr2 arr, ArrowApply arr, ArrowApply arr') => arr2 (a, b) c -> arr' (f a, g b) (f (g c))

class FunctorArrow f arr arr => InitialAlgebraA arr a f | a -> f where

Methods

deconstructA :: arr a (f a)

class Category a => LoopArrow a prod where

Methods

loop_secondSA :: a (prod b d) (prod c d) -> a b c

loop_firstSA :: a (prod d b) (prod d c) -> a b c

class Category a => MessagingArrow a where

Methods

splitA :: a (Either b c) (Either d e) -> (a b d, a c e)

class Category a => UnifyingArrow a where

Methods

joinA :: a (Either b b) b

class Category a => DoublingArrow a where

Methods

doubleA :: a b (b, b)

class Category a => CircuitArrow a where

Methods

delayA :: b -> a b b

class Category a => StorageArrow a where

Methods

storageA :: a b c -> (a b (), a () c)

class Category a => ArrowUnit a where

Methods

aunit :: a b ()

class Category a => ArrowFix a prod where

Methods

fixArrow :: a (prod b c) c -> a b c

class Category a => ArrowCircuit a where

Methods

delay :: b -> a b b

class Category a => ArrowDual a d where

Methods

collapseArrow :: d (d a) b c -> a b c

class ArrowDual a d => RecArrow a d where

Methods

recArrow :: a b c -> d a b c

class ArrowDual a d => ArrowDualApp a d where

Methods

appArrow :: a (a b c, d a b c) c

module Math.Tools.Adjunction

module Math.Tools.CoFunctor

module Math.Tools.CoMonad

module Math.Tools.Complex

module Math.Tools.Expression

class BiFunctor op

Minimal complete definition

bimap

class Functor f => InterleaveFunctor f where

Methods

interleave :: f a -> f a -> f a infixr 8

(<&>) :: Applicative f => f a -> f b -> f (a, b) infixl 8

unzipWith :: Functor f => (a -> (b, c)) -> f a -> (f b, f c)

fst3 :: (a, b, c) -> a

snd3 :: (a, b, c) -> b

funzip :: Functor f => f (a, b) -> (f a, f b)

thrd3 :: (a, b, c) -> c

funzip3 :: Functor f => f (a, b, c) -> (f a, f b, f c)

fzipWith3 :: Applicative f => (a -> b -> c -> d) -> f a -> f b -> f c -> f d

fzip :: Applicative f => f a -> f b -> f (a, b)

fzip3 :: Applicative f => f a -> f b -> f c -> f (a, b, c)

fzip4 :: Applicative f => f a -> f b -> f c -> f d -> f (a, b, c, d)

liftA4 :: Applicative f => (a -> b -> c -> d -> e) -> f a -> f b -> f c -> f d -> f e

module Math.Tools.Identity

module Math.Tools.Id

module Math.Tools.I

module Math.Tools.Integer

module Math.Tools.Integer

module Math.Tools.Isomorphism

module Math.Tools.LineInfo

module Math.Tools.List

module Math.Tools.Map

module Math.Tools.Map

module Math.Tools.Maybe

module Math.Tools.Median

module Math.Tools.Monad

module Math.Tools.Monoid

module Math.Tools.NaturalNumber

module Math.Tools.NaturalTransformation

module Math.Tools.Orthogonal

module Math.Tools.ParseMonad

module Math.Tools.ParserInterface

module Math.Tools.Parser

module Math.Tools.ParsingCombinators

module Math.Tools.PrettyP

newtype Prop a

We must represent classical logic with characteristic functions, because of the excluded middle axiom, and because characteristic functions (as properties) have complex relationship with possible elements of a. See the Universe type class for a solution. For exposition of classical logic used here, see "Lawvere,Rosebrugh: Sets for mathematics".

Constructors

Characteristic

Fields

runCharacteristic :: a -> Bool

subseteq :: Universe a => Prop a -> Prop a -> Bool

intersects :: Universe a => Prop a -> Prop a -> Bool

splitProp :: Universe a => Prop a -> ([a], [a])

See "Soare: Recursively enumerable sets and degrees", this is split of recursive into two recursively enumerable sets.

propositionalTable :: Universe a => Prop a -> [(a, Bool)]

fromBool :: Bool -> Prop a

propSubstitute :: Contravariant p => p a -> p (a, b)

propExists :: (Adjunction f p, Contravariant p) => f (p a) -> (a, b)

propExistsSnd :: (Adjunction f p, Contravariant p) => f (p b) -> (a, b)

propAlways :: (Adjunction p g, Contravariant p) => a -> g (p (a, b))

propAlwaysSnd :: (Adjunction p g, Contravariant p) => b -> g (p (a, b))

axiomN :: (ModalLogic p, ImplicativeLogic p) => p a -> p a

https://en.wikipedia.org/wiki/Modal_logic

axiomK :: (ModalLogic p, ImplicativeLogic p) => p a -> p a -> p a

https://en.wikipedia.org/wiki/Modal_logic

axiom4 :: (ModalLogic p, ImplicativeLogic p) => p a -> p a

https://en.wikipedia.org/wiki/Modal_logic

axiomB :: (ModalLogic p, ImplicativeLogic p) => p a -> p a

https://en.wikipedia.org/wiki/Modal_logic

axiomD :: (ModalLogic p, ImplicativeLogic p) => p a -> p a

https://en.wikipedia.org/wiki/Modal_logic

axiom5 :: (ModalLogic p, ImplicativeLogic p) => p a -> p a

https://en.wikipedia.org/wiki/Modal_logic

axiomT :: (ModalLogic p, ImplicativeLogic p) => p a -> p a

https://en.wikipedia.org/wiki/Modal_logic

systemK :: (ModalLogic p, ImplicativeLogic p) => p a -> p a -> p a

systemT :: (ModalLogic p, ImplicativeLogic p) => p a -> p a -> p a

systemS4 :: (ModalLogic p, ImplicativeLogic p) => p a -> p a -> p a

systemS5 :: (ModalLogic p, ImplicativeLogic p) => p a -> p a -> p a

systemD :: (ModalLogic p, ImplicativeLogic p) => p a -> p a -> p a

vectorEqualizer :: (Eq (n a), LinearTransform m n a) => (m :*: n) a -> (m :*: n) a -> Prop (m a)

list_truthvalues :: Universe a => Prop a -> [(a, Bool)]

It is impossible to convert characteristic index into recursively enumerable index (indexing of r.e. sets), or into canonical index (which has one bit per element describing membership). Thus we must use the Universe constraint.

apply :: Prop (Prop b, b)

result :: Eq a => a -> Prop a

result and the bind operations provide Monad instance, except for the | fact that the Universe and Eq constraints are required. Thus an adjunction instance also exists.

existentialBind :: Universe a => Prop a -> (a -> Prop b) -> Prop b

existentialJoin :: (Universe b, Eq b) => Prop (Prop b) -> Prop b

universalBind :: Universe a => Prop a -> (a -> Prop b) -> Prop b

universalJoin :: (Universe b, Eq b) => Prop (Prop b) -> Prop b

propFixedpoint :: (Prop a -> Prop a) -> Prop a

truthvalue :: Bool -> Prop a

satisfiable :: Universe a => Prop a -> Bool

Satisfiable returns true if there exists an element of the universe such that the proposition is satisfied. Note symmetry with axiom.

axiom :: Universe a => Prop a -> Bool

axiom returns true, if the proposition holds for all elements of the universe.

notMember :: a -> Prop a -> Bool

subseteqProp :: Prop a -> Prop a -> Prop a

\[x \in subseteq i j \iff x \in i \implies x \in j\]

isSection :: Prop a -> (Bool -> a) -> Bool

doublePropJoin :: Prop (Prop (Prop (Prop a))) -> Prop (Prop a)

fromAssoc :: Eq a => [(a, Bool)] -> Prop a

fromAssocWith :: Eq a => [(a, b)] -> Prop b -> Prop a

prop_fromList :: Eq a => [a] -> Prop a

prop_fromSet :: Ord a => Set a -> Prop a

prop_fromMap :: (Ord i, Eq e) => Map i e -> Prop (i, e)

map_prop :: (Eq b, Universe a) => (a -> b) -> Prop a -> Prop b

Due to constraints, Prop cannot be made as instance of Functor.

equalizer_fixedpoint :: HasEqualizers p a => (a -> a) -> p a

invertible :: HasEqualizers p a => (b -> a) -> (a -> b) -> p a

kernel :: (Num c, HasEqualizers p c) => (a -> c) -> p a

ternary_op :: (Bool -> Bool -> Bool -> Bool) -> Prop a -> Prop a -> Prop a -> Prop a

binary_op :: (Bool -> Bool -> Bool) -> Prop a -> Prop a -> Prop a

gen_binary_op :: (b -> b' -> c) -> (a -> b) -> (a -> b') -> a -> c

binary :: Relational p => (a -> b -> p c) -> p (a, (b, c))

evaluate :: a -> Prop (Prop a)

evaluate' :: (Universe a, Eq a) => a -> Prop (Prop a)

integrate_prop :: Universe a => Prop a -> Prop (Prop a)

intersect :: Prop a -> Prop a -> Prop a

intersect3 :: Prop a -> Prop a -> Prop a -> Prop a

median3 :: Prop a -> Prop a -> Prop a -> Prop a

union :: Prop a -> Prop a -> Prop a

union3 :: Prop a -> Prop a -> Prop a -> Prop a

excluded_middle :: Universe b => Prop b -> Bool

noncontradiction :: Universe b => Prop b -> Bool

demorgan_union :: Universe a => Prop a -> Prop a -> Bool

demorgan_intersect :: Universe a => Prop a -> Prop a -> Bool

disjunction_commutative :: Universe a => Prop a -> Prop a -> Bool

conjunction_commutative :: Universe a => Prop a -> Prop a -> Bool

graph :: Eq b => (a -> b) -> Prop (a, b)

\[graph(f) = [<x,y> | y = f(x)]\]

image_factorization :: (Eq b, Universe a) => (a -> b) -> Prop b

bind :: Universe b => (a -> Prop b) -> (b -> Prop c) -> a -> Prop c

prop_compose :: Universe b => Prop (a, b) -> Prop (b, c) -> Prop (a, c)

transitive :: Universe a => Prop (a, a) -> Bool

\[<x,z> \in transitive \mathbb{R} \iff (\forall y. <x,y> \in \mathbb{R} & <y,z> \in \mathbb{R} \implies <x,z> \in \mathbb{R})\]

antisymmetric :: Eq a => Prop (a, a) -> Prop (a, a)

propIsoTerminal :: Prop ((), b) -> Prop b

propIsoTerminalInverse :: Prop b -> Prop ((), b)

fiber :: HasEqualizers Prop c => (b -> c) -> c -> Prop b

\[fiber(p,y) = { x | p(x) = y } = p^{-1}({y})\]

fiber_ :: Eq b => (a -> b) -> b -> Prop a

equivalenceRelation :: HasEqualizers Prop b => (a -> b) -> Prop (a, a)

reflexive :: Eq a => Prop (a, a)

reflexiveClosure :: Eq a => Prop (a, a) -> Prop (a, a)

symmetricClosure :: Prop (a, a) -> Prop (a, a)

transitiveStep :: Universe a => Prop (a, a) -> Prop (a, a)

transitiveClosure :: Universe a => Prop (a, a) -> Prop (a, a)

equivalenceGeneratedBy :: (Eq a, Universe a) => Prop (a, a) -> Prop (a, a)

opposite :: Prop (a, b) -> Prop (b, a)

leftAdjunctProp :: (a -> Prop b) -> b -> Prop a

Adjunction \[P^{op} \dashv P\] between functors \[P\] and \[P^{op}\], where \[P : C^{op} \Rightarrow C\] and \[P^{op} : C \Rightarrow C^{op}\]. This is based on symmetry of product in relations.

unitProp :: b -> Prop (Prop b)

relProp :: (a -> Bool) -> Prop a

rightAdjunct of that adjunction is difficult, because it would be in C^{op}, and we don't have that category here. So how could we implement that? We'd need opposite arrow, but then the implementation would be exactly same as leftAdjunct_prop and unit_prop, but lifted to opposite category. I'm not implementing that now.

binaryRelProp :: (a -> a -> Bool) -> Prop (a, a)

binaryRel :: (a -> a -> Bool) -> a -> Prop a

http://nlab.mathforge.org/nlab/show/well-founded+relation

equalFunctions :: (Universe a, Eq b) => Prop (a -> b, a -> b)

testEqualityAt :: a -> Prop (Prop a, Prop a)

equalizerProp :: Prop a -> Prop a -> Prop a

pullbackProp :: Prop a -> Prop b -> Prop (a, b)

sum :: Prop a -> Prop b -> Prop (Either a b)

propAnd :: Prop (Either a b) -> Prop (a, b)

propOr :: Prop (Either a b) -> Prop (a, b)

propIf :: Prop a -> Prop (Either a a) -> Prop a

propIfGeneral :: Prop a -> Prop (Either b c) -> Prop (a, b, c)

equal :: Eq a => Prop (a, a)

notEqual :: Eq a => Prop (a, a)

lessthan :: Ord a => Prop (a, a)

lesseq :: Ord a => Prop (a, a)

orListProp :: Prop a -> Prop [a]

andListProp :: Prop a -> Prop [a]

enumerate :: Universe a => Prop a -> [a]

existentialMap :: (Universe a, Eq b) => (a -> b) -> Prop a -> Prop b

universalMap :: (Universe a, Eq b) => (a -> b) -> Prop a -> Prop b

\[universal\_map( f, p ) = \{ b | p <= f^{-1}(\{b\}) \}\]

propExistential :: Universe b => Prop (a, b) -> Prop a

prop_existential is basically the same as 'existential_map fst', except for performance. 'existential_map fst' will need to iterate over all pairs \[<a,b>\], whereas prop_existential will only iterate through all elements of b only. Similarly for prop_universal.

propUniversal :: Universe b => Prop (a, b) -> Prop a

image :: Universe a => Prop (a, b) -> Prop b

propUniversalImage :: (Universe a, Universe b, Eq b) => Prop (a, b) -> Prop b

forevery :: Universe b => Prop b -> Prop ()

newtype b :\: a

Constructors

PSub

Fields

runpsub :: PProp a -> PProp b

data PProp a where

Constructors

PProp :: Prop a -> PProp a

PNeg :: (Prop a -> Prop b) -> PProp (b :\: a)

pnegFunc :: (a -> b) -> PProp (a :\: b)

module Math.Tools.Tree

module Math.Tools.Universe

module Math.Tools.Visitor