Math.Tools

Documentation

module Math.Tools.Adjunction

class ArrowDual a d => ArrowDualApp a d where

Minimal complete definition

appArrow

Methods

appArrow :: a (a b c, d a b c) c

class ArrowDual a d => RecArrow a d where

Minimal complete definition

recArrow

Methods

recArrow :: a b c -> d a b c

class ArrowDual a d where

Minimal complete definition

collapseArrow

Methods

collapseArrow :: d (d a) b c -> a b c

class Category a => ArrowCircuit a where

Minimal complete definition

delay

Methods

delay :: b -> a b b

class Category a => ArrowFix a where

Minimal complete definition

fixArrow

Methods

fixArrow :: a (b, c) c -> a b c

class Category a => ArrowUnit a where

Minimal complete definition

aunit

Methods

aunit :: a b ()

class Category a => StorageArrow a where

Minimal complete definition

storageA

Methods

storageA :: a b c -> (a b (), a () c)

class Category a => CircuitArrow a where

Minimal complete definition

delayA

Methods

delayA :: b -> a b b

class Category a => DoublingArrow a where

Minimal complete definition

doubleA

Methods

doubleA :: a b (b, b)

class Category a => UnifyingArrow a where

Minimal complete definition

joinA

Methods

joinA :: a (Either b b) b

class Category a => MessagingArrow a where

Minimal complete definition

splitA

Methods

splitA :: a (Either b c) (Either d e) -> (a b d, a c e)

class LoopArrow a where

Minimal complete definition

loop_secondSA, loop_firstSA

Methods

loop_secondSA :: a (b, d) (c, d) -> a b c

loop_firstSA :: a (d, b) (d, c) -> a b c

class FunctorArrow f arr => InitialAlgebraA arr a f | a -> f where

Minimal complete definition

deconstructA

Methods

deconstructA :: arr a (f a)

class (Category arr, Category arr') => AdjunctionArrow arr arr' f g | f -> g, g -> f, arr -> arr', arr' -> arr where

Minimal complete definition

ladA, radA

Methods

ladA :: arr (f a) b -> arr' a (g b)

radA :: arr' a (g b) -> arr (f a) b

unitA :: arr' a (g (f a))

counitA :: arr (f (g b)) b

class CoFunctorArrow (f :: * -> *) (arr :: * -> * -> *) where

Minimal complete definition

inverseImage

Methods

inverseImage :: arr c d -> arr (f d) (f c)

class FunctorArrow f arr => ApplicativeArrow (f :: * -> *) (arr :: * -> * -> *) where

Arrow version of Applicative from Haskell Prelude

Minimal complete definition

apure, aapply

Methods

apure :: arr a (f a)

aapply :: f (arr a b) -> arr (f a) (f b)

class FunctorArrow (f :: * -> *) (arr :: * -> * -> *) where

Arrow version of Functor from Haskell Prelude

Minimal complete definition

amap

Methods

amap :: arr c d -> arr (f c) (f d)

class Category arr => CPSArrow arr where

Minimal complete definition

callcc

Methods

callcc :: (forall bot. arr c bot -> arr b c) -> arr b c

class Category arr => BiArrow arr where

Minimal complete definition

(<->)

Methods

(<->) :: (a -> b) -> (b -> a) -> arr a b

class Category arr => CoArrow arr where

Minimal complete definition

coarr, coleft, coright

Methods

coarr :: (b -> a) -> arr a b

coleft :: arr a b -> arr (Either a d) (Either b d)

coright :: arr a b -> arr (Either d a) (Either d b)

class OpArrow p arr arr' where

Minimal complete definition

inverse_imageA

Methods

inverse_imageA :: arr a b -> arr' (p b) (p a)

class Category arr => Groupoid arr where

Minimal complete definition

invertA

Methods

invertA :: arr a b -> arr b a

class Category arr => FailureArrow arr err where

Minimal complete definition

failA, catchA

Methods

failA :: arr err b

catchA :: arr a b -> arr err b -> arr a b

class (Category arr, Category arr') => OppositeCategoryAction arr arr' where

Minimal complete definition

categoryA

Methods

categoryA :: arr a' a -> arr' x a -> arr' x a'

class Category c => OverCategory c y where

Minimal complete definition

overmap

Methods

overmap :: c a b -> c b y -> c a y

class MonoidalCategory arr => BraidedCategory arr where

Minimal complete definition

braiding

Methods

braiding :: arr (Prod arr a b) (Prod arr b a)

class Category arr => MonoidalCategory arr where

https://en.wikipedia.org/wiki/Monoidal_category

Minimal complete definition

(-*-), assoc, deassoc, leftunitor, unleftunitor, rightunitor, unrightunitor

Associated Types

type Prod arr a b

type MUnit arr

Methods

(-*-) :: arr a b -> arr a' b' -> arr (Prod arr a a') (Prod arr b b')

deassoc :: arr (Prod arr a (Prod arr b c)) (Prod arr (Prod arr a b) c)

leftunitor :: arr (Prod arr (MUnit arr) a) a

unleftunitor :: arr a (Prod arr (MUnit arr) a)

rightunitor :: arr (Prod arr a (MUnit arr)) a

unrightunitor :: arr a (Prod arr a (MUnit arr))

class (Category arr, Monoid m) => MonoidArrow arr m a where

Minimal complete definition

monoidA

Methods

monoidA :: m -> arr a a

class (Category ar, Category ar') => ArrowTransformation ar ar' where

Minimal complete definition

mapA

Methods

mapA :: ar a b -> ar' a b

outerA :: (FunctorArrow f arr, FunctorArrow g arr, ArrowApply arr) => arr (a, b) c -> arr (f a, g b) (f (g c))

class CoFunctor p => Universal p where

Minimal complete definition

exist, univ

Methods

exist :: Universe b => p (a, b) -> p a

univ :: Universe b => p (a, b) -> p a

class CoFunctor p => UniversalLogic p where

Minimal complete definition

singleton, all_members

Methods

singleton :: Eq a => a -> p a

all_members :: Universe a => p a

class CoFunctor p => BooleanLogic p where

Minimal complete definition

true

Methods

true :: p a

class CoFunctor p => ImplicativeLogic p where

Minimal complete definition

(-=>-)

Methods

(-=>-) :: p a -> p a -> p a

class CoFunctor p => SubtractiveLogic p where

Minimal complete definition

(-\-)

Methods

(-\-) :: p a -> p a -> p a

class CoFunctor p => Continuous p where

Minimal complete definition

expand

Methods

expand :: p (Either a b) -> (p a, p b)

class PropositionalLogic p => ModalLogic p where

https://en.wikipedia.org/wiki/Modal_logic

Methods

eventually :: p a -> p a

always :: p a -> p a

class CoFunctor p => PropositionalLogic p where

Minimal complete definition

(-&-), (-|-), invert

Methods

(-&-) :: p a -> p a -> p a

(-|-) :: p a -> p a -> p a

invert :: p a -> p a

class CoFunctor p => HasVariables p t where

Minimal complete definition

variable

Methods

variable :: String -> p t

class HasTerminalObject t

Minimal complete definition

terminal

class Functor p => Topology p where

Minimal complete definition

arbitrary_union, finite_intersect

Methods

arbitrary_union :: [p a] -> p a

finite_intersect :: [p a] -> p a

class (Eq c, CoFunctor p) => HasEqualizers p c where

Minimal complete definition

equalizer

Methods

equalizer :: (a -> c) -> (a -> c) -> p a

pullback :: (a -> c) -> (b -> c) -> p (a, b)

class CoFunctor p => Difference p minus | p -> minus where

Minimal complete definition

difference, undifference

Methods

difference :: (p a -> p b) -> p (b `minus` a)

undifference :: p (b `minus` a) -> p a -> p b

class CoFunctor p => Existential p where

Minimal complete definition

direct_image, universal_image

Methods

direct_image :: (Universe a, Eq b) => (a -> b) -> p a -> p b

universal_image :: (Universe a, Eq b) => (a -> b) -> p a -> p b

class CoFunctor p => Relational p where

Minimal complete definition

relation, unrelation

Methods

relation :: (a -> p b) -> p (a, b)

unrelation :: p (a, b) -> a -> p b

class Propositional p => BinaryLogic p where

Minimal complete definition

toBool

Methods

toBool :: p () -> Bool

class Propositional p where

Minimal complete definition

runProposition

Methods

runProposition :: p a -> a -> p ()

class Representable p where

Minimal complete definition

represent, (!>)

Associated Types

type Representation p :: * -> *

Methods

represent :: Representation p a -> p a

(!>) :: p a -> (Representation p a -> p a) -> p a

class CoFunctor p where

Minimal complete definition

inverse_image

Methods

inverse_image :: (a -> b) -> p b -> p a

(|>>) :: p b -> (a -> b) -> p a

member :: Propositional p => a -> p a -> p ()

isIn :: BinaryLogic p => a -> p a -> Bool

isNotIn :: BinaryLogic p => a -> p a -> Bool

terminal_part :: Propositional p => p () -> p ()

propositional_implication :: PropositionalLogic p => p a -> p a -> p a

relative_complement :: PropositionalLogic p => p a -> p a -> p a

intersect_list :: (BooleanLogic p, PropositionalLogic p) => [p a] -> p a

union_list :: (BooleanLogic p, PropositionalLogic p) => [p a] -> p a

characteristic :: BooleanLogic p => (a -> p ()) -> p a

false_prop :: (BooleanLogic p, PropositionalLogic p) => p a

(-<=>-) :: (ImplicativeLogic p, PropositionalLogic p) => p a -> p a -> p a

symmetric_difference :: (PropositionalLogic p, SubtractiveLogic p) => p a -> p a -> p a

boundary :: (PropositionalLogic p, BooleanLogic p, SubtractiveLogic p) => p a -> p a

(|<<) :: CoFunctor p => p b -> (a -> b) -> p a infixl 1

curry_subst :: CoFunctor p => ((a, b) -> c) -> p (b -> c) -> p a

uncurry_subst :: CoFunctor p => (a -> b -> c) -> p c -> p (a, b)

dualize :: Difference p minus => (a -> b) -> p (a `minus` b)

coin1 :: CoFunctor p => p a -> p (a, b)

coin2 :: CoFunctor p => p b -> p (a, b)

cofst :: CoFunctor p => p (Either a b) -> p a

cosnd :: CoFunctor p => p (Either a b) -> p b

cocase :: CoFunctor p => (Either a b -> c) -> p c -> (p a, p b)

copair :: (Propositional p, BooleanLogic p) => p a -> p b -> p (Either a b)

pair_subst :: CoFunctor p => (a -> b) -> (a -> c) -> p (b, c) -> p a

graph_subst :: CoFunctor p => (a -> b) -> p (a, b) -> p a

module Math.Tools.CoMonad

module Math.Tools.Complex

module Math.Tools.Expression

module Math.Tools.FixedPoint

module Math.Tools.Functor

class Monad g => Group g where

Methods

greturn :: top -> g top

gjoin :: g (g top) -> g top

glift :: (a -> b) -> a -> g b

gbind :: (a -> g b) -> g a -> g b

gjoin3 :: Group g => g (g (g top)) -> g top

gmap :: Group g => (a -> b) -> g (g a) -> g b

join_nested :: (Group g1, Group g) => (g1 a -> g (g b)) -> a -> g b

group_leftAdjunct :: (Group g, Group g1) => (g1 a -> c) -> a -> g c

group_flip :: (Group f, Group g) => (f a -> g (g b)) -> f (f a) -> g b

group_join_self :: Group g => g a -> g a

module Math.Tools.Identity

module Math.Tools.Id

module Math.Tools.I

newton_square_root_list :: Integral a => a -> [a]

Some numbers may cause newton_square_root_list to produce an infinite sequence such as newton_square_root_list 48 == [1,24,13,8,7,8,7,...] We guard against this by terminating if the difference between two adjacent results is less than 2.

gcd_ :: Integral a => a -> a -> a

rising_factorial_power :: (Num a, Eq a) => a -> a -> a

See Knuth: TAOCP, or http://www.math.harvard.edu/~mazur/preprints/when_is_one.pdf

falling_factorial_power :: (Num a, Eq a) => a -> a -> a

See Knuth: TAOCP or http://www.math.harvard.edu/~mazur/preprints/when_is_one.pdf

binomial :: (Enum a, Fractional a) => a -> a -> a

http://en.wikipedia.org/wiki/Binomial_coefficient

module Math.Tools.Isomorphism

module Math.Tools.LineInfo

interleave2 :: [[a]] -> [a]

shift :: Int -> [a] -> [a]

shift rotates the first k elements (and leaves the rest of the list intact)

selections :: [a] -> [(a, [a])]

From discussion in Haskell-Cafe mailing list by Koen Claessen: https://mail.haskell.org/pipermail/haskell-cafe/2002-June/003122.html

twice_selections :: [a] -> [(a, a, [a])]

permutations :: [a] -> [[a]]

From discussion in Haskell-Cafe mailing list by Koen Claessen: https://mail.haskell.org/pipermail/haskell-cafe/2002-June/003122.html

all_maps :: Ord k => [k] -> [a] -> [Map k a]

all_total_maps :: Ord k => [k] -> [a] -> [Map k a]

lookupMapM :: (Monad m, Show i, Ord i) => Map i a -> i -> m a

unzip :: Map i (a, b) -> (Map i a, Map i b)

unzip3 :: Map i (a, b, c) -> (Map i a, Map i b, Map i c)

foldM :: Monad m => (a -> b -> m b) -> m b -> Map i a -> m b

foldKeyM :: Monad m => (i -> a -> b -> m b) -> m b -> Map i a -> m b

mapM :: (Monad m, Ord i) => (a -> m b) -> Map i a -> m (Map i b)

sequence_ :: (Monad m, Ord i) => Map i (m a) -> m ()

sequenceByKey :: (Monad m, Ord i) => Map i (i -> m a) -> m (Map i a)

joinMap :: Map String (Map String a) -> Map String a

splitMapByKey :: (Ord s, Ord m) => Map (m, s) a -> Map m (Map s a)

partitionMap :: Ord k => Map k a -> (a -> Either b c) -> (Map k b, Map k c)

intersectZipMapA :: (Ord i, ArrowChoice arr) => arr (Map i a, Map i b) (Map i (a, b))

intersectMapA :: (Ord i, ArrowChoice arr) => arr (a, b) c -> arr (Map i a, Map i b) (Map i c)

zipWithMapM :: (Monad m, Ord i) => (a -> b -> m c) -> Map i a -> Map i b -> m (Map i c, Map i a, Map i b)

intersectMapKeyA :: (Ord i, ArrowChoice arr) => arr (i, (a, b)) c -> arr (Map i a, Map i b) (Map i c)

mapMapWithKey :: (ArrowChoice arr, Eq i) => arr (i, e) (i, f) -> arr (Map i e) (Map i f)

unionMapA :: (Ord i, ArrowChoice arr) => arr (a, a) a -> arr (Map i a, Map i a) (Map i a)

sequenceMap :: Ord i => Map i (IO b) -> IO (Map i b)

sequenceMapA :: (Arrow arr, Ord i) => Map i (arr a b) -> arr a (Map i b)

sequenceMapWithKeyA :: (Arrow arr, Ord i) => Map i (arr (a, i) b) -> arr a (Map i b)

foldMapA :: ArrowApply arr => arr (a, b) b -> arr (Map i a, b) b

foldMapKeyA :: ArrowApply arr => arr (i, a, b) b -> arr (Map i a, b) b

composeMapA :: (Arrow arr, Ord i, Ord a) => arr (Map i a, Map a b) (Map i b)

module Math.Tools.Maybe

module Math.Tools.Median

module Math.Tools.Monad

module Math.Tools.Monoid

class Naturalistic n where

Minimal complete definition

zero_natu, plus_natu

Methods

zero_natu :: n

plus_natu :: n -> n -> n

module Math.Tools.NaturalNumber

module Math.Tools.NaturalTransformation

new_function :: Functor f => (a -> f a) -> (f c -> c) -> a -> c

fold_func :: Functor f => (a -> f a) -> (f c -> f c -> c) -> a -> a -> c

fbind :: Functor f => (f c -> c) -> (a -> f a) -> a -> c

outer'' :: (Functor f1, Functor f2) => (t -> a -> b) -> f1 t -> f2 a -> f1 (f2 b)

outer :: (Functor f, Functor g) => (a -> b -> c) -> f a -> g b -> f (g c)

outer_left :: (Functor f, Functor g) => f (a -> b) -> g a -> f (g b)

outer_right :: (Functor f, Functor g) => f a -> g (a -> b) -> f (g b)

list_dotproduct :: Num a => [a] -> [a] -> a

outer' :: Functor f => (a -> f b -> b) -> f a -> f (f b)

full_outer :: (Functor f, Functor g) => f a -> (a -> g b) -> (a -> b -> c) -> f (g c)

self :: Functor f => (a -> a -> b) -> f a -> f (f b)

mapf :: Functor f => f a -> (a -> b) -> f b

outer_product :: (Functor f, Functor g) => f a -> g b -> f (g (a, b))

outer_apply :: (Functor f, Functor g) => f (b -> c) -> g b -> f (g c)

outer4 :: (Functor f, Functor g, Functor h, Functor i) => (a -> b -> c -> d -> e) -> f a -> g b -> h c -> i d -> f (g (h (i e)))

module Math.Tools.ParseMonad

module Math.Tools.ParserInterface

module Math.Tools.Parser

module Math.Tools.ParsingCombinators

module Math.Tools.PrettyP

module Math.Tools.Tree

module Math.Tools.Universe

module Math.Tools.Visitor