Math.Tools

Documentation

module Math.Tools.Adjunction

module Math.Tools.CoFunctor

module Math.Tools.CoMonad

module Math.Tools.Complex

module Math.Tools.Expression

module Math.Tools.Functor

module Math.Tools.Identity

module Math.Tools.Id

module Math.Tools.I

module Math.Tools.Integer

module Math.Tools.Integer

module Math.Tools.Isomorphism

module Math.Tools.LineInfo

module Math.Tools.List

module Math.Tools.Map

module Math.Tools.Map

module Math.Tools.Maybe

module Math.Tools.Median

module Math.Tools.Monad

module Math.Tools.Monoid

module Math.Tools.NaturalNumber

module Math.Tools.NaturalTransformation

module Math.Tools.Orthogonal

module Math.Tools.ParseMonad

module Math.Tools.ParserInterface

module Math.Tools.Parser

module Math.Tools.ParsingCombinators

module Math.Tools.PrettyP

newtype Prop a

We must represent classical logic with characteristic functions, because of the excluded middle axiom, and because characteristic functions (as properties) have complex relationship with possible elements of a. See the Universe type class for a solution. For exposition of classical logic used here, see "Lawvere,Rosebrugh: Sets for mathematics".

Constructors

Characteristic

Fields

runCharacteristic :: a -> Bool

subseteq :: Universe a => Prop a -> Prop a -> Bool

intersects :: Universe a => Prop a -> Prop a -> Bool

split_prop :: Universe a => Prop a -> ([a], [a])

See "Soare: Recursively enumerable sets and degrees", this is split of recursive into two recursively enumerable sets.

propositional_table :: Universe a => Prop a -> [(a, Bool)]

fromBool :: Bool -> Prop a

axiom_N :: (ModalLogic p, ImplicativeLogic p) => p a -> p a

https://en.wikipedia.org/wiki/Modal_logic

axiom_K :: (ModalLogic p, ImplicativeLogic p) => p a -> p a -> p a

https://en.wikipedia.org/wiki/Modal_logic

axiom_4 :: (ModalLogic p, ImplicativeLogic p) => p a -> p a

https://en.wikipedia.org/wiki/Modal_logic

axiom_B :: (ModalLogic p, ImplicativeLogic p) => p a -> p a

https://en.wikipedia.org/wiki/Modal_logic

axiom_D :: (ModalLogic p, ImplicativeLogic p) => p a -> p a

https://en.wikipedia.org/wiki/Modal_logic

axiom_5 :: (ModalLogic p, ImplicativeLogic p) => p a -> p a

https://en.wikipedia.org/wiki/Modal_logic

axiom_T :: (ModalLogic p, ImplicativeLogic p) => p a -> p a

https://en.wikipedia.org/wiki/Modal_logic

system_K :: (ModalLogic p, ImplicativeLogic p) => p a -> p a -> p a

system_T :: (ModalLogic p, ImplicativeLogic p) => p a -> p a -> p a

system_S4 :: (ModalLogic p, ImplicativeLogic p) => p a -> p a -> p a

system_S5 :: (ModalLogic p, ImplicativeLogic p) => p a -> p a -> p a

system_D :: (ModalLogic p, ImplicativeLogic p) => p a -> p a -> p a

vector_equalizer :: (Eq (n a), LinearTransform m n a) => (m :*: n) a -> (m :*: n) a -> Prop (m a)

matrix_equalizer :: (Eq (Scalar (h a)), InnerProductSpace (h a), Transposable h f, Applicative g, Applicative f, Foldable g, Foldable f) => (g :*: h) a -> (g :*: h) a -> Prop ((h :*: f) a)

list_truthvalues :: Universe a => Prop a -> [(a, Bool)]

It is impossible to convert characteristic index into recursively enumerable index (indexing of r.e. sets), or into canonical index (which has one bit per element describing membership). Thus we must use the Universe constraint.

apply :: Prop (Prop b, b)

result :: Eq a => a -> Prop a

result and the bind operations provide Monad instance, except for the | fact that the Universe and Eq constraints are required. Thus an adjunction instance also exists.

existential_bind :: Universe a => Prop a -> (a -> Prop b) -> Prop b

existential_join :: (Universe b, Eq b) => Prop (Prop b) -> Prop b

universal_bind :: Universe a => Prop a -> (a -> Prop b) -> Prop b

universal_join :: (Universe b, Eq b) => Prop (Prop b) -> Prop b

prop_fixedpoint :: (Prop a -> Prop a) -> Prop a

truthvalue :: Bool -> Prop a

satisfiable :: Universe a => Prop a -> Bool

Satisfiable returns true if there exists an element of the universe such that the proposition is satisfied. Note symmetry with axiom.

axiom :: Universe a => Prop a -> Bool

axiom returns true, if the proposition holds for all elements of the universe.

notMember :: a -> Prop a -> Bool

subseteq_prop :: Prop a -> Prop a -> Prop a

\[x \in subseteq i j \iff x \in i \implies x \in j\]

is_section :: Prop a -> (Bool -> a) -> Bool

doublePropJoin :: Prop (Prop (Prop (Prop a))) -> Prop (Prop a)

fromAssoc :: Eq a => [(a, Bool)] -> Prop a

fromAssocWith :: Eq a => [(a, b)] -> Prop b -> Prop a

prop_fromList :: Eq a => [a] -> Prop a

prop_fromSet :: Ord a => Set a -> Prop a

prop_fromMap :: (Ord i, Eq e) => Map i e -> Prop (i, e)

map_prop :: (Eq b, Universe a) => (a -> b) -> Prop a -> Prop b

Due to constraints, Prop cannot be made as instance of Functor.

equalizer_fixedpoint :: HasEqualizers p a => (a -> a) -> p a

invertible :: HasEqualizers p a => (b -> a) -> (a -> b) -> p a

kernel :: (Num c, HasEqualizers p c) => (a -> c) -> p a

ternary_op :: (Bool -> Bool -> Bool -> Bool) -> Prop a -> Prop a -> Prop a -> Prop a

binary_op :: (Bool -> Bool -> Bool) -> Prop a -> Prop a -> Prop a

gen_binary_op :: (b -> b' -> c) -> (a -> b) -> (a -> b') -> a -> c

binary :: Relational p => (a -> b -> p c) -> p (a, (b, c))

evaluate :: a -> Prop (Prop a)

evaluate' :: (Universe a, Eq a) => a -> Prop (Prop a)

integrate_prop :: Universe a => Prop a -> Prop (Prop a)

intersect :: Prop a -> Prop a -> Prop a

intersect3 :: Prop a -> Prop a -> Prop a -> Prop a

median3 :: Prop a -> Prop a -> Prop a -> Prop a

union :: Prop a -> Prop a -> Prop a

union3 :: Prop a -> Prop a -> Prop a -> Prop a

excluded_middle :: Universe b => Prop b -> Bool

noncontradiction :: Universe b => Prop b -> Bool

demorgan_union :: Universe a => Prop a -> Prop a -> Bool

demorgan_intersect :: Universe a => Prop a -> Prop a -> Bool

disjunction_commutative :: Universe a => Prop a -> Prop a -> Bool

conjunction_commutative :: Universe a => Prop a -> Prop a -> Bool

graph :: Eq b => (a -> b) -> Prop (a, b)

\[graph(f) = [<x,y> | y = f(x)]\]

image_factorization :: (Eq b, Universe a) => (a -> b) -> Prop b

bind :: Universe b => (a -> Prop b) -> (b -> Prop c) -> a -> Prop c

prop_compose :: Universe b => Prop (a, b) -> Prop (b, c) -> Prop (a, c)

transitive :: Universe a => Prop (a, a) -> Bool

\[<x,z> \in transitive \mathbb{R} \iff (\forall y. <x,y> \in \mathbb{R} & <y,z> \in \mathbb{R} \implies <x,z> \in \mathbb{R})\]

antisymmetric :: Eq a => Prop (a, a) -> Prop (a, a)

prop_iso_terminal :: Prop ((), b) -> Prop b

prop_iso_terminal_inverse :: Prop b -> Prop ((), b)

fiber :: HasEqualizers Prop c => (b -> c) -> c -> Prop b

\[fiber(p,y) = { x | p(x) = y } = p^{-1}({y})\]

fiber_ :: Eq b => (a -> b) -> b -> Prop a

equivalence_relation :: HasEqualizers Prop b => (a -> b) -> Prop (a, a)

reflexive :: Eq a => Prop (a, a)

reflexive_closure :: Eq a => Prop (a, a) -> Prop (a, a)

symmetric_closure :: Prop (a, a) -> Prop (a, a)

transitive_step :: Universe a => Prop (a, a) -> Prop (a, a)

transitive_closure :: Universe a => Prop (a, a) -> Prop (a, a)

equivalence_generated_by :: (Eq a, Universe a) => Prop (a, a) -> Prop (a, a)

opposite :: Prop (a, b) -> Prop (b, a)

leftAdjunct_prop :: (a -> Prop b) -> b -> Prop a

Adjunction \[P^{op} \dashv P\] between functors \[P\] and \[P^{op}\], where \[P : C^{op} \Rightarrow C\] and \[P^{op} : C \Rightarrow C^{op}\]. This is based on symmetry of product in relations.

unit_prop :: b -> Prop (Prop b)

rel_prop :: (a -> Bool) -> Prop a

rightAdjunct of that adjunction is difficult, because it would be in C^{op}, and we don't have that category here. So how could we implement that? We'd need opposite arrow, but then the implementation would be exactly same as leftAdjunct_prop and unit_prop, but lifted to opposite category. I'm not implementing that now.

binary_rel_prop :: (a -> a -> Bool) -> Prop (a, a)

binary_rel :: (a -> a -> Bool) -> a -> Prop a

http://nlab.mathforge.org/nlab/show/well-founded+relation

equal_functions :: (Universe a, Eq b) => Prop (a -> b, a -> b)

test_equality_at :: a -> Prop (Prop a, Prop a)

equalizer_prop :: Prop a -> Prop a -> Prop a

pullback_prop :: Prop a -> Prop b -> Prop (a, b)

sum :: Prop a -> Prop b -> Prop (Either a b)

prop_and :: Prop (Either a b) -> Prop (a, b)

prop_or :: Prop (Either a b) -> Prop (a, b)

prop_if :: Prop a -> Prop (Either a a) -> Prop a

prop_if_general :: Prop a -> Prop (Either b c) -> Prop (a, b, c)

equal :: Eq a => Prop (a, a)

not_equal :: Eq a => Prop (a, a)

lessthan :: Ord a => Prop (a, a)

lesseq :: Ord a => Prop (a, a)

or_list_prop :: Prop a -> Prop [a]

and_list_prop :: Prop a -> Prop [a]

enumerate :: Universe a => Prop a -> [a]

existential_map :: (Universe a, Eq b) => (a -> b) -> Prop a -> Prop b

universal_map :: (Universe a, Eq b) => (a -> b) -> Prop a -> Prop b

\[universal\_map( f, p ) = \{ b | p <= f^{-1}(\{b\}) \}\]

prop_existential :: Universe b => Prop (a, b) -> Prop a

prop_existential is basically the same as 'existential_map fst', except for performance. 'existential_map fst' will need to iterate over all pairs \[<a,b>\], whereas prop_existential will only iterate through all elements of b only. Similarly for prop_universal.

prop_universal :: Universe b => Prop (a, b) -> Prop a

image :: Universe a => Prop (a, b) -> Prop b

prop_universal_image :: (Universe a, Universe b, Eq b) => Prop (a, b) -> Prop b

forevery :: Universe b => Prop b -> Prop ()

newtype b :\: a

Constructors

PSub

Fields

runpsub :: PProp a -> PProp b

data PProp a where

Constructors

PProp :: Prop a -> PProp a

PNeg :: (Prop a -> Prop b) -> PProp (b :\: a)

pneg_func :: (a -> b) -> PProp (a :\: b)

module Math.Tools.Tree

module Math.Tools.Universe

module Math.Tools.Visitor