Math.Tools.Adjunction

Documentation

class (Functor f, Functor g) => Adjunction f g | f -> g, g -> f where

The problem with this type class is that categories are not represented at all. This assumes all functors are F : C -> C. This is a standard definition of adjunction from category theory.

Minimal complete definition

Nothing

Methods

leftAdjunct :: (f a -> b) -> a -> g b

unit :: a -> g (f a)

rightAdjunct :: (a -> g b) -> f a -> b

counit :: f (g b) -> b

embed_adjunction :: (Adjunction f g, Adjunction g f) => (f a -> f b) -> a -> b

zipR :: Adjunction f u => (u a, u b) -> u (a, b)

zipR and cozipL are from Edward Kmett's Adjunction package http://hackage.haskell.org/packages/archive/adjunctions/0.9.0.4/doc/html/src/Data-Functor-Adjunction.html

cozipL :: Adjunction f u => f (Either a b) -> Either (f a) (f b)

zipR and cozipL are from Edward Kmett's Adjunction package http://hackage.haskell.org/packages/archive/adjunctions/0.9.0.4/doc/html/src/Data-Functor-Adjunction.html

swap :: (b, a) -> (a, b)

data AdjM f g a

Constructors

AdjM

Fields

runAdjM :: g (f a)