Math.Tools.CoMonad

Documentation

class Functor w => Coapplicative w where

Methods

coeval :: w a -> a

colambda :: (w a -> w b) -> w (a -> b)

class Functor w => Comonad w where

From youtube video "Category theory II 7.2: Comonads categorically and examples" by Bartosz Milewski, comonad has a "focal point" defined by extract operation, and a mechanism for applying a cokleisli arrow to the whole environment [e.g. if you give average to extend, you'd smooth the whole data structure].

Minimal complete definition

extract

Methods

extract :: w a -> a

duplicate :: w a -> w (w a)

extend :: (w a -> b) -> w a -> w b

(=>>) :: w a -> (w a -> b) -> w b

(.>>) :: w a -> b -> w b

class Comonad w => CircularComonad w where

Methods

rotate :: w a -> w a

class CircularComonad w => FiniteComonad w where

Methods

inverseRotate :: w a -> w a

class Comonad w => BinaryTreeComonad w where

Methods

nextLevel :: w a -> (w a, w a)

class Comonad w => TernaryTreeComonad w where

Methods

nextLevel3 :: w a -> (w a, w a, w a)

class Comonad w => ChainComonad w where

Methods

coMonadChain :: w a -> Maybe (w a)

class Comonad w => TreeComonad w where

Methods

children :: w a -> [w a]

class Comonad w => DenseComonad w where

Methods

split :: w a -> w (a, a)

class Comonad w => NonemptyComonad w where

Methods

nonzero :: w a -> Either a (w a)

class CircularComonad w => InfiniteComonad w where

Methods

pre :: a -> w a -> w a

class UncertainComonad w where

Methods

perhaps :: w a -> w (Maybe a)

preList :: (InfiniteComonad w, Foldable t) => t a -> w a -> w a

indexComonad :: InfiniteComonad w => Integer -> w a -> a

permute :: Comonad w => w a -> w a

wmap :: Comonad w => (a -> b) -> w a -> w b

class Reference w where

Methods

dereference :: w a -> a

invoke :: w a -> (w a -> b) -> w b

new_reference :: w a -> b -> w b

class Unique w where

Methods

copy :: a -> (w a, a)

class Continuation m where

Methods

escape :: ((a -> m b) -> m a) -> m a

data CoState s a

Constructors

CoState (s -> a) s

newtype CoKleisli w a b

Constructors

CoKleisli

Fields

unCoKleisli :: w a -> b

data Neg o a

Constructors

Neg ((a -> o) -> o)

evalC :: Neg o o -> o

shiftC :: ((a -> Neg o o) -> Neg o o) -> Neg o a

shift and reset are from Wadler: Monads and composable continuations http://citeseer.nj.nec.com/wadler93monads.html

resetC :: Neg o o -> Neg o o

http://citeseer.nj.nec.com/wadler93monads.html

liftC :: (a -> o) -> Neg o a -> Neg o o