Math.Tools.Integer

Documentation

square_root :: Integer -> Integer

newton_square_root_list :: Integral a => a -> [a]

Some numbers may cause newton_square_root_list to produce an infinite sequence such as newton_square_root_list 48 == [1,24,13,8,7,8,7,...] We guard against this by terminating if the difference between two adjacent results is less than 2.

gcd_ :: Integral a => a -> a -> a

rising_factorial_power :: (Num a, Eq a) => a -> a -> a

See Knuth: TAOCP, or http://www.math.harvard.edu/~mazur/preprints/when_is_one.pdf

falling_factorial_power :: (Num a, Eq a) => a -> a -> a

See Knuth: TAOCP or http://www.math.harvard.edu/~mazur/preprints/when_is_one.pdf

binomial :: (Enum a, Fractional a) => a -> a -> a

http://en.wikipedia.org/wiki/Binomial_coefficient