Math.Tools.Isomorphism

Documentation

data a :==: b

See http://twanvl.nl/blog/haskell/isomorphism-lenses I also saw somewhere (can't remember where) a blog post for an idea for combinator library using isomorphisms. This is some sort of implementation of that idea.

Constructors

Iso

Fields

isomorphism_epimorphism :: a -> b

isomorphism_section :: b -> a

type Iso a b = a :==: b

type Aut a = a :==: a

leftIdempotent :: (a :==: b) -> Endo a

https://en.wikipedia.org/Galois_Connection Note that since we don't check the equations for isomorphisms, this function need not produce identity.

rightIdempotent :: (a :==: b) -> Endo b

https://en.wikipedia.org/Galois_Connection Note that since we don't check the equations for isomorphisms, this function need not produce an identity.

automorphism :: Endo a -> Endo a -> Aut a

visit_iso :: ComposableVisitor v => (v :==: a) -> v -> a

adjoint_iso :: Adjunction f g => (f a -> b) :==: (a -> g b)

class (BiArrow arr, Groupoid arr) => Isomorphic arr where

Methods

iso :: arr a b -> a :==: b

class IsomorphicFunctor f where

Associated Types

data IsoA f a b

Methods

transformIso :: IsoA f a b -> f a :==: f b

appIsoF :: IsomorphicFunctor f => IsoA f a b -> f a -> f b

toArrow :: BiArrow arr => (a :==: b) -> arr a b

abstract :: Isomorphic arr => (a :==: b) :==: arr a b

runIso :: (a :==: b) -> a -> b

type FRepresentable f a = forall r. (a -> r) -> f r

type FCorepresentable f a = forall r. (r -> a) -> f r

representable :: FRepresentable f a -> f a

corepresentable :: FCorepresentable f a -> f a

inverse_representable :: Functor f => f a -> FRepresentable f a

coinverse_representable :: CoFunctor f => f a -> FCorepresentable f a

(=<) :: (a :==: b) -> a -> b

liftIso :: Isomorphic arr => ((a :==: b) -> c :==: d) -> arr a b -> arr c d

runIsoInverse :: (a :==: b) -> b -> a

embedIso :: Groupoid cat => cat a b -> cat b b -> cat a a

invertIso :: (a :==: b) -> b :==: a

equalIso :: Eq a => (a, a) :==: Either a (a, a)

inIso :: Rec f :==: f (Rec f)

mapIso :: Functor f => (a :==: b) -> f a :==: f b

inverseImageIso :: CoFunctor p => (a :==: b) -> p b :==: p a

bimapIso :: (Functor f, CoFunctor f) => (a :==: b) -> (a :==: c) -> f c :==: f b

orderedMapIso :: (Monad m, Ord i, Ord a) => Map i a -> Map a i -> m (i :==: a)

constraint: all elements of the map must be present in both maps.

curryIso :: ((a, b) -> c) :==: (a -> b -> c)

intIso :: Aut a -> (Integer -> a) :==: a

pairingIso :: (Integer, Integer) :==: Integer

Cantor pairing function. Algorithm for the inverse is from https://secure.wikimedia.org/wikipedia/en/wiki/Pairing_function

squareIso :: Aut Integer

absIso :: (Num a, Eq a) => a :==: Either a a

plusminusIso :: Integral a => Aut (a, a)

timesDivideIso :: Floating a => Aut (a, a)

polarIso :: RealFloat a => Complex a :==: (a, a)

ratioIso :: Integral a => (a, a) :==: (Ratio a, a)

injectIso :: (a :==: b) -> (a :==: Either a b, b :==: Either a b)

leftRightIso :: (a :==: Either a a, a :==: Either a a)

type Bot = forall a. a

andIso :: (Bool, Bool) :==: (Bool, Ordering)

orIso :: (Bool, Bool) :==: (Bool, Ordering)

initialIso :: (BiArrow arr, Arrow arr') => arr () (arr' Bot a)

terminalIso :: (Isomorphic arr, Arrow arr') => arr () (arr' a ())

productIso :: (Isomorphic arr, ArrowApply arr') => arr (arr' c (a, b)) (arr' c a, arr' c b)

coproductIso :: (Isomorphic arr, ArrowChoice arr', ArrowApply arr') => arr (arr' (Either a b) c) (arr' a c, arr' b c)

exponentialIso :: ArrowApply arr' => arr' (a, b) c :==: arr' a (arr' b c)

eitherIso :: Isomorphic arr => arr a c1 -> arr b c2 -> arr (Either c1 c2) c -> arr (Either a b) c

notIso :: BiArrow arr => arr Bool Bool

boolIso :: BiArrow arr => arr Bool (Either () ())

assocIso :: BiArrow arr => arr ((a, b), c) (a, (b, c))

tripleIso :: BiArrow arr => arr (a, b, c) ((a, b), c)

quadIso :: BiArrow arr => arr (a, b, c, d) ((a, b), (c, d))

pentaIso :: BiArrow arr => arr (a, b, c, d, e) ((a, b), c, (d, e))

zipEither :: BiArrow arr => arr (Either x y, Either a b) (Either (x, a) (Either (x, b) (Either (y, a) (y, b))))

distIso :: BiArrow arr => arr (Either (a, b) (a, c)) (a, Either b c)

assocEitherIso :: BiArrow arr => arr (Either a (Either b c)) (Either (Either a b) c)

terminalSecondIso :: BiArrow arr => arr (a, ()) a

terminalThirdIso :: BiArrow arr => arr (a, b, ()) (a, b)

terminalFourthIso :: BiArrow arr => arr (a, b, c, ()) (a, b, c)

swapIso :: BiArrow arr => arr (a, b) (b, a)

swapEitherIso :: BiArrow arr => arr (Either a b) (Either b a)

swapEither :: Either a b -> Either b a

(<**>) :: (a :==: b) -> (c :==: d) -> (a, c) :==: (b, d)

encodeDecode :: Binary a => a :==: ByteString

encoding :: Binary a => Aut ByteString -> Aut a

decoding :: Binary a => Aut a -> Aut ByteString

product3 :: (a :==: b) -> (c :==: d) -> (e :==: f) -> (a, c, e) :==: (b, d, f)

first3 :: (a :==: a') -> (a, b, c) :==: (a', b, c)

second3 :: (b :==: b') -> (a, b, c) :==: (a, b', c)

third3 :: (c :==: c') -> (a, b, c) :==: (a, b, c')

(<||>) :: Isomorphic arr => arr a b -> arr c d -> arr (Either a c) (Either b d)

ifIso :: Isomorphic arr => arr a b -> arr a c -> arr (a, Bool) (Either b c)

branch :: Isomorphic arr => arr a b -> arr a c -> arr b d -> arr c d -> arr (a, Bool) (d, Bool)

returnIso :: BiArrow arr => a -> arr () a

fixIso :: Isomorphic arr => arr a a -> arr () (a, a)

foldrIso :: ((a, [a]) :==: b) -> (Bool :==: b) -> [a] :==: b

appendIso :: BiArrow arr => arr ([a], [a]) ([a], Int)

listIso :: BiArrow arr => arr [a] (Either (a, [a]) ())

foldListIso :: (a :==: b) -> (() :==: d) -> (Either (b, c) d :==: c) -> [a] :==: c

mapListIso :: (a :==: b) -> [a] :==: [b]

This instance of map for list doesn't need fmap:

maybeIso :: BiArrow arr => arr (Maybe a) (Either a ())

data IsomorphismA arr a b

Constructors

IsoA

Fields

epimorphismA :: arr a b

sectionA :: arr b a