Math.Tools.Map

Documentation

all_maps :: Ord k => [k] -> [a] -> [Map k a]

all_total_maps :: Ord k => [k] -> [a] -> [Map k a]

lookupMapM :: (Monad m, Show i, Ord i) => Map i a -> i -> m a

unzip :: Map i (a, b) -> (Map i a, Map i b)

unzip3 :: Map i (a, b, c) -> (Map i a, Map i b, Map i c)

foldM :: Monad m => (a -> b -> m b) -> m b -> Map i a -> m b

foldKeyM :: Monad m => (i -> a -> b -> m b) -> m b -> Map i a -> m b

mapM :: (Monad m, Ord i) => (a -> m b) -> Map i a -> m (Map i b)

sequence_ :: (Monad m, Ord i) => Map i (m a) -> m ()

sequence :: (Monad m, Ord i) => Map i (m a) -> m (Map i a)

sequenceByKey :: (Monad m, Ord i) => Map i (i -> m a) -> m (Map i a)

joinMap :: Map String (Map String a) -> Map String a

splitMapByKey :: (Ord s, Ord m) => Map (m, s) a -> Map m (Map s a)

partitionMap :: Ord k => Map k a -> (a -> Either b c) -> (Map k b, Map k c)

intersectZipMapA :: (Ord i, Arrow arr, FunctorArrow [] arr) => arr (Map i a, Map i b) (Map i (a, b))

intersectMapA :: (Ord i, Arrow arr, FunctorArrow [] arr) => arr (a, b) c -> arr (Map i a, Map i b) (Map i c)

zipWithMapM :: (Monad m, Ord i) => (a -> b -> m c) -> Map i a -> Map i b -> m (Map i c, Map i a, Map i b)

intersectMapKeyA :: (Ord i, Arrow arr, FunctorArrow [] arr) => arr (i, (a, b)) c -> arr (Map i a, Map i b) (Map i c)

mapMapWithKey :: (ArrowChoice arr, Eq i) => arr (i, e) (i, f) -> arr (Map i e) (Map i f)

unionMapA :: (Ord i, Arrow arr, FunctorArrow [] arr) => arr (a, a) a -> arr (Map i a, Map i a) (Map i a)

sequenceMap :: Ord i => Map i (IO b) -> IO (Map i b)

sequenceMapA :: (Arrow arr, Ord i) => Map i (arr a b) -> arr a (Map i b)

sequenceMapWithKeyA :: (Arrow arr, Ord i) => Map i (arr (a, i) b) -> arr a (Map i b)

foldMapA :: ArrowApply arr => arr (a, b) b -> arr (Map i a, b) b

foldMapKeyA :: ArrowApply arr => arr (i, a, b) b -> arr (Map i a, b) b

composeMapA :: (Arrow arr, Ord i, Ord a) => arr (Map i a, Map a b) (Map i b)