Math.Tools.Median

Documentation

class Ord m => MedianAlgebra m where

Rules expected of MedianAlgebra.

From TAOCP section 7.1.1 http://www-cs-faculty.stanford.edu/~uno/taocp.html

"median/majority"

med x x y == x

"median/commutative"

med x y z == med x z y == med y z x == med z x y == med z y x

"median/associative"

med x w (med y w z) == med (med x w y) w z

"median/distributive"

med (med x y z) u v == med x (med y u v) (med z u v)

Methods

med :: m -> m -> m -> m

median5 :: MedianAlgebra a => a -> a -> a -> a -> a -> a

median :: Ord a => a -> a -> a -> a

data Interval a

Constructors

Interval

Fields

interval_startpoint :: a

interval_endpoint :: a

in_interval :: (Eq a, MedianAlgebra a) => a -> Interval a -> Bool

interval :: (Monad m, MedianAlgebra a) => Interval a -> m a -> m a

median_homomorphism :: MedianAlgebra a => Interval a -> a -> a

andor :: Ord a => (a, a) -> (a, a)

andor is same as (x,y) -> (x && y, x || y), thus the name, except that the type b is more general. Thus if both and and or are combined, sorting of the pair occurs.

andor' :: Ord a => (a, a) -> (a, a)

andor3 :: Ord a => (a, a, a) -> (a, a, a)