Math.Tools.Monad

Documentation

class Monad m => ModalMonad m where

https://ncatlab.org/nlab/show/idemponent+monad

Methods

splitM :: m a -> m (m a)

class Monad m => NegatableMonad m where

Methods

negateM :: m a -> m a

class Monad m => KnowledgeMonad m where

Methods

knownM :: m a -> a

data AdjM f g a

Constructors

AdjM

Fields

runAdjM :: g (f a)

appM2 :: Monad m => (a -> b -> m c) -> m a -> m b -> m c

appM3 :: Monad m => (a -> b -> c -> m d) -> m a -> m b -> m c -> m d

(<=<) :: Monad m => (b -> m c) -> (a -> m b) -> a -> m c

type from tweet by Chris Martin, but well known.

fromMaybeM :: Monad m => Maybe a -> m a

bind2 :: Monad m => m (m a) -> (a -> m b) -> m b

join3 :: Monad m => m (m (m b)) -> m b

reduceM :: Monad m => m (a -> m b) -> a -> m b

This is strange, you can remove arbitrary monad with this.

mapK :: Monad m => m (a -> m b) -> m a -> m b